% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{add_table_theme}
\alias{add_table_theme}
\title{Add a standard theme to a \code{gt} table.}
\usage{
add_table_theme(x)
}
\arguments{
\item{x}{\code{data.frame} A data.frame that will be converted to a \code{gt} table.}
}
\description{
Add a standard theme to a \code{gt} table.
}
\keyword{internal}
