% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Assess.R
\name{AE_Assess}
\alias{AE_Assess}
\title{AE Assessment}
\usage{
AE_Assess(
  dfInput,
  vThreshold = NULL,
  cLabel = "",
  cMethod = "poisson",
  strValueColumn = NULL,
  bDataList = FALSE
)
}
\arguments{
\item{dfInput}{input data}

\item{vThreshold}{numeric vector with 2 threshold values.  Defaults to c(-5,5) for method = "poisson" and c(.0001,NA) for method = Wilcoxon}

\item{cLabel}{Assessment label}

\item{cMethod}{valid methods are "poisson" (the default), or  "wilcoxon"}

\item{strValueColumn}{Optional, Name of the Column to use for sign of Flag. Only used if cMethod = 'wilcoxon'.
If value for that row is higher than median of strValueColumn then Flag = 1, if lower then Flag = -1.}

\item{bDataList}{Should all assessment datasets be returned as a list? If False (the default), only the finding data frame is returned}
}
\value{
Finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
AE Assessment
}
\examples{
dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )
SafetyAE <- AE_Assess( dfInput )
SafetyAE_Wilk <- AE_Assess( dfInput, cMethod="wilcoxon")

}
