% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Visualize.R
\name{AE_Visualize}
\alias{AE_Visualize}
\title{Safety Assessment - Site-level plot}
\usage{
AE_Visualize(dfFlagged, title = "", unit = "weeks")
}
\arguments{
\item{dfFlagged}{Flagged data frame generated by \code{\link{AE_Assess}} when \code{bDataList} is FALSE.}

\item{title}{Title for the plot}

\item{unit}{exposure time unit. Defaults to "weeks".}
}
\value{
site level plot object
}
\description{
Safety Assessment - Site-level plot
}
