% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Wilcoxon.R
\name{AE_Wilcoxon_Analyze}
\alias{AE_Wilcoxon_Analyze}
\title{AE_Wilcoxon_Analyze}
\usage{
AE_Wilcoxon_Analyze(dfTransformed)
}
\arguments{
\item{dfTransformed}{paramDescription}
}
\description{
Creates Analyis results data for Adverse Event assessment using the Wilcoxon sign-ranked test
}
