% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{analyticsBounds}
\alias{analyticsBounds}
\title{analyticsBounds Dataset}
\format{
A data frame with 12667 rows and 8 columns:
\describe{
\item{Threshold}{number of standard deviations that the upper and lower bounds are based on}
\item{Denominator}{calculated denominator value}
\item{LogDenominator}{calculated log denominator value}
\item{Numerator}{calculated numerator value}
\item{Metric}{calculated rate/metric value}
\item{MetricID}{unique metric identifier}
\item{StudyID}{unique study identifier}
\item{SnapshotDate}{date of the snapshot}
}
}
\source{
Generated from \code{analyticsBounds.csv} dataset in the \code{gsm} package.
}
\usage{
analyticsBounds
}
\description{
analyticsBounds Dataset
}
\keyword{datasets}
