% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{analyticsInput}
\alias{analyticsInput}
\title{analyticsInput Dataset}
\format{
A data frame with 25 rows and 5 columns:
\describe{
\item{SubjectID}{unique subject identifier}
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Numerator}{number of flags in group}
\item{Denominator}{total subjects in group}
}
}
\source{
Generated from \code{analyticsInput.csv} dataset in the \code{gsm} package.
}
\usage{
analyticsInput
}
\description{
analyticsInput Dataset
}
\keyword{datasets}
