% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Binary.R
\name{Analyze_Binary}
\alias{Analyze_Binary}
\title{Funnel Plot Analysis with Normal Approximation for Binary Outcomes}
\usage{
Analyze_Binary(dfTransformed, bQuiet = TRUE)
}
\arguments{
\item{dfTransformed}{\code{data.frame} in format produced by \code{\link{Transform_Rate}}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one row per site with columns: GroupID, Numerator, Numerator_Other, Denominator, Denominator_Other, Prop, Prop_Other, Metric, Estimate, Score.
}
\description{
Funnel Plot Analysis with Normal Approximation for Binary Outcomes
}
\details{
Creates analysis results data for percentage data using funnel plot method with normal approximation.
}
\section{Statistical Methods}{

This function applies funnel plots using asymptotic limits based on normal approximation of binomial distribution for
the binary outcome with the sample sizes of the sites to assess data quality and safety.
}

\section{Data Specification}{

The input data (\code{dfTransformed}) for Analyze_Binary is typically created using \code{\link{Transform_Rate}} and should be one record per site with required columns for:
\itemize{
\item \code{GroupID} - Site ID
\item \code{Numerator} - Total number of participants at site with event of interest
\item \code{Denominator} - Total number of participants at site
\item \code{Metric} - Proportion of participants at site with event of interest
}
}

\examples{
dfInput <- Disp_Map_Raw()

dfTransformed <- Transform_Rate(
  dfInput,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Total"
)

dfAnalyzed <- Analyze_Binary(dfTransformed)

}
