% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Binary_PredictBounds.R
\name{Analyze_Binary_PredictBounds}
\alias{Analyze_Binary_PredictBounds}
\title{Funnel Plot Analysis for Binary Outcomes - Predicted Boundaries}
\usage{
Analyze_Binary_PredictBounds(
  dfTransformed,
  vThreshold = c(-3, -2, 2, 3),
  bQuiet = TRUE
)
}
\arguments{
\item{dfTransformed}{\code{data.frame} data.frame in format produced by
\code{\link{Transform_Rate}}. Must include GroupID, N, Numerator and Denominator}

\item{vThreshold}{\code{numeric} upper and lower boundaries in residual space. Should be identical to
the thresholds used AE_Assess().}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} containing predicted boundary values with upper and lower bounds across the
range of observed values.
}
\description{
Funnel Plot Analysis for Binary Outcomes - Predicted Boundaries
}
\details{
Apply funnel plot analysis to site level data and then calculates predicted percentages and upper- and
lower- bounds for across the full range of sample sizes.
}
\section{Statistical Methods}{

This function applies funnel plots analysis to site-level data and then calculates predicted percentages
and upper- and lower- bounds (funnels) for across the full range of sample sizes.
}

\section{Data Specification}{

The input data (\code{dfTransformed}) for Analyze_Poisson is typically created using
\code{\link{Transform_Rate}} and should be one record per site with columns for:
\itemize{
\item \code{GroupID} - Unique subject ID
\item \code{Numerator} - Number of Events
\item \code{Denominator} - Number of days of exposure
}
}

\examples{
dfInput <- Disp_Map_Raw()

dfTransformed <- Transform_Rate(
  dfInput,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Total"
)

dfAnalyzed <- Analyze_Binary(dfTransformed)

dfBounds <- Analyze_Binary_PredictBounds(dfTransformed, c(-3, -2, 2, 3))

}
