% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Chisq.R
\name{Analyze_Chisq}
\alias{Analyze_Chisq}
\title{Chi-squared Test Analysis}
\usage{
Analyze_Chisq(dfTransformed, strOutcome = "TotalCount")
}
\arguments{
\item{dfTransformed}{data.frame in format produced by \code{\link{Transform_EventCount}}}

\item{strOutcome}{required, name of column in dfTransformed dataset to perform the chi-squared test on}

\item{...}{additional arguments to pass to \code{\link{chisq.test}}}
}
\value{
data.frame with one row per site, columns:   SiteID, N , ..., SiteProp, OtherProp, Statistic, PValue
}
\description{
Creates Analysis results data for count data using the chi-squared test
}
\details{
@details

Analyzes count data using the chi-squared test
}
\section{Statistical Methods}{


TODO Coming soon ...
}

\section{Data Specification}{


The input data (\code{ dfTransformed}) for the Analyze_Chisq is typically created using \code{\link{Transform_EventCount}}  and should be one record per Site with columns for:
\itemize{
\item \code{SiteID} - Site ID
\item \code{N} - Total number of participants at site
\item \code{Count} - Total number of participants at site with event of interest
}
}

\examples{
dfInput <- Disp_Map(dfDisp = safetyData::adam_adsl, strCol = "DCREASCD",strReason = "Adverse Event")
dfTransformed <- Transform_EventCount( dfInput, cCountCol = 'Count' )
dfAnalyzed <- Analyze_Chisq( dfTransformed )

}
