% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_MissingInvalid.R
\name{Analyze_MissingInvalid}
\alias{Analyze_MissingInvalid}
\title{Inclusion/Exclusion or Consent Assessment.}
\usage{
Analyze_MissingInvalid(dfTransformed)
}
\arguments{
\item{dfTransformed}{data.frame in format produced by \code{\link{Transform_EventCount}}. Must include: SubjectID, SiteID, TotalCount.}
}
\value{
input data frame with columns added for "PValue" and "Estimate" (with all rows in PValue = NA).
}
\description{
Analysis of Inclusion/Exclusion or Consent.
}
\details{
This analysis is for IE and Consent Assessments. It simply duplicates the TotalCount column of the input as Estimate which keeps the naming convention consistent for the subsequent Flag
and summary functions as well as overall reporting consistency. PValue is not used for this assessment so all rows for PVAlue are set to NA.
}
\section{Data Specification}{


The input data (\code{ dfTransformed}) for IE_Analyze or Consent_Analyze is typically created using \code{\link{Transform_EventCount}} and should be one record per Site with columns for:
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{TotalCount} - Number of Missing or Invalid
}
}

\examples{
dfInput <- IE_Map_Raw(clindata::raw_ie_a2 )

dfTransformed <- gsm::Transform_EventCount( dfInput, cCountCol = "Count")
dfAnalyzed <- gsm::Analyze_MissingInvalid( dfTransformed ) 



}
