% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_QTL.R
\name{Analyze_QTL}
\alias{Analyze_QTL}
\title{QTL Analysis for Binary and Rate Outcomes.}
\usage{
Analyze_QTL(dfTransformed, nConfLevel = 0.95, strOutcome = "binary")
}
\arguments{
\item{dfTransformed}{\code{data.frame} Transformed data for analysis. Data should have one record per site with expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, and \code{Metric}. For more details see the Data Pipeline vignette: \code{vignette("DataPipeline", package = "gsm")}.
For this function, \code{dfTransformed} should typically be created using \code{\link[=Transform_Rate]{Transform_Rate()}}.}

\item{nConfLevel}{\code{numeric} specified confidence interval for QTL analysis.}

\item{strOutcome}{\code{character} indicates statistical test used for QTL analysis. One of \code{rate} or \code{binary}.}
}
\value{
\code{data.frame} with one row with columns: GroupID, Numerator, Denominator, Metric, Method, ConfLevel, Estimate, LowCI, UpCI and Score.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Creates confidence intervals for the observed proportion of participants for the event of interest. Uses the exact binomial (\code{stats::binom.test}) or poisson test (\code{stats::poisson.test}).
}
\section{Statistical Methods}{


TODO Coming soon ...
}

\examples{
# binary
dfTransformed <- tibble::tibble(
  GroupID      = c("AA-AA-000-0000"),
  Numerator    = c(122),
  Denominator  = c(1301),
  Metric       = c(0.0938)
)

dfAnalyzed <- Analyze_QTL(dfTransformed, strOutcome = "binary")

# rate
dfTransformedRate <- tibble::tibble(
  GroupID      = c("AA-AA-000-0000"),
  Numerator    = c(4473),
  Denominator  = c(6.19),
  Metric       = c(723)
)
dfAnalyzed <- Analyze_QTL(dfTransformedRate, strOutcome = "rate")

}
