% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Rate.R
\name{Analyze_Rate}
\alias{Analyze_Rate}
\title{Funnel Plot Analysis with Normal Approximation for Rate Outcomes}
\usage{
Analyze_Rate(dfTransformed, bQuiet = TRUE)
}
\arguments{
\item{dfTransformed}{\code{data.frame} in format produced by \code{\link{Transform_Rate}}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one row per site with columns: GroupID, Numerator, Numerator_Other, Denominator, Denominator_Other, Prop, Prop_Other, Metric, Estimate, Score.
}
\description{
Funnel Plot Analysis with Normal Approximation for Rate Outcomes
}
\details{
Creates analysis results data for rate data using funnel plot method with normal approximation.
}
\section{Statistical Methods}{

This function applies funnel plots using asymptotic limits based on normal approximation of Poisson distribution
the rate outcome verses the total exposure of the sites under investigation to assess data quality and safety.
}

\section{Data Specification}{

The input data (\code{dfTransformed}) for Analyze_Rate is typically created using \code{\link{Transform_Rate}} and should be one record per site with required columns for:
\itemize{
\item \code{GroupID} - Site ID
\item \code{Numerator} - Total number of events of interest at site
\item \code{Denominator} - Total number of days of exposure at site
\item \code{Metric} - Rate of events at site (Numerator / Denominator)
}
}

\examples{
dfInput <- AE_Map_Raw() \%>\% na.omit()

dfTransformed <- Transform_Rate(
  dfInput,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)

dfAnalyzed <- Analyze_Rate(dfTransformed)

}
