% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Wilcoxon.R
\name{Analyze_Wilcoxon_SummarizeCtsData}
\alias{Analyze_Wilcoxon_SummarizeCtsData}
\title{SummarizeCtsData}
\usage{
Analyze_Wilcoxon_SummarizeCtsData(vData, nDigits = 3)
}
\arguments{
\item{vData}{data vector}

\item{nDigits}{number of digits}
}
\description{
Helper function for Wilcoxon sign-ranked test
}
