% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{AppendDroppedWorkflows}
\alias{AppendDroppedWorkflows}
\title{Appends the previous snapshot lStudyAssessResults to the current snapshot lStudyAssessResults}
\usage{
AppendDroppedWorkflows(lPrevSnapshot, lResults)
}
\arguments{
\item{lPrevSnapshot}{\code{list} the previous Snapshot object}

\item{lResults}{\code{list} the current lResults created in \code{Make_Snapshot}}
}
\value{
Appended lSnapshot object
}
\description{
Appends the previous snapshot lStudyAssessResults to the current snapshot lStudyAssessResults
}
\keyword{internal}
