% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{AppendLogs}
\alias{AppendLogs}
\title{Appends the previous snapshot logs to the current snapshot logs}
\usage{
AppendLogs(lPrevSnapshot, lSnapshot, files = names(lPrevSnapshot$lSnapshot))
}
\arguments{
\item{lPrevSnapshot}{\code{list} the previous Snapshot object}

\item{lSnapshot}{\code{list} the current Snapshot object}

\item{files}{\code{vector} Optional vector of desired files to append, defaults to all files within the previous snapshot}
}
\value{
Appended lSnapshot object
}
\description{
Appends the previous snapshot logs to the current snapshot logs
}
\keyword{internal}
