% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ApplySpec.R
\name{ApplySpec}
\alias{ApplySpec}
\title{Apply Data Specification}
\usage{
ApplySpec(dfSource, columnSpecs, domain)
}
\arguments{
\item{dfSource}{\code{data.frame} Data frame to apply the column specification to.}

\item{columnSpecs}{\code{list} Column specification.}

\item{domain}{\code{character} Domain name.}
}
\value{
\code{data.frame} \code{dfSource} with the column specification applied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Apply a column specification to a data frame. The column specification is a
named list where the names are the target column names and the values are
lists of column specifications. The column specifications can include:
\itemize{
\item \code{required}: Whether the column is required.
\item \code{type}: The data type to convert the column to.
\item \code{source_col}: The name of the source column to map to the target column.
}
}
