% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Helpers.R
\name{AssessStatus}
\alias{AssessStatus}
\title{Create Study Results table for Report}
\usage{
AssessStatus(assessment, strType)
}
\arguments{
\item{assessment}{\code{list} a snapshot list containing the parameters to assess}

\item{strType}{\code{string} a string defining what report to define: "kri", "cou", or "qtl"}
}
\description{
Create Study Results table for Report
}
\keyword{internal}
