% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Augment_Snapshot.R
\name{Augment_Snapshot}
\alias{Augment_Snapshot}
\title{Attach longitudinal content to snapshot}
\usage{
Augment_Snapshot(lSnapshot, cPath, bAppendTimeSeriesCharts = TRUE)
}
\arguments{
\item{lSnapshot}{\code{list} The output of \code{\link[=Make_Snapshot]{Make_Snapshot()}} where \code{bReturnStudyObject = TRUE}.}

\item{cPath}{\code{character} Path to longitudinal snapshots, returned by \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}

\item{bAppendTimeSeriesCharts}{\code{logical} Append time series charts to KRIs? Default: \code{TRUE}.}
}
\value{
\code{list} The output of \code{\link[=Make_Snapshot]{Make_Snapshot()}} with longitudinal data and widgets appended.
}
\description{
Attach stacked, longitudinal snapshot data and time series widgets to output of
\code{\link[=Make_Snapshot]{Make_Snapshot()}}.
}
\examples{
\dontrun{
# Path to study
study_path <- here::here("data-raw/AA-AA-000-0000")

# Output current snapshot to dated folder within study folder.
#snapshot_path <- paste(study_path, Sys.Date(), sep = '/'); dir.create(snapshot_path)
snapshot <- Make_Snapshot(
  #cPath = snapshot_path
)

augmented_snapshot <- Augment_Snapshot(
  lSnapshot = snapshot,
  cPath = study_path
)
}

}
