% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barChart.R
\name{barChart}
\alias{barChart}
\title{KRI bar chart}
\usage{
barChart(
  data = NULL,
  config = list(),
  threshold = NULL,
  yaxis = "score",
  selectedGroupIDs = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{data with column names:
\itemize{
\item{\code{studyid}}
\item{\code{workflowid}}
\item{\code{groupid}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{metric}}
\item{\code{score}}
\item{\code{flag}}
}}

\item{config}{configuration data with columns:
\itemize{
\item{\code{workflow}}
\item{\code{gsm_version}}
\item{\code{group}}
\item{\code{metric}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{outcome}}
\item{\code{model}}
\item{\code{score}}
\item{\code{data_inputs}}
\item{\code{data_filters}}
}}

\item{threshold}{a one row data frame containing columns:
\itemize{
\item{\code{workflowid}}
\item{\code{gsm_version}}
\item{\code{param}}
\item{\code{index}}
\item{\code{default}}
\item{\code{configurable}}
}}

\item{yaxis}{either \code{'score'} or \code{'metric'}}

\item{selectedGroupIDs}{group IDs to highlight, \code{NULL} by default, can be a single site or a vector.}

\item{width}{width of widget, full screen by default}

\item{height}{height of widget, calculated based on width}

\item{elementId}{ID of container HTML element}
}
\description{
<Add Description>
}
