% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-BindResults.R
\name{BindResults}
\alias{BindResults}
\title{Helper function to bind results from multiple workflows}
\usage{
BindResults(lResults, strName, lColValues = NULL)
}
\arguments{
\item{lResults}{List of results in the format returned by RunWorkflows()}

\item{strName}{Name of the object to stack. Pulled from lResults$lData.}

\item{lColValues}{List of column names/values to add to all rows in the final data frame. e.g. l(SnapshotDate = sys.date(), StudyID = "ABC-123"),  Default is NULL.}
}
\value{
A data frame.
}
\description{
Used to stack results (e.g. dfSummary) from a list of analysis pipeline output formatted like the result of RunWorklows(). Also adds study level metadata when provided.
}
