% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Rate.R
\name{calculateCol_bySubj}
\alias{calculateCol_bySubj}
\title{Count or Sum a Column by Subject}
\usage{
calculateCol_bySubj(
  df,
  strSubjectCol = "SubjectID",
  strCol,
  strColOriginal,
  strMethod = c("Count", "Sum")
)
}
\arguments{
\item{df}{The df to modify.}

\item{strSubjectCol}{\code{character} Column name in \code{dfSubjects} to use for
subject ID. Default: "SubjectID"}

\item{strCol}{The column to produce.}

\item{strColOriginal}{The name of \code{strCol} in \code{df}.}

\item{strMethod}{Whether to "Count" or "Sum" the values in the column.}
}
\value{
The modified df.
}
\description{
Count or Sum a Column by Subject
}
\keyword{internal}
