% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-CalculatePercentage.R
\name{CalculatePercentage}
\alias{CalculatePercentage}
\title{Helper function to compile "long" group metadata}
\usage{
CalculatePercentage(
  data,
  strCurrentCol,
  strTargetCol,
  strPercVal,
  strPercStrVal
)
}
\arguments{
\item{data}{The input dataframe}

\item{strCurrentCol}{Column that represents site count or participant count}

\item{strTargetCol}{Column that represents the target count for site/participants}

\item{strPercVal}{Name of column that will contain the numeric percentage value on enrollment}

\item{strPercStrVal}{Name of column that will contain the n/N (xx.x\%) string}
}
\value{
A data frame containing two additional columns for the precentage value and associated string
}
\description{
Used to create percentage values on enrollment data, such as site activation and
and participantment enrollment, and pasting together an appropriate n/N (xx.x\%) value
as well.
}
