% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-CheckClindataMeta.R
\name{CheckClindataMeta}
\alias{CheckClindataMeta}
\title{Check gsm_version in clindata metadata}
\usage{
CheckClindataMeta(
  config = list(config_param = clindata::config_param, config_workflow =
    clindata::config_workflow),
  ci_check = FALSE
)
}
\arguments{
\item{config}{\code{list} Named list of metadata to check from the \code{{clindata}} package.}

\item{ci_check}{\code{logical} If \code{TRUE}, return data.frame with results of the version check. If \code{FALSE} (default), only console messages are returned.}
}
\value{
Console message indicating if \code{{clindata}} and \code{{gsm}} versions are compatible.
}
\description{
Check gsm_version in clindata metadata
}
\examples{
\dontrun{
CheckClindataMeta()
}

}
