% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-gt.R
\name{cols_label_month}
\alias{cols_label_month}
\title{Month and year columns for gt tables}
\usage{
cols_label_month(data, columns = gt::everything())
}
\arguments{
\item{data}{A \code{gt_tbl} object.}

\item{columns}{Columns to target for formatting.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Split date columns in the style "YYYY-MM" or "YYYY-MM-DD" into month columns
with year spanners.
}
\keyword{internal}
