% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uti-CombineSpecs.R
\name{combine_specs}
\alias{combine_specs}
\title{Combine Specifications}
\usage{
combine_specs(lSpecs)
}
\arguments{
\item{lSpecs}{A list of specifications.}
}
\value{
A list representing the combined specification.
}
\description{
This function combines a list of specifications into a single specification.
}
\examples{
combine_specs(list(list(domain1 = "spec1"), list(domain2 = "spec2")))
}
