% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{CompileResultsSummary}
\alias{CompileResultsSummary}
\title{Compile the results summary of all kri's in a snapshot}
\usage{
CompileResultsSummary(lResults)
}
\arguments{
\item{lResults}{\code{list} the output of \code{Study_Assess()} containing results of kri analysis}
}
\description{
Compile the results summary of all kri's in a snapshot
}
\keyword{internal}
