% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{config_workflow}
\alias{config_workflow}
\title{Workflow Configuration Data}
\format{
A data frame with 8 rows and 12 variables:
\describe{
\item{studyid}{Study ID}
\item{workflowid}{Workflow ID}
\item{gsm_version}{Version of \code{gsm} that the workflow will be run on.}
\item{active}{Boolean indicating if the workflow for the specified workflowid is active.}
}
}
\source{
\url{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}
}
\usage{
config_workflow
}
\description{
A dataset containing a sample workflow configuration file for running GSM with data from the {clindata} package.
}
\keyword{datasets}
