% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consent_Analyze.R
\name{Consent_Analyze}
\alias{Consent_Analyze}
\title{Inclusion/Exclusion Assessment - Analyze}
\usage{
Consent_Analyze(dfTransformed)
}
\arguments{
\item{dfTransformed}{}
}
\value{

}
\description{
Analysis of Inclusion/Exclusion
}
