% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consent_Assess.R
\name{Consent_Assess}
\alias{Consent_Assess}
\title{Consent Assessment}
\usage{
Consent_Assess(dfInput, nThreshold = 0.5, cLabel = "", bDataList = FALSE)
}
\arguments{
\item{dfInput}{input dataset with subject information containing : (required variables)}

\item{nThreshold}{integer threshold values Flagging, integer values greater than will be flagged.}

\item{cLabel}{Assessment label}

\item{bDataList}{Should all assessment datasets be returned as a list? If False (the default), only the finding data frame is returned}
}
\value{
Finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
Consent Assessment
}
