% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consent_Transform.R
\name{Consent_Transform}
\alias{Consent_Transform}
\title{Consent - Transformed Input Data}
\usage{
Consent_Transform(dfInput)
}
\arguments{
\item{dfInput}{frame in format produced by \code{\link{Consent_Map_Raw}}}
}
\description{
DESCRIPTION
}
