% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabAbnorm_Map_Raw_orig.R
\name{CountLabAbnormalities}
\alias{CountLabAbnormalities}
\title{Count Lab Abnormalities from Raw Lab Dataset}
\usage{
CountLabAbnormalities(dfLab, dfTrt, strToxVar, optGrade = 1)
}
\arguments{
\item{dfLab}{lab dataset with columns SUBJID and rows for each lab abonrmality record}

\item{strToxVar}{directly input an ex dataset with columns SUBJID EXSTDAT EXENDAT}
}
\value{
Data frame of detected lab abnormalities
}
\description{
Counts the number of lab abnormalities in a raw lab dataset based on specified tox grading variable
Requires the following columns: SUBJID LBDTM strToxVar
And requries additional dfTrt dataframe from LabAbnorm_Map_Raw
}
