% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Country_Map_Raw.R
\name{Country_Map_Raw}
\alias{Country_Map_Raw}
\title{Count Participants by Country}
\usage{
Country_Map_Raw(dfSite = Site_Map_Raw())
}
\arguments{
\item{dfSite}{\code{data.frame} Data frame returned by \code{\link[=Site_Map_Raw]{Site_Map_Raw()}}.}
}
\value{
\code{data.frame} Data frame with one record per country.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
