% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Country_Report.R
\name{Country_Report}
\alias{Country_Report}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
Country_Report(lAssessments, strOutpath = NULL)
}
\arguments{
\item{lAssessments}{\code{list} The results of multiple assessments run using \code{Study_Assess}.}

\item{strOutpath}{\code{character} File path; location where the report will be saved.}
}
\value{
HTML report of study data.
}
\description{
Country Report
}
\details{
Create country-level HTML summary report using the results of \code{Study_Assess}, including tables, charts, and error checking.
}
