% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate_Chart.R
\name{covariate_chart_site}
\alias{covariate_chart_site}
\title{Title}
\usage{
covariate_chart_site(dfCovariate, strType = "total")
}
\arguments{
\item{strType}{one of 'total' or 'percent'}
}
\description{
Title
}
\keyword{internal}
