% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate_Chart.R
\name{covariate_chart_study}
\alias{covariate_chart_study}
\title{Make covariate study chart.}
\usage{
covariate_chart_study(dfCovariate, strType)
}
\arguments{
\item{dfCovariate}{\code{data.frame} study-level data.frame returned by \link{Covariate_Map}.}

\item{strType}{\code{character} one of "total" or "percent".}
}
\value{
\code{plotly} chart.
}
\description{
Make covariate study chart.
}
\keyword{internal}
