% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate_Charts.R
\name{Covariate_Charts}
\alias{Covariate_Charts}
\title{Create charts for distribution of kri metrics}
\usage{
Covariate_Charts(lInput)
}
\arguments{
\item{lInput}{\code{list} output from \code{Distribution_Map()} function}
}
\description{
Create charts for distribution of kri metrics
}
\keyword{internal}
