% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate_Report.R
\name{Covariate_Report}
\alias{Covariate_Report}
\title{Create and output Covariate Report}
\usage{
Covariate_Report(lSnapshot, lCovariateTables, lCovariateCharts)
}
\arguments{
\item{snapshot}{\code{list} snapshot object output from \code{Make_Snapshot()}}
}
\description{
Create and output Covariate Report
}
\keyword{Internal}
