% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate_Workflow.R
\name{Covariate_Workflow}
\alias{Covariate_Workflow}
\title{Make dataframe of covariate workflow mapping}
\usage{
Covariate_Workflow(kri = sprintf("kri00\%02d", 1:12))
}
\arguments{
\item{kri}{\code{vector} vector containing kri's to include in data frame ex. \code{c("kri0001", "kri0002")} defaults to kri's 1:12}
}
\description{
Make dataframe of covariate workflow mapping
}
