% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate_Charts.R
\name{CovariateSiteChart}
\alias{CovariateSiteChart}
\title{Create site chart for distribution of kri metrics}
\usage{
CovariateSiteChart(lInput)
}
\arguments{
\item{lInput}{\code{list} output from \code{Distribution_Map()} function}
}
\description{
Create site chart for distribution of kri metrics
}
\keyword{internal}
