% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sidebar.R
\name{create_sidebar}
\alias{create_sidebar}
\title{creates sidebar info for covariate report}
\usage{
create_sidebar(study, date, total)
}
\arguments{
\item{study}{\code{string} study name}

\item{date}{\code{string} study date}

\item{total}{\code{string} total patients in study}
}
\description{
creates sidebar info for covariate report
}
\keyword{internal}
