% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataChg_Assess.R
\name{DataChg_Assess}
\alias{DataChg_Assess}
\title{Data Change Rate Assessment}
\usage{
DataChg_Assess(
  dfInput,
  vThreshold = NULL,
  strMethod = "NormalApprox",
  lMapping = yaml::read_yaml(system.file("mappings", "DataChg_Assess.yaml", package =
    "gsm")),
  strGroup = "Site",
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data, a data frame with one record per data point.}

\item{vThreshold}{\code{numeric} Threshold specification, a vector of length 2 or 4 that defaults to \code{c(-3, -2, 2, 3)} for a Normal Approximation (\code{strMethod = "NormalApprox"}),
\code{c(.01, .05)} for Fisher's exact test (\code{strMethod = "fisher"}), and \code{c(3.491, 5.172)} for a nominal assessment (\code{strMethod = "identity"}).}

\item{strMethod}{\code{character} Statistical method. Valid values:
\itemize{
\item \code{"NormalApprox"} (default)
\item \code{"fisher"}
\item \code{"identity"}
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column. Default: package-defined Labs Assessment mapping.}

\item{strGroup}{\code{character} Grouping variable. \code{"Site"} (the default) uses the column named in \code{mapping$strSiteCol}. Other valid options using the default mapping are \code{"Study"} and \code{"CustomGroup"}.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} \code{lData}, a named list with:
\itemize{
\item each data frame in the data pipeline
\itemize{
\item \code{dfTransformed}, returned by \code{\link[=Transform_Rate]{Transform_Rate()}}
\item \code{dfAnalyzed}, returned by \code{\link[=Analyze_NormalApprox]{Analyze_NormalApprox()}}, \code{\link[=Analyze_Fisher]{Analyze_Fisher()}}, or \code{\link[=Analyze_Identity]{Analyze_Identity()}}
\item \code{dfFlagged}, returned by \code{\link[=Flag_NormalApprox]{Flag_NormalApprox()}}, \code{\link[=Flag_Fisher]{Flag_Fisher()}}, or \code{\link[=Flag]{Flag()}}
\item \code{dfSummary}, returned by \code{\link[=Summarize]{Summarize()}}
\item \code{dfBounds}, returned by \code{\link[=Analyze_NormalApprox_PredictBounds]{Analyze_NormalApprox_PredictBounds()}} when \code{strMethod == "NormalApprox"}
}
\item \code{list} \code{lCharts}, a named list with:
\itemize{
\item \code{scatter}, a ggplot2 object returned by \code{\link[=Visualize_Scatter]{Visualize_Scatter()}} only when strMethod != "identity"
\item \code{barMetric}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}}
\item \code{barScore}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}}
}
\item \code{list} \code{lChecks}, a named list with:
\itemize{
\item \code{dfInput}, a named list returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{status}, a boolean returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{mapping}, a named list that is provided as an argument to the \code{lMapping} parameter in \code{\link[=DataChg_Assess]{DataChg_Assess()}}
\item \code{spec}, a named list used to define variable specifications
}
}
}
\description{
Evaluates rate of reported data point with >1 changes.
}
\details{
The Data Change Rate Assessment uses the standard \href{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}{GSM data pipeline} to flag possible outliers. Additional details regarding the data pipeline and statistical
methods are described below.
}
\section{Data specification}{
\tabular{llll}{
   \strong{Domain} \tab \strong{Column Key} \tab \strong{Default Value} \tab \strong{Required?} \cr
   dfInput \tab strIDCol \tab SubjectID \tab TRUE \cr
   dfInput \tab strCountCol \tab Count \tab TRUE \cr
   dfInput \tab strGroupCol \tab SiteID \tab TRUE \cr
}
}

\section{Statistical assumptions}{
The default function \code{Analyze_NormalApprox} applies funnel plots using
asymptotic limits based on normal approximation of binomial distribution
for the binary outcome, or normal approximation of Poisson distribution
for the rate outcome with volume (the sample sizes or total exposure of
the sites) to assess data quality and safety.

Alternatively, the Fisher’s exact test, or percentage value (identity)
is used to generate estimates and p-values for each site (as specified
with the \code{strMethod} parameter). Those outputs are then used to flag
possible outliers using the thresholds specified in \code{vThreshold}. In the
Fisher’s exact test, sites with $p<0.05$ are flagged as 1 by default.

See \code{\link[=Analyze_NormalApprox]{Analyze_NormalApprox()}} or
\code{\link[=Analyze_Fisher]{Analyze_Fisher()}} or
\code{\link[=Analyze_Identity]{Analyze_Identity()}} for additional
details about the statistical methods and their assumptions.
}

\examples{
dfInput <- DataChg_Map_Raw()
DataChg_assessment_NormalApprox <- DataChg_Assess(dfInput, strMethod = "NormalApprox")
DataChg_assessment_fisher <- DataChg_Assess(dfInput, strMethod = "fisher")
DataChg_assessment_identity <- DataChg_Assess(dfInput, strMethod = "identity")

}
