% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataEntry_Map_Raw.R
\name{DataEntry_Map_Raw}
\alias{DataEntry_Map_Raw}
\title{Data Entry Lag - Raw Mapping}
\usage{
DataEntry_Map_Raw(
  dfs = list(dfSUBJ = clindata::rawplus_dm, dfDataEntry = clindata::edc_data_entry_lag),
  lMapping = yaml::read_yaml(system.file("mappings", "mapping_edc.yaml", package =
    "gsm")),
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frame:
\itemize{
\item \code{dfSUBJ}: \code{data.frame} Subject-level data with one record per participant.
\item \code{dfDataEntry}: \code{data.frame} Data-Point-level data with one record per data entry.
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{\link[=is_mapping_valid]{is_mapping_valid()}}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one record per subject, the input to \code{\link[=DataEntry_Assess]{DataEntry_Assess()}}. If
\code{bReturnChecks} is \code{TRUE} \code{DataEntry_Map_Raw} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert raw data entry data to formatted input data to \code{\link[=DataEntry_Assess]{DataEntry_Assess()}}.
}
\details{
\code{DataEntry_Map_Raw} creates an input dataset for the Data Entry Lag Assessment \code{\link[=DataEntry_Assess]{DataEntry_Assess()}} by adding
Data Entry Lag Counts (derived from \code{dfDataEntry}) to basic subject-level data (from \code{dfSUBJ}).
}
\section{Data specification}{
\tabular{lllll}{
   \strong{Domain} \tab \strong{Column Key} \tab \strong{Default Value} \tab \strong{Required?} \tab \strong{Require Unique Values?} \cr
   dfSUBJ \tab strSiteCol \tab siteid \tab TRUE \tab FALSE \cr
   dfSUBJ \tab strIDCol \tab subjid \tab TRUE \tab TRUE \cr
   dfDataChg \tab strIDCol \tab  \tab TRUE \tab FALSE \cr
   dfDataChg \tab strDataPointsCol \tab  \tab TRUE \tab FALSE \cr
   dfDataChg \tab strDataPointsChangeCol \tab  \tab TRUE \tab FALSE \cr
}
}

\examples{
# Run with defaults.
dfInput <- DataEntry_Map_Raw()

# Run with error checking and message log.
dfInput <- DataEntry_Map_Raw(bReturnChecks = TRUE, bQuiet = FALSE)

}
