% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disc_Bar_Plot.R
\name{Disc_Bar_Plot}
\alias{Disc_Bar_Plot}
\title{generates plotly bar plot of study discontinuation information for discontinuation report}
\usage{
Disc_Bar_Plot(dfInput, strGroup)
}
\arguments{
\item{strGroup}{\code{string} type of bar plot to output option either 'Study' or 'Site}

\item{dfInput_study}{\code{data.frame} output from \code{Disc_Study_Map()}}
}
\description{
generates plotly bar plot of study discontinuation information for discontinuation report
}
\keyword{internal}
