% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disc_Report.R
\name{Disc_Report}
\alias{Disc_Report}
\title{Create and output Discontinuation Report}
\usage{
Disc_Report(
  dfSUBJ = clindata::rawplus_dm,
  dfSTUDCOMP = clindata::rawplus_studcomp,
  dfSDRGCOMP = clindata::rawplus_sdrgcomp,
  lMapping = gsm::Read_Mapping("rawplus"),
  strMethod = "NormalApprox"
)
}
\arguments{
\item{dfSUBJ}{\code{data.frame} raw data of patient demographics defaults to \code{clindata::rawplus_dm}}

\item{dfSTUDCOMP}{\code{data.frame} raw data of study completion defaults to \code{clindata::rawplus_studcomp}}

\item{dfSDRGCOMP}{\code{data.frame} raw data of treatment completion defaults to \code{clindata::rawplus_sdrgcomp}}

\item{lMapping}{\code{list} the mapping to use for the data frames provided defaults to \code{gsm::Read_Mapping("rawplus")}}

\item{strMethod}{\code{string} the method to use to calculate KRI metrics options: 'NormalApprox' (default), 'Fisher', 'Identity', or 'QTL'}
}
\description{
Create and output Discontinuation Report
}
