% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disc_Study_Map.R
\name{Disc_Study_Map}
\alias{Disc_Study_Map}
\title{Map study completion data to output for discontinuation report}
\usage{
Disc_Study_Map(dfSTUDCOMP, lMapping)
}
\arguments{
\item{dfSTUDCOMP}{\code{data.frame} study completion data to map}

\item{lMapping}{\code{list} mapping variables to use}
}
\description{
Map study completion data to output for discontinuation report
}
\keyword{internal}
