% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disc_Treatment_Map.R
\name{Disc_Treatment_Map}
\alias{Disc_Treatment_Map}
\title{Map treatment completion data to output for discontinuation report}
\usage{
Disc_Treatment_Map(dfSDRGCOMP, lMapping)
}
\arguments{
\item{dfSDRGCOMP}{\code{data.frame} treatment completion data to map}

\item{lMapping}{\code{list} mapping variables to use}
}
\description{
Map treatment completion data to output for discontinuation report
}
\keyword{internal}
