% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disp_Map.R
\name{Disp_Map}
\alias{Disp_Map}
\title{Disposition Map}
\usage{
Disp_Map(dfDisp = NULL, strCol = NULL, strReason = "any")
}
\arguments{
\item{dfDisp}{demographics data with the following required columns: SUBJID and SITEID.}

\item{strCol}{column name containing discontinuation reason.}

\item{strReason}{character vector of length 1 containing a single reason to run Disposisition Assessment on.}
}
\value{
Data frame with one record per person with columns: SubjectID, SiteID, Count, and the value passed to strCol.

#' @examples
df <- Disp_Map(dfDisp = safetyData::adam_adsl, strCol = "DCREASCD", strReason = "adverse event")
}
\description{
Convert from ADaM or raw format to input format for Disposition Assessment.
}
\section{Data Specification}{


This function creates an input dataset for the Disposition Assessment (link to code) by adding Discontinuation Reason Counts to basic subject-level data.

The following columns are required:
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{SITEID} - Site ID
}
}

