% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Map.R
\name{Distribution_Map}
\alias{Distribution_Map}
\title{Map through raw data for distribution of causes checked in kri's}
\usage{
Distribution_Map(lSnapshot, dist_list, lMapping)
}
\arguments{
\item{lSnapshot}{\code{list} a snapshot object from \code{Make_Snapshot()}}

\item{dist_list}{\code{list} a list of data and linked category columns}

\item{lMapping}{\code{list} a list of predefined columns to map over}
}
\description{
Map through raw data for distribution of causes checked in kri's
}
\keyword{internal}
