% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\name{example_groups}
\alias{example_groups}
\title{Load Default Clindata Groups}
\usage{
example_groups()
}
\value{
A data frame with group data.
}
\description{
Load and clean group data for use as default values.
}
\examples{
example_groups()
}
