% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\name{example_metrics}
\alias{example_metrics}
\title{Load the meta_workflow Dataset}
\usage{
example_metrics()
}
\value{
The \code{meta_workflow} dataset.
}
\description{
A convenience function to load the \code{\link{meta_workflow}} dataset.
}
\examples{
example_metrics()
}
