% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ExtractCharts.R
\name{ExtractCharts}
\alias{ExtractCharts}
\title{Extract charts from \verb{gsm v1.8.4} and move to top-level.}
\usage{
ExtractCharts(lResults)
}
\description{
Extract charts from \verb{gsm v1.8.4} and move to top-level.
}
\keyword{internal}
