% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ExtractDirectoryPaths.R
\name{ExtractDirectoryPaths}
\alias{ExtractDirectoryPaths}
\title{Extract a vector of directory paths that contain a specified file name.}
\usage{
ExtractDirectoryPaths(
  cPath,
  file,
  include.partial.match = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cPath}{\code{character} Path to longitudinal data folders.}

\item{file}{\code{character} File name that must be in the folders to qualify for selection}

\item{include.partial.match}{\code{logical} TRUE = include directory path if any files of interest are found, FALSE = only include directories that contain all files of interest.}

\item{verbose}{\code{logical} TRUE = display warnings, FALSE = don't display warnings}
}
\value{
\code{vector} containing paths to directories that contain the file of interest.
}
\description{
Extract a vector of directory paths that contain a specified file name.
}
\keyword{internal}
