% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{ExtractFlags}
\alias{ExtractFlags}
\title{Extract flag counts by site or KRI}
\usage{
ExtractFlags(lResults, group)
}
\arguments{
\item{lResults}{\code{list} the output of \code{Study_Assess()} containing results of kri analysis}

\item{group}{\code{character} a character field to specify what to use to group flags by. options = "site", "kri"}
}
\description{
Extract flag counts by site or KRI
}
\keyword{internal}
