% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ExtractSnapshotDate.R
\name{ExtractSnapshotDate}
\alias{ExtractSnapshotDate}
\title{Extract the snapshot dates of snapshot directories.}
\usage{
ExtractSnapshotDate(snapshot_paths)
}
\arguments{
\item{snapshot_paths}{\code{vector} a vector of Paths that contain snapshots, output of \code{ExtractDirectoryPaths}.}
}
\value{
\code{data.frame} a data.frame containing the foldername and the associated snapshot date.
}
\description{
Extract the snapshot dates of snapshot directories.
}
