% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-Report.R
\name{filter_by_latest_SnapshotDate}
\alias{filter_by_latest_SnapshotDate}
\title{Filter by Latest Snapshot Date}
\usage{
filter_by_latest_SnapshotDate(df, strSnapshotDate = NULL)
}
\arguments{
\item{df}{A data frame containing the results.}

\item{strSnapshotDate}{A character string representing the snapshot date.}
}
\value{
A data frame containing the results for the most recent snapshot date.
}
\description{
Filter a data frame to the most recent snapshot date.
}
\examples{
reportingResults_latest <- filter_by_latest_SnapshotDate(reportingResults)

}
