% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Rate.R
\name{filterByStr}
\alias{filterByStr}
\title{Filter by a String Expression}
\usage{
filterByStr(df, strFilter)
}
\arguments{
\item{df}{The df to filter.}

\item{strFilter}{The filter expression as a comma-separated string or a
character vector.}
}
\value{
The filtered df.
}
\description{
Filter by a String Expression
}
\keyword{internal}
