% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-FilterData.R
\name{FilterData}
\alias{FilterData}
\title{Subset a data frame}
\usage{
FilterData(dfInput, strCol, anyVal, bReturnChecks = FALSE, bQuiet = TRUE)
}
\arguments{
\item{dfInput}{\code{data.frame} A data.frame to be filtered, likely within a mapping function.}

\item{strCol}{\code{character} Domain in \code{lMapping} that references the column to filter on.}

\item{anyVal}{\code{character} Domain in \code{lMapping} that references the value to filter on.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{is_mapping_valid}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame provided as \code{dfInput} and filtered on \code{strCol} == \code{anyVal}.
}
\description{
\code{FilterData} is used to subset strata for a stratified KRI. This function is called within \code{RunStep()} to create
separate strata for a categorical variable (e.g, levels of \code{aetoxgr} for an Adverse Events assessment).
}
\examples{

subset <- FilterData(
  dfInput = clindata::rawplus_ae,
  strCol = "aeser",
  anyVal = "N"
)

}
