% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-FilterDomain.R
\name{FilterDomain}
\alias{FilterDomain}
\title{Utility function for basic filtering}
\usage{
FilterDomain(
  df,
  strDomain,
  lMapping,
  strColParam,
  strValParam,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{df}{data.frame be filtered}

\item{lMapping}{mapping}

\item{strColParam}{column to filter}

\item{strValParam}{value or values to keep}

\item{bReturnChecks}{checks}

\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}
}
\description{
Utility function for basic filtering
}
\examples{
te_ae <- FilterDomain(df = dfAE,
                      lMapping = lMapping,
                      strDomain = "dfAE",
                      strColParam = "strTreatmentEmergentCol",
                      strValParam = "strTreatmentEmergentVal")

}
