% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag_Fisher.R
\name{Flag_Fisher}
\alias{Flag_Fisher}
\title{Flag_Fisher}
\usage{
Flag_Fisher(dfAnalyzed, vThreshold = NULL)
}
\arguments{
\item{dfAnalyzed}{data.frame where flags should be added.}

\item{vThreshold}{Vector of 2 numeric values representing lower and upper p-value thresholds.}
}
\value{
\code{data.frame} with "Flag" column added
}
\description{
Add columns flagging sites that represent possible statistical outliers.
}
\details{
This function provides a generalized framework for flagging sites as part of the \href{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}{GSM data pipeline}.
}
\section{Data Specification}{

\code{Flag} is designed to support the input data (\code{dfAnalyzed}) from many different
\code{Analyze} functions. At a minimum, the input data must have a \code{SiteID} column and a column
of numeric values (identified by the \code{strColumn} parameter) that will be compared to the
specified thresholds (\code{vThreshold}) to calculate a new \code{Flag} column. Optionally, a second column
of numeric values (identified by \code{strValueColumn}) can be specified to set the directionality of
the \code{Flag}.

In short, the following columns are considered:
\itemize{
\item \code{GroupID} - Group ID (required)
\item \code{strColumn} - A column to use for Thresholding (required)
\item 'strValueColumn' - A column to be used for the sign of the flag (optional)
}
}

\examples{
dfInput <- AE_Map_Adam()

dfTransformed <- Transform_Rate(dfInput,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)

dfAnalyzed <- Analyze_Poisson(dfTransformed)

dfFlagged <- Flag(dfAnalyzed, vThreshold = c(-5, 5))


}
