% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag_HighLow.R
\name{Flag_HighLow}
\alias{Flag_HighLow}
\title{Safety Assessment - Make data frame with flagged values}
\usage{
Flag_HighLow(dfAnalyzed, dThresholdHi, dThresholdLo)
}
\arguments{
\item{dThresholdHi}{Higher threshold to be compared}

\item{dThresholdLo}{Lower threshold to be compared}

\item{dfFlagged}{data frame in format produced by \code{\link{AE_Poisson_Analyze}}}
}
\value{
input data frame with the following columns added for "Threshold", "FlagOver" and "FlagUnder"
}
\description{
Adds columns flagging sites that represent possible statistical outliers
}
