% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag_QTL.R
\name{Flag_QTL}
\alias{Flag_QTL}
\title{Flag QTL}
\usage{
Flag_QTL(dfAnalyzed, vThreshold = NULL)
}
\arguments{
\item{dfAnalyzed}{data.frame where flags should be added.}

\item{vThreshold}{Numeric value representing the threshold. NA and NaN values in strColumn are given NA flag values.}
}
\value{
\code{data.frame} with "Flag" column added
}
\description{
Flag QTL
}
\details{
This function flags a study-level QTL metric by comparing the mean and lower confidence bound for the metric to a single threshold.
}
\section{Data Specification}{

\code{Flag_NormalApprox} is designed to support the input data (\code{dfAnalyzed}) from \code{Analyze_QTL} function.
At a minimum, the input data must have numeric \code{Score} and \code{LowCI} columns that will be compared to the specified threshold (\code{vThreshold}) to
calculate a new \code{Flag} column.
}

\examples{
dfInput <- Disp_Map_Raw()
dfTransformed <- Transform_Rate(
  dfInput,
  strGroupCol = "StudyID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Total"
)
dfAnalyzed <- Analyze_QTL(dfTransformed)
dfFlagged <- Flag_QTL(dfAnalyzed, vThreshold = 0.02)

}
