% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{Flags_by_site}
\alias{Flags_by_site}
\title{Create Study Results table for Report}
\usage{
Flags_by_site(lResults)
}
\arguments{
\item{lResults}{\code{list} the output of \code{Study_Assess()} containing results of kri analysis}
}
\description{
Create Study Results table for Report
}
\keyword{internal}
