% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Enrolled.R
\name{Get_Enrolled}
\alias{Get_Enrolled}
\title{Get_Enrolled}
\usage{
Get_Enrolled(dfSUBJ, dfConfig, lMapping, strUnit, strBy)
}
\arguments{
\item{strBy}{}
}
\description{
Derive number of enrolled participants or sites by a given denominator (study or site).
}
\examples{

Get_Enrolled(dfSUBJ = clindata::rawplus_dm,
             dfConfig = config_param,
             lMapping = yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package = "gsm")),
             strUnit = "participant",
             strBy = "site")

}
