% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClass.R
\name{GetClass}
\alias{GetClass}
\title{A helper function for ensuring testing and joining operations between data frames are identical in terms of column and classes}
\usage{
GetClass(df)
}
\arguments{
\item{df}{the data frame to analyze}
}
\value{
a \code{data.frame} of column names and associated data classes
}
\description{
A helper function for ensuring testing and joining operations between data frames are identical in terms of column and classes
}
\keyword{internal}
