% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-Logger.R
\name{GetLogger}
\alias{GetLogger}
\title{Custom logger retriever
Function that retries a custom logger object}
\usage{
GetLogger(strName = "gsm")
}
\arguments{
\item{strName}{string of logger object created from \code{set_logger()}}
}
\value{
logger object
}
\description{
Custom logger retriever
Function that retries a custom logger object
}
