% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Helpers.R
\name{GetSnapshotDate}
\alias{GetSnapshotDate}
\title{Extrapolate study snapshot date and number of patients in study}
\usage{
GetSnapshotDate(status_study)
}
\arguments{
\item{status_study}{\code{df} a dataframe containing status of study pulled from \code{params$status_study} in report}
}
\description{
Extrapolate study snapshot date and number of patients in study
}
\keyword{internal}
