% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-func-namespace.R
\name{GetStrFunctionIfNamespaced}
\alias{GetStrFunctionIfNamespaced}
\title{GetFunctionIfNamespaced}
\usage{
GetStrFunctionIfNamespaced(strFunction)
}
\arguments{
\item{strFunction}{the function to be called}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function looks to see if a strFunction is namespaced and looks it up
allowing the do.call in run step to process correctly
this will either return a function, or the original string if not namespaced
}
\examples{
fn <- GetStrFunctionIfNamespaced("dplyr::glimpse")
fn(head(Theoph))
}
