% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-gt.R
\name{gt-shared}
\alias{gt-shared}
\title{Shared gt arguments}
\arguments{
\item{data}{A \code{gt_tbl} object.}

\item{columns}{Columns to target for formatting.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Reusable definitions of common gt arguments.
}
\keyword{internal}
