% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-gt.R
\name{gt_style}
\alias{gt_style}
\title{Shared formats for gt tables}
\usage{
gt_style(data)
}
\arguments{
\item{data}{A \code{gt_tbl} object.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Apply standardized formatting to gt tables.
}
\keyword{internal}
