% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE_Analyze.R
\name{IE_Analyze}
\alias{IE_Analyze}
\title{Inclusion/Exclusion Assessment - Analyze}
\usage{
IE_Analyze(dfTransformed)
}
\arguments{
\item{dfTransformed}{}
}
\value{

}
\description{
Analysis of Inclusion/Exclusion
}
