% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-Ingest.R
\name{Ingest}
\alias{Ingest}
\title{Ingests data from a source based on a given specification.}
\usage{
Ingest(lSource, lSpec)
}
\arguments{
\item{lSource}{A list of source data frames.}

\item{lSpec}{A list of column specifications.}
}
\value{
A list of data frames, where each data frame corresponds to a domain in the specification.
}
\description{
Ingests data from a source based on a given specification.
}
\details{
This function ingests data from a source based on a given specification. It iterates over each domain in the specification and checks if the columns exist in the source data. It then constructs a query to select the columns from the source and calls the \code{RunQuery} function to retrieve the data. The resulting data frames are stored in a list, where each data frame corresponds to a domain in the specification.
}
