% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{input_data_schema}
\alias{input_data_schema}
\title{Input Data Schema}
\format{
A data frame with 77 rows and 8 variables.
#' \describe{
\item{Source}{clinical, ctms, or edc.}
\item{Domain}{Name of the Raw+ data domain.}
\item{Database Table}{Name of Database Table.}
\item{{clindata} Name}{Name of the data source as it is referenced in \code{clindata}, e.g., "dfNAME".}
\item{GSM_Domain_Key}{Name of \code{gsm} key that is paired with \code{column}.}
\item{Column Description}{Description of \code{Column}.}
\item{Column}{Raw+ column name.}
\item{GSM_Column_Key}{Column key as it is referred to in mapping.}
}
}
\usage{
input_data_schema
}
\description{
A dataset containing key/column pairs for clinical, ctms, and edc mappings.
}
\keyword{datasets}
