% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_Rate.R
\name{Input_Rate}
\alias{Input_Rate}
\title{Input_Rate}
\usage{
Input_Rate(
  dfs = NULL,
  strNumeratorMethod = "Count",
  strDenominatorMethod = "Count",
  strNumeratorCol = NULL,
  strDenominatorCol = NULL
)
}
\arguments{
\item{dfs}{\code{list} of \code{data.frame}s including dfSUBJ, dfNumerator, and dfDenominator. Default: NULL}

\item{strNumeratorMethod}{\code{character} Method to calculate numerator. Default: "Count"}

\item{strDenominatorMethod}{\code{character} Method to calculate denominator. Default: "Count"}

\item{strNumeratorCol}{\code{character} Column name in \code{dfNumerator} to use for numerator calculation. Default: NULL}

\item{strDenominatorCol}{\code{character} Column name in \code{dfDenominator} to use for denominator calculation. Default: NULL}
}
\value{
\code{data.frame} with the following columns:
\itemize{
\item SubjectID: The subject ID
\item SiteID: The site ID
\item StudyID: The study ID
\item CountryID: The country ID
\item Numerator: The calculated numerator value
\item Denominator: The calculated denominator value
\item Rate: The calculated input rate
}
}
\description{
Calculate a subject level rate from numerator and denominator data

This function takes in a list of data frames including dfSUBJ, dfNumerator, and dfDenominator,
and calculates a subject level rate based on the specified numerator and denominator methods (either
"Count" or "Sum"). If the method is "Count", the function simply counts the number of rows in the
provided data frame. If the numerator method is "Sum", the function takes the sum of the values in
the specified column (strNumeratorCol or strDenominatorCol). The function returns a data frame with
the calculated input rate for each subject.

The data requirements for the function are as follows:
\itemize{
\item dfSubjects: A data frame with columns for SubjectID, SiteID, StudyID, and CountryID
\item dfNumerator: A data frame with a column for SubjectID and \code{strNumeratorCol} if \code{strNumeratorMethod} is "Sum"
\item dfDenominator: A data frame with a column for SubjectID and \code{strDenominatorCol} if \code{strDenominatorMethod} is "Sum"
}

All other columns are dropped from the output data frame. Note that if no values for a subject are
found in dfNumerator/dfDenominator numerator and denominator values are filled with 0 in the output data frame.
}
\keyword{interal}
