% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Report.R
\name{KRI_Report}
\alias{KRI_Report}
\title{KRI_Report function}
\usage{
KRI_Report(
  lCharts = NULL,
  dfSummary = NULL,
  dfStudy = NULL,
  dfSite = NULL,
  dfMetrics = NULL,
  strOutpath = NULL
)
}
\arguments{
\item{lCharts}{A list of charts to include in the report.}

\item{dfSummary}{A data frame containing summary information.}

\item{dfStudy}{A data frame containing study metadata (e.g., from CTMS).}

\item{dfSite}{A data frame containing site metadata (e.g., from CTMS).}

\item{dfMetrics}{A data frame containing metric metadata (e.g., from workflows).}

\item{strOutpath}{The output path for the generated report. If not provided, the report will be saved in the current working directory with the name "kri_report.html".}
}
\value{
None
}
\description{
This function generates a KRI report based on the provided inputs.
}
\keyword{KRI}
\keyword{report}
