% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabAbnorm_Assess.R
\name{LabAbnorm_Assess}
\alias{LabAbnorm_Assess}
\title{Lab Abnormality Assessment}
\usage{
LabAbnorm_Assess(
  dfInput,
  vThreshold = NULL,
  cLabel = "",
  cMethod = "poisson",
  bDataList = FALSE
)
}
\arguments{
\item{dfInput}{input data}

\item{cLabel}{Assessment label}

\item{bDataList}{Should all assessment datasets be returned as a list? If False (the default), only the finding data frame is returned}

\item{lThreshold}{list of threshold values}

\item{nCutoff}{optional parameter to control the auto-thresholding}
}
\value{
Finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
Lab Abnormality Assessment
}
