% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabAbnorm_Map_Adam.R
\name{LabAbnorm_Map_Adam}
\alias{LabAbnorm_Map_Adam}
\title{LabAbnorm Assessment - ADaM Mapping}
\usage{
LabAbnorm_Map_Adam(dfADSL, dfADLB, mapping = NULL)
}
\arguments{
\item{dfADSL}{ADaM demographics data with the following required columns:  USUBJID, SITEID, TRTEDT (end date), TRTSDT (start date)}

\item{dfADLB}{ADaM Lab data with the following required columns: USUBJID}

\item{mapping}{List containing expected columns in each data set. By default, mapping for dfADLB is: \code{strIDCol} = "SUBJID". By default, mapping for dfRDSL is: \code{strIDCol} = "SubjectID", \code{strSiteCol} = "SiteID", and \code{strExposureCol} = "TimeOnTreatment". TODO: add more descriptive info or reference to mapping.}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count (Number of Adverse Events), Exposure (Time on Treatment in Days), Rate (AEs/Day)
}
\description{
Convert from ADaM format to input format for Lab event Assessment.
}
\details{
This function maps from ADaM ADSL and ADLB data to the required input for \code{\link{LabAbnorm_Assess}}.
}
\section{Data Specification}{


This function creates an input dataset for the Lab Measurements Assessment (\code{\link{LabAbnorm_Assess}}) by adding Lab Measurements Counts (from \code{dfADLB}) to basic subject-level data (from \code{dfADSL}).

The following columns are required:
\itemize{
\item \code{dfADSL}
\itemize{
\item \code{USUBJID} - Unique subject ID
\item \code{SITEID} - Site ID
\item \code{TRTEDT} - Treatment End date
\item \code{TRTSDT} - Treatment Start date
}
\item \code{dfADLB}
\itemize{
\item \code{USUBJID} - Unique subject ID
}
}

Note that the function can generate data summaries for specific types of Lab Measurements by passing filtered ADAE data to dfADAE.
}

\examples{
dfInput <- LabAbnorm_Map_Adam( safetyData::adam_adsl, safetyData::adam_adlbc )

}
