% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabAbnorm_Map_Raw_orig.R
\name{LabAbnorm_Map_Raw_orig}
\alias{LabAbnorm_Map_Raw_orig}
\title{Lab Abnormality Assessment Mapping from Raw Data- Make Input Data}
\usage{
LabAbnorm_Map_Raw_orig(
  strPath = NULL,
  dfSubid = NULL,
  dfEx = NULL,
  dfDa = NULL,
  dfLab = NULL,
  strToxVar = "TOXGRG",
  optTE = FALSE,
  optGrade = 1
)
}
\arguments{
\item{strPath}{path to the raw datasets}

\item{dfSubid}{directly input a subid dataset with columns SUBJID INVID}

\item{dfEx}{directly input an ex dataset with columns SUBJID EXSTDAT EXENDAT}

\item{dfDa}{directly input an ex dataset with columns SUBJID DADPDAT DARETDAT}

\item{dfLab}{directly input an lab dataset with columns SUBJID and rows for each lab abonrmality record}

\item{strToxVar}{variable to use for toxicity grading info in lab dataset}

\item{optTE}{option to restrict lab dataset to TE lab abnormalities defined as occuring on or after first dose date}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count, Exposure, Rate, Unit
}
\description{
Convert from raw data format to needed input format for Safety Assessment
Requires the following raw datasets: subid, ex, lab dataset, da (optional)
}
