% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabAbnorm_Validate_Raw_orig.R
\name{LabAbnorm_Validate_Raw}
\alias{LabAbnorm_Validate_Raw}
\title{Lab abnormal Assessment - check validity of raw data}
\usage{
LabAbnorm_Validate_Raw(raw)
}
\arguments{
\item{raw}{a list of dataframe including the following data frames: subid, ex, covlab}
}
\value{
a list containing validation results
}
\description{
Lab abnormal Assessment - check validity of raw data
}
