% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB_Assess.R
\name{LB_Assess}
\alias{LB_Assess}
\title{LB_Assess Function}
\usage{
LB_Assess(
  dfInput,
  vThreshold = NULL,
  strMethod = "NormalApprox",
  lMapping = yaml::read_yaml(system.file("mappings", "LB_Assess.yaml", package = "gsm")),
  strGroup = "Site",
  nMinDenominator = NULL,
  bMakeCharts = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{The input data frame containing the data to be assessed.}

\item{vThreshold}{The threshold values used for flagging abnormality.}

\item{strMethod}{The method used for analysis, which can be "NormalApprox", "Fisher", or "Identity".}

\item{lMapping}{A mapping object used for data checking.}

\item{strGroup}{The grouping variable used for analysis.}

\item{nMinDenominator}{The minimum denominator value for calculating summary statistics.}

\item{bMakeCharts}{\code{logical} Boolean value indicating whether to create charts.}

\item{bQuiet}{A logical value indicating whether to display progress messages.}
}
\value{
A list containing the transformed data, analyzed data, flagged data, summary data, and charts.
}
\description{
The LB_Assess function performs an assessment on a given input data frame. It checks the validity of the input parameters, sets thresholds and flagging parameters based on the chosen method, and then runs the assessment using the specified method. It returns a list containing the transformed data, analyzed data, flagged data, summary data, and charts.
}
\examples{
df <- data.frame(
  Site = c("A", "A", "B", "B"),
  Count = c(10, 5, 8, 4),
  Total = c(100, 200, 150, 300)
)

result <- LB_Assess(df, vThreshold = c(-2, 2), strMethod = "NormalApprox", strGroup = "Site")

print(result$lData$dfTransformed)
print(result$lData$dfAnalyzed)
print(result$lData$dfFlagged)
print(result$lData$dfSummary)
print(result$lCharts$scatter)
print(result$lCharts$barMetric)
print(result$lCharts$barScore)

}
