% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log.R
\name{Log}
\alias{Log}
\title{Log output.}
\usage{
Log(strFileName = NULL)
}
\arguments{
\item{strFileName}{\code{character} File name or full path to a directory to save the log file.}
}
\description{
Log output.
}
\keyword{internal}
