% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cli_logger_wrappers.R
\name{LogMessage}
\alias{LogMessage}
\title{Custom logging function that wraps cli messaging}
\usage{
LogMessage(level, message, cli_detail)
}
\arguments{
\item{level}{logger levels}

\item{message}{message to display}

\item{cli_detail}{for cli style alerts the detail for info}
}
\description{
Custom logging function that wraps cli messaging
}
