% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Covariate.R
\name{Make_Covariate}
\alias{Make_Covariate}
\title{Make Covariate Data}
\usage{
Make_Covariate(lSnapshot)
}
\arguments{
\item{lSnapshot}{\code{list} a snapshot object from \code{Make_Snapshot()}}
}
\description{
Make Covariate Data
}
