% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Input_Rate.R
\name{Make_Input_Rate}
\alias{Make_Input_Rate}
\title{Make_Input_Rate}
\usage{
Make_Input_Rate(
  dfs = NULL,
  strNumeratorMethod = "Count",
  strDenominatorMethod = "Count",
  strNumeratorCol = NULL,
  strDenominatorCol = NULL,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} of \code{data.frame}s including dfSUBJ, dfNumerator, and dfDenominator. Default: NULL}

\item{strNumeratorMethod}{\code{character} Method to calculate numerator. Default: "Count"}

\item{strDenominatorMethod}{\code{character} Method to calculate denominator. Default: "Count"}

\item{strNumeratorCol}{\code{character} Column name in \code{dfNumerator} to use for numerator calculation. Default: NULL}

\item{strDenominatorCol}{\code{character} Column name in \code{dfDenominator} to use for denominator calculation. Default: NULL}

\item{strIDCol}{\code{character} Column name in to use for subject ID - should be included in all \code{dfs}. Default: NULL}
}
\value{
\code{data.frame} with calculated input rate
}
\description{
Calculate a subject level rate from numerator and denominator data
}
\examples{
# Run for AE KRI
dfInput <- Make_Input_Rate(
 dfs = list(
  dfSubjects = clindata::rawplus_dm,
  dfNumerator = clindata::rawplus_ae,
  dfDenominator = clindata::rawplus_dm
),
strIDCol = "subjid",
strNumeratorMethod = "Count",
strDenominatorMethod = "Sum",
strDenominatorCol = "timeontreatment"
)

}
