% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plot_Data.R
\name{Make_Plot_Data}
\alias{Make_Plot_Data}
\title{make plot data for bar chart}
\usage{
Make_Plot_Data(dfInput, strGroup)
}
\arguments{
\item{dfInput}{\code{data.frame} data created with}

\item{strGroup}{\code{string} Site or Study level specification}
}
\description{
make plot data for bar chart
}
\keyword{internal}
