% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Snapshot.R
\name{Make_Snapshot}
\alias{Make_Snapshot}
\title{Make Snapshot}
\usage{
Make_Snapshot(
  lAssessment,
  lSnapshotPrevious = NULL,
  strAnalysisDate = NULL,
  lMeta = list(config_param = gsm::config_param, config_workflow = gsm::config_workflow,
    meta_params = gsm::meta_param, meta_site = clindata::ctms_site, meta_study =
    clindata::ctms_study, meta_workflow = gsm::meta_workflow),
  lData = list(dfSUBJ = clindata::rawplus_dm, dfAE = clindata::rawplus_ae, dfPD =
    clindata::ctms_protdev, dfCONSENT = clindata::rawplus_consent, dfIE =
    clindata::rawplus_ie, dfLB = clindata::rawplus_lb, dfSTUDCOMP =
    clindata::rawplus_studcomp, dfSDRGCOMP = clindata::rawplus_sdrgcomp \%>\%
    filter(.data$phase == "Blinded Study Drug Completion"), dfDATACHG =
    clindata::edc_data_points, dfDATAENT = clindata::edc_data_pages, dfQUERY =
    clindata::edc_queries, dfENROLL = clindata::rawplus_enroll),
  lMapping = Read_Mapping(),
  lAssessments = MakeWorkflowList(lMeta = lMeta)
)
}
\arguments{
\item{lAssessment}{\code{list}}

\item{lSnapshotPrevious}{\code{list}}

\item{lMeta}{\code{list}}

\item{lData}{\code{list}}
}
\value{
\code{list}
}
\description{
Make Snapshot
}
\examples{
\dontrun{
lSnapshot <- Make_Snapshot()
}

}
