% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timeline.R
\name{Make_Timeline}
\alias{Make_Timeline}
\title{Create interactive timeline for study data}
\usage{
Make_Timeline(study, n_breaks = 10, date_format = "\%b\\n\%Y")
}
\arguments{
\item{study}{the table containing study data of interest}

\item{n_breaks}{the number of breaks to include in the x-axis. NOTE pretty breaks are used to assign break locations and may give more or less that the number of breaks specified.}

\item{date_format}{The format to apply to the x-axis labels}
}
\value{
\code{plot} interactive timeline plot.
}
\description{
Create interactive timeline for study data
}
\examples{
Make_Timeline(clindata::ctms_study)

}
