% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeAssessmentList.R
\name{makeAssessmentList}
\alias{makeAssessmentList}
\title{Load assessments from a package/directory}
\usage{
makeAssessmentList(path = "assessments", package = "gsm")
}
\arguments{
\item{path}{location of assessment yaml files. If package is specified looks in /inst folder.}

\item{package}{package with assessments}
}
\value{
A list of assessments
}
\description{
Load assessments from a package/directory
}
\details{
Coming soon
}
\examples{
 makeAssessmentList(path="assessments", package="gsm")

}
