% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeBounds.R
\name{MakeBounds}
\alias{MakeBounds}
\title{Calculate Bounds for display in scatterplots}
\usage{
MakeBounds(
  dfResults,
  dfMetrics,
  strMetrics = unique(dfResults$MetricID),
  dSnapshotDate = unique(dfResults$SnapshotDate),
  strStudyID = unique(dfResults$StudyID)
)
}
\arguments{
\item{dfResults}{Results data.}

\item{dfMetrics}{Metric metadata including thresholds and analysis type.}

\item{strMetrics}{Character vector of \code{MetricID}s to include in \code{dfBounds}.
All unique values from \code{dfResults$MetricID} used by default.}

\item{dSnapshotDate}{Snapshot date. Uses \code{dfResults$SnapshotDate} by default.
If more than one snapshot date is found in \code{dfResults}, a warning is thrown
and this function returns \code{NULL}.}

\item{strStudyID}{Study ID. Uses \code{dfResults$StudyID} by default. If more than
one snapshot date is found in \code{dfResults}, a warning is thrown and this
function returns \code{NULL}.}
}
\value{
A data frame.
}
\description{
Calculate bounds across a set of metrics.
}
\examples{
dfBounds <- MakeBounds(
  dfResults = reportingResults, dfMetrics = reportingMetrics
)

}
