% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeCharts.R
\name{MakeCharts}
\alias{MakeCharts}
\title{Helper function create charts for multiple metrics}
\usage{
MakeCharts(dfResults, dfBounds, dfGroups, dfMetrics)
}
\arguments{
\item{dfResults}{A data frame with the results of the simulation.}

\item{dfBounds}{A data frame with the bounds of the simulation.}

\item{dfGroups}{A data frame with the groups of the simulation.}

\item{dfMetrics}{A data.frame of metric metadata, captured at the top of metric workflows and returned by \code{\link[=MakeMetric]{MakeMetric()}}}
}
\value{
A list of charts for each metric.
}
\description{
Helper function create charts for multiple metrics
}
