% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-CovariateReportHelpers.R
\name{MakeCovariateDashboard}
\alias{MakeCovariateDashboard}
\title{Helper functions for the covariate report.}
\usage{
MakeCovariateDashboard(
  lSnapshot,
  lCovariateTables,
  lCovariateCharts,
  strKRI,
  elementId = NULL
)
}
\arguments{
\item{lSnapshot}{\code{list} a snapshot object from \code{Make_Snapshot()}}

\item{lCovariateTables}{\code{list} Output of \link{Make_Covariate}$lCovariateTables}

\item{lCovariateCharts}{\code{list} Output of \link{Make_Covariate}$lCovariateCharts}

\item{strKRI}{\code{character} KRI ID.}
}
\description{
Helper functions for the covariate report.
}
\keyword{internal}
