% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeDfConfig.R
\name{MakeDfConfig}
\alias{MakeDfConfig}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
MakeDfConfig(
  strMethod,
  strGroup,
  strAbbreviation,
  strMetric,
  strNumerator,
  strDenominator,
  vThreshold
)
}
\arguments{
\item{strMethod}{\code{character} Statistical method used in the \verb{*_Assess()} function.}

\item{strGroup}{\code{character} Grouping variable used in the \verb{*_Assess()} function.}

\item{strAbbreviation}{\code{character} Abbreviation for the assessment being run. (E.g., Adverse Event = 'AE'; Protocol Deviation = 'PD').}

\item{strMetric}{\code{character} Brief description of the \code{Metric} being returned from the assessment being run. (E.g., "AE Reporting (Rate)").}

\item{strNumerator}{\code{character} Brief description of the \code{Numerator} being returned from the assessment being run; used as axis label.}

\item{strDenominator}{\code{character} Brief description of the \code{Denominator} being returned from the assessment being run; used as axis label.}

\item{vThreshold}{\code{numeric} Numeric vector of thresholds provided to the \verb{*_Assess()} function; used for flagging.}
}
\value{
\code{data.frame} containing all metadata needed to create an interactive visualization using \code{{rbm-viz}} htmlwidget.
}
\description{
\code{MakeDfConfig} is a utility function that is used within \verb{*_Assess()} functions in \code{{gsm}}.

The output of \code{MakeDfConfig} is used as a configuration dataset for a \code{{rbm-viz}} htmlwidget.
}
\details{
Make dfConfig for use in {rbm-viz}.
}
\examples{
\dontrun{
dfConfig <- MakeDfConfig(
  #   strMethod = strMethod,
  #   strGroup = strGroup,
  #   strAbbreviation = "AE",
  #   strMetric = "AE Reporting (Rate)",
  #   strNumerator = "AEs",
  #   strDenominator = "Days in Study",
  #   vThreshold = vThreshold
)
}

}
