% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Helpers.R
\name{MakeErrorLog}
\alias{MakeErrorLog}
\title{Extrapolate study snapshot date and number of patients in study}
\usage{
MakeErrorLog(data)
}
\arguments{
\item{data}{\code{list} a list containing active assessments}
}
\description{
Extrapolate study snapshot date and number of patients in study
}
\keyword{internal}
