% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeFlagHistory.R
\name{MakeFlagHistory}
\alias{MakeFlagHistory}
\title{MakeFlagHistory}
\usage{
MakeFlagHistory(lStackedSnapshots)
}
\arguments{
\item{lStackedSnapshots}{\code{list} the return from calling \link{gsm}}
}
\value{
\code{data.frame} of flag and score results by date of snapshot.
}
\description{
MakeFlagHistory
}
\examples{

lSnapshot <- Make_Snapshot()

lStackedSnapshots <- StackSnapshots(
  cPath = system.file("snapshots", "AA-AA-000-0000", package = "gsm"),
  lSnapshot = lSnapshot,
)

dfFlagHistory <- MakeFlagHistory(lStackedSnapshots = lStackedSnapshots)

}
