% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeGroupInfo.R
\name{MakeGroupInfo}
\alias{MakeGroupInfo}
\title{Helper function to compile "long" group metadata}
\usage{
MakeGroupInfo(data, strGroupLevel, strGroupCols = "GroupID")
}
\arguments{
\item{data}{The input data frame.}

\item{strGroupLevel}{A string specifying the group type.}

\item{strGroupCols}{A string specifying the group columns.}
}
\value{
A long format data frame.
}
\description{
Used to format group-level metadata (dfGroups) for use in charts and reports. This function takes a data frame and a string specifying the group columns, and returns a long format data frame.
}
\examples{
df <- data.frame(GroupID = c(1, 2, 3), Param1 = c(10, 20, 30), Param2 = c(100, 200, 300))
MakeGroupInfo(df, "GroupID")

}
