% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{MakeKRIGlossary}
\alias{MakeKRIGlossary}
\title{Create KRI metadata table in KRIReport.Rmd}
\usage{
MakeKRIGlossary(
  dfMetaWorkflow = gsm::meta_workflow,
  strWorkflowIDs = NULL,
  DroppedWorkflowIDs = NULL
)
}
\arguments{
\item{dfMetaWorkflow}{\code{data.frame} Workflow metadata from \code{params} within \code{KRIReport.Rmd}}

\item{strWorkflowIDs}{todo...}

\item{DroppedWorkflowIDs}{todo...}
}
\description{
Create KRI metadata table in KRIReport.Rmd
}
\keyword{internal}
