% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeLAssess.R
\name{MakeLAssess}
\alias{MakeLAssess}
\title{Make lAssess within an Assess function.}
\usage{
MakeLAssess(dfInput = NULL, lTags = NULL)
}
\value{
A list containing key/value pairs for parameters and their respective arguments
}
\description{
Utility function to standardize the output for lAssess within an Assess function. This function captures all parameters and
returns the values supplied when running the Assess function.
}
