% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeMetric.R
\name{MakeMetric}
\alias{MakeMetric}
\title{Parse meta data from workflows to a data frame}
\usage{
MakeMetric(lWorkflows)
}
\arguments{
\item{lWorkflows}{A list of workflows, like the one returned by MakeWorkflowList().}
}
\value{
A data frame.
}
\description{
Used to format metric metadata (dfMetrics) for use in charts and reports. This function takes a list of workflows and returns a data frame with one row per MetricID.
}
\examples{
lWorkflows <- MakeWorkflowList(strNames = "kri")
dfMetrics <- MakeMetricInfo(lWorkflows)

}
