% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{MakeOverviewMessage}
\alias{MakeOverviewMessage}
\title{Create message describing study summary for Report}
\usage{
MakeOverviewMessage(
  report,
  study_id,
  snapshot_date,
  subjects,
  overview_raw_table,
  red_kris
)
}
\arguments{
\item{report}{\code{string} type of report being run}

\item{study_id}{\code{string} a string representing the study id}

\item{snapshot_date}{\code{string} a string representing snapshot date}

\item{subjects}{\code{string} a string or number containing the count of total subjects}

\item{overview_raw_table}{\code{data.frame} raw overview table output with \code{Overview_Table(bInteractive = FALSE)}}

\item{red_kris}{\code{string} a string or number containing the count of red flags in kri's}
}
\description{
Create message describing study summary for Report
}
\keyword{internal}
