% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeParamLabels.R
\name{MakeParamLabels}
\alias{MakeParamLabels}
\title{Create Labels for Parameters}
\usage{
MakeParamLabels(dfGroups, lParamLabels = NULL)
}
\arguments{
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. Created by passing CTMS site and study data to \code{\link[=MakeLongMeta]{MakeLongMeta()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Param}, \code{Value}.}

\item{lParamLabels}{\code{list} Labels for parameters, with the parameters as names, and the label as value.}
}
\value{
\code{dfGroups} with an added \code{Label} column.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Used to create components of the group metadata dictionary (dfGroups) for use
in charts and reports. This function takes a data frame and a string
specifying the group columns, and returns a long format data frame.
}
\examples{
head(gsm::reportingGroups)
MakeParamLabels(head(gsm::reportingGroups))
MakeParamLabels(
  head(gsm::reportingGroups),
  list(ParticipantCount = "Number of Participants")
)

}
