% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeQTLCharts.R
\name{MakeQTLCharts}
\alias{MakeQTLCharts}
\title{MakeQTLCharts}
\usage{
MakeQTLCharts(strQtlName, lStackedSnapshots)
}
\arguments{
\item{strQtlName}{\code{character} Name of QTL workflow, e.g., "qtl0004".}

\item{lStackedSnapshots}{\code{list} List of stacked snapshot data - cross-sectional / longitudinal data.}
}
\value{
\code{list} of HTML widgets.
}
\description{
Make cross-sectional timeseries charts in \verb{[gsm::Make_Snapshot()]}.
}
