% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeQualificationSpec.R
\name{MakeQualificationSpec}
\alias{MakeQualificationSpec}
\title{MakeQualificationSpec}
\usage{
MakeQualificationSpec(strFunctionName)
}
\arguments{
\item{strFunctionName}{\code{character} Name of function to create specification for.}
}
\value{
\code{list} Named list of qualification tests and relevant test data:
\itemize{
\item \code{ID}: Specification ID - prefixed with "S" (e.g., S1_1)
\item \code{Description}: Description of specification.
\item \code{Risk}: One of \code{High}, \code{Medium}, or \code{Low}.
\item \code{Impact}: One of \code{High}, \code{Medium}, or \code{Low}.
\item \code{Tests}: \verb{character vector} All Test IDs (e.g., T1_1, T1_2) that apply to the specification.
}
}
\description{
MakeQualificationSpec
}
\examples{
\dontrun{
MakeQualificationSpec("AE_Assess")
}

}
