% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{MakeReportSetup}
\alias{MakeReportSetup}
\title{Create Study Results table for Report}
\usage{
MakeReportSetup(assessment, dfSite, strType)
}
\arguments{
\item{assessment}{\code{list} a snapshot list containing the parameters to assess}

\item{dfSite}{\code{data.frame} Site-level metadata containing within \code{params$status_site} of report}
}
\description{
Create Study Results table for Report
}
\keyword{internal}
