% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeResultsAnalysis.R
\name{MakeResultsAnalysis}
\alias{MakeResultsAnalysis}
\title{Make \code{results_analysis}}
\usage{
MakeResultsAnalysis(lResults)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{\link[=Study_Assess]{Study_Assess()}}.}
}
\value{
\code{data.frame} QTL results stacked data from \code{\link[=Study_Assess]{Study_Assess()}}
}
\description{
Make \code{results_analysis}
}
\examples{
\dontrun{
study <- Study_Assess()
results_analysis <- MakeResultsAnalysis(study)
}

}
