% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeResultsAnalysis.R
\name{MakeResultsAnalysisLog}
\alias{MakeResultsAnalysisLog}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
MakeResultsAnalysisLog(lResults)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{\link[=Study_Assess]{Study_Assess()}}.}
}
\value{
\code{data.frame} QTL results stacked data from \code{\link[=Study_Assess]{Study_Assess()}}
}
\description{
Make \code{results_analysis_log}
}
\examples{
\dontrun{
study <- Study_Assess()
results_analysis <- MakeResultsAnalysisLog(study)
}

}
