% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeResultsBounds.R
\name{MakeResultsBounds}
\alias{MakeResultsBounds}
\title{Make \code{results_bounds}}
\usage{
MakeResultsBounds(lResults, dfConfigWorkflow)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{\link[=Study_Assess]{Study_Assess()}}.}

\item{dfConfigWorkflow}{\code{data.frame} Workflow configuration data.}
}
\value{
\code{data.frame} Containing stacked results from \code{dfBounds}.
}
\description{
Make \code{results_bounds}
}
\examples{
\dontrun{
study <- Study_Assess()
results_bounds <- MakeResultsBounds(lResults = study, dfConfigWorkflow = gsm::config_workflow)
}

}
