% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeResultsBounds.R
\name{MakeResultsBoundsLog}
\alias{MakeResultsBoundsLog}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
MakeResultsBoundsLog(lResults)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{\link[=Study_Assess]{Study_Assess()}}.}

\item{dfConfigWorkflow}{\code{data.frame} Workflow configuration data.}
}
\value{
\code{data.frame} Containing stacked results from \code{dfBounds}.
}
\description{
Make \code{results_bounds_log}
}
\examples{
\dontrun{
study <- Study_Assess()
results_bounds <- MakeResultsBoundsLog(lResults = study)
}

}
