% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeResultsSummaryLog.R
\name{MakeResultsSummaryLog}
\alias{MakeResultsSummaryLog}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
MakeResultsSummaryLog(lResults)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{\link[=Study_Assess]{Study_Assess()}}.}
}
\value{
\code{data.frame} Stacked results (a la \code{dplyr::bind_rows()}) of the \code{dfSummary} data.frames, from the
output of \code{\link[=Study_Assess]{Study_Assess()}}
}
\description{
Make \code{results_summary_log}
}
\examples{
\dontrun{
study <- Study_Assess()
results_summary <- MakeResultsSummaryLog(study, gsm::config_workflow)
}

}
