% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{MakeResultsTable}
\alias{MakeResultsTable}
\title{Create Study Results table for Report}
\usage{
MakeResultsTable(assessment, summary_table)
}
\arguments{
\item{assessment}{\code{list} a snapshot list containing the parameters to assess}

\item{summary_table}{\code{data.frame} a summary table created from \code{MakeSummayTable}}
}
\description{
Create Study Results table for Report
}
\keyword{internal}
