% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{MakeRptKRIBoundsDetails}
\alias{MakeRptKRIBoundsDetails}
\title{Create rpt_kri_bounds_details output for \code{Make_Snapshot()}}
\usage{
MakeRptKRIBoundsDetails(lResults, config_workflow, gsm_analysis_date)
}
\arguments{
\item{lResults}{\code{list} the output from \code{Study_Assess()}}

\item{config_workflow}{\code{data.frame} configuration workflow in lMeta argument of \code{Make_Snapshot()}}

\item{gsm_analysis_date}{\code{string} Date of snapshot}
}
\description{
Create rpt_kri_bounds_details output for \code{Make_Snapshot()}
}
\keyword{internal}
