% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{MakeRptKRIDetail}
\alias{MakeRptKRIDetail}
\title{Create rpt_kri_detail output for \code{Make_Snapshot()}}
\usage{
MakeRptKRIDetail(lResults, status_site, meta_workflow, gsm_analysis_date)
}
\arguments{
\item{lResults}{\code{list} the output from \code{Study_Assess()}}

\item{meta_workflow}{\code{string} the meta_workflow stated in lMeta argument of \code{Make_Snapshot()}}

\item{gsm_analysis_date}{\code{string} the gsm analysis date calculated in \code{Make_Snapshot()}}

\item{status_study}{\code{data.frame} the output from \code{Site_Map_Raw()}}
}
\description{
Create rpt_kri_detail output for \code{Make_Snapshot()}
}
\keyword{internal}
