% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{MakeRptQtlAnalysis}
\alias{MakeRptQtlAnalysis}
\title{Create rpt_qtl_analysis output for \code{Make_Snapshot()}}
\usage{
MakeRptQtlAnalysis(lResults, gsm_analysis_date)
}
\arguments{
\item{lResults}{\code{list} the output from \code{Study_Assess()}}

\item{gsm_analysis_date}{\code{string} Date of snapshot}
}
\description{
Create rpt_qtl_analysis output for \code{Make_Snapshot()}
}
\keyword{internal}
