% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{MakeRptQtlDetails}
\alias{MakeRptQtlDetails}
\title{Make QTL Details Table}
\usage{
MakeRptQtlDetails(lResults, dfMetaWorkflow, dfConfigParam, gsm_analysis_date)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{\link[=Study_Assess]{Study_Assess()}}.}

\item{dfMetaWorkflow}{\code{data.frame} Workflow metadata. See \link{meta_workflow}.}

\item{gsm_analysis_date}{\code{date} Date that \code{{gsm}} snapshot was run.}
}
\description{
Make QTL Details Table
}
