% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{MakeRptSiteKriDetails}
\alias{MakeRptSiteKriDetails}
\title{Create rpt_kri_site_detail output for \code{Make_Snapshot()}}
\usage{
MakeRptSiteKriDetails(
  lResults,
  status_site,
  meta_workflow,
  meta_param = NULL,
  gsm_analysis_date
)
}
\arguments{
\item{lResults}{\code{list} the output from \code{Study_Assess()}}

\item{status_site}{\code{data.frame} the output from \code{Site_Map_Raw()}}

\item{meta_workflow}{\code{string} the meta_workflow stated in lMeta argument of \code{Make_Snapshot()}}

\item{meta_param}{\code{string} the meta_param stated in lMeta argument of \code{Make_Snapshot()}}

\item{gsm_analysis_date}{\code{string} the gsm analysis date calculated in \code{Make_Snapshot()}}
}
\description{
Create rpt_kri_site_detail output for \code{Make_Snapshot()}
}
\keyword{internal}
