% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{MakeRptStudySnapshot}
\alias{MakeRptStudySnapshot}
\title{MakeRptStudySnapshot}
\usage{
MakeRptStudySnapshot(lMeta, gsm_analysis_date)
}
\arguments{
\item{lMeta}{A list containing metadata information about the study.}

\item{strAnalysisDate}{A string representing the analysis date.}
}
\value{
A data frame with columns study_id, snapshot_date, is_latest, next_snapshot_date, pt_cycle_id, and pt_data_dt.
}
\description{
The function MakeRptStudySnapshot takes two inputs, lMeta and strAnalysisDate, and returns a data frame as output.
}
\keyword{internal}
