% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{MakeRptThresholdParam}
\alias{MakeRptThresholdParam}
\title{Create rpt_qtl_threshold_param output for \code{Make_Snapshot()}}
\usage{
MakeRptThresholdParam(
  meta_param,
  status_param,
  gsm_analysis_date,
  type,
  verbose = FALSE
)
}
\arguments{
\item{meta_param}{\code{data.frame} the meta_param defined in lMeta argument of \code{Make_Snapshot()} Default: gsm::meta_param}

\item{status_param}{\code{data.frame} the config_param defined in lMeta argument of \code{Make_Snapshot()}}

\item{gsm_analysis_date}{\code{string} Date of snapshot}

\item{type}{\code{string} type of threshold to output}

\item{verbose}{\code{logical} whether or not to display function messages}
}
\description{
Create rpt_qtl_threshold_param output for \code{Make_Snapshot()}
}
\keyword{internal}
