% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStatusWorkflow.R
\name{MakeStatusWorkflow}
\alias{MakeStatusWorkflow}
\title{Make \code{status_workflow}}
\usage{
MakeStatusWorkflow(lResults, dfConfigWorkflow)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{\link[=Study_Assess]{Study_Assess()}}.}

\item{dfConfigWorkflow}{\code{data.frame} Workflow configuration data.}
}
\value{
\code{data.frame} with columns \code{studyid}, \code{workflowid}, \code{gsm_version}, \code{active}, \code{status}, and \code{notes}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
study <- Study_Assess()
status_workflow <- MakeStatusWorkflow(
  lResults = study,
  dfConfigWorkflow = gsm::config_workflow
)
}

}
