% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeStrata.R
\name{MakeStrata}
\alias{MakeStrata}
\title{Utility function for creating analysis strata}
\usage{
MakeStrata(df, strDomain, strCol, strVal, bReturnChecks = FALSE, bQuiet = TRUE)
}
\arguments{
\item{df}{\code{data.frame} A data.frame to be filtered, likely within a mapping function.}

\item{strDomain}{\code{character} Domain step that is being stratified.}

\item{strCol}{\code{character} Domain in \code{lMapping} that references the column to filter on.}

\item{strVal}{\code{character} Domain in \code{lMapping} that references the value to filter on.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{is_mapping_valid}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame provided as \code{df} and filtered on \code{strCol} == \code{strVal}.
If \code{bReturnChecks} is \code{TRUE}, a \code{list} is returned with a filtered \code{df}, and a list of checks run on input data (\code{lChecks}).
}
\description{
Utility function for creating analysis strata
}
\examples{

strata1 <- MakeStrata(
  df = clindata::rawplus_ae,
  strDomain = 'dfAE'
  strCol = "AE_SERIOUS",
  strVal = "No"
)

}
