% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeStratifiedAssessment.R
\name{MakeStratifiedAssessment}
\alias{MakeStratifiedAssessment}
\title{Create multiple workflows from a single stratified workflow}
\usage{
MakeStratifiedAssessment(lWorkflow, lData, lMapping, bQuiet = TRUE)
}
\arguments{
\item{lWorkflow}{\code{list} A named list of metadata defining how an workflow should be run.}

\item{lData}{\code{list} A named list of domain-level data frames.}

\item{lMapping}{\code{list} A named list identifying the columns needed in each data domain.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} A list of workflows for each specified strata
}
\description{
Create multiple workflows from a single stratified workflow
}
\examples{

# Adverse Events by Grade
StratifiedAE <- MakeStratifiedAssessment(
  lData = list(
    dfSUBJ = clindata::rawplus_subj,
    dfAE = clindata::rawplus_ae
  ),
  lMapping = yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package = "gsm")),
  lWorkflow = MakeAssessmentList()$aeGrade
)

# Protocol Deviations by Category
lMapping <- yaml::read_yaml(
  system.file("mappings", "mapping_rawplus.yaml", package = "gsm")
)
StratifiedPD <- MakeStratifiedAssessment(
   lData = list(
     dfSUBJ = clindata::rawplus_subj,
     dfPD = clindata::rawplus_pd
   ),
   lMapping = lMapping,
   lWorkflow = MakeAssessmentList()$pdCategory
 )

StratifiedPD \%>\%
  purrr::map(~.x \%>\%
    RunAssessment(
      lData = list(
        dfSUBJ = clindata::rawplus_subj,
        dfPD = clindata::rawplus_pd
      ),
      lMapping = lMapping
    )
  )

}
