% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{MakeSummaryTable}
\alias{MakeSummaryTable}
\title{Create Summary table in KRIReport.Rmd for each KRI}
\usage{
MakeSummaryTable(assessment)
}
\arguments{
\item{assessment}{\code{data.frame} from \code{params} within \code{KRIReport.Rmd}}
}
\description{
Create Summary table in KRIReport.Rmd for each KRI
}
\keyword{internal}
