% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-MakeTargetData.R
\name{MakeTargetData}
\alias{MakeTargetData}
\title{Helper function to compile "long" group metadata}
\usage{
MakeTargetData(data, current_col, target_col, perc_val, perc_str_val)
}
\arguments{
\item{data}{The input dataframe}

\item{current_col}{Column that represents site count or participant count}

\item{target_col}{Column that represents the target count for site/participants}

\item{perc_val}{Name of column that will contain the numeric percentage value on enrollment}

\item{perc_str_val}{Name of column that will contain the n/N (xx.x\%) string}
}
\value{
A data frame containing two additional columns for the precentage value and associated string
}
\description{
Used to create percentage values on enrollment data, such as site activation and
and participantment enrollment, and pasting together an appropriate n/N (xx.x\%) value
as well.
}
