% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTimeSeriesLongitudinal.R
\name{MakeTimeSeriesLongitudinal}
\alias{MakeTimeSeriesLongitudinal}
\title{Create longitudinal snapshot from results_summary.}
\usage{
MakeTimeSeriesLongitudinal(cDirectory)
}
\arguments{
\item{cDirectory}{\code{character} Path to longitudinal data folders.}
}
\value{
\code{data.frame} containing longitudinal snapshots of \code{{gsm}} analyses.
}
\description{
Create longitudinal snapshot from results_summary.
}
\examples{

\dontrun{
sim <- clindata::run_simulation(n_sites = 20, n_subjects = 250)
dir <- here::here("simulation", "study-20_sites-250_subjects")
results_summary_over_time <- MakeTimeSeriesLongitudinal(dir = dir)
}

}
