% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTrendData.R
\name{MakeTrendData}
\alias{MakeTrendData}
\title{Make Time Series Trend Data}
\usage{
MakeTrendData(lSnapshot, cDirectory, bAppendTimeSeriesCharts = TRUE)
}
\arguments{
\item{lSnapshot}{\code{list} The output of \code{\link[=Make_Snapshot]{Make_Snapshot()}} where \code{bReturnStudyObject = TRUE}.}

\item{cDirectory}{\code{character} Path to longitudinal snapshots, returned by \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}

\item{bAppendTimeSeriesCharts}{\code{logical} Append time series charts to KRIs? Default: \code{TRUE}.}
}
\value{
\code{list} The output of \code{\link[=Study_Assess]{Study_Assess()}} with longitudinal data appended.
}
\description{
Make Time Series Trend Data
}
\examples{
\dontrun{
study <- Study_Assess()
directory_location <- here::here("snapshots")

trend_data <- MakeTrendData(lStudyAssess = study, cDirectory = directory_location)

}

}
