% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeWideGroups.R
\name{MakeWideGroups}
\alias{MakeWideGroups}
\title{Helper function to compile "wide" group metadata}
\usage{
MakeWideGroups(dfMeta, strGroupLevel)
}
\arguments{
\item{dfMeta}{The meta data dictionary with one row per GroupID per Param.
Must have columns GroupID, GroupLevel, Param, and Value.}

\item{strGroupLevel}{A string specifying the group level; used to filter
dfMeta$GroupLevel.}
}
\value{
A long format data frame.
}
\description{
Used to convert metadata dictionary (dfMeta) to a wide format data frame for
use in charts and reports.
}
\examples{
MakeWideGroups(reportingGroups, "Site")

}
