% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeWorkflowList.R
\name{MakeWorkflowList}
\alias{MakeWorkflowList}
\title{Load assessments from a package/directory}
\usage{
MakeWorkflowList(
  strNames = NULL,
  strPath = "workflow",
  strPackage = "gsm",
  bRecursive = FALSE
)
}
\arguments{
\item{strNames}{\verb{array of character} List of workflows to include. NULL (the default) includes all workflows in the specified locations.}

\item{strPath}{\code{character} The location of assessment YAML files. If package is specified, function will look in \verb{/inst} folder.}

\item{strPackage}{\code{character} The name of the package with assessments.}

\item{bRecursive}{\code{logical} Find files in nested folders? Default FALSE.}
}
\value{
\code{list} A list of assessments with workflow and parameter metadata.
}
\description{
Load assessments from a package/directory
}
\details{
\code{MakeWorkflowList()} is a utility function that creates a workflow mapping for assessments used in \code{Study_Assess()}.
}
\examples{
workflow <- MakeWorkflowList(strPath = "workflow", strPackage = "gsm")

# get specific workflow files
workflow <- MakeWorkflowList(strNames = c("kri0001", "kri0005", "cou0003"))

}
