% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeWorkflowYaml.R
\name{MakeWorkflowYaml}
\alias{MakeWorkflowYaml}
\title{Make a workflow YAML file.}
\usage{
MakeWorkflowYaml(dfMetaWorkflow, strFileName, bSave = TRUE)
}
\arguments{
\item{dfMetaWorkflow}{File containing metadata for all supported KRIs.}

\item{strFileName}{File name for workflow YAML file.}

\item{bSave}{Save as file in current working directory? Default: \code{TRUE}, else returns a list.}
}
\value{
Saves a YAML file in current working directory.
}
\description{
Make a workflow YAML file.
}
\examples{
library(dplyr)

meta_df <- left_join(meta_workflow, meta_workflow_params, by = "workflowid") \%>\% split(.$workflowid)
\dontrun{
purrr::map2(meta_df, names(meta_df), ~MakeWorkflowYaml(dfMetaWorkflow = .x, strFileName = .y))
}

}
