% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MapMaker.R
\name{MapMaker}
\alias{MapMaker}
\title{Read YAML mapping in Assess functions without params}
\usage{
MapMaker(strFunctionName = NULL)
}
\arguments{
\item{strFunctionName}{\code{character} Default function name derived within an Assess function.}
}
\value{
YAML mapping
}
\description{
Read YAML mapping in Assess functions without params
}
