% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{Match_Class}
\alias{Match_Class}
\title{Augment the previous lSnapshot classes to be inline with the current lSnapshot classes}
\usage{
Match_Class(lPrevSnapshot, lSnapshot)
}
\arguments{
\item{lPrevSnapshot}{\code{list} the previous Snapshot object}

\item{lSnapshot}{\code{list} the current Snapshot object}
}
\value{
Augmented previous snapshot object
}
\description{
Augment the previous lSnapshot classes to be inline with the current lSnapshot classes
}
\keyword{internal}
