% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mergeSubjects.R
\name{mergeSubjects}
\alias{mergeSubjects}
\title{Merge Domain data with subject-level data shell}
\usage{
mergeSubjects(
  dfDomain,
  dfSubjects,
  strIDCol = "SubjectID",
  vFillZero = NULL,
  bQuiet = TRUE
)
}
\arguments{
\item{dfDomain}{Subject-level domain data containing one record per participant.}

\item{dfSubjects}{Subject level data often using ADSL-like data. Should include one record per participant for each participant included in the analysis population (all other participants should be dropped before calling mergeSubjects)}

\item{strIDCol}{name of ID Column - default='SubjectID'}

\item{vFillZero}{Columns from dfDomain to fill with zeros when no matching row is found in for an ID in dfSubject}

\item{bQuiet}{print messages?}
}
\value{
data set with one record per IDCol
}
\description{
Merge Domain data with subject-level data shell
}
