% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{meta_param}
\alias{meta_param}
\title{meta_param Dataset}
\format{
A data frame with 176 rows and 6 columns:
\describe{
\item{workflowid}{workflow ID, chr, kri value "kri0001"}
\item{gsm_version}{gsm version used, chr, most recent version installed i.e. "1.9.2"}
\item{param}{parameter name, chr, i.e. "vThreshold"}
\item{index}{index value for parameter, int}
\item{default}{default value for parameter at index, int, i.e. -3}
\item{configurable}{is parameter configuable, lgl}
}
}
\source{
Generated from \code{meta_param.csv} dataset in the \code{gsm} package.
}
\usage{
meta_param
}
\description{
meta_param Dataset
}
\keyword{datasets}
