% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{meta_workflow}
\alias{meta_workflow}
\title{meta_workflow Dataset}
\format{
A data frame with 26 rows and 12 columns:
\describe{
\item{workflowid}{workflow ID, chr, kri value "kri0001"}
\item{gsm_version}{gsm version used, chr, most recent version installed i.e. "1.9.2"}
\item{group}{grouping used for workflow, chr, i.e. "Site"}
\item{abbreviation}{abbraviated name of workflowid (max 3-letters), chr, i.e. "AE"}
\item{metric}{metric label, chr, i.e. "AE Reporting Rate"}
\item{numerator}{Numerator Label, chr, i.e. "AEs"}
\item{denominator}{Denominator Label, chr, i.e. "Days on Study"}
\item{outcome}{Metric outcome type, chr, i.e. "Rate"}
\item{model}{Statistical model used to create score, chr, i.e. "Normal Approximation (Rate)"}
\item{score}{Score Label, chr, i.e. "Adjusted Z-Score"}
\item{data_inputs}{All data inputs for workflow, chr, i.e. "rawplus.ae"}
\item{data_filters}{filters applied in workflow, chr}
}
}
\source{
Generated from \code{meta_workflow.csv} dataset in the \code{gsm} package.
}
\usage{
meta_workflow
}
\description{
meta_workflow Dataset
}
\keyword{datasets}
