% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overview_Table_Country.R
\name{Overview_Table_Country}
\alias{Overview_Table_Country}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
Overview_Table_Country(lAssessments, bInteractive = TRUE)
}
\arguments{
\item{lAssessments}{\code{list} The output of running \code{\link[=Study_Assess]{Study_Assess()}}.}

\item{bInteractive}{\code{logical} Display interactive widget? Default: \code{TRUE}.}
}
\description{
Overview Table by Country - Create country-level summary of red and amber KRIs for a study.
}
\examples{
\dontrun{
lAssessments <- Study_Assess()
Overview_Table_Country(lAssessments)
}

}
