% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-parse_data_mapping.R
\name{parse_data_mapping}
\alias{parse_data_mapping}
\title{Parse Data Mapping.}
\usage{
parse_data_mapping(content = NULL, file = NULL)
}
\arguments{
\item{content}{\code{list} data mapping}

\item{file}{\code{character} file path of .yaml file}
}
\description{
Transform nested data mapping to tabular structure for use in documentation.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
