% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-parse_data_spec.R
\name{parse_data_spec}
\alias{parse_data_spec}
\title{Parse Data Specification}
\usage{
parse_data_spec(content = NULL, file = NULL)
}
\arguments{
\item{content}{\code{list} data specification}

\item{file}{\code{character} file path of .yaml file}
}
\description{
Transform nested data specification to tabular structure
}
