% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ParseWarnings.R
\name{ParseWarnings}
\alias{ParseWarnings}
\title{Parse warnings from the result of Study_Assess.}
\usage{
ParseWarnings(lResults)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{gsm::Study_Assess()}.}
}
\value{
\code{data.frame} With columns \code{notes} and \code{workflowid}.
}
\description{
Parse warnings from the result of Study_Assess.
}
\examples{

study <- Study_Assess()

warnings <- ParseWarnings(study)

}
