% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseWorkflow.R
\name{parseWorkflow}
\alias{parseWorkflow}
\title{Parse Workflow}
\usage{
parseWorkflow(workflow)
}
\arguments{
\item{workflow}{The workflow to be parsed.}
}
\value{
The parsed result.
}
\description{
This function parses the workflow and returns the parsed result.
}
\examples{

wf <- yaml::read_yaml(system.file("workflow", "kri0001.yaml", package = "gsm"))
parseWorkflow(workflow = wf)

wf_text<-'
steps:
  - name: RunQuery
    output: dfSubjects
    params:
      df: dfSUBJ
      lMapping: lMapping$dfSUBJ
      strQuery: "SELECT * FROM df WHERE {`strEnrollCol`} == {`strEnrollVal`}"
  - name: RunQuery
    output: dfDenominator
    params:
      df: dfLB
      lMapping: lMapping$dfLB
      strQuery: "SELECT * FROM df WHERE {`strGradeCol`} IN ({`strGradeAnyVal`*})"
  - name: RunQuery
    output: dfNumerator
    params:
      df: dfLB
      lMapping: lMapping$dfLB
      strQuery: "SELECT * FROM df WHERE {`strGradeCol`} IN ({`strGradeHighVal`*})"  
  - name: Make_Input_Rate
    output: dfInput
    params:
      dfs:
        - dfSubjects
        - dfNumerator
        - dfDenominator
      lMapping: lMapping
      lDomains: 
        dfSubjects: dfSUBJ
        dfNumerator: dfLB
        dfDenominator: dfLB
      strNumeratorMethod: Count
      strDenominatorMethod: Count
  - name: Transform_Rate
    output: dfTransformed
    params:
      dfInput: dfInput
  - name: Analyze_NormalApprox
    output: dfAnalyzed
    params:
      dfTransformed: dfTransformed 
  - name: Analyze_NormalApprox_PredictBounds
    output: dfBounds
    params:
      dfTransformed: dfTransformed
  - name: Flag_NormalApprox
    output: dfFlagged
    params:
      dfAnalyzed: dfAnalyzed
      vThreshold: 
        - -3
        - -2
        - 2
        - 3
  - name: Summarize
    output: dfSummary
    params: 
      dfFlagged: dfFlagged
  - name: MakeKRICharts
    output: lCharts
    params:
      dfSummary: dfSummary
      dfBounds: dfBounds
      lLabels: lMeta'

parseWorkflow(yaml::yaml.load(wf_text))

}
