% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Assess_Binary.R
\name{PD_Assess_Binary}
\alias{PD_Assess_Binary}
\title{Protocol Deviation Assessment (Binary)}
\usage{
PD_Assess_Binary(
  dfInput,
  vThreshold = NULL,
  strMethod = "NormalApprox",
  lMapping = yaml::read_yaml(system.file("mappings", "PD_Assess_Binary.yaml", package =
    "gsm")),
  lLabels = list(workflowid = "", group = strGroup, abbreviation = "PD", metric =
    "Protocol Deviation Rate", numerator = "Protocol Deviations", denominator =
    "Days on Study", model = "Normal Approximation", score = "Adjusted Z-Score"),
  strGroup = "Site",
  nMinDenominator = NULL,
  nConfLevel = NULL,
  bMakeCharts = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data, a data frame with one record per subject.}

\item{vThreshold}{\code{numeric} Threshold specification, a numeric vector that defaults to \code{c(-3, -2, 2, 3)} for a Normal Approximation (\code{strMethod = "NormalApprox"}),
\code{c(-7, -5, 5, 7)} for a Poisson model (\code{strMethod = "Poisson"}), \code{c(0.000895, 0.003059)} for a nominal assessment (\code{strMethod = "Identity"}), and \code{c(0.01)} for a QTL (\code{strMethod = "QTL"}).}

\item{strMethod}{\code{character} Statistical method. Valid values:
\itemize{
\item \code{"NormalApprox"} (default)
\item \code{"Poisson"}
\item \code{"Identity"}
\item \code{"QTL"}
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column. Default: package-defined Protocol Deviation Assessment mapping.}

\item{lLabels}{\code{list} Labels used to populate chart labels.}

\item{strGroup}{\code{character} Grouping variable. \code{"Site"} (the default) uses the column named in \code{mapping$strSiteCol}. Other valid options using the default mapping are \code{"Study"}, \code{"Country"}, and \code{"CustomGroup"}.}

\item{nMinDenominator}{\code{numeric} Specifies the minimum denominator required to return a \code{score} and calculate a \code{flag}. Default: NULL}

\item{nConfLevel}{\code{numeric} Confidence level for QTL analysis.}

\item{bMakeCharts}{\code{logical} Boolean value indicating whether to create charts.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} \code{lData}, a named list with:
\itemize{
\item each data frame in the data pipeline
\itemize{
\item \code{dfTransformed}, returned by \code{\link[=Transform_Rate]{Transform_Rate()}}
\item \code{dfAnalyzed}, returned by \code{\link[=Analyze_NormalApprox]{Analyze_NormalApprox()}}, \code{\link[=Analyze_Poisson]{Analyze_Poisson()}} or \code{\link[=Analyze_Identity]{Analyze_Identity()}}
\item \code{dfFlagged}, returned by \code{\link[=Flag_NormalApprox]{Flag_NormalApprox()}}, \code{\link[=Flag_Poisson]{Flag_Poisson()}}, or \code{\link[=Flag]{Flag()}}
\item \code{dfSummary}, returned by \code{\link[=Summarize]{Summarize()}}
\item \code{dfBounds}, returned by \code{\link[=Analyze_NormalApprox_PredictBounds]{Analyze_NormalApprox_PredictBounds()}} when \code{strMethod == "NormalApprox"} or \code{\link[=Analyze_Poisson_PredictBounds]{Analyze_Poisson_PredictBounds()}} when \code{strMethod == "Poisson"}
}
\item \code{list} \code{lCharts}, a named list with:
\itemize{
\item \code{scatter}, a ggplot2 object returned by \code{\link[=Visualize_Scatter]{Visualize_Scatter()}} only when strMethod != "Identity"
\item \code{barMetric}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}}
\item \code{barScore}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}}
}
\item \code{list} \code{lChecks}, a named list with:
\itemize{
\item \code{dfInput}, a named list returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{status}, a boolean returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{mapping}, a named list that is provided as an argument to the \code{lMapping} parameter in \code{\link[=PD_Assess_Rate]{PD_Assess_Rate()}}
\item \code{spec}, a named list used to define variable specifications
}
}
}
\description{
Evaluates Protocol Deviation (PD) rates to identify sites that may be over- or under-reporting PDs.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The Protocol Deviation Assessment uses the standard \href{https://gilead-biostats.github.io/gsm/articles/DataPipeline.html}{GSM data pipeline} to flag possible outliers. Additional details regarding the data pipeline and statistical
methods are described below.
}
\examples{
dfInput <- PD_Map_Raw_Binary()
pd_assessment_NormalApprox <- PD_Assess_Binary(dfInput)
pd_assessment_poisson <- PD_Assess_Binary(dfInput, strMethod = "Poisson")
pd_assessment_identity <- PD_Assess_Binary(dfInput, strMethod = "Identity")

}
