% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Map_Raw.R
\name{PD_Map_Raw}
\alias{PD_Map_Raw}
\title{Protocol Deviation Assessment Mapping from Raw Data- Make Input Data}
\usage{
PD_Map_Raw(dfSubid = NULL, dfEx = NULL, dfPd = NULL)
}
\arguments{
\item{dfSubid}{directly input a subid dataset with columns SUBJID INVID}

\item{dfEx}{directly input an ex dataset with columns SUBJID EXSTDAT EXENDAT}

\item{dfPd}{directly input a PD dataset with columns SUBJID and rows for each Protocol Deviation}

\item{strPath}{path to the raw datasets}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count, Exposure, Rate, Unit
}
\description{
Convert from raw data format to needed input format for Safety Assessment
Requires the following raw datasets: subid, ex, pd
}
