% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Map_Raw_Binary.R
\name{PD_Map_Raw_Binary}
\alias{PD_Map_Raw_Binary}
\title{Protocol Deviation Assessment (Binary Outcome) - Raw Mapping}
\usage{
PD_Map_Raw_Binary(
  dfs = gsm::UseClindata(list(dfSUBJ = "clindata::rawplus_dm", dfPD =
    "clindata::ctms_protdev")),
  lMapping = gsm::Read_Mapping(c("ctms", "rawplus")),
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfPD}: \code{data.frame} Event-level data with one record per PD.
\item \code{dfSUBJ}: \code{data.frame} Subject-level data with one record per subject.
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column. Default: package-defined mapping for raw+.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{\link[=is_mapping_valid]{is_mapping_valid()}}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame with one record per subject, the input to \code{\link[=PD_Assess_Binary]{PD_Assess_Binary()}}. If
\code{bReturnChecks} is \code{TRUE} \code{PD_Map_Raw_Binary} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert raw protocol deviation (PD) data, typically processed case report form data, to formatted
input data to \code{\link[=PD_Assess_Binary]{PD_Assess_Binary()}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{PD_Map_Raw_Binary} combines PD data with subject-level study duration data to create formatted
input data to \code{\link[=PD_Assess_Binary]{PD_Assess_Binary()}}. This function creates an input dataset for the PD Assessment
(\code{\link[=PD_Assess_Binary]{PD_Assess_Binary()}}) by binding subject-level PD counts (derived from \code{dfPD}) to subject-level
data (from \code{dfSUBJ}). Note that the function can generate data summaries for specific types of
PDs by passing filtered PD data to \code{dfPD}.
}
\section{Data specification}{
\tabular{llllll}{
   \strong{Domain} \tab \strong{Column Key} \tab \strong{Default Value} \tab \strong{Required?} \tab \strong{Require Unique Values?} \tab \strong{Accept NA/Empty Values?} \cr
   dfSUBJ \tab strSiteCol \tab siteid \tab TRUE \tab FALSE \tab FALSE \cr
   dfSUBJ \tab strIDCol \tab subjid \tab TRUE \tab TRUE \tab FALSE \cr
   dfSUBJ \tab strTimeOnStudyCol \tab timeonstudy \tab TRUE \tab FALSE \tab TRUE \cr
   dfPD \tab strIDCol \tab subjid \tab TRUE \tab FALSE \tab FALSE \cr
}
}

\examples{
# Run with defaults.
dfInput <- PD_Map_Raw_Binary()

# Run with error checking and message log.
dfInput <- PD_Map_Raw_Binary(bReturnChecks = TRUE, bQuiet = FALSE)

}
