% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Helpers.R
\name{qtl_analysis}
\alias{qtl_analysis}
\title{Compile QTL analysis results into data frame}
\usage{
qtl_analysis(lAssessments, results_summary)
}
\arguments{
\item{lAssessments}{\code{list} a list containing active assessments}

\item{results_summary}{compiled results summary from \code{qtl_summary()}}
}
\description{
Compile QTL analysis results into data frame
}
\keyword{internal}
