% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{qtl_results}
\alias{qtl_results}
\title{Compile QTL analysis results into data frame}
\usage{
qtl_results(lAssessments)
}
\arguments{
\item{lAssessments}{\code{list} a list containing active assessments}
}
\description{
Compile QTL analysis results into data frame
}
\keyword{internal}
