% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Helpers.R
\name{qtl_summary}
\alias{qtl_summary}
\title{Compile QTL summary results into data frame}
\usage{
qtl_summary(lAssessments)
}
\arguments{
\item{lAssessments}{\code{list} a list containing active assessments}
}
\description{
Compile QTL summary results into data frame
}
\keyword{internal}
