% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-quietDown.R
\name{quietDown}
\alias{quietDown}
\title{Utility function to suppress all messages in unit tests}
\usage{
quietDown(action = "start")
}
\arguments{
\item{action}{'start' (the default) or 'stop'. Used in \code{testthat/setup.R} to temporarily divert messages to a log file.}
}
\description{
Utility function to suppress all messages in unit tests
}
