% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{rank_chg}
\alias{rank_chg}
\title{insert icon for status in gt table}
\usage{
rank_chg(status)
}
\arguments{
\item{status}{boolean status}
}
\description{
from https://themockup.blog/posts/2020-10-31-embedding-custom-features-in-gt-tables/
}
