% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{rbm_data_spec}
\alias{rbm_data_spec}
\title{rbm_data_spec Dataset}
\format{
A data frame with 479 rows and 5 columns:
\describe{
\item{System}{System that uses data, chr, i.e. "Gismo"}
\item{Table}{domain of data, chr, i.e. "status_study"}
\item{Column}{column names found in table, chr, i.e. "studyid"}
\item{Description}{Description of what the column represents, chr, i.e. "Unique Study ID"}
\item{Order}{index of the column column location in dataset, int}
}
}
\source{
Generated from \code{rbm_data_spec.csv} dataset in the \code{gsm} package.
}
\usage{
rbm_data_spec
}
\description{
rbm_data_spec Dataset
}
\keyword{datasets}
