% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Mapping.R
\name{Read_Mapping}
\alias{Read_Mapping}
\title{Read Mapping}
\usage{
Read_Mapping(strDomain = NULL, package = "gsm")
}
\arguments{
\item{strDomain}{\code{character} vector of length 1 or greater. Valid values include: \code{adam}, \code{ctms}, \code{edc}, and \code{rawplus}. Additionally, for very specific use-cases, \code{domain} is a valid option.}

\item{package}{\code{character} a string of the package name to define the path to the mapping yaml files}
}
\value{
\code{list} of mappings for each available domain: \code{ADaM}, \code{CTMS}, \code{EDC}, and \code{rawplus}.
}
\description{
Read default mapping files from \code{inst/mappings}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
mappings <- Read_Mapping(c("ctms", "rawplus"))

}
