% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatStackedSnapshots.R
\name{ReformatStackedSnapshots}
\alias{ReformatStackedSnapshots}
\title{Reformat \code{lStackedSnapshots} from \verb{gsm v1.8.4} to \verb{gsm v1.9.0+}.}
\usage{
ReformatStackedSnapshots(lStackedSnapshots)
}
\arguments{
\item{lStackedSnapshots}{}
}
\description{
Reformat \code{lStackedSnapshots} from \verb{gsm v1.8.4} to \verb{gsm v1.9.0+}.
}
\keyword{internal}
