% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemapLog.R
\name{RemapLog}
\alias{RemapLog}
\title{Remap the log output of Make_Snapshot to new col names and indexes}
\usage{
RemapLog(table, table_name = NULL)
}
\arguments{
\item{table}{\code{data.frame} the table of interest from \code{Make_Snapshot()} lSnapshot log}

\item{table_name}{\code{character} optional argument to define table to remap by name, when the table object name doesn't match}
}
\description{
Remap the log output of Make_Snapshot to new col names and indexes
}
\keyword{internal}
