% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-RemoveInvalidSubjectIDs.R
\name{RemoveInvalidSubjectIDs}
\alias{RemoveInvalidSubjectIDs}
\title{Remove invalid Subject IDs}
\usage{
RemoveInvalidSubjectIDs(
  df,
  strDomain = "input",
  strIDCol = "subjid",
  strInvalidValues = "",
  bQuiet = TRUE
)
}
\arguments{
\item{df}{\code{data.frame} data frame with column for subject ID}

\item{strDomain}{\code{character} domain name of data frame}

\item{strIDCol}{\code{character} name of subject ID Column; Default: \code{"SubjectID"}}

\item{strInvalidValues}{\code{character} array of invalid subject ID values; Default: \code{""}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} input data frame with invalid subject ID values removed
}
\description{
\code{RemoveInvalidSubjectIDs} is a helper function that allows a user to specify a character vector of \code{SubjectID}s to remove from input data.
}
\examples{
RemoveInvalidSubjectIDs(
  clindata::rawplus_protdev,
  strIDCol = "subjid",
  bQuiet = FALSE
)

}
