% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-render_rmd.R
\name{RenderRmd}
\alias{RenderRmd}
\title{Custom Rmarkdown render function}
\usage{
RenderRmd(input, output_file, params)
}
\arguments{
\item{input}{path to the template Rmd file}

\item{output_file}{path to the location where output will be saved}

\item{params}{list of params to pass to the template Rmd file}
}
\value{
Rendered Rmarkdown file
}
\description{
Rmarkdown render function that defaults to rendering intermediate rmd files in a temporary directory
}
