% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_FormatFlag.R
\name{Report_FormatFlag}
\alias{Report_FormatFlag}
\title{KRI Directionality Logo.}
\usage{
Report_FormatFlag(flag_value, title = NULL)
}
\arguments{
\item{flag_value}{\code{numeric} Values between -2 and 2.}

\item{title}{\code{character} Data that will be passed to the \if{html}{\out{<title>}} tag.}
}
\value{
A character vector of fontawesome icon SVGs.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Reporting utility function to create fontawesome directionality logos`
}
