% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_MetricCharts.R
\name{Report_MetricCharts}
\alias{Report_MetricCharts}
\title{Render charts for a given metric to markdown}
\usage{
Report_MetricCharts(lCharts)
}
\arguments{
\item{lCharts}{A list of charts for the selected metric.}
}
\value{
Markdown content with charts and a summary table for the metric
}
\description{
This function generates a markdown framework for charts
}
