% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_MetricTable.R
\name{Report_MetricTable}
\alias{Report_MetricTable}
\title{Generate a summary table for a report}
\usage{
Report_MetricTable(
  dfResults,
  dfGroups,
  strSnapshotDate = NULL,
  strGroupLevel = c("Site", "Study", "Country"),
  strGroupDetailsParams = NULL
)
}
\arguments{
\item{dfResults}{The summary data frame}

\item{dfGroups}{A data frame containing the group-level metadata.}

\item{strSnapshotDate}{user specified snapshot date as string}

\item{strGroupLevel}{group level for the table}

\item{strGroupDetailsParams}{one or more parameters from dfGroups to be added as columns in the table}
}
\value{
A datatable containing the summary table
}
\description{
This function generates a summary table for a report by joining the provided
results data frame with the site-level metadata from dfGroups. It then filters and arranges the
data based on certain conditions and displays the result in a datatable.
}
