% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_OverviewText.R
\name{Report_OverviewText}
\alias{Report_OverviewText}
\title{Create message describing study summary for Report}
\usage{
Report_OverviewText(lSetup, dfResults, lStudy)
}
\arguments{
\item{lSetup}{\code{list} that is produced by \code{Report_StudyInfo()}.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[=BindResults]{BindResults()}}.}

\item{lStudy}{\code{list} contains study-level metadata.}
}
\description{
Create message describing study summary for Report
}
\keyword{internal}
