% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Setup.R
\name{Report_Setup}
\alias{Report_Setup}
\title{Calculate needed values for report}
\usage{
Report_Setup(dfStudy = NULL, dfMetrics = NULL, dfSummary = NULL)
}
\arguments{
\item{dfStudy}{\code{data.frame} Site-level metadata.}
}
\description{
Calculates the following:
\itemize{
\item Snapshot date: dfStudy$SnapshotDate (if available) or Sys.Date()
\item Study ID: dfStudy$StudyID (if available)  or "Unknown"
}
}
\keyword{internal}
