% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_StudyInfo.R
\name{Report_StudyInfo}
\alias{Report_StudyInfo}
\title{Report Study Information}
\usage{
Report_StudyInfo(dfStudy, dfStudyLabels = NULL)
}
\arguments{
\item{dfStudy}{A data frame containing study information.}

\item{dfStudyLabels}{A data frame containing study labels. Default is NULL.}
}
\value{
None
}
\description{
This function generates a table summarizing study metadata.
}
\keyword{internal}
