% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{reportingGroups}
\alias{reportingGroups}
\title{reportingGroups Dataset}
\format{
A data frame with 3903 rows and 4 columns:
\describe{
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Param}{Parameter of interest for the group}
\item{Value}{Value of specified Param}
}
}
\source{
Generated from \code{reportingGroups.csv} dataset in the \code{gsm} package.
}
\usage{
reportingGroups
}
\description{
reportingGroups Dataset
}
\keyword{datasets}
