% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{reportingMetrics_country}
\alias{reportingMetrics_country}
\title{reportingMetrics_country Dataset}
\format{
A data frame with 12 rows and 17 columns:
\describe{
\item{MetricID}{unique metric identifier}
\item{File}{yaml file for workflow}
\item{GroupLevel}{level of grouping variable}
\item{Abbreviation}{abbreviation for the metric}
\item{Metric}{name of the metric}
\item{Numerator}{data source for the numerator}
\item{Denominator}{data source for the denominator}
\item{Model}{model used to calculate metric}
\item{Score}{type of score reported}
\item{Type}{statistical outcome type}
\item{Threshold}{thresholds to be used for bounds and flags}
\item{nMinDenominator}{minimum denominator required to return a score and calculate a flag}
}
}
\source{
Generated from \code{reportingMetrics_country.csv} dataset in the \code{gsm} package.
}
\usage{
reportingMetrics_country
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{datasets}
