% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevertSnapshotLogs.R
\name{RevertSnapshotLogs}
\alias{RevertSnapshotLogs}
\title{Function to revert to 1.8 data model}
\usage{
RevertSnapshotLogs(
  lSnapshot,
  lMeta = NULL,
  lData = NULL,
  lMapping = NULL,
  bQuiet = TRUE
)
}
\arguments{
\item{lSnapshot}{\code{list} the snapshot object to revert}

\item{lMeta}{\code{list} the meta data to use for reversion}

\item{lData}{\code{list} the data to use for reversion}

\item{lMapping}{\code{list} the mapping to use for reversion}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\description{
Function to revert to 1.8 data model
}
\keyword{internal}
