% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-RunAssessment.R
\name{RunAssessment}
\alias{RunAssessment}
\title{Run a single assessment}
\usage{
RunAssessment(lAssessment, lData, lMapping, lTags = NULL, bQuiet = FALSE)
}
\arguments{
\item{lAssessment}{a named list of metadata defining how each assessment should be run. Properties should include: \code{label}, \code{tags} and \code{workflow}}

\item{lData}{a named list of domain level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{lMapping}{a named list identifying the columns needed in each data domain.}

\item{lTags}{tags}

\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}
}
\value{
Returns \code{lAssessment} with \code{lData}, \code{lResults}, \code{bStatus} and \code{lChecks} added based on the results of the execution of \code{assessment$workflow}.
}
\description{
Atempts to run a single assessments (\code{lAssessment}) using shared data (\code{lData}) and metadata (\code{lMapping}). Calls \code{RunStep} for each item in \code{lAssessment$Workflow} and saves the results to \code{lAssessment}
}
