% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunQTL.R
\name{RunQTL}
\alias{RunQTL}
\title{{experimental} Run QTL - Runs QTL based on {gsm}-defined workflow, or by using a custom workflow.}
\usage{
RunQTL(strName = NULL, lWorkflow = NULL, lData = NULL, lMapping = NULL)
}
\arguments{
\item{strName}{\code{character} Name of QTL YAML file found in \code{inst/workflow/experimental}, e.g. "qtl0003"}

\item{lWorkflow}{\code{character} Optional. Custom workflow to use in place of pre-defined workflow.}

\item{lData}{\code{list} List of data needed to run QTL analysis.}

\item{lMapping}{\code{list} List of data-domain mappings.}
}
\description{
Run a QTL based on a pre-defined or custom workflow outside of \code{Make_Snapshot()} or \code{Study_Assess()}. This is a helper function that makes it possible to
run a QTL for any supported \verb{*_Assess} function.
}
\examples{
qtl <- RunQTL('qtl0003')

}
