% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{sampleBounds}
\alias{sampleBounds}
\title{sampleBounds Dataset}
\format{
A data frame with 12667 rows and 8 columns:
\describe{
\item{Threshold}{The number of standard deviations that the upper and lower bounds are based on}
\item{Denominator}{The calculated denominator value}
\item{LogDenominator}{The calculated log denominator value}
\item{Numerator}{The calculated numerator value}
\item{Metric}{The calculated rate/metric value}
\item{MetricID}{The Metric ID}
\item{StudyID}{The Study ID}
\item{SnapshotDate}{The Date of the snapshot}
}
}
\source{
Generated from \code{sampleBounds.csv} dataset in the \code{gsm} package.
}
\usage{
sampleBounds
}
\description{
sampleBounds Dataset
}
\keyword{datasets}
