% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{sampleGroups}
\alias{sampleGroups}
\title{sampleGroups Dataset}
\format{
A data frame with 3903 rows and 4 columns:
\describe{
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Param}{Parameter of interest for the group}
\item{Value}{Value of specified Param}
}
}
\source{
Generated from \code{sampleGroups.csv} dataset in the \code{gsm} package.
}
\usage{
sampleGroups
}
\description{
sampleGroups Dataset
}
\keyword{datasets}
