% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{sampleMetrics}
\alias{sampleMetrics}
\title{sampleMetrics Dataset}
\format{
A data frame with 12 rows and 11 columns:
\describe{
\item{MetricID}{unique metric identifier}
\item{File}{yaml file for workflow}
\item{GroupLevel}{level of grouping variable}
\item{Abbreviation}{abbreviation for the metric}
\item{Metric}{name of the metric}
\item{Numerator}{data source for the numerator}
\item{Denominator}{data source for the denominator}
\item{Model}{model used to calculate metric}
\item{Score}{type of score reported}
\item{strThreshold}{thresholds to be used for bounds and flags}
\item{nMinDenominator}{minimum denominator required to return a score and calculate a flag}
}
}
\source{
Generated from \code{sampleMetrics.csv} dataset in the \code{gsm} package.
}
\usage{
sampleMetrics
}
\description{
sampleMetrics Dataset
}
\keyword{datasets}
