% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{sampleResults}
\alias{sampleResults}
\title{sampleResults Dataset}
\format{
A data frame with 2574 rows and 10 columns:
\describe{
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Numerator}{number of flags in group}
\item{Denominator}{total subjects in group}
\item{Metric}{The calculated rate/metric value}
\item{Score}{The statistical score}
\item{Flag}{The ordinal flag to be applied}
\item{MetricID}{The Metric ID}
\item{StudyID}{The Study ID}
\item{SnapshotDate}{The Date of the snapshot}
}
}
\source{
Generated from \code{sampleResults.csv} dataset in the \code{gsm} package.
}
\usage{
sampleResults
}
\description{
sampleResults Dataset
}
\keyword{datasets}
