% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{sampleSummary}
\alias{sampleSummary}
\title{sampleSummary Dataset}
\format{
A data frame with 176 rows and 7 columns:
\describe{
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Numerator}{number of flags in group}
\item{Denominator}{total subjects in group}
\item{Metric}{calculated rate/metric value}
\item{Score}{statistical score}
\item{Flag}{ordinal flag to be applied}
}
}
\source{
Generated from \code{sampleSummary.csv} dataset in the \code{gsm} package.
}
\usage{
sampleSummary
}
\description{
sampleSummary Dataset
}
\keyword{datasets}
