% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveQTL.R
\name{SaveQTL}
\alias{SaveQTL}
\title{SaveQTL}
\usage{
SaveQTL(
  lSnapshot,
  strPath = "./inst/qtl_dummy_data/dummyqtldata.csv",
  bQuiet = TRUE
)
}
\arguments{
\item{lSnapshot}{\code{list} List returned by \code{\link[=RunQTL]{RunQTL()}}}

\item{strPath}{\code{character} Path to historical QTL data.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\description{
Save QTL analysis results to a directory location. The \code{strPath} argument specifies the filepath, including filename, of the current QTL analysis.
Running \code{SaveQTL} will save a new file with an added row for the single specified QTL analysis. The new file will be named the same as the original file,
but will incorporate the run date for version control.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
dispQTL <- RunQTL("qtl0006")
SaveQTL(dispQTL)
}

}
