% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screening_Map_Raw.R
\name{Screening_Map_Raw}
\alias{Screening_Map_Raw}
\title{Screening Assessment - Raw Mapping}
\usage{
Screening_Map_Raw(
  dfs = gsm::UseClindata(list(dfENROLL = "clindata::rawplus_enroll")),
  lMapping = gsm::Read_Mapping("rawplus"),
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfENROLL}: \code{data.frame} Subject-level study enrollment data with one record per subject. Default: \code{clindata::rawplus_enroll}
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column. Default: package-defined mapping for raw+.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{\link[=is_mapping_valid]{is_mapping_valid()}}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one record per subject, the input to \code{\link[=Screening_Assess]{Screening_Assess()}}. If
\code{bReturnChecks} is \code{TRUE} \code{Screening_Map_Raw} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert raw screening data to formatted input data to \code{\link[=Screening_Assess]{Screening_Assess()}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{Screening_Map_Raw} creates an input dataset for the Screening Assessment
\code{\link[=Screening_Assess]{Screening_Assess()}} by identifying Screen Failures (derived from \code{dfENROLL}).
}
\section{Data specification}{
\tabular{lllll}{
   \strong{Domain} \tab \strong{Column Key} \tab \strong{Default Value} \tab \strong{Required?} \tab \strong{Accept NA/Empty Values?} \cr
   dfENROLL \tab strSiteCol \tab siteid \tab TRUE \tab FALSE \cr
   dfENROLL \tab strIDCol \tab subjectid \tab TRUE \tab FALSE \cr
   dfENROLL \tab strScreenFailCol \tab enrollyn \tab TRUE \tab TRUE \cr
   dfENROLL \tab strScreenFailReasonCol \tab sfreas \tab TRUE \tab TRUE \cr
}
}

\examples{
# Run with defaults.
dfInput <- Screening_Map_Raw()

# Run with error checking and message log.
dfInput <- Screening_Map_Raw(bReturnChecks = TRUE, bQuiet = FALSE)

}
