% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-logger.R
\name{set_logger}
\alias{set_logger}
\title{Title}
\usage{
set_logger(
  strName = NULL,
  strOutputTarget = "file",
  strLogLevel = "INFO",
  lMetadata = list()
)
}
\arguments{
\item{strName}{User defined name for the logger/log}

\item{strOutputTarget}{Destination to sending logging to}

\item{strLogLevel}{level of logs to include}

\item{lMetadata}{}
}
\value{
logger object
}
\description{
Title
}
