% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-Logger.R
\name{SetLogger}
\alias{SetLogger}
\title{Set a Custom Logger Object}
\usage{
SetLogger(logger = logger())
}
\arguments{
\item{logger}{A logger object}
}
\value{
an option holding the logger
}
\description{
Set a Custom Logger Object
}
