% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{dfMetrics}{A data.frame of metric metadata, captured at the top of metric workflows and returned by \code{\link[=MakeMetric]{MakeMetric()}}}
}
\description{
Reused parameter definitions are gathered here for easier usage. Edit the
definitions in \verb{man/glossary/\{term\}.Rmd} (one file per term).
}
\keyword{internal}
