% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_Map_Raw.R
\name{Site_Map_Raw}
\alias{Site_Map_Raw}
\title{Map input site data to expected structure.}
\usage{
Site_Map_Raw(
  dfs = list(dfSITE = clindata::ctms_site, dfSUBJ = clindata::rawplus_dm),
  lMapping = gsm::Read_Mapping(c("ctms", "rawplus")),
  dfConfig = clindata::config_param
)
}
\arguments{
\item{dfs}{\code{list} Named list of data frames:
\itemize{
\item dfSUBJ: subject-level clinical
\item dfSITE: site-level CTMS data
}}

\item{lMapping}{\code{list} Named list of domain-level column mappings.}

\item{dfConfig}{\code{data.frame} Study-specific parameter configuration data.}
}
\description{
Map input site data to expected structure.
}
\keyword{internal}
