% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-StackSnapshots.R
\name{StackSnapshots}
\alias{StackSnapshots}
\title{Create longitudinal snapshot from results_summary.}
\usage{
StackSnapshots(cPath, lSnapshot = NULL)
}
\arguments{
\item{cPath}{\code{character} Path to longitudinal data folders.}

\item{lSnapshot}{\code{list} Optional. Pass the output of \code{\link[=Make_Snapshot]{Make_Snapshot()}} to be appended to historical results.}
}
\value{
\code{data.frame} containing longitudinal snapshots of \code{{gsm}} analyses.
}
\description{
Create longitudinal snapshot from results_summary.
}
\examples{

\dontrun{
sim <- clindata::run_simulation(n_sites = 20, n_subjects = 250)
dir <- here::here("simulation", "study-20_sites-250_subjects")
results_summary_over_time <- StackSnapshots(cPath = dir)
}

}
