% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cli_logger_wrappers.R
\name{stop_if}
\alias{stop_if}
\title{Custom stop message}
\usage{
stop_if(cnd, message)
}
\arguments{
\item{cnd}{condition for stopping}

\item{message}{message to display; may contain glue-style placeholders}
}
\description{
Custom stop message
}
