% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Assess.R
\name{Study_Assess}
\alias{Study_Assess}
\title{Run Multiple Assessments on a Study}
\usage{
Study_Assess(
  lData = NULL,
  lMapping = NULL,
  lAssessments = NULL,
  lSubjFilters = NULL,
  lTags = list(Study = "myStudy"),
  bQuiet = FALSE
)
}
\arguments{
\item{lData}{a named list of domain level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{lMapping}{a named list identifying the columns needed in each data domain.}

\item{lAssessments}{a named list of metadata defining how each assessment should be run. By default, \code{MakeAssessmentList()} imports YAML specifications from \code{inst/assessments}.}

\item{lSubjFilters}{Optionally specify Subject-level filters that will be applied to all assessments. For example \code{list(strRandFlagCol="Y")} could be used to subset to Randomized Participants.}

\item{lTags}{a named list of Tags to be passed to each assessment. Default is \code{list(Study="myStudy")} could be expanded to include other important metadata such as analysis population or study phase.}

\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}
}
\value{
A list of assessments containing status information and results.
}
\description{
\code{Study_Assess()} attempts to run multiple assessments using a shared set of data (\code{lData}) and an associated data mapping (\code{lMapping}). By default, the \code{rawplus} data and associated mapping from the {clindata} package is used, and all assessments defined in \code{inst/assessments} are evaluated. Individual assessments are run using \code{gsm::RunAssessment()}
}
\examples{
Study_Assess() # run using defaults

}
