% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_AssessmentReport.R
\name{Study_AssessmentReport}
\alias{Study_AssessmentReport}
\title{{experimental} Make Summary of 1 or more assessment Data checks}
\usage{
Study_AssessmentReport(lAssessments, bViewReport = FALSE)
}
\arguments{
\item{lAssessments}{\code{list} List of 1+ assessments like those created by \code{RunWorkflow()} or \code{Study_Assess()}}

\item{bViewReport}{\code{logical} HTML table of dfSummary that can be viewed in most IDEs. Default: \code{FALSE}}
}
\value{
\code{list} containing a \code{data.frame} summarizing the checks \code{dfSummary} and a \code{data.frame} listing all checks (\code{dfAllChecks}).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Make overview table with one row per assessment and one column per site showing flagged assessments.
}
\examples{

assessment <- Study_Assess(lData = list(
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::rawplus_protdev,
  dfSUBJ = clindata::rawplus_dm
))

report <- Study_AssessmentReport(lAssessments = assessment)

}
