% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Map_Raw.R
\name{Study_Map_Raw}
\alias{Study_Map_Raw}
\title{Map input study data to expected structure.}
\usage{
Study_Map_Raw(
  dfs = list(dfSTUDY = clindata::ctms_study, dfSUBJ = clindata::rawplus_dm),
  lMapping = gsm::Read_Mapping(c("ctms", "rawplus")),
  dfConfig = clindata::config_param
)
}
\arguments{
\item{dfs}{\code{list} Named list of data frames:
\itemize{
\item dfSUBJ: subject-level clinical
\item dfSTUDY: study-level CTMS data
}}

\item{lMapping}{\code{list} Named list of domain-level column mappings.}

\item{dfConfig}{\code{data.frame} Study-specific parameter configuration data.}
}
\description{
Map input study data to expected structure.
}
\keyword{internal}
