% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Report.R
\name{Study_Report}
\alias{Study_Report}
\title{Study Report}
\usage{
Study_Report(
  lAssessments,
  lMeta = list(Project = "My Project"),
  strOutpath = NULL
)
}
\arguments{
\item{lAssessments}{character vector listing assessments}

\item{lMeta}{list of metadata related to study}

\item{strOutpath}{path to save the report}
}
\value{
HTML report of study data.
}
\description{
Pulls needed study data and runs one or more assessments
}
\examples{
\dontrun{
assessment <- Study_Assess()
Study_Report(assessment, lMeta = list(study = "my study name"))
}

}
