% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{SubsetStackedSnapshots}
\alias{SubsetStackedSnapshots}
\title{SubsetStackedSnapshots}
\usage{
SubsetStackedSnapshots(strWorkflowId, lStackedSnapshots)
}
\arguments{
\item{strWorkflowId}{\code{character} workflow ID.}
}
\description{
This function is used to subset a list of stacked snapshots within \code{\link[=Make_Snapshot]{Make_Snapshot()}}.
}
\keyword{internal}
