% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_title.R
\name{tab_title}
\alias{tab_title}
\title{function for adding a title to a tabbed flexdashboard section}
\usage{
tab_title(title)
}
\arguments{
\item{title}{title to add above tabs}
}
\description{
function for adding a title to a tabbed flexdashboard section
}
\keyword{internal}
