% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesContinuous.R
\name{timeSeriesContinuous}
\alias{timeSeriesContinuous}
\title{Time-Series Continuous Plot}
\usage{
timeSeriesContinuous(
  kri,
  raw_results,
  raw_workflow,
  raw_param,
  selectedGroupIDs = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  addSiteSelect = TRUE
)
}
\arguments{
\item{kri}{selected workflow to filter data and display}

\item{raw_results}{results_summary_over_time}

\item{raw_workflow}{meta_workflow}

\item{raw_param}{meta_param}

\item{selectedGroupIDs}{specific group to highlight in the chart}

\item{width}{width of widget}

\item{height}{height of widget}

\item{elementId}{optional elementId of widget}

\item{addSiteSelect}{add a dropdown to highlight sites?}

\item{raw_param_over_time}{status_param_over_time}
}
\description{
A widget that displays a time-series plot based on longitudinal "snapshots" using \code{{gsm}}
}
