% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_Rate.R
\name{Transform_Rate}
\alias{Transform_Rate}
\title{Transform Rate}
\usage{
Transform_Rate(
  dfInput,
  strNumeratorCol = "Numerator",
  strDenominatorCol = "Denominator"
)
}
\arguments{
\item{dfInput}{A data.frame with one record per subject.}

\item{strNumeratorCol}{Required. Numerical or logical. Column to be counted.}

\item{strDenominatorCol}{\code{numeric} Required. Numerical \code{Exposure} column.}

\item{strGroupCol}{\code{character} Required. Name of column for grouping variable. Default: \code{"SiteID"}}
}
\value{
\code{data.frame} with one row per site with columns \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, and \code{Metric}.
}
\description{
Convert from input data format to needed input format to derive KRI for an Assessment. Calculate a site-level rate.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function transforms data to prepare it for the analysis step. It is currently only sourced for the Adverse Event, Disposition, Labs, and Protocol Deviations Assessments.
}
\section{Data Specification}{


(\code{dfInput}) must include the columns specified by \code{strNumeratorCol}, \code{strDenominatorCol} and \code{strGroupCol}.
Required columns include:
\itemize{
\item \code{SiteID} - Site ID
\item \code{StudyID} - Study ID
\item \code{CustomGroupID} - Custom Group ID
\item \code{Count} - Number of events of interest; the actual name of this column is specified by the parameter \code{strNumeratorCol}
\item \code{Exposure} - Number of days on treatment; the actual name of this column is specified by the parameter \code{strDenominatorCol}
}
}

\examples{
dfInput <- tibble::tibble(
 GroupID = c("G1", "G1", "G2", "G2"),
 GroupLevel = rep("site",4),
 Numerator = c(1, 2, 3, 4),
 Denominator = c(10, 20, 30, 40)
)

dfTransformed <- Transform_Rate(
  dfInput,
  strNumeratorCol = "Numerator",
  strDenominatorCol = "Denominator",
)

}
