% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-CombineSpecs.R
\name{update_column}
\alias{update_column}
\title{Update Column Specification}
\usage{
update_column(existing_col, new_col, col_name)
}
\arguments{
\item{existing_col}{A list representing the existing column specification (can be \code{NULL} if the column does not yet exist).}

\item{new_col}{A list representing the new column specification to be merged with the existing one.}
}
\value{
A list containing the updated column specification.
}
\description{
This function updates a column specification by handling deduplication, resolving conflicts in the \code{required} field, and checking for type mismatches.
}
\keyword{internal}
