% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDataLogs.R
\name{UpdateDataLogs}
\alias{UpdateDataLogs}
\title{Update archived snapshot logs pre v1.9.0 to the data model of post v1.9.0}
\usage{
UpdateDataLogs(lPrevSnapshot, lMeta = NULL, lData = NULL, lMapping = NULL)
}
\arguments{
\item{lPrevSnapshot}{\code{list} snapshot.rds file of archived snapshot}

\item{lMeta}{\code{list} lMeta data used in previous snapshot defaults to \code{Make_Snapshot()} lMeta default arguments if NULL}

\item{lData}{\code{list} lData argument of previous snapshot defaults to \code{Make_Snapshot()} lData default arguments if NULL}

\item{lMapping}{\code{list} lMapping argument of previous snapshot. defaults to \code{Read_Mapping()} output if NULL}
}
\value{
\code{list} list of corrected lSnapshot logs and lStackedSnapshots logs if available
}
\description{
Update archived snapshot logs pre v1.9.0 to the data model of post v1.9.0
}
\examples{
lPrevSnapshot <- readRDS(system.file("data-longitudinal", "AA-AA-000-0000", "2019-12-01", "snapshot.rds", package = "clindata"))
updated_lPrevSnapshot <- UpdateDataLogs(lPrevSnapshot)
}
