% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-UpdateGSMVersion.R
\name{UpdateGSMVersion}
\alias{UpdateGSMVersion}
\title{Update GSM Version}
\usage{
UpdateGSMVersion(version = NULL)
}
\arguments{
\item{version}{If \code{NULL} (the default), updates metadata to current version as indicated in the DESCRIPTION file.}
}
\value{
Updated metadata. Currently \link{meta_param} and \link{meta_workflow}
}
\description{
Automatically updates metadata that relies on the current \code{{gsm}} version.
}
\examples{
\dontrun{
UpdateGSMVersion()
}

}
