% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-UpdateLabels.R
\name{UpdateLabels}
\alias{UpdateLabels}
\title{Update visualization labels based on metadata.}
\usage{
UpdateLabels(lStudyAssessResults, dfMetaWorkflow)
}
\arguments{
\item{lStudyAssessResults}{\code{list} The output of \code{\link[=Study_Assess]{Study_Assess()}}.}

\item{dfMetaWorkflow}{\code{data.frame} The metadata for a workflow or set of workflows.}
}
\value{
\code{list} \code{lStudyAssessResults} with updated JavaScript visualizations, annotated with the values provided by \code{dfMetaWorkflow}.
}
\description{
If visualizations are created outside of a standard snapshot (e.g., \code{\link[=Make_Snapshot]{Make_Snapshot()}}), there is not enough descriptive metadata to assign
meaningful labels. \code{UpdateLabels} updates labels for JavaScript visualizations (from \code{rbm-viz}) using metadata provided by \code{dfMetaWorkflow.}
}
\examples{
\dontrun{
study <- Study_Assess()
updated_study <- UpdateLabels(study, gsm::meta_workflow)
}

}
