% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-testSkeleton.R
\name{use_gsm_test}
\alias{use_gsm_test}
\title{Create testthat skeleton}
\usage{
use_gsm_test(type = "map")
}
\arguments{
\item{type}{Choose "map" or "assess", depending on the type of test file you need.}
}
\description{
Create testthat skeleton
}
\examples{
\dontrun{
use_gsm_test() # default is "map"
use_gsm_test("map")
use_gsm_test("assess")
}
}
