% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_filter_df.R
\name{util_filter_df}
\alias{util_filter_df}
\title{Utility Function for Filtering Dataframes}
\usage{
util_filter_df(df, strCol = NULL, strValue = NULL)
}
\arguments{
\item{df}{dataframe, input}

\item{strCol}{character, column to be filtered on.}

\item{strValue}{value to be filtered on ()}
}
\value{
Filtered dataframe if strCol != NULL and strValue != NULL, otherwise input dataframe will be returned.
}
\description{
Utility Function for Filtering Dataframes
}
\examples{
util_filter_df(clindata::rawplus_covlab, strCol = "TOXFLG", strValue = 1 )


}
