% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-validate-input.R
\name{validate_df}
\alias{validate_df}
\title{Validate df}
\usage{
validate_df(
  df,
  allow_null = FALSE,
  required_colnames = character(),
  df_arg = rlang::caller_arg(df),
  call = rlang::caller_env()
)
}
\arguments{
\item{df}{\code{data.frame} The df to validate.}

\item{allow_null}{\code{logical} Whether the df can be empty.}

\item{required_colnames}{\code{character} 0 or more required columns.}

\item{df_arg}{\code{character} The name of the \code{df} in the calling function.}

\item{call}{\code{environment} Where the call came from, for error messaging.}
}
\value{
The validated df.
}
\description{
Validate df
}
\keyword{internal}
