% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Count.R
\name{Visualize_Count}
\alias{Visualize_Count}
\title{Site-level visualization of site-level Inclusion/Exclusion results}
\usage{
Visualize_Count(
  dfAnalyzed,
  strTotalCol = "N",
  strCountCol = "TotalCount",
  strTitle = ""
)
}
\arguments{
\item{dfAnalyzed}{Map results from IE or Consent assessments.}

\item{strTotalCol}{Column containing total of site-level participants. Default is "N" from \code{\link{Transform_EventCount}}.}

\item{strCountCol}{Column containing total number of site-level occurances. Default is "TotalCount" from \code{\link{Transform_EventCount}}.}

\item{strTitle}{Title of plot. NULL by default.}
}
\value{
site level plot object
}
\description{
Site-level visualization of site-level Inclusion/Exclusion results
}
\examples{
ie_input <- IE_Map_Raw(
   clindata::raw_ie_all ,
   clindata::rawplus_rdsl,
   strCategoryCol = 'IECAT_STD',
   vCategoryValues= c("EXCL","INCL"),
   strResultCol = 'IEORRES',
   vExpectedResultValues=c(0,1)
)

ie_assess <- IE_Assess(ie_input)
Visualize_Count(ie_assess$dfAnalyzed)

consent_input <- Consent_Map_Raw(
  dfConsent = clindata::raw_consent,
  dfRDSL = clindata::rawplus_rdsl,
  strConsentReason = NULL
)

consent_assess <- Consent_Assess(consent_input)
Visualize_Count(consent_assess$dfAnalyzed)

}
