% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Metric.R
\name{Visualize_Metric}
\alias{Visualize_Metric}
\title{Visualize_Metric Function}
\usage{
Visualize_Metric(
  dfSummary,
  dfBounds = NULL,
  dfGroups = NULL,
  dfMetrics = NULL,
  strMetricID = NULL,
  strSnapshotDate = NULL,
  bDebug = FALSE
)
}
\arguments{
\item{dfSummary}{\code{data.frame} A data.frame returned by \code{\link[=Summarize]{Summarize()}}.}

\item{dfBounds}{\code{data.frame}, A data.frame returned by \code{\link[=Analyze_NormalApprox_PredictBounds]{Analyze_NormalApprox_PredictBounds()}} or \code{\link[=Analyze_Poisson_PredictBounds]{Analyze_Poisson_PredictBounds()}}}

\item{dfGroups}{\code{data.frame} Site metadata.}

\item{dfMetrics}{\code{data.frame} Metrics metadata.}

\item{strMetricID}{\code{character} MetricID to subset the data.}

\item{strSnapshotDate}{\code{character} Snapshot date to subset the data.}
}
\value{
A list containing the following charts:
\itemize{
\item scatterJS: A scatter plot using JavaScript.
\item scatter: A scatter plot using ggplot2.
\item barMetricJS: A bar chart using JavaScript with metric on the y-axis.
\item barScoreJS: A bar chart using JavaScript with score on the y-axis.
\item barMetric: A bar chart using ggplot2 with metric on the y-axis.
\item barScore: A bar chart using ggplot2 with score on the y-axis.
\item timeSeriesContinuousScoreJS: A time series chart using JavaScript with score on the y-axis.
\item timeSeriesContinuousMetricJS: A time series chart using JavaScript with metric on the y-axis.
\item timeSeriesContinuousNumeratorJS: A time series chart using JavaScript with numerator on the y-axis.
}
}
\description{
The function creates all available charts for a metric using the data provided
}
