% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Poisson.R
\name{Visualize_Poisson}
\alias{Visualize_Poisson}
\title{Site-level visualization of site-level Poisson Model results}
\usage{
Visualize_Poisson(dfFlagged, dfBounds = NULL, strUnit = "days")
}
\arguments{
\item{dfFlagged}{analyze_poisson results with flags added.}

\item{dfBounds}{data.frame giving prediction bounds for range of dfFlagged}

\item{strUnit}{exposure time unit. Defaults to "days".}
}
\value{
site level plot object
}
\description{
Site-level visualization of site-level Poisson Model results
}
\examples{
dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )
dfTransformed <- Transform_EventCount( dfInput, strCountCol = 'Count', strExposureCol = "Exposure" )
dfAnalyzed <- Analyze_Poisson( dfTransformed)
dfFlagged <- Flag( dfAnalyzed , strColumn = 'Residuals', vThreshold =c(-5,5))

}
