% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_QTL.R
\name{Visualize_QTL}
\alias{Visualize_QTL}
\title{Visualize_QTL}
\usage{
Visualize_QTL(strQtlName, dfSummary, dfParams, dfAnalysis, lLabels)
}
\arguments{
\item{strQtlName}{\code{character} Name of QTL workflow, e.g., "qtl0004".}

\item{dfSummary}{\code{data.frame} Longitudinal data.}

\item{dfParams}{\code{data.frame} Longitudinal parameter/configuration data.}

\item{dfAnalysis}{\code{data.frame} Longitudinal QTL analysis results.}

\item{lLabels}{\code{list} Longitudinal workflow/metadata.}
}
\value{
\code{list} of HTML widgets.
}
\description{
Make cross-sectional timeseries charts.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
