% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Scatter.R
\name{Visualize_Scatter}
\alias{Visualize_Scatter}
\title{Group-level visualization of group-level results}
\usage{
Visualize_Scatter(
  dfResults,
  dfBounds = NULL,
  strGroupCol = NULL,
  strGroupLabel = NULL,
  strUnit = "days",
  vColors = c("#999999", "#FADB14", "#FF4D4F")
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[=BindResults]{BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID}, \code{SnapshotDate}.}

\item{dfBounds}{\code{data.frame} data.frame giving prediction bounds for range of denominator in dfResults.}

\item{strGroupCol}{\code{character} name of stratification column for facet wrap Default: \code{NULL}}

\item{strGroupLabel}{\code{character} name of group, used for labeling axes. Default: \code{NULL}}

\item{strUnit}{\code{character} exposure time unit. Default: \code{days}}

\item{vColors}{\code{character} vector of hex colors for plotting boundaries/thresholds. Index 1: mean; index 2: first threshold boundary; index 3: second threshold boundary.}
}
\value{
group-level plot object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
Visualize_Scatter(
  dfResults = reportingResults,
  dfBounds = reportingBounds
)

## Create Faceted output
Visualize_Scatter(
  dfResults = reportingResults,
  dfBounds = reportingBounds,
  strGroupCol = SnapshotDate,
  strGroupLabel = "Snapshot Date"
)

## Custom Colors
Visualize_Scatter(
  dfResults = reportingResults,
  dfBounds = reportingBounds,
  vColors = c("#F4E7E7", "#C17070", "#981212")
)
}

}
