% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Workflow.R
\name{Visualize_Workflow}
\alias{Visualize_Workflow}
\title{Visualize Workflow}
\usage{
Visualize_Workflow(lAssessment)
}
\arguments{
\item{lAssessment}{}
}
\value{
A flowchart.
}
\description{
Show the data structure of subject and site-level data in an assessment.
}
