% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_FlagOverTime.R
\name{Widget_FlagOverTime-shiny}
\alias{Widget_FlagOverTime-shiny}
\alias{Widget_FlagOverTimeOutput}
\alias{renderWidget_FlagOverTime}
\title{Shiny bindings for Widget_FlagOverTime}
\usage{
Widget_FlagOverTimeOutput(outputId, width = "100\%", height = "400px")

renderWidget_FlagOverTime(
  dfResults,
  dfMetrics,
  strGroupLevel = c("Site", "Study", "Country")
)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to
\code{\link[gsm.mapping:BindResults]{gsm.mapping::BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator},
\code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID},
\code{SnapshotDate}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting. Created by passing an \code{lWorkflow} object to \code{\link[gsm.mapping:MakeMetric]{gsm.mapping::MakeMetric()}}.
Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric},
\code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{Threshold}. For more
details see the Data Model vignette: \code{vignette("DataModel", package = "gsm")}.}

\item{strGroupLevel}{\code{character} Value for the group level. Default: "Site".}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Output and render functions for using Widget_FlagOverTime within
Shiny applications and interactive Rmd documents.
}
