% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_FlagOverTime.R
\name{Widget_FlagOverTime}
\alias{Widget_FlagOverTime}
\title{Flag Over Time Widget}
\usage{
Widget_FlagOverTime(
  dfResults,
  dfMetrics,
  strGroupLevel = c("Site", "Study", "Country"),
  strFootnote = NULL,
  width = "100\%",
  height = "400px"
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[=BindResults]{BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID}, \code{SnapshotDate}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and reporting. Created by passing an \code{lWorkflow} object to \code{\link[=MakeMetric]{MakeMetric()}}. Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric}, \code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{Threshold}. For more details see the Data Model vignette: \code{vignette("DataModel", package = "gsm")}.}

\item{strGroupLevel}{\code{character} Value for the group level. Default: "Site".}

\item{strFootnote}{\code{character} Text to insert for figure}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A widget that generates a table of flags over time using
\code{\link[=Report_FlagOverTime]{Report_FlagOverTime()}}.
}
\examples{
reportingResultsSubset <- dplyr::filter(
  reportingResults,
  GroupID \%in\% head(unique(reportingResults$GroupID))
)
Widget_FlagOverTime(
  dfResults = reportingResultsSubset,
  dfMetrics = reportingMetrics
)
}
