% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_ScatterPlot.R
\name{Widget_ScatterPlot-shiny}
\alias{Widget_ScatterPlot-shiny}
\alias{Widget_ScatterPlotOutput}
\alias{renderWidget_ScatterPlot}
\title{Shiny bindings for Widget_ScatterPlot}
\usage{
Widget_ScatterPlotOutput(outputId, width = "100\%", height = "400px")

renderWidget_ScatterPlot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a Widget_ScatterPlot}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Output and render functions for using Widget_ScatterPlot within Shiny
applications and interactive Rmd documents.
}
