% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_SiteOverview.R
\name{Widget_SiteOverview-shiny}
\alias{Widget_SiteOverview-shiny}
\alias{Widget_SiteOverviewOutput}
\alias{renderWidget_SiteOverview}
\title{Shiny bindings for Widget_SiteOverview}
\usage{
Widget_SiteOverviewOutput(outputId, width = "100\%", height = "400px")

renderWidget_SiteOverview(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a Widget_SiteOverview}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Output and render functions for using Widget_SiteOverview within Shiny
applications and interactive Rmd documents.
}
