% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_SiteOverview.R
\name{Widget_SiteOverview}
\alias{Widget_SiteOverview}
\title{Site Overview Table}
\usage{
Widget_SiteOverview(
  dfSummary,
  dfSite = clindata::ctms_site,
  dfMetrics = gsm::meta_workflow
)
}
\arguments{
\item{dfSummary}{\code{data.frame} Analysis results data with columns:
\itemize{
\item{\code{GroupID}}
\item{\code{WorkflowID}}
\item{\code{Numerator}}
\item{\code{Denominator}}
\item{\code{Metric}}
\item{\code{Score}}
\item{\code{Flag}}
}}

\item{dfSite}{\code{data.frame} Site metadata with columns:
\itemize{
\item{\code{studyid}}
\item{\code{siteid} (derived from \code{pi_number})}
\item{\code{account}}
\item{\code{pi_last_name}}
\item{\code{status}}
\item{\code{country}}
\item{\code{enrolled_participants} (derived)}
}}

\item{dfMetrics}{\code{data.frame} Workflow metadata with columns:
\itemize{
\item{\code{workflowid}}
\item{\code{abbreviation}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{metric}}
\item{\code{score}}
}}
}
\description{
A widget that displays a site overview table based on the output of KRI pipelines.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
data_raw <- gsm::UseClindata(
    list(
        "dfSUBJ" = "clindata::rawplus_dm",
        "dfAE" = "clindata::rawplus_ae",
        "dfPD" = "clindata::ctms_protdev",
        "dfLB" = "clindata::rawplus_lb",
        "dfSTUDCOMP" = "clindata::rawplus_studcomp",
        "dfSDRGCOMP" = "clindata::rawplus_sdrgcomp \%>\%
      dplyr::filter(.data$phase == 'Blinded Study Drug Completion')",
        "dfDATACHG" = "clindata::edc_data_points",
        "dfDATAENT" = "clindata::edc_data_pages",
        "dfQUERY" = "clindata::edc_queries",
        "dfENROLL" = "clindata::rawplus_enroll"
    )
)

workflows <- MakeWorkflowList()

data_mapped <- RunWorkflow(workflows$mapping, data_raw)$lData

results <- workflows[ grepl('^kri', names(workflows) ) ] \%>\%
    map(\(workflow) RunWorkflow(workflow, data_mapped))

dfSummary <- results \%>\% map_dfr(~ {
    data <- .x$lData$dfSummary
    data$workflowid <- .x$name
    data
})

counts <- RunWorkflow(workflows$counts, data_mapped)

dfSite <- clindata::ctms_site \%>\%
    mutate(
        siteid = .data$site_num,
        invname = paste(.data$pi_last_name, .data$pi_first_name, sep = ', ')
    ) \%>\%
    left_join(
        counts$lData$dfSiteCounts,
        c('site_num' = 'GroupID')
    ) \%>\%
    rename(
        status = site_status,
        enrolled_participants = nSubjects
    )

dfMetrics <- results \%>\% map_dfr(~ {
    metric <- .x$meta
    metric$workflowid <- metric$MetricID
    metric$vThreshold <- paste(metric$vThreshold, collapse = ',')
    metric
})

Widget_SiteOverview(
  dfSummary,
  dfSite,
  dfMetrics
)
}
}
