% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_TimeSeriesQTL.R
\name{Widget_TimeSeriesQTL}
\alias{Widget_TimeSeriesQTL}
\title{Time Series QTL}
\usage{
Widget_TimeSeriesQTL(
  qtl,
  dfSummary,
  lLabels,
  dfParams,
  dfAnalysis,
  selectedGroupIDs = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{qtl}{specific qtl to filter to}

\item{dfSummary}{\code{data.frame} Longitudinal data, typically \code{rpt_site_kri_details} from \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}

\item{lLabels}{\code{list} Longitudinal workflow/metadata, typically \code{rpt_kri_details} from \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}

\item{dfParams}{\code{data.frame} Longitudinal parameter/configuration data, typically \code{rpt_kri_threshold_param} from \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}

\item{dfAnalysis}{\code{data.frame} Longitudinal QTL analysis results, typically \code{rpt_qtl_analysis} from  \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}

\item{selectedGroupIDs}{TODO}

\item{width}{the width of the widget}

\item{height}{the height of the widget}

\item{elementId}{id of widget, automatically generated if not supplied}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
A Time Series graphic for qtl data
}
