% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Assess.R
\name{AE_Assess}
\alias{AE_Assess}
\title{AE Assessment using Poisson Regression}
\usage{
AE_Assess(
  dfInput,
  dThreshold = NULL,
  cLabel = "",
  method = "poisson",
  bDataList = FALSE
)
}
\arguments{
\item{dfInput}{input data}

\item{cLabel}{Assessment label}

\item{method}{valid methods are "poisson" (the default), "wilcoxon" and "negativebinomial"}

\item{bDataList}{Should all assessment datasets be returned as a list? If False (the default), only the finding data frame is returned}

\item{lThreshold}{list of threshold values}

\item{nCutoff}{optional parameter to control the auto-thresholding}
}
\value{
Finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
AE Assessment using Poisson Regression
}
