% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Assess.R
\name{AE_Assess}
\alias{AE_Assess}
\title{AE Assessment}
\usage{
AE_Assess(
  dfInput,
  vThreshold = NULL,
  cLabel = "",
  method = "poisson",
  bDataList = FALSE
)
}
\arguments{
\item{dfInput}{input data}

\item{vThreshold}{list of threshold values default c(-5,5) for method = "poisson", c(.0001,NA) for method = Wilcoxon}

\item{cLabel}{Assessment label}

\item{method}{valid methods are "poisson" (the default), or  "wilcoxon"}

\item{bDataList}{Should all assessment datasets be returned as a list? If False (the default), only the finding data frame is returned}

\item{nCutoff}{optional parameter to control the auto-thresholding}
}
\value{
Finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
AE Assessment
}
\examples{
dfInput <- AE_Map( safetyData::adam_adsl, safetyData::adam_adae )
SafetyAE <- AE_Assess( dfInput )
SafetyAE_Wilk <- AE_Assess( dfInput, method="wilcoxon")

}
