% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Assess.R
\name{AE_Assess}
\alias{AE_Assess}
\title{Adverse Event Assessment}
\usage{
AE_Assess(
  dfInput,
  vThreshold = NULL,
  strMethod = "poisson",
  lTags = list(Assessment = "AE"),
  bChart = TRUE,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data, a data frame with one record per subject.}

\item{vThreshold}{\code{numeric} Threshold specification, a vector of length 2 that defaults to \code{c(-5, 5)} for \code{strMethod} = "poisson" and \code{c(.0001, NA)} for \code{strMethod} = "wilcoxon".}

\item{strMethod}{\code{character} Statistical model. Valid values include "poisson" (default) and  "wilcoxon".}

\item{lTags}{\code{list} Assessment tags, a named list of tags describing the assessment that defaults to \code{list(Assessment="AE")}. \code{lTags} is returned as part of the assessment (\code{lAssess$lTags}) and each tag is added as a column in \code{lAssess$dfSummary}.}

\item{bChart}{\code{logical} Generate data visualization? Default: \code{TRUE}}

\item{bReturnChecks}{\code{logical} Return input checks from \code{is_mapping_valid}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} Assessment, a named list with:
\itemize{
\item each data frame in the data pipeline
\itemize{
\item \code{dfInput}
\item \code{dfTransformed}, returned by \code{\link{Transform_EventCount}}
\item \code{dfAnalyzed}, returned by \code{\link{dfAnalyzed}}
\item \code{dfFlagged}, returned by \code{\link{dfFlagged}}
\item \code{dfSummary}, returned by \code{\link{dfSummary}}
}
\item assessment metadata
\itemize{
\item \code{strFunctionName}
\item \code{lParams}
\item \code{lTags}
}
\item output(s)
\itemize{
\item \code{chart}
}
}
}
\description{
Flag sites that may be over- or under-reporting adverse events (AEs).
}
\details{
The Adverse Event Assessment uses the standard \href{https://github.com/Gilead-BioStats/gsm/wiki/Data-Pipeline-Vignette}{GSM data pipeline} to flag possible outliers. Additional details regarding the data pipeline and statistical
methods are described below.
}
\section{Data specification}{
\if{html}{\out{
<table>
<thead>
<tr>
<th style="text-align:left;">
domain
</th>
<th style="text-align:left;">
col_key
</th>
<th style="text-align:left;">
col_value
</th>
<th style="text-align:left;">
vRequired
</th>
<th style="text-align:left;">
vUniqueCols
</th>
</tr>
</thead>
<tbody>
<tr>
<td style="text-align:left;">
dfInput
</td>
<td style="text-align:left;">
strIDCol
</td>
<td style="text-align:left;">
SubjectID
</td>
<td style="text-align:left;">
TRUE
</td>
<td style="text-align:left;">
TRUE
</td>
</tr>
<tr>
<td style="text-align:left;">
dfInput
</td>
<td style="text-align:left;">
strSiteCol
</td>
<td style="text-align:left;">
SiteID
</td>
<td style="text-align:left;">
TRUE
</td>
<td style="text-align:left;">
FALSE
</td>
</tr>
<tr>
<td style="text-align:left;">
dfInput
</td>
<td style="text-align:left;">
strCountCol
</td>
<td style="text-align:left;">
Count
</td>
<td style="text-align:left;">
TRUE
</td>
<td style="text-align:left;">
FALSE
</td>
</tr>
<tr>
<td style="text-align:left;">
dfInput
</td>
<td style="text-align:left;">
strExposureCol
</td>
<td style="text-align:left;">
Exposure
</td>
<td style="text-align:left;">
TRUE
</td>
<td style="text-align:left;">
FALSE
</td>
</tr>
<tr>
<td style="text-align:left;">
dfInput
</td>
<td style="text-align:left;">
strRateCol
</td>
<td style="text-align:left;">
Rate
</td>
<td style="text-align:left;">
TRUE
</td>
<td style="text-align:left;">
FALSE
</td>
</tr>
</tbody>
</table>
}}
}

\section{Statistical assumptions}{
A Poisson or Wilcoxon model is used to generate estimates and p-values
for each site (as specified with the \code{strMethod} parameter). Those model
outputs are then used to flag possible outliers using the thresholds
specified in \code{vThreshold}. In the Poisson model, sites with an estimate
less than -5 are flagged as -1 and greater than 5 are flagged as 1 by
default. For Wilcoxon, sites with p-values less than 0.0001 are flagged
by default.

See \code{\link[=Analyze_Poisson]{Analyze_Poisson()}} and
\code{\link[=Analyze_Wilcoxon]{Analyze_Wilcoxon()}} for additional
details about the statistical methods and their assumptions.
}

\examples{
dfInput <- AE_Map_Raw()
ae_assessment_poisson <- AE_Assess(dfInput)
ae_assessment_wilcoxon <- AE_Assess(dfInput, strMethod = "wilcoxon")

}
