% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Assess.R
\name{AE_Assess}
\alias{AE_Assess}
\title{AE Assessment}
\usage{
AE_Assess(dfInput, vThreshold = NULL, strLabel = "", strMethod = "poisson")
}
\arguments{
\item{dfInput}{input data with one record per person and the following required columns: SubjectID, SiteID, Count, Exposure}

\item{vThreshold}{numeric vector with 2 threshold values.  Defaults to c(-5,5) for method = "poisson" and c(.0001,NA) for method = Wilcoxon.}

\item{strLabel}{Assessment label}

\item{strMethod}{valid methods are "poisson" (the default), or  "wilcoxon"}
}
\value{
A list containing all data and metadata in the standard data pipeline (\code{dfInput}, \code{dfTransformed}, \code{dfAnalyzed}, \code{dfFlagged}, \code{dfSummary}, \code{strFunctionName}, and \code{lParams}) is returned.
}
\description{
The Adverse Event Assessment flags sites that may be over- or under-reporting adverse events.
}
\details{
The Adverse Event Assessment uses the standard GSM data pipeline (TODO add link to data vignette) to flag possible outliers. More details regarding the data pipeline and statistical methods are described below.
}
\section{Data Specification}{


The input data (\code{dfInput}) for the AE Assessment is typically created using \code{\link{AE_Map_Raw}} or \code{\link{AE_Map_Adam}} and should be one record per person with columns for:
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{Count} - Number of Adverse Events
\item \code{Exposure} - Number of days of exposure
\item \code{Rate} - Rate of Exposure (Count / Exposure)
}

The Assessment
\itemize{
\item \code{\link{Transform_EventCount}} creates \code{dfTransformed}.
\item \code{\link{Analyze_Poisson}} or \code{\link{Analyze_Wilcoxon}} creates \code{dfAnalyzed}.
\item \code{\link{Flag}} creates \code{dfFlagged}.
\item \code{\link{Summarize}} creates \code{dfSummary}.
}
}

\section{Statistical Assumptions}{


A Poisson or Wilcoxon model is used to generate estimates and p-values for each site (as specified with the \code{strMethod} parameter). Those model outputs are then used to flag possible outliers using the thresholds specified in \code{vThreshold}. In the Poisson model, sites with an estimand less than -5 are flagged as -1 and greater than 5 are flagged as 1 by default. For Wilcoxon, sites with p-values less than 0.0001 are flagged by default.

See \code{\link{Analyze_Poisson}} and \code{\link{Analyze_Wilcoxon}} for additional details about the statistical methods and thier assumptions.
}

\examples{
dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )
SafetyAE <- AE_Assess( dfInput )
SafetyAE_Wilk <- AE_Assess( dfInput, strMethod="wilcoxon")$dfSummary

}
