% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Map_Adam.R
\name{AE_Map_Adam}
\alias{AE_Map_Adam}
\title{Adverse Event Assessment - ADaM Mapping}
\usage{
AE_Map_Adam(
  dfs = list(dfADSL = safetyData::adam_adsl, dfADAE = safetyData::adam_adae),
  lMapping = NULL,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfADAE}: \code{data.frame} Event-level data with one record per AE.
\item \code{dfADSL}: \code{data.frame} Subject-level data with one record per subject.
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name of the column.}

\item{bReturnChecks}{\code{logical} Return input checks from {gsm::is_mapping_valid()}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame with one record per subject, the input to {gsm::AE_Assess()}. If
\code{bReturnChecks} is \code{TRUE} \code{AE_Map_Adam} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert analysis adverse event (AE) data, by default ADaM data, to formatted input data to
{gsm::AE_Assess()}.
}
\details{
\code{AE_Map_Adam} combines AE data with subject-level treatment exposure data to create formatted
input data to {gsm::AE_Assess()}. This function creates an input dataset for the AE Assessment
({gsm::AE_Assess()}) by binding subject-level AE counts (derived from \code{dfADAE}) to subject-level
data (from \code{dfADSL}). Note that the function can generate data summaries for specific types of
AEs by passing filtered AE data to \code{dfADAE}.
}
\section{Data specification}{
\tabular{llll}{
   domain \tab col_key \tab col_value \tab vRequired \cr
   dfADSL \tab strIDCol \tab USUBJID \tab TRUE \cr
   dfADSL \tab strSiteCol \tab SITEID \tab TRUE \cr
   dfADSL \tab strStartCol \tab TRTSDT \tab TRUE \cr
   dfADSL \tab strEndCol \tab TRTEDT \tab TRUE \cr
   dfADAE \tab strIDCol \tab USUBJID \tab TRUE \cr
}
}

\examples{
# Run with defaults
dfInput <- AE_Map_Adam()

# Run with error checking and message log
dfInput <- AE_Map_Adam(bReturnChecks = TRUE, bQuiet = FALSE)

}
