% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Map_Adam.R
\name{AE_Map_Adam}
\alias{AE_Map_Adam}
\title{AE Assessment - ADaM Mapping}
\usage{
AE_Map_Adam(dfADSL, dfADAE)
}
\arguments{
\item{dfADSL}{ADaM demographics data with the following required columns:  USUBJID, SITEID, TRTEDT (end date), TRTSDT (start date)}

\item{dfADAE}{ADaM AE data with the following required columns: USUBJID}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count (Number of Adverse Events), Exposure (Time on Treatment in Days), Rate (AEs/Day)
}
\description{
Convert from ADaM format to input format for Safety Assessment.
}
\details{
This function maps from ADaM ADSL and ADAE data to the required input for \code{\link{AE_Assess}}.
}
\section{Data Specification}{


This function creates an input dataset for the Adverse Event Assessment (\code{\link{AE_Assess}}) by adding Adverse Event Counts (from \code{dfADAE}) to basic subject-level data (from \code{dfADSL}).

The following columns are required:
\itemize{
\item \code{dfADSL}
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{SITEID} - Site ID
\item \code{TRTEDT} - Treatment End date
\item \code{TRTSDT} - Treatment Start date
}
\item \code{dfADAE}
\itemize{
\item \code{USUBJID} - Unique subject ID
}
}

Note that the function can generate data summaries for specific types of AEs by passing filtered ADAE data to dfADAE.
}

\examples{
dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )

}
