% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Map_Adam.R
\name{AE_Map_Adam}
\alias{AE_Map_Adam}
\title{AE Assessment - ADaM Mapping}
\usage{
AE_Map_Adam(
  dfs = list(dfADSL = safetyData::adam_adsl, dfADAE = safetyData::adam_adae),
  lMapping = NULL,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfADAE}: \code{data.frame} One record per AE
\item \code{dfADSL}: \code{data.frame} One record per subject
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name of the column.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{is_mapping_valid}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame with one record per subject and these columns:\tabular{ll}{
   Name \tab Description \cr
   \code{SubjectID} \tab Unique Subject Identifier \cr
   \code{SiteID} \tab Site Identifier \cr
   \code{Count} \tab Number of Adverse Events \cr
   \code{Exposure} \tab Number of Exposure Days \cr
   \code{Rate} \tab Exposure Rate (\code{Count} / \code{Exposure}) \cr
}


If \code{bReturnChecks} is \code{TRUE} \code{AE_Map_Adam} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert analysis adverse event (AE) data, by default ADaM data, to formatted input data to AE
Assessment.
}
\details{
Combines AE data with subject-level treatment exposure data to create formatted input data to
\code{\link{AE_Assess}}.
}
\section{Data Specification}{


This function creates an input dataset for the AE Assessment (\code{\link{AE_Assess}}) by binding
subject-level adverse event counts (from \code{dfADAE}) to subject-level data (from \code{dfADSL}).\tabular{lllll}{
   Domain \tab Key \tab Value \tab Description \tab Required? \cr
   \code{dfADAE} \tab \code{strIDCol} \tab USUBJID \tab Unique Subject Identifier \tab Yes \cr
   \code{dfADSL} \tab \code{strIDCol} \tab USUBJID \tab Unique Subject Identifier \tab Yes \cr
   \code{dfADSL} \tab \code{strSiteCol} \tab SITEID \tab Site Identifier \tab Yes \cr
   \code{dfADSL} \tab \code{strStartCol} \tab TRTSDT \tab Treatment Start Date \tab Yes \cr
   \code{dfADSL} \tab \code{strEndCol} \tab TRTEDT \tab Treatment End Date \tab Yes \cr
}


Note that the function can generate data summaries for specific types of AEs by passing filtered
adverse event data to \code{dfADAE}.
}

\examples{
# Run with defaults
dfInput <- AE_Map_Adam()

# Run with error checking and message log
dfInput <- AE_Map_Adam(bReturnChecks = TRUE, bQuiet = FALSE)

}
