% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Map_Adam.R
\name{AE_Map_Adam}
\alias{AE_Map_Adam}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
AE_Map_Adam(
  dfs = list(dfADSL = safetyData::adam_adsl, dfADAE = safetyData::adam_adae),
  lMapping = gsm::Read_Mapping("adam"),
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfADAE}: \code{data.frame} Event-level data with one record per AE. Default: \code{safetyData::adam_adsl}
\item \code{dfADSL}: \code{data.frame} Subject-level data with one record per subject. Default: \code{safetyData::adam_adae}
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name of the column. Default: package-defined mapping for ADaM.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{\link[=is_mapping_valid]{is_mapping_valid()}}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame with one record per subject, the input to \code{\link[=AE_Assess]{AE_Assess()}}. If
\code{bReturnChecks} is \code{TRUE} \code{AE_Map_Adam} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert analysis adverse event (AE) data, by default ADaM data, to formatted input data to
\code{\link[=AE_Assess]{AE_Assess()}}.
}
\details{
Adverse Event Assessment - ADaM Mapping

\code{AE_Map_Adam} combines AE data with subject-level treatment exposure data to create formatted
input data to \code{\link[=AE_Assess]{AE_Assess()}}. This function creates an input dataset for the AE Assessment
(\code{\link[=AE_Assess]{AE_Assess()}}) by binding subject-level AE counts (derived from \code{dfADAE}) to subject-level
data (from \code{dfADSL}). Note that the function can generate data summaries for specific types of
AEs by passing filtered AE data to \code{dfADAE}.
}
\examples{
# Run with defaults
dfInput <- AE_Map_Adam()

# Run with error checking and message log
dfInput <- AE_Map_Adam(bReturnChecks = TRUE, bQuiet = FALSE)

}
