% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Map_Raw.R
\name{AE_Map_Raw}
\alias{AE_Map_Raw}
\title{AE Assessment - Raw Mapping}
\usage{
AE_Map_Raw(
  dfs = list(dfAE = clindata::rawplus_ae, dfSUBJ = clindata::rawplus_subj),
  lMapping = clindata::mapping_rawplus,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfAE}: \code{data.frame} One record per AE
\item \code{dfSUBJ}: \code{data.frame} One record per subject
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name of the column.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{is_mapping_valid}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame with one record per subject and these columns:\tabular{ll}{
   Name \tab Description \cr
   \code{SubjectID} \tab Unique Subject Identifier \cr
   \code{SiteID} \tab Site Identifier \cr
   \code{Count} \tab Number of Adverse Events \cr
   \code{Exposure} \tab Number of Exposure Days \cr
   \code{Rate} \tab Exposure Rate (\code{Count} / \code{Exposure}) \cr
}


If \code{bReturnChecks} is \code{TRUE} \code{AE_Map_Raw} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert raw adverse event (AE) data, typically processed case report form data, to formatted
input data to AE Assessment.
}
\details{
Combines AE data with subject-level treatment exposure data to create formatted input data to
\code{\link{AE_Assess}}.
}
\section{Data Specification}{


This function creates an input dataset for the AE Assessment (\code{\link{AE_Assess}}) by binding
subject-level adverse event counts (derived from \code{dfAE}) to subject-level data (from \code{dfSUBJ}).\tabular{lllll}{
   Domain \tab Key \tab Value \tab Description \tab Required? \cr
   \code{dfAE} \tab \code{strIDCol} \tab SubjectID \tab Unique Subject Identifier \tab Yes \cr
   \code{dfSUBJ} \tab \code{strIDCol} \tab SubjectID \tab Unique Subject Identifier \tab Yes \cr
   \code{dfSUBJ} \tab \code{strSiteCol} \tab SiteID \tab Site Identifier \tab Yes \cr
   \code{dfSUBJ} \tab \code{strTimeOnTreatmentCol} \tab TimeOnTreatment \tab Number of Exposure Days \tab Yes \cr
}


Note that the function can generate data summaries for specific types of AEs by passing filtered
adverse event data to \code{dfAE}.
}

\examples{
# Run with defaults
dfInput <- AE_Map_Raw()

# Run with error checking and message log
dfInput <- AE_Map_Raw(bReturnChecks = TRUE, bQuiet = FALSE)

}
