% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Map_Raw.R
\name{AE_Map_Raw}
\alias{AE_Map_Raw}
\title{Safety Assessment Mapping from Raw Data- Make Input Data}
\usage{
AE_Map_Raw(dfAE = NULL, dfExposure = NULL)
}
\arguments{
\item{dfExposure}{exposure dataset calculated via \code{\link{Transform_Exposure}}}

\item{dfAe}{AE dataset with columns SUBJID and rows for each AE record}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count, Exposure, Rate, Unit
}
\description{
Convert from raw data format to needed input format for Safety Assessment
Requires the following raw datasets: subid, ex, ae, da (optional)
}
