% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Map_Raw.R
\name{AE_Map_Raw}
\alias{AE_Map_Raw}
\title{AE Assessment - Raw Mapping}
\usage{
AE_Map_Raw(dfAE = NULL, dfRDSL = NULL, strExposureCol = "TimeOnTreatment")
}
\arguments{
\item{dfAE}{AE dataset with columns SUBJID and rows for each AE record}

\item{dfRDSL}{Subject-level Raw Data (RDSL) required columns: SubjectID, SiteID, value specified in strExposureCol}

\item{strExposureCol}{Name of exposure column. 'TimeOnTreatment' by default}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count (number of AEs), Exposure (Time on Treatment in Days), Rate (AE/Day)
}
\description{
Convert Raw data, typically processed case report form data - to input format for Safety Assessment.
}
\details{
This function combines AE data with treatment exposure from subject-level Raw Data (RDSL) to create the required input for \code{\link{AE_Assess}}.
}
\section{Data Specification}{


This function creates an input dataset for the Adverse Event Assessment (\code{\link{AE_Assess}}) by adding Adverse Event Counts to basic subject-level treatment exposure data from \code{clindata::TreatmentExposure}.

The following columns are required:
\itemize{
\item \code{dfAE}
\itemize{
\item \code{SUBJID} - Unique subject ID
}
\item \code{dfRDSL}
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item Value specified in strExposureCol - Treatment Exposure in days; "TimeOnTreatment" by default
}
}

Note that the function can generate data summaries for specific types of AEs, but passing filtered ADAE data to dfADAE.
}

\examples{
 dfInput <- AE_Map_Raw(clindata::raw_ae, clindata::rawplus_rdsl)

}
