% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Map_Raw.R
\name{AE_Map_Raw}
\alias{AE_Map_Raw}
\title{Adverse Event Assessment - Raw Mapping}
\usage{
AE_Map_Raw(
  dfs = gsm::UseClindata(list(dfSUBJ = "clindata::rawplus_dm", dfAE =
    "clindata::rawplus_ae")),
  lMapping = gsm::Read_Mapping("rawplus"),
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfAE}: \code{data.frame} Event-level data with one record per AE. Default: \code{clindata::rawplus_ae}
\item \code{dfSUBJ}: \code{data.frame} Subject-level data with one record per subject. Default: \code{clindata::rawplus_dm}
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column. Default: package-defined mapping for raw+.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{\link[=is_mapping_valid]{is_mapping_valid()}}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame with one record per subject, the input to \code{\link[=AE_Assess]{AE_Assess()}}. If
\code{bReturnChecks} is \code{TRUE} \code{AE_Map_Raw} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert raw adverse event (AE) data, typically processed case report form data, to formatted
input data to \code{\link[=AE_Assess]{AE_Assess()}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{AE_Map_Raw} combines AE data with subject-level treatment exposure data to create formatted
input data to \code{\link[=AE_Assess]{AE_Assess()}}. This function creates an input dataset for the AE Assessment
(\code{\link[=AE_Assess]{AE_Assess()}}) by binding subject-level AE counts (derived from \code{dfAE}) to subject-level
data (from \code{dfSUBJ}). Note that the function can generate data summaries for specific types of
AEs by passing filtered AE data to \code{dfAE}.
}
\examples{
# Run with defaults.
dfInput <- AE_Map_Raw()

# Run with error checking and message log.
dfInput <- AE_Map_Raw(bReturnChecks = TRUE, bQuiet = FALSE)

}
