% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE_Poisson_PredictBounds.R
\name{AE_Poisson_PredictBounds}
\alias{AE_Poisson_PredictBounds}
\title{Safety Assessment - Calculate prediction bounds}
\usage{
AE_Poisson_PredictBounds(dfOutput)
}
\arguments{
\item{dfOutput}{data frame in format produced by \code{\link{AE_Map_Adam}} or \code{\link{AE_Map_Raw}}.}
}
\value{
dataframe with prediction boundary curves
}
\description{
Safety Assessment - Calculate prediction bounds
}
