% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Chisq.R
\name{Analyze_Chisq}
\alias{Analyze_Chisq}
\title{Chi-squared Test Analysis}
\usage{
Analyze_Chisq(dfTransformed, strOutcome = "TotalCount", bQuiet = TRUE)
}
\arguments{
\item{dfTransformed}{\code{data.frame} in format produced by \code{\link{Transform_EventCount}}}

\item{strOutcome}{\code{character} required, name of column in dfTransformed dataset to perform the chi-squared test on. Default is "TotalCount".}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one row per site with columns: SiteID, TotalCount, TotalCount_Other, N, N_Other, Prop, Prop_Other, Statistic, PValue.
}
\description{
Creates Analysis results data for count data using the chi-squared test
}
\details{
Analyzes count data using the chi-squared test
}
\section{Statistical Methods}{


TODO Coming soon ...
}

\section{Data Specification}{


The input data (\code{dfTransformed}) for Analyze_Chisq is typically created using \code{\link{Transform_EventCount}} and should be one record per site with required columns for:
\itemize{
\item \code{SiteID} - Site ID
\item \code{N} - Total number of participants at site
\item \code{TotalCount} - Total number of participants at site with event of interest
}
}

\examples{
dfInput <- Disp_Map(dfDisp = safetyData::adam_adsl,
                    strCol = "DCREASCD",
                    strReason = "Adverse Event")

dfTransformed <- Transform_EventCount(dfInput,
                                      strCountCol = "Count",
                                      strKRILabel = "Discontinuations due to AE/Month")

dfAnalyzed <- Analyze_Chisq(dfTransformed)

}
