% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Fisher.R
\name{Analyze_Fisher}
\alias{Analyze_Fisher}
\title{Fisher's Exact Test Analysis}
\usage{
Analyze_Fisher(dfTransformed, strOutcome = "TotalCount", bQuiet = TRUE)
}
\arguments{
\item{dfTransformed}{\code{data.frame} in format produced by \code{\link{Transform_EventCount}}}

\item{strOutcome}{\code{character} required, name of column in dfTransformed dataset to perform Fisher test on. Default is "TotalCount".}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one row per site with columns: SiteID, TotalCount, TotalCount_Other, N, N_Other, Prop, Prop_Other, Estimate, PValue.
}
\description{
Creates Analysis results data for count data using the Fisher's exact test
}
\details{
Analyzes count data using the Fisher's exact test
}
\section{Statistical Methods}{


TODO Coming soon ...
}

\section{Data Specification}{


The input data (\code{dfTransformed}) for Analyze_Fisher is typically created using \code{\link{Transform_EventCount}} and should be one record per site with required columns for:
\itemize{
\item \code{SiteID} - Site ID
\item \code{N} - Total number of participants at site
\item \code{TotalCount} - Total number of participants at site with event of interest
}
}

\examples{
dfInput <- Disp_Map_Raw()
dfTransformed <- Transform_EventCount(
                   dfInput,
                   strCountCol = "Count",
                   strKRILabel = "Discontinuations due to AE"
                 )
dfAnalyzed <- Analyze_Fisher(dfTransformed)

}
