% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Identity.R
\name{Analyze_Identity}
\alias{Analyze_Identity}
\title{Analyze Identity}
\usage{
Analyze_Identity(dfTransformed, strValueCol = "KRI", strLabelCol = "KRIColumn")
}
\arguments{
\item{dfTransformed}{\code{data.frame} created by \code{Transform_EventCount()}}

\item{strValueCol}{\code{character} Name of column that will be copied as \code{Score}}

\item{strLabelCol}{\code{character} Name of column that will be copied as \code{ScoreLabel}}
}
\value{
\code{data.frame} that adds two columns to \code{dfTransformed}: \code{Score} and \code{ScoreLabel}
}
\description{
Used in the data pipeline between \code{Transform} and \code{Flag} to rename KRI and Score columns.
}
