% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Identity.R
\name{Analyze_Identity}
\alias{Analyze_Identity}
\title{Identity Analysis.}
\usage{
Analyze_Identity(dfTransformed, strValueCol = "Metric")
}
\arguments{
\item{dfTransformed}{\code{data.frame} created by \code{\link{Transform_Count}}}

\item{strValueCol}{\code{character} Name of column that will be copied as \code{Score}}
}
\value{
\code{data.frame} with one row per site with columns: GroupID, TotalCount, Metric, and Score.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Used in the data pipeline between \code{Transform} and \code{Flag} to rename KRI and Score columns.

More information can be found in \href{https://gilead-biostats.github.io/gsm/articles/KRI\%20Method.html#the-identity-method}{The Identity Method}
of the KRI Method vignette.
}
\section{Data Specification}{


The input data (\code{dfTransformed}) for \code{Analyze_Identity} is typically created using \code{\link{Transform_Rate}} and should be one record per site with required columns for:
\itemize{
\item \code{GroupID} - Site ID
\item \code{Numerator} - Total number of participants at site with event of interest.
\item \code{Denominator} - Total number of participants at site/Total number of days of exposure at site.
\item \code{Metric} - Proportion of participants at site with event of interest/Rate of events at site (Numerator / Denominator).
}
}

\examples{
dfInput <- tibble::tribble(
  ~SubjectID, ~SiteID, ~StudyID, ~CountryID, ~CustomGroupID, ~Exposure, ~Count, ~Rate,
  "0496", "5", "AA-AA-000-0000", "US", "0X167", 730, 5, 5/720,
  "1350", "78", "AA-AA-000-0000", "US", "0X002", 50, 2, 2/50,
  "0539", "139", "AA-AA-000-0000", "US", "0X052", 901, 5, 5/901,
  "0329", "162", "AA-AA-000-0000", "US", "0X049", 370, 3, 3/370,
  "0429", "29", "AA-AA-000-0000", "Japan", "0X116", 450, 2, 2/450,
  "1218", "143", "AA-AA-000-0000", "US", "0X153", 170, 3, 3/170,
  "0808", "173", "AA-AA-000-0000", "US", "0X124", 680, 6, 6/680,
  "1314", "189", "AA-AA-000-0000", "US", "0X093", 815, 4, 4/815,
  "1236", "58", "AA-AA-000-0000", "China", "0X091", 225, 1, 1/225,
  "0163", "167", "AA-AA-000-0000", "US", "0X059", 360, 3, 3/360
)
dfTransformed <- Transform_Count(dfInput, strCountCol = "Count")
dfAnalyzed <- Analyze_Identity(dfTransformed)

}
