% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Identity.R
\name{Analyze_Identity}
\alias{Analyze_Identity}
\title{Analyze Identity}
\usage{
Analyze_Identity(dfTransformed, strValueCol = "Metric", bQuiet = TRUE)
}
\arguments{
\item{dfTransformed}{\code{data.frame} created by \code{\link{Transform_Count}}}

\item{strValueCol}{\code{character} Name of column that will be copied as \code{Score}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one row per site with columns: SiteID, N, Metric,Score
}
\description{
Analyze Identity
}
\details{
Used in the data pipeline between \code{Transform} and \code{Flag} to rename KRI and Score columns.
}
\examples{
dfInput <- Consent_Map_Raw()
dfTransformed <- Transform_Count(dfInput, strCountCol = "Count")
dfAnalyzed <- Analyze_Identity(dfTransformed)

}
