% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_NormalApprox.R
\name{Analyze_NormalApprox}
\alias{Analyze_NormalApprox}
\title{Funnel Plot Analysis with Normal Approximation for Binary and Rate Outcomes.}
\usage{
Analyze_NormalApprox(dfTransformed, strType = "binary")
}
\arguments{
\item{dfTransformed}{\code{data.frame} in format produced by \code{\link{Transform_Rate}}}

\item{strType}{\code{character} Statistical outcome type. Valid values:
\itemize{
\item \code{"binary"} (default)
\item \code{"rate"}
}}
}
\value{
\code{data.frame} with one row per site with columns: GroupID, Numerator, Denominator, Metric, OverallMetric, Factor, and Score.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Creates analysis results data for percentage/rate data using funnel plot method with normal approximation.

More information can be found in \href{https://gilead-biostats.github.io/gsm/articles/KRI\%20Method.html#the-normal-approximation-method}{The Normal Approximation Method}
of the KRI Method vignette.
}
\section{Statistical Methods}{

This function applies funnel plots using asymptotic limits based on the normal approximation of a binomial distribution for
the binary outcome, or normal approximation of a Poisson distribution for the rate outcome with volume (the sample sizes
or total exposure of the sites) to assess data quality and safety.
}

\section{Data Specification}{


The input data (\code{dfTransformed}) is typically created using \code{\link{Transform_Rate}} and should be one record per site with required columns for:
\itemize{
\item \code{GroupID} - Site ID
\item \code{Numerator} - Total number of participants at site with event of interest/Total number of events of interest at site
\item \code{Denominator} - Total number of participants at site/Total number of days of exposure at site
\item \code{Metric} - Proportion of participants at site with event of interest/Rate of events at site (Numerator / Denominator)
}
}

\examples{
# Binary
dfTransformed <- tibble::tribble(
  ~GroupID,  ~Numerator,  ~Denominator,  ~Metric,
  139, 5, 901, 0.00555,
  143, 3, 170, 0.0176,
  162, 3, 370, 0.00811,
  167, 3, 360, 0.00833,
  173, 6, 680, 0.00882,
  189, 4, 815, 0.00491,
  29,  2, 450, 0.00444,
  5, 5, 730, 0.00685,
  58, 1, 225, 0.00444,
  78, 2, 50, 0.04
)

dfAnalyzed <- Analyze_NormalApprox(dfTransformed, strType = "binary")

# Rate
dfAnalyzed <- Analyze_NormalApprox(dfTransformed, strType = "rate")

}
