% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Poisson.R
\name{Analyze_Poisson}
\alias{Analyze_Poisson}
\title{AE Poisson Assessment - Analysis}
\usage{
Analyze_Poisson(dfTransformed)
}
\arguments{
\item{dfTransformed}{data.frame in format produced by \code{\link{Transform_EventCount}}}
}
\value{
input data frame with the  columns added for "Residuals", "PredictedCount" and "PValue"
}
\description{
Adds columns for site-level statistical assessment of distribution of reported safety outcomes
}
\examples{
dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )
dfTransformed <- Transform_EventCount( dfInput, cCountCol = 'Count', cExposureCol = "Exposure" )
dfAnalyzed <- Analyze_poisson( dfTransformed ) 

}
