% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Poisson.R
\name{Analyze_Poisson}
\alias{Analyze_Poisson}
\title{Poisson Analysis - Site Residuals}
\usage{
Analyze_Poisson(dfTransformed)
}
\arguments{
\item{dfTransformed}{data.frame in format produced by \code{\link{Transform_EventCount}}. Must include SubjectID, SiteID, TotalCount and TotalExposure.}
}
\value{
input data frame with columns added for "Residuals" and "PredictedCount"
}
\description{
Poisson Analysis - Site Residuals
}
\details{
Fits a Poisson model to site level data and adds columns capturing Residual and Predicted Count for each site.
}
\section{Statistical Methods}{


This function fits a poisson model to site-level data and then calculates residuals for each site. The poisson model is run using standard methods in the \code{stats} package by fitting a \code{glm} model with family set to \code{poisson} using a "log" link. Site-level residuals are calculated  \code{stats::predict.glm} via \code{broom::augment}.
}

\section{Data Specification}{


The input data (\code{ dfTransformed}) for the Analyze_Poisson is typically created using \code{\link{Transform_EventCount}} and should be one record per Site with columns for:
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{TotalCount} - Number of Events
\item \code{TotalExposure} - Number of days of exposure
}
}

\examples{
dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )
dfTransformed <- Transform_EventCount( dfInput, strCountCol = 'Count', strExposureCol = "Exposure" )
dfAnalyzed <- Analyze_Poisson( dfTransformed )

}
