% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Poisson.R
\name{Analyze_Poisson}
\alias{Analyze_Poisson}
\title{AE Poisson Assessment - Analysis}
\usage{
Analyze_Poisson(dfTransformed)
}
\arguments{
\item{dfTransformed}{frame in format produced by \code{\link{AE_Transform}}}
}
\value{
input data frame with the  columns added for "Residuals", "PredictedCount" and "PValue"
}
\description{
Adds columns for site-level statistical assessment of distribution of reported safety outcomes
}
