% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Poisson.R
\name{Analyze_Poisson}
\alias{Analyze_Poisson}
\title{Poisson Analysis - Site Residuals.}
\usage{
Analyze_Poisson(dfTransformed)
}
\arguments{
\item{dfTransformed}{data.frame in format produced by \code{\link{Transform_Rate}}. Must include GroupID, Numerator, Denominator and Metric.}
}
\value{
\code{data.frame} with one row per site with columns: GroupID, Numerator, Denominator, Metric, Score, and PredictedCount.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Fits a Poisson model to site-level data and adds columns capturing Residual and Predicted Count for each site.

More information can be found in \href{https://gilead-biostats.github.io/gsm/articles/KRI\%20Method.html#the-poisson-regression-method}{The Poisson Regression Method}
of the KRI Method vignette.
}
\section{Statistical Methods}{


This function fits a Poisson model to site-level data and then calculates residuals for each site.
The Poisson model is run using standard methods in the \code{stats} package by fitting a \code{glm} model with family
set to \code{poisson} using a "log" link. Site-level residuals are calculated using \code{stats::predict.glm} via \code{broom::augment}.
}

\section{Data Specification}{


The input data (\code{dfTransformed}) for \code{Analyze_Poisson} is typically created using \code{\link{Transform_Rate}} and should be one record per site with required columns for:
\itemize{
\item \code{GroupID} - Site ID
\item \code{Numerator} - Number of Adverse Events
\item \code{Denominator} - Number of days of exposure
\item \code{Metric} - Rate of exposure (Numerator / Denominator)
}
}

\examples{
dfTransformed <- Transform_Rate(sampleInput)

dfAnalyzed <- Analyze_Poisson(dfTransformed)

}
