% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Poisson.R
\name{Analyze_Poisson}
\alias{Analyze_Poisson}
\title{Poisson Analysis - Site Residuals}
\usage{
Analyze_Poisson(dfTransformed, bQuiet = TRUE)
}
\arguments{
\item{dfTransformed}{data.frame in format produced by \code{\link{Transform_Rate}}. Must include GroupID, Numerator, Denominator and Metric.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with columns added for "Score" and "PredictedCount".
}
\description{
Poisson Analysis - Site Residuals
}
\details{
Fits a Poisson model to site level data and adds columns capturing Residual and Predicted Count for each site.
}
\section{Statistical Methods}{


This function fits a poisson model to site-level data and then calculates residuals for each site. The poisson model is run using standard methods in the \code{stats} package by fitting a \code{glm} model with family set to \code{poisson} using a "log" link. Site-level residuals are calculated  \code{stats::predict.glm} via \code{broom::augment}.
}

\section{Data Specification}{


The input data (\code{dfTransformed}) for Analyze_Poisson is typically created using \code{\link{Transform_Rate}} and should be one record per site with required columns for:
\itemize{
\item \code{GroupID} - Site ID
\item \code{Numerator} - Number of Adverse Events
\item \code{Denominator} - Number of days of exposure
\item \code{Metric} - Rate of exposure (Numerator / Denominator)
}
}

\examples{
dfInput <- AE_Map_Raw() \%>\% na.omit()  #na.omit is placeholder for now
dfTransformed <- Transform_Rate(dfInput,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)

dfAnalyzed <- Analyze_Poisson(dfTransformed)

}
