% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Poisson.R
\name{Analyze_Poisson}
\alias{Analyze_Poisson}
\title{AE Poisson Assessment - Analysis}
\usage{
Analyze_Poisson(dfTransformed)
}
\arguments{
\item{dfTransformed}{data.frame in format produced by \code{\link{Transform_EventCount}}. Must include}
}
\value{
input data frame with columns added for "Residuals", "PredictedCount" and "PValue"
}
\description{
Adds columns for site-level statistical assessment of distribution of reported safety outcomes
}
\details{
Fits a Poisson Model to site-level data.
}
\section{Statistical Methods}{


TODO Coming soon ...
}

\section{Data Specification}{


The input data (\code{ dfTransformed}) for the Analyze_Poisson is typically created using \code{\link{Transform_EventCount}} and should be one record per Site with columns for:
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{Count} - Number of Adverse Events
\item \code{Exposure} - Number of days of exposure
}
}

\examples{
dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )
dfTransformed <- Transform_EventCount( dfInput, cCountCol = 'Count', cExposureCol = "Exposure" )
dfAnalyzed <- Analyze_Poisson( dfTransformed ) 

}
