% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Poisson_PredictBounds.R
\name{Analyze_Poisson_PredictBounds}
\alias{Analyze_Poisson_PredictBounds}
\title{Safety Assessment - Calculate prediction bounds}
\usage{
Analyze_Poisson_PredictBounds(dfTransformed, vThreshold)
}
\arguments{
\item{dfTransformed}{data frame in format produced by \code{\link{AE_Map_Adam}} or \code{\link{AE_Map_Raw}}.}
}
\value{
dataframe with prediction boundary curves
}
\description{
Safety Assessment - Calculate prediction bounds
}
