% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Poisson_PredictBounds.R
\name{Analyze_Poisson_PredictBounds}
\alias{Analyze_Poisson_PredictBounds}
\title{Poisson Analysis - Predicted Boundaries.}
\usage{
Analyze_Poisson_PredictBounds(
  dfTransformed,
  vThreshold = c(-5, 5),
  nStep = NULL
)
}
\arguments{
\item{dfTransformed}{\code{data.frame} data.frame in format produced by
\code{\link{Transform_Rate}}. Must include GroupID, N, Numerator and Denominator.}

\item{vThreshold}{\code{numeric} upper and lower boundaries in residual space. Should be identical to
the thresholds used AE_Assess().}

\item{nStep}{\code{numeric} step size of imputed bounds.}
}
\value{
\code{data.frame} containing predicted boundary values with upper and lower bounds across the
range of observed values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Fits a Poisson model to site-level data and then calculates predicted count values and upper- and
lower- bounds for across the full range of exposure values.
}
\section{Statistical Methods}{

This function fits a Poisson model to site-level data and then calculates residuals for each
site. The Poisson model is run using standard methods in the \code{stats} package by fitting a \code{glm}
model with family set to \code{poisson} using a "log" link. Upper and lower boundary values are then
calculated using the method described here TODO: Add link.
}

\section{Data Specification}{


The input data (\code{dfTransformed}) for \code{Analyze_Poisson_PredictBounds} is typically created using
\code{\link{Transform_Rate}} and should be one record per site with columns for:
\itemize{
\item \code{GroupID} - Unique subject ID
\item \code{Numerator} - Number of Events
\item \code{Denominator} - Number of days of exposure
}
}

\examples{
dfInput <- tibble::tribble(
  ~SubjectID, ~SiteID, ~StudyID, ~CountryID, ~CustomGroupID, ~Exposure, ~Count, ~Rate,
  "0496", "5", "AA-AA-000-0000", "US", "0X167", 730, 5, 5/720,
  "1350", "78", "AA-AA-000-0000", "US", "0X002", 50, 2, 2/50,
  "0539", "139", "AA-AA-000-0000", "US", "0X052", 901, 5, 5/901,
  "0329", "162", "AA-AA-000-0000", "US", "0X049", 370, 3, 3/370,
  "0429", "29", "AA-AA-000-0000", "Japan", "0X116", 450, 2, 2/450,
  "1218", "143", "AA-AA-000-0000", "US", "0X153", 170, 3, 3/170,
  "0808", "173", "AA-AA-000-0000", "US", "0X124", 680, 6, 6/680,
  "1314", "189", "AA-AA-000-0000", "US", "0X093", 815, 4, 4/815,
  "1236", "58", "AA-AA-000-0000", "China", "0X091", 225, 1, 1/225,
  "0163", "167", "AA-AA-000-0000", "US", "0X059", 360, 3, 3/360
)

dfTransformed <- Transform_Rate(dfInput,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)

dfBounds <- Analyze_Poisson_PredictBounds(dfTransformed, c(-5, 5))

}
