% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_QTL.R
\name{Analyze_QTL}
\alias{Analyze_QTL}
\title{QTL Analysis for Binary and Rate Outcomes.}
\usage{
Analyze_QTL(dfTransformed, nConfLevel = 0.95, strOutcome = "binary")
}
\arguments{
\item{dfTransformed}{\code{data.frame} in format produced by \code{\link{Transform_Rate}}}

\item{nConfLevel}{\code{numeric} specified confidence interval for QTL analysis.}

\item{strOutcome}{\code{character} indicates statistical test used for QTL analysis. One of \code{rate} or \code{binary}.}
}
\value{
\code{data.frame} with one row with columns: GroupID, Numerator, Denominator, Metric, Method, ConfLevel, Estimate, LowCI, UpCI and Score.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Creates confidence intervals for the observed proportion of participants for the event of interest. Uses the exact binomial (\code{stats::binom.test}) or poisson test (\code{stats::poisson.test}).
}
\section{Statistical Methods}{


TODO Coming soon ...
}

\section{Data Specification}{


The input data (\code{dfTransformed}) for \code{Analyze_QTL} is typically created using \code{\link{Transform_Rate}} and should be one record for the entire study with required columns for:
\itemize{
\item \code{GroupID} - GroupID should be the StudyID
\item \code{N} - Total number of participants at site
\item \code{Numerator} - Total number of participants at site with event of interest
\item \code{Denominator} - Total number of participants at a site
\item \code{Metric} - Proportion of participants at site with event of interest
}
}

\examples{
# binary
dfTransformed <- tibble::tibble(
  GroupID      = c("AA-AA-000-0000"),
  Numerator    = c(122),
  Denominator  = c(1301),
  Metric       = c(0.0938)
)

dfAnalyzed <- Analyze_QTL(dfTransformed, strOutcome = "binary")

# rate
dfTransformedRate <- tibble::tibble(
  GroupID      = c("AA-AA-000-0000"),
  Numerator    = c(4473),
  Denominator  = c(6.19),
  Metric       = c(723)
)
dfAnalyzed <- Analyze_QTL(dfTransformedRate, strOutcome = "rate")

}
