% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_QTL.R
\name{Analyze_QTL}
\alias{Analyze_QTL}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
Analyze_QTL(
  dfTransformed,
  nConfLevel = 0.95,
  strOutcome = "binary",
  bQuiet = TRUE
)
}
\arguments{
\item{dfTransformed}{\code{data.frame} in format produced by \code{\link{Transform_Rate}}}

\item{nConfLevel}{\code{numeric} specified confidence interval for QTL analysis.}

\item{strOutcome}{\code{character} indicates statistical test used for QTL analysis. One of \code{rate} or \code{binary}.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one row with columns: GroupID, N, Numerator, Denominator, Metric, Method, ConfLevel, Estimate, LowCI, UpCI and Score.
}
\description{
QTL Analysis
}
\details{
Creates QTL Analysis results data for binary event (e.g. Yes/No) of interest either as a proportion or as a rate

Creates confidence intervals for the observed proportion of participants with event of interest using the exact binomial or poisson test
}
\section{Statistical Methods}{


TODO Coming soon ...
}

\section{Data Specification}{


The input data (\code{dfTransformed}) for Analyze_QTL is typically created using \code{\link{Transform_Rate}} and should be one record for the entire study with required columns for:
\itemize{
\item \code{GroupID} - GroupID should be the StudyID
\item \code{N} - Total number of participants at site
\item \code{Numerator} - Total number of participants at site with event of interest
\item \code{Denominator} - Total number of participants at a site
\item \code{Metric} - Proportion of participants at site with event of interest
}
}

\examples{
dfInput <- Disp_Map_Raw()
dfTransformed <- Transform_Rate(dfInput,
  strNumeratorCol = "Count",
  strDenominatorCol = "Total",
  strGroupCol = "StudyID"
)

dfAnalyzed <- Analyze_QTL(dfTransformed, strOutcome = "binary")
dfFlagged <- Flag_QTL(dfAnalyzed, vThreshold = 0.2)



dfInput <- PD_Map_Raw_Binary()
dfTransformed <- Transform_Rate(dfInput,
  strGroupCol = "StudyID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Total"
)

dfAnalyzed <- Analyze_QTL(dfTransformed, strOutcome = "rate")
dfFlagged <- Flag_QTL(dfAnalyzed, vThreshold = 0.01)

}
