% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Wilcoxon.R
\name{Analyze_Wilcoxon}
\alias{Analyze_Wilcoxon}
\title{Analyze_Wilcoxon}
\usage{
Analyze_Wilcoxon(dfTransformed)
}
\arguments{
\item{dfTransformed}{data.frame in format produced by \code{\link{Transform_EventCount}}}
}
\description{
Creates Analyis results data for Adverse Event assessment using the Wilcoxon sign-ranked test
}
