% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Wilcoxon.R
\name{Analyze_Wilcoxon}
\alias{Analyze_Wilcoxon}
\title{AE Wilcoxon Assessment - Analysis}
\usage{
Analyze_Wilcoxon(dfTransformed, strOutcome = "Rate")
}
\arguments{
\item{dfTransformed}{data.frame in format produced by \code{\link{Transform_EventCount}}}

\item{strOutcome}{required, name of column in dfTransformed dataset to perform Wilcoxon test on. Default="Rate"}
}
\value{
data.frame with one row per site, columns: SiteID, N, TotalCount, TotalExposure, Rate, Estimate, PValue
}
\description{
Creates Analysis results data for Adverse Event assessment using the Wilcoxon sign-ranked test
}
\details{
@details

Fits a Wilcox Model to site-level data.
}
\section{Statistical Methods}{


A Wilcoxon model is used to generate estimates and p-values for each site (as specified with the \code{strOutcome} parameter).
}

\section{Data Specification}{


The input data (dfTransformed) for Analyze_Wilcoxon is typically created using \code{\link{Transform_EventCount}} and should be one record per site with required columns for:
\itemize{
\item \code{SiteID} - Site ID
\item \code{N} - Number of participants
\item \code{TotalCount} - Number of Adverse Events
\item \code{TotalExposure} - Number of days of exposure
\item \code{Rate} - Rate of exposure (TotalCount / TotalExposure)
}
}

\examples{
dfInput <- AE_Map_Raw(clindata::rawplus_ae, clindata::rawplus_subj)
dfTransformed <- Transform_EventCount( dfInput, strCountCol = 'Count', strExposureCol = "Exposure" )
dfAnalyzed <- Analyze_Wilcoxon( dfTransformed , strOutcome ="Rate")

}
