% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-BindResults.R
\name{BindResults}
\alias{BindResults}
\title{Helper function to bind results from multiple workflows}
\usage{
BindResults(lResults, strName, dSnapshotDate = sys.date(), strStudyID)
}
\arguments{
\item{lResults}{List of results in the format returned by RunWorkflows()}

\item{strName}{Name of the object to stack. Pulled from lResults$lData.}

\item{dSnapshotDate}{Date of the snapshot. Default is sys.date().}

\item{strStudyID}{study ID. Default is 'ABC-123'.}
}
\value{
A data frame.
}
\description{
Used to stack results (e.g. dfSummary) from a list of analysis pipeline output formatted like the result of RunWorklows(). Also adds study level metadata when provided.
}
