% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-CheckInputs.R
\name{CheckInputs}
\alias{CheckInputs}
\title{Check mapping inputs}
\usage{
CheckInputs(context, dfs, mapping = NULL, bQuiet = TRUE)
}
\arguments{
\item{yaml}{}
}
\value{
list
}
\description{
Check mapping inputs
}
