% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-CheckInputs.R
\name{CheckInputs}
\alias{CheckInputs}
\title{Check mapping inputs}
\usage{
CheckInputs(context, dfs, mapping = NULL, bQuiet = TRUE)
}
\arguments{
\item{context}{Description of the data pipeline "step" that is being checked, i.e., "AE_Map_Raw" or "PD_Assess".}

\item{dfs}{list of data frames.}

\item{mapping}{YAML mapping for a given context.}

\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}
}
\value{
list
}
\description{
Check mapping inputs
}
