% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-CombineSpecs.R
\name{combine_domain}
\alias{combine_domain}
\title{Combine Domain Specifications}
\usage{
combine_domain(domain_specs)
}
\arguments{
\item{domain_specs}{A list of lists, where each sublist represents the specifications for a domain across multiple specs.}
}
\value{
A list representing the combined specifications for the domain.
}
\description{
This function combines multiple column specifications for a single domain by applying deduplication and resolving conflicts.
}
\keyword{internal}
