% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uti-CombineSpecs.R, R/util-CombineSpecs.R
\name{CombineSpecs}
\alias{CombineSpecs}
\title{Combine Specifications}
\usage{
CombineSpecs(lSpecs)

CombineSpecs(lSpecs)
}
\arguments{
\item{lSpecs}{A list of specifications.}
}
\value{
A list representing the combined specification.

A list representing the combined specification.
}
\description{
Combine a list of specifications into a single specification.

This function combines a list of specifications into a single specification.
}
\examples{
all_wf <- MakeWorkflowList()
all_specs <- CombineSpecs(all_wf)

all_wf <- gsm::MakeWorkflowList()
all_specs <- CombineSpecs(all_wf)

}
