% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-CombineSpecs.R
\name{CombineSpecs}
\alias{CombineSpecs}
\title{Combine Specifications}
\usage{
CombineSpecs(lSpecs)
}
\arguments{
\item{lSpecs}{A list of specifications.}
}
\value{
A list representing the combined specification.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Combine a list of specifications into a single specification.
}
\examples{
all_wf <- MakeWorkflowList()
all_specs <- CombineSpecs(all_wf)

}
