% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{config_param}
\alias{config_param}
\title{Configuration Parameters}
\format{
a data frame with 124 rows and 6 columns
}
\source{
./data-raw/config/param.csv
}
\usage{
config_param
}
\description{
Configuration Parameters (param) data CTMS data
}
\keyword{datasets}
