% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{config_workflow}
\alias{config_workflow}
\title{Configuration Workflow}
\format{
a data frame with 26 rows and 4 columns
}
\source{
./data-raw/config/workflow.csv
}
\usage{
config_workflow
}
\description{
Configuration Workflow (workflow) data CTMS data
}
\keyword{datasets}
