% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specs.R
\docType{data}
\name{config_workflow}
\alias{config_workflow}
\title{config_workflow Dataset}
\format{
A data frame with 26 rows and 4 columns:
\describe{
\item{studyid}{Study ID, chr, constant value "AA-AA-000-0000"}
\item{workflowid}{workflow ID, chr, kri value "kri0001"}
\item{gsm_version}{gsm version used, chr, most recent version installed i.e. "1.9.2"}
\item{active}{active status of workflow, lgl}
}
}
\source{
Generated from \code{config_workflow.csv} dataset in the \code{gsm} package.
}
\usage{
config_workflow
}
\description{
config_workflow Dataset
}
\keyword{datasets}
