% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consent_Assess.R
\name{Consent_Assess}
\alias{Consent_Assess}
\title{Consent Assessment}
\usage{
Consent_Assess(
  dfInput,
  nThreshold = 0.5,
  lTags = list(Assessment = "Consent"),
  bChart = TRUE,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data, a data frame with one record per subject.}

\item{nThreshold}{\code{numeric} Threshold specification. Default: \code{0.5}}

\item{lTags}{\code{list} Assessment tags, a named list of tags describing the assessment that defaults to \code{list(Assessment="IE")}. \code{lTags} is returned as part of the assessment (\code{lAssess$lTags}) and each tag is added as a column in \code{lAssess$dfSummary}.}

\item{bChart}{\code{logical} Generate data visualization? Default: \code{TRUE}}

\item{bReturnChecks}{\code{logical} Return input checks from \code{is_mapping_valid}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} Assessment, a named list with:
\itemize{
\item each data frame in the data pipeline
\itemize{
\item \code{dfInput}
\item \code{dfTransformed}, returned by \code{\link[=Transform_EventCount]{Transform_EventCount()}}
\item \code{dfAnalyzed}, a copy of \code{dfTransformed} and input to \verb{[gsm::Flag()]}
\item \code{dfFlagged}, returned by \code{\link[=Flag]{Flag()}}
\item \code{dfSummary}, returned by \code{\link[=Summarize]{Summarize()}}
}
\item assessment metadata
\itemize{
\item \code{strFunctionName}
\item \code{lParams}
\item \code{lTags}
}
\item output(s)
\itemize{
\item \code{chart}
}
}
}
\description{
Flag sites where subject consent was:
\itemize{
\item not given
\item never obtained
\item not followed by subject randomization
\item obtained after subject randomization
}
}
\details{
The Consent Assessment uses the standard \href{https://github.com/Gilead-BioStats/gsm/wiki/Data-Pipeline-Vignette}{GSM data pipeline} to flag sites with consent issues. This assessment detects sites with subjects who participated
in study activities before consent was finalized. The count returned in the summary represents
the number of subjects at a given site for whom:
\itemize{
\item consent was not given
\item consent was not obtained
\item consent did not result in randomization
\item consent was obtained after randomization
}

Additional details regarding the data pipeline and statistical methods are described below.
}
\section{Data specification}{
\tabular{lllll}{
   \strong{Domain} \tab \strong{Column Key} \tab \strong{Default Value} \tab \strong{Required?} \tab \strong{Require Unique Values?} \cr
   dfInput \tab strIDCol \tab SubjectID \tab TRUE \tab TRUE \cr
   dfInput \tab strSiteCol \tab SiteID \tab TRUE \tab FALSE \cr
   dfInput \tab strCountCol \tab Count \tab TRUE \tab FALSE \cr
}
}

\examples{
dfInput <- Consent_Map_Raw()
consent_assessment <- Consent_Assess(dfInput)

}
