% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consent_Assess.R
\name{Consent_Assess}
\alias{Consent_Assess}
\title{Consent Assessment}
\usage{
Consent_Assess(dfInput, nThreshold = 0.5, lTags = list(Assessment = "Consent"))
}
\arguments{
\item{dfInput}{input data with one record per person and the following required columns: SubjectID, SiteID, Count.}

\item{nThreshold}{Any sites where 'N' is greater than nThreshold will be flagged. Default value is 0.5, which flags any site with one or more subjects meeting any of the criteria.}

\item{lTags}{named list of tags describing the assessment. \code{lTags} is returned as part of the assessment (\code{lAssess$lTags}) and each tag is added as columns in \code{lassess$dfSummary}. Default is \code{list(Assessment="Consent")}}
}
\value{
A list containing all data and metadata in the standard data pipeline (\code{dfInput}, \code{dfTransformed}, \code{dfAnalyzed}, \code{dfFlagged}, \code{dfSummary}, \code{strFunctionName}, \code{lParams} and \code{lTags}) is returned.
}
\description{
Consent Assessment
}
\details{
The Consent Assessment uses the standard GSM data pipeline (TODO add link to data vignette) to flag sites with participants who started study activities before consent was finalized. More details regarding the data pipeline and statistical methods are described below.
}
\section{Data Specification}{


The input data (\code{dfInput}) for Consent Assessment is typically created using \code{\link{Consent_Map_Raw}} and should be one record per person with columns for:
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{Count} - Number of findings of errors/outliers.
}

The Assessment
\itemize{
\item \code{\link{Transform_EventCount}} creates \code{dfTransformed}.
\item \code{\link{Flag}} creates \code{dfFlagged}.
\item \code{\link{Summarize}} creates \code{dfSummary}.
}
}

\section{Statistical Assumptions}{


This Assessment finds any sites where one or more subjects meets any of the following criteria: No Consent, Missing Consent, Missing Randomization Date, or
Consent date later in time than the Randomization Date. 'N' in the summary represents the number of subjects in a study that meet one or more criteria. Sites
With N greater than user specified \code{nThreshold} will be flagged.
}

\examples{

library(dplyr)
raw_consent <- clindata::raw_ic_elig \%>\%
select(SUBJID, CONSDAT = DSSTDAT_RAW) \%>\%
 mutate(
     CONSCAT_STD = "MAINCONSENT",
     CONSYN = "Y",
     CONSDAT = as.Date(CONSDAT, format = "\%d \%B \%Y")
   ) \%>\%
   filter(SUBJID != "")

dfInput <- Consent_Map_Raw(
   dfCONSENT = clindata::rawplus_consent,
   dfSUBJ = clindata::rawplus_subj,
   strConsentTypeValue = "MAINCONSENT",
   strConsentStatusValue="Y"
)

Consent_Summary <- Consent_Assess(dfInput)$dfSummary

}
