% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consent_Map_Raw.R
\name{Consent_Map_Raw}
\alias{Consent_Map_Raw}
\title{Consent Assessment Mapping from Raw Data- Make Input Data}
\usage{
Consent_Map_Raw(dfConsent = NULL, dfIxrsrand = NULL)
}
\arguments{
\item{dfConsent}{consent data frame}

\item{dfIxrsrand}{Ixrsand data frame}
}
\description{
Convert from raw data format to needed input format for Consent Assessment
}
