% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consent_Map_Raw.R
\name{Consent_Map_Raw}
\alias{Consent_Map_Raw}
\title{Consent Assessment Mapping from Raw Data- Make Input Data}
\usage{
Consent_Map_Raw(dfConsent, dfRDSL, strConsentReason = "mainconsent")
}
\arguments{
\item{dfConsent}{consent data frame with columns: SUBJID, CONSCAT_STD , CONSYN , CONSDAT.}

\item{dfRDSL}{Subject-level Raw Data (RDSL) required columns: SubjectID SiteID RandDate.}

\item{strConsentReason}{default = "mainconsent", filters on CONSCAT_STD of dfConsent, if NULL no filtering is done.}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count.
}
\description{
Convert from raw data format to needed input format for \code{\link{Consent_Assess}}
}
\details{
This function uses raw Consent and RDSL data to create the required input for \code{\link{Consent_Assess}}.
}
\section{Data Specification}{


The following columns are required:
\itemize{
\item \code{dfConsent}
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{CONSCAT_STD} - Type of Consent_Coded value
\item \code{CONSYN} - Did the subject give consent? Yes / No.
\item \code{CONSDAT} - If yes, provide date consent signed
}
\item \code{dfRDSL}
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{RandData} - Randomization Date
}
}
}

\examples{

input <- Consent_Map_Raw(dfConsent = clindata::raw_consent, dfRDSL = clindata::rawplus_rdsl, strConsentReason = NULL)

}
