% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consent_Map_Raw.R
\name{Consent_Map_Raw}
\alias{Consent_Map_Raw}
\title{Consent Assessment Mapping from Raw Data- Make Input Data}
\usage{
Consent_Map_Raw(dfConsent = NULL, dfIxrsrand = NULL)
}
\arguments{
\item{strPath}{path to the raw datasets}

\item{dfSubid}{directly input a subid dataset with columns SUBJID INVID}

\item{dfIe}{directly input a ie dataset with columns SUBJID IECAT IETEST IEORRES}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Total, Valid, Invalid, Missing, Details
}
\description{
Convert from raw data format to needed input format for Consent Assessment
}
