% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consent_Map_Raw.R
\name{Consent_Map_Raw}
\alias{Consent_Map_Raw}
\title{Consent Assessment Mapping from Raw Data}
\usage{
Consent_Map_Raw(
  dfCONSENT,
  dfSUBJ,
  mapping = NULL,
  strConsentTypeValue = "MAINCONSENT",
  strConsentStatusValue = "Y"
)
}
\arguments{
\item{dfCONSENT}{consent data frame with columns: SUBJID, CONSCAT_STD , CONSYN , CONSDAT.}

\item{dfSUBJ}{Subject-level Raw Data required columns: SubjectID SiteID RandDate.}

\item{mapping}{List containing expected columns in each data set.}

\item{strConsentTypeValue}{default = "mainconsent", filters on CONSCAT_STD of dfCONSENT, if NULL no filtering is done.}

\item{strConsentStatusValue}{default = "Yes", expected Status value for valid consent.}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count.
}
\description{
Convert from raw data format to needed input format for \code{\link{Consent_Assess}}
}
\details{
This function uses raw Consent and Subject data to create the required input for \code{\link{Consent_Assess}}.
}
\section{Data Specification}{

The following columns are required:
\itemize{
\item \code{dfCONSENT}
\itemize{
\item \code{SubjectID} - Subject ID
\item \code{CONSCAT_STD} - Type of Consent_Coded value
\item \code{CONSYN} - Did the subject give consent? Yes / No.
\item \code{CONSDAT} - If yes, provide date consent signed
}
\item \code{dfSUBJ}
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{RandDate} - Randomization Date
}
}
}

\examples{

input <- Consent_Map_Raw(
   dfCONSENT = clindata::rawplus_consent,
   dfSUBJ = clindata::rawplus_subj
)

}
