% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ConsolidateStrata.R
\name{ConsolidateStrata}
\alias{ConsolidateStrata}
\title{Consolidate stratified assessment outputs}
\usage{
ConsolidateStrata(lOutput, lStratifiedOutput, bQuiet = TRUE)
}
\arguments{
\item{lOutput}{\code{list} The output from \code{\link[=RunWorkflow]{RunWorkflow()}}}

\item{lStratifiedOutput}{\code{list} Multiple outputs from  \code{\link[=RunWorkflow]{RunWorkflow()}}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} containing \code{lWorkflow} with \code{tags}, \code{workflow}, \code{path}, \code{name}, \code{lData},
\code{lChecks}, \code{bStatus}, \code{checks}, and \code{lResults} added based on the results of the execution of
\code{assessment$workflow}.
}
\description{
Consolidates multiple stratified assessment outputs by stacking the data frames returned by an
\verb{*_Assess} function and generating a paneled data visualization.
}
\examples{
lData <- list(
  dfSUBJ = clindata::rawplus_dm,
  dfAE = clindata::rawplus_ae
)
lMapping <- yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package = "gsm"))

lWorkflow <- MakeWorkflowList(bRecursive = TRUE, strNames = "aeGrade")$aeGrade
lOutput <- RunWorkflow(lWorkflow, lData = lData, lMapping = lMapping)

lStratifiedWorkflow <- MakeStratifiedAssessment(
  lWorkflow,
  lData,
  lMapping
)
lStratifiedOutput <- lStratifiedWorkflow \%>\%
  purrr::map(~ RunWorkflow(
    .x,
    lData,
    lMapping
  ))

lConsolidatedOutput <- ConsolidateStrata(
  lOutput,
  lStratifiedOutput
)

}
