% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate_Chart.R
\name{covariate_chart_site}
\alias{covariate_chart_site}
\title{Make covariate site chart.}
\usage{
covariate_chart_site(dfCovariate, strType = "total")
}
\arguments{
\item{dfCovariate}{\code{data.frame} site-level data.frame returned by \link{Covariate_Map}.}

\item{strType}{\code{character} one of "total" or "percent".}
}
\value{
\code{plotly} chart.
}
\description{
Make covariate site chart.
}
\keyword{internal}
