% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate_Map.R
\name{Covariate_Map}
\alias{Covariate_Map}
\title{Covariate Map}
\usage{
Covariate_Map(dfCovariate, strCovariteColName, strWorkflowId)
}
\arguments{
\item{dfCovariate}{\code{data.frame} dataframe of kri covariate of interest}

\item{strCovariteColName}{\code{string} name of covariate column of interest}

\item{strWorkflowId}{\code{string} name of kri workflowid}
}
\description{
Covariate Map
}
