% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate_Map.R
\name{Covariate_Map}
\alias{Covariate_Map}
\title{Map through raw data for distribution of causes checked in kri's}
\usage{
Covariate_Map(lSnapshot, dist_list)
}
\arguments{
\item{lSnapshot}{\code{list} a snapshot object from \code{Make_Snapshot()}}

\item{dist_list}{\code{list} a list of data and linked category columns}
}
\description{
Map through raw data for distribution of causes checked in kri's
}
\keyword{internal}
