% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sidebar.R,
%   R/util-CovariateReportHelpers.R
\name{create_sidebar}
\alias{create_sidebar}
\title{creates sidebar info for covariate report}
\usage{
create_sidebar(study, date, total, var, domain)

create_sidebar(study, date, total, var, domain)
}
\arguments{
\item{study}{\code{character} a snapshot object from \code{Make_Snapshot()}}

\item{date}{\code{character} a snapshot object from \code{Make_Snapshot()}}

\item{total}{\code{character} a snapshot object from \code{Make_Snapshot()}}

\item{var}{\code{character} a snapshot object from \code{Make_Snapshot()}}

\item{domain}{\code{character} a snapshot object from \code{Make_Snapshot()}}
}
\description{
creates sidebar info for covariate report

Create sidebar component for Covariate Report.
}
\keyword{internal}
