% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disc_Report.R
\name{Disc_Report}
\alias{Disc_Report}
\title{Create and output Discontinuation Report}
\usage{
Disc_Report(
  dfSTUDCOMP = clindata::rawplus_studcomp,
  dfSDRGCOMP = clindata::rawplus_sdrgcomp,
  lMapping = gsm::Read_Mapping("rawplus")
)
}
\arguments{
\item{dfSTUDCOMP}{\code{data.frame} raw data of study completion defaults to \code{clindata::rawplus_studcomp}}

\item{dfSDRGCOMP}{\code{data.frame} raw data of treatment completion defaults to \code{clindata::rawplus_sdrgcomp}}

\item{lMapping}{\code{list} the mapping to use for the data frames provided defaults to \code{gsm::Read_Mapping("rawplus")}}
}
\description{
Create and output Discontinuation Report
}
