% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disp_Assess.R
\name{Disp_Assess}
\alias{Disp_Assess}
\title{Disposition Assessment}
\usage{
Disp_Assess(
  dfDisp,
  vThreshold = NULL,
  cLabel = "",
  cMethod = "wilcoxon",
  bDataList = FALSE
)
}
\arguments{
\item{dfDisp}{input data with one record per person and the following required columns: SubjectID, SiteID, Count}

\item{vThreshold}{numeric vector with 2 threshold values.  Defaults to c(-5,5) for method = "poisson" and c(.0001,NA) for method = "wilcoxon".}

\item{cLabel}{Assessment label}

\item{cMethod}{valid methods are "poisson" (the default), or  "wilcoxon"}

\item{bDataList}{Should all assessment datasets be returned as a list? If False (the default), only the Summary data frame is returned}
}
\value{
If \code{bDataList} is false (the default), the summary data frame (\code{dfSummary}) is returned. If \code{bDataList} is true, a list containing all data in the standard data pipeline (\code{dfInput}, \code{dfTransformed}, \code{dfAnalyzed}, \code{dfFlagged} and \code{dfSummary}) is returned.
}
\description{
The Disposition Assessment flags sites that may be misrepresenting discontinuation reasons.
}
\details{
The Disposition Assessment uses the standard GSM data pipeline (TODO add link to data vignette) to flag possible outliers. More details regarding the data pipeline and statistical methods are described below.
}
\section{Data Specification}{


The input data (\code{dfDisp}) for the Disposition Assessment is typically created using \code{\link{Disp_Map}} and should be one record per person with columns for:
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{Count} - Number of Discontinuation Reasons
}

The Assessment
\itemize{
\item \code{\link{Transform_EventCount}} creates \code{dfTransformed}.
\item \code{\link{Analyze_Poisson}} or \code{\link{Analyze_Wilcoxon}} creates \code{dfAnalyzed}.
\item \code{\link{Flag}} creates \code{dfFlagged}.
\item \code{\link{Summarize}} creates \code{dfSummary}.
}
}

\section{Statistical Assumptions}{


A Poisson or Wilcoxon model is used to generate estimates and p-values for each site (as specified with the \code{cMethod} parameter). Those model outputs are then used to flag possible outliers using the thresholds specified in \code{vThreshold}. In the Poisson model, sites with an estimand less than -5 are flagged as -1 and greater than 5 are flagged as 1 by default. For Wilcoxon, sites with p-values less than 0.0001 are flagged by default.

See \code{\link{Analyze_Poisson}} and \code{\link{Analyze_Wilcoxon}} for additional details about the statistical methods and thier assumptions.
}

\examples{
dfInput <- Disp_Map(dfDisp = safetyData::adam_adsl,strCol = "DCREASCD",strReason = "adverse event")
SafetyAE <- Disp_Assess( dfInput )
SafetyAE_Wilk <- Disp_Assess( dfInput, cMethod="wilcoxon")

}
