% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disp_Assess.R
\name{Disp_Assess}
\alias{Disp_Assess}
\title{Disposition Assessment}
\usage{
Disp_Assess(
  dfInput,
  vThreshold = NULL,
  strMethod = "chisq",
  strKRILabel = "\% Discontinuation/Treatment",
  strGroupCol = "SiteID",
  lTags = list(Assessment = "Disposition"),
  bChart = TRUE,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data, a data frame with one record per subject.}

\item{vThreshold}{\code{numeric} Threshold specification, a vector of length 2 that defaults to
\code{c(.05, NA)} for both Chi-square test (\code{strMethod} = "chisq") and Fisher's exact test (\code{strMethod} = "fisher").}

\item{strMethod}{\code{character} Statistical method. Valid values:
\itemize{
\item \code{"chisq"} (default)
\item \code{"fisher"}
}}

\item{strKRILabel}{\code{character} KRI description. Default: \code{"DCs/Week"}}

\item{strGroupCol}{\code{character} Name of column for grouping variable. Default: \code{"SiteID"}}

\item{lTags}{\code{list} Assessment tags, a named list of tags describing the assessment that defaults
to \code{list(Assessment = "AE")}. \code{lTags} is returned as part of the assessment (\code{lAssess$lTags})
and each tag is added as a column in \code{lAssess$dfSummary}.}

\item{bChart}{\code{logical} Generate data visualization? Default: \code{TRUE}}

\item{bReturnChecks}{\code{logical} Return input checks from \code{\link[=is_mapping_valid]{is_mapping_valid()}}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} Assessment, a named list with:
\itemize{
\item each data frame in the data pipeline
\itemize{
\item \code{dfInput}
\item \code{dfTransformed}, returned by \code{\link[=Transform_EventCount]{Transform_EventCount()}}
\item \code{dfAnalyzed}, returned by \code{\link[=Analyze_Chisq]{Analyze_Chisq()}} or \code{\link[=Analyze_Fisher]{Analyze_Fisher()}}
\item \code{dfFlagged}, returned by \code{\link[=Flag]{Flag()}}
\item \code{dfSummary}, returned by \code{\link[=Summarize]{Summarize()}}
}
\item assessment metadata
\itemize{
\item \code{strFunctionName}
\item \code{lParams}
\item \code{lTags}
}
\item output(s)
\itemize{
\item \code{chart}
}
}
}
\description{
Disposition Assessment
}
\examples{
dfInput <- Disp_Map_Raw()
disp_assessment_chisq <- Disp_Assess(dfInput)
disp_assessment_fisher <- Disp_Assess(dfInput, strMethod = "fisher")

}
