% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disp_Map.R
\name{Disp_Map}
\alias{Disp_Map}
\title{Disposition Map}
\usage{
Disp_Map(
  dfDisp,
  strCol,
  strReason = "any",
  vReasonIgnore = c("", " ", "completed", NA)
)
}
\arguments{
\item{dfDisp}{disposition data with the following required columns: SUBJID and SITEID. Must also include the value specified in the \code{strCol} parameter.}

\item{strCol}{column name containing discontinuation reason.}

\item{strReason}{character string containing reason for discontinuation. Can be a value found in \code{dfDisp$strCol} or "any" (the default), which selects all reasons not included in \code{vReasonIgnore}}

\item{vReasonIgnore}{character vector containing reasons to ignore when counting Discontinuation Reason (i.e., "Completed", "", etc.)}
}
\value{
Data frame with one record per person with columns: SubjectID, SiteID, Count, and the value passed to strCol.

#' @examples
df <- Disp_Map(dfDisp = safetyData::adam_adsl, strCol = "DCREASCD", strReason = "adverse event")
}
\description{
Convert from ADaM or raw format to input format for Disposition Assessment.
}
\section{Data Specification}{


This function creates an input dataset for the Disposition Assessment (link to code) by adding Discontinuation Reason Counts to basic subject-level data.

The following columns are required:
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{SITEID} - Site ID
}
}

