% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disp_Map_Raw.R
\name{Disp_Map_Raw}
\alias{Disp_Map_Raw}
\title{Disposition Assessment - Raw Mapping}
\usage{
Disp_Map_Raw(
  dfs = gsm::UseClindata(list(dfSUBJ = "clindata::rawplus_dm", dfSTUDCOMP =
    "clindata::rawplus_studcomp", dfSDRGCOMP =
    "clindata::rawplus_sdrgcomp \%>\% filter(.data$phase == 'Blinded Study Drug Completion')")),
  lMapping = gsm::Read_Mapping("rawplus"),
  strContext = "Study",
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfSTUDCOMP}: \code{data.frame} Subject-level study disposition data with one record per discontinuation reason. Default: \code{clindata::rawplus_studcomp}
\item \code{dfSDRGCOMP}: \code{data.frame} Subject-level treatment disposition data with one record per discontinuation reason. Default: \code{clindata::rawplus_sdrgcomp}
\item \code{dfSUBJ}: \code{data.frame} Subject-level data with one record per participant.Default: \code{clindata::rawplus_dm}
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column. Default: package-defined mapping for raw+.}

\item{strContext}{\code{character} Disposition Context - "Study" (default) or "Treatment".}

\item{bReturnChecks}{\code{logical} Return input checks from \code{\link[=is_mapping_valid]{is_mapping_valid()}}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one record per subject, the input to \code{\link[=Disp_Assess]{Disp_Assess()}}. If
\code{bReturnChecks} is \code{TRUE} \code{Disp_Map_Raw} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert raw disposition data to formatted input data to to formatted
input data to \code{\link[=Disp_Assess]{Disp_Assess()}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{Disp_Map_Raw} creates an input dataset for the Disposition Assessment \code{\link[=Disp_Assess]{Disp_Assess()}} by adding
Discontinuation Reason Counts (derived from \code{dfDISP}) to basic subject-level data (from \code{dfSUBJ}).
}
\examples{
# Run with defaults.
dfInput <- Disp_Map_Raw()

# Run with error checking and message log.
dfInput <- Disp_Map_Raw(bReturnChecks = TRUE, bQuiet = FALSE)

}
