% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ExtractDirectoryPaths.R
\name{ExtractDirectoryPaths}
\alias{ExtractDirectoryPaths}
\title{Extract a vector of directory paths that contain a specified file name.}
\usage{
ExtractDirectoryPaths(cPath, file)
}
\arguments{
\item{cPath}{\code{character} Path to longitudinal data folders.}

\item{file}{\code{character} File name that must be in the folders to qualify for selection}
}
\value{
\code{vector} containing paths to directories that contain the file of interest.
}
\description{
Extract a vector of directory paths that contain a specified file name.
}
