% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{ExtractStudyAge}
\alias{ExtractStudyAge}
\title{Create Study Results table for Report}
\usage{
ExtractStudyAge(fpfv, snapshot_date)
}
\arguments{
\item{fpfv}{\code{date} the date of the first patient visit of the study}

\item{snapshot_date}{\code{date} the date of the snapshot to derive the period of time from the fpfv}
}
\description{
Create Study Results table for Report
}
\keyword{internal}
