% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_covariate.R
\name{filter_covariate}
\alias{filter_covariate}
\title{filter covariate by group}
\usage{
filter_covariate(df, strGroup = "study")
}
\arguments{
\item{df}{\code{data.frame} data frame to filter}

\item{strGroup}{\code{string} group to filter by. options include: "study", "site", "subject}
}
\description{
filter covariate by group
}
\keyword{internal}
