% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-FilterData.R
\name{FilterData}
\alias{FilterData}
\title{Subset a data frame}
\usage{
FilterData(dfInput, strCol, anyVal, bReturnChecks = FALSE, bQuiet = TRUE)
}
\arguments{
\item{dfInput}{\code{data.frame} A data.frame to be filtered, likely within a mapping function.}

\item{strCol}{\code{character} Domain in \code{lMapping} that references the column to filter on.}

\item{anyVal}{\code{character} Domain in \code{lMapping} that references the value to filter on.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{is_mapping_valid}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame provided as \code{dfInput} and filtered on \code{strCol} == \code{anyVal}.
}
\description{
\code{FilterData} is used to select rows of a data.frame that satisfy a specified condition. \code{FilterData} is called within the \code{RunStep} function to subset available data
for a stratified KRI on an existing categorical variable. For example, we can filter Adverse Event assessment data on the \code{aetoxgr} variable,
which specifies if the adverse event was mild, moderate, or severe.
}
\examples{

subset <- FilterData(
  dfInput = clindata::rawplus_ae,
  strCol = "aeser",
  anyVal = "N"
)

}
