% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-filterDomain.R
\name{FilterDomain}
\alias{FilterDomain}
\title{Utility function for basic filtering}
\usage{
FilterDomain(df, col, vals, bQuiet = TRUE)
}
\arguments{
\item{df}{data.frame be filtered}

\item{col}{column to filter}

\item{vals}{value or values to keep}

\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}
}
\description{
Utility function for basic filtering
}
\examples{
te_ae<-FilterDomain(clindata::rawplus_ae, "AE_TE_FLAG", TRUE)
grade34_ae<-FilterDomain(clindata::rawplus_ae, "AE_GRADE", c(3,4))

}
