% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-FilterDomain.R
\name{FilterDomain}
\alias{FilterDomain}
\title{Utility function for basic filtering}
\usage{
FilterDomain(
  df,
  strDomain,
  lMapping,
  strColParam,
  strValParam,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{df}{\code{data.frame} A data.frame to be filtered, likely within a mapping function.}

\item{strDomain}{\code{character} Domain step that is being filtered.}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name of the column.}

\item{strColParam}{\code{character} Domain in \code{lMapping} that references the column to filter on.}

\item{strValParam}{\code{character} Domain in \code{lMapping} that references the value to filter on.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{is_mapping_valid}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame provided as \code{df} and filtered on \code{strColParam} == \code{strValParam}.
If \code{bReturnChecks} is \code{TRUE}, a \code{list} is returned with a filtered \code{df}, and a list of checks run on input data (\code{lChecks}).
}
\description{
Utility function for basic filtering
}
\examples{
lMapping <- list(dfAE = list(
  strIDCol = "SubjectID", strTreatmentEmergentCol = "AE_TE_FLAG",
  strTreatmentEmergentVal = TRUE, strGradeCol = "AE_GRADE",
  strSeriousCol = "AE_SERIOUS", strSeriousVal = "Yes"
))

te_ae <- FilterDomain(
  df = clindata::rawplus_ae,
  lMapping = lMapping,
  strDomain = "dfAE",
  strColParam = "strSeriousCol",
  strValParam = "strSeriousVal"
)

}
