% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag.R
\name{Flag}
\alias{Flag}
\title{Make data frame with flagged values}
\usage{
Flag(dfAnalyzed, strColumn = "PValue", vThreshold = c(0.05, NA))
}
\arguments{
\item{dfAnalyzed}{data frame in format produced by \code{\link{AE_Poisson_Analyze}}}

\item{strColumn}{Name of the Column to use for thresholding}

\item{vThreshold}{vector of 2 numeric values representing lower and upper threshold values. If NA is provided for either threshold value it is ignored, and no values are flagged based on the threshold.}
}
\value{
input data frame with the columns added for "ThresholdLow","ThresholdHigh","ThresholdCol" and "Flag"
}
\description{
Adds columns flagging sites that represent possible statistical outliers. Rows with PValue less than 0.05 are flagged by default.
}
