% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag.R
\name{Flag}
\alias{Flag}
\title{Make data frame with flagged values}
\usage{
Flag(
  dfAnalyzed,
  strColumn = "PValue",
  vThreshold = c(0.05, NA),
  strValueColumn = NULL
)
}
\arguments{
\item{dfAnalyzed}{data frame where flags should be added}

\item{strColumn}{Name of the Column to use for thresholding}

\item{vThreshold}{vector of 2 numeric values representing lower and upper threshold values. All values in strColumn are compared to vThreshold using strict comparisons. Values less than the lower threshold or greater than the upper threshold are flagged as -1 and 1 respectively. Values equal to the threshold values are set to 0 (i.e. not flagged). If NA is provided for either threshold value it is ignored, and no values are flagged based on the threshold. NA and NaN values in strColumn are given NA flag values.}

\item{strValueColumn}{Optional, Name of the Column to use for sign of Flag. If value for that row is higher than median of strValueColumn then Flag = 1, if lower then Flag = -1.}
}
\value{
input data frame with the columns added for "ThresholdLow","ThresholdHigh","ThresholdCol" and "Flag"
}
\description{
Adds columns flagging sites that represent possible statistical outliers. Rows with PValue less than 0.05 are flagged by default.
}
\section{Data Pipeline}{


The input data (\code{ dfAnalyzed}) for Flag() is typically created using any of these analysis functions:
\code{\link{Analyze_Wilcoxon}}, \code{\link{Analyze_Poisson}}, \code{\link{IE_Analyze}}, \code{\link{Consent_Analyze}}
\itemize{
\item \code{SiteID} - Site ID (required)
\item \code{strColumn} - A column to use for Thresholding (required)
\item 'strValueColumn' - A column to be used for the sign of the flag (optional)
}
}

\examples{
dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )
dfTransformed <- Transform_EventCount( dfInput, cCountCol = 'Count', cExposureCol = "Exposure" )
dfAnalyzed <- Analyze_Wilcoxon( dfTransformed ) 
dfFlagged <- Flag( dfAnalyzed ,  strColumn = 'PValue', strValueColumn = 'Statistic')

}
