% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag.R
\name{Flag}
\alias{Flag}
\title{Flag}
\usage{
Flag(dfAnalyzed, strColumn = "Score", vThreshold = NULL, strValueColumn = NULL)
}
\arguments{
\item{dfAnalyzed}{\code{data.frame} where flags should be added.}

\item{strColumn}{\code{character} Name of the column to use for thresholding. Default: \code{"Score"}}

\item{vThreshold}{\code{numeric} Vector of 2 numeric values representing lower and upper threshold values. All
values in \code{strColumn} are compared to \code{vThreshold} using strict comparisons. Values less than
the lower threshold or greater than the upper threshold are flagged. Values equal to the
threshold values are set to 0 (i.e., not flagged). If NA is provided for either threshold value,
it is ignored and no values are flagged based on the threshold. NA and NaN values in \code{strColumn}
are given NA flag values.}

\item{strValueColumn}{\code{character} Name of the column to use for sign of \code{Flag.} If the value for
that row is higher than the median of \code{strValueColumn}, then \code{Flag} is set to 1. Similarly, if
the value for that row is lower than the median of \code{strValueColumn}, then Flag is set to -1.}
}
\value{
\code{data.frame} with one row per site with columns: \code{GroupID}, \code{TotalCount}, \code{Metric}, \code{Score}, \code{Flag}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Add columns flagging sites that represent possible statistical outliers when the Identity statistical
method is used.
}
\details{
This function provides a generalized framework for flagging sites as part of the
GSM data model (see \code{vignette("DataModel")}).
}
\section{Data Specification}{

\code{Flag} is designed to support the input data (\code{dfAnalyzed}) from the \code{Analyze_Identity()}
function. At a minimum, the input data must have a \code{strGroupCol} parameter and a numeric
\code{strColumn} parameter defined. \code{strColumn} will be compared to the specified thresholds in
\code{vThreshold} to define a new \code{Flag} column, which identifies possible statistical outliers. If a
user would like to see the directionality of those identified points, they can define the
\code{strValueColumn} parameter, which will assign a positive or negative indication to already
flagged points.

The following columns are considered required:
\itemize{
\item \code{GroupID} - Group ID; default is \code{SiteID}
\item \code{GroupLevel} - Group Type
\item \code{strColumn} - A column to use for thresholding
}

The following column is considered optional:
\itemize{
\item \code{strValueColumn} - A column to be used for the sign/directionality of the flagging
}
}

\examples{
dfTransformed <- Transform_Count(analyticsInput, strCountCol = "Numerator")

dfAnalyzed <- Analyze_Identity(dfTransformed)

dfFlagged <- Flag(dfAnalyzed, vThreshold = c(0.001, 0.01))

}
