% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag_Fisher.R
\name{Flag_Fisher}
\alias{Flag_Fisher}
\title{Flag_Fisher}
\usage{
Flag_Fisher(dfAnalyzed, vThreshold = NULL)
}
\arguments{
\item{dfAnalyzed}{data.frame where flags should be added.}

\item{vThreshold}{Vector of 2 numeric values representing lower and upper p-value thresholds.}
}
\value{
\code{data.frame} with one row per site with columns: \code{GroupID}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{PredictedCount}, and \code{Flag}.
}
\description{
Add columns flagging sites that represent possible statistical outliers when the Fisher's Exact Test is used.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function flags sites based on the Fisher's Exact Test result as part of the \href{https://gilead-biostats.github.io/gsm/articles/DataPipeline.html}{GSM data pipeline}.
}
\section{Data Specification}{

\code{Flag_Fisher} is designed to support the input data (\code{dfAnalyzed}) generated from the \code{Analyze_Fisher} function. At a minimum, the input must define a \code{dfAnalyzed} data frame with \code{Score}, \code{Prop}, and \code{Prop_Other} variables included and a \code{vThreshold}. These inputs will be used to identify possible statistical outliers in a new \code{Flag} column by comparing \code{Score}, \code{Prop}, and \code{Prop_Other} values to the specified thresholds.

The following columns are considered required:
\itemize{
\item \code{GroupID} - Group ID; default is \code{SiteID}
\item \code{Score} - P-value calculated from the rates of exposure provided to \code{Analyze_Fisher()}
\item \code{Prop} - Proportion of events of interest over days of exposure
\item \code{Prop_Other} - Cumulative proportion of events of interest over days of exposure
}
}

\examples{
dfTransformed <- tibble::tribble(
  ~GroupID,  ~Numerator,  ~Denominator,  ~Metric,
  139, 5, 901, 0.00555,
  143, 3, 170, 0.0176,
  162, 3, 370, 0.00811,
  167, 3, 360, 0.00833,
  173, 6, 680, 0.00882,
  189, 4, 815, 0.00491,
  29,  2, 450, 0.00444,
  5, 5, 730, 0.00685,
  58, 1, 225, 0.00444,
  78, 2, 50, 0.04
)

dfAnalyzed <- Analyze_Poisson(dfTransformed)

dfFlagged <- Flag(dfAnalyzed, vThreshold = c(-5, 5))

}
