% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag_Poisson.R
\name{Flag_Poisson}
\alias{Flag_Poisson}
\title{Flag_Poisson}
\usage{
Flag_Poisson(dfAnalyzed, vThreshold = NULL)
}
\arguments{
\item{dfAnalyzed}{data.frame where flags should be added.}

\item{vThreshold}{Vector of 4 numeric values representing lower and upper threshold values. All
values in strColumn are compared to vThreshold using strict comparisons. Values less than the lower threshold or greater than the upper threshold are flagged as -1 and 1 respectively. Values equal to the threshold values are set to 0 (i.e. not flagged). If NA is provided for either threshold value it is ignored, and no values are flagged based on the threshold. NA and NaN values in strColumn are given NA flag values.}
}
\value{
\code{data.frame} with \code{Flag} column added
}
\description{
Add columns flagging sites that represent possible statistical outliers.
}
\details{
This function provides a generalized framework for flagging sites as part of the \href{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}{GSM data pipeline}.
}
\section{Data Specification}{

\code{Flag} is designed to support the input data (\code{dfAnalyzed}) from many different
\code{Analyze} functions.  dfTransformed must have a \code{score} column.
}

\examples{
dfInput <- AE_Map_Adam()

dfTransformed <- Transform_Rate(dfInput,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)

dfAnalyzed <- Analyze_Poisson(dfTransformed)

dfFlagged <- Flag_Poisson(dfAnalyzed, vThreshold = c(-7, -5, 5, 7))

}
