% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag.R
\name{Flag_Poisson}
\alias{Flag_Poisson}
\title{Flag_Poisson}
\usage{
Flag_Poisson(
  dfAnalyzed,
  strColumn = "Score",
  vThreshold = c(-3, -2, 2, 3),
  vFlag = c(-2, -1, 0, 1, 2),
  vFlagOrder = c(2, -2, 1, -1, 0)
)
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Alias for `Flag()``
}
