% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flag_Poisson.R
\name{Flag_Poisson}
\alias{Flag_Poisson}
\title{Flag_Poisson}
\usage{
Flag_Poisson(dfAnalyzed, vThreshold = NULL)
}
\arguments{
\item{dfAnalyzed}{data.frame where flags should be added.}

\item{vThreshold}{Vector of 4 numeric values representing lower and upper threshold values. All values in the \code{Score} column are compared to \code{vThreshold} using strict comparisons. Values less than the lower thresholds or greater than the upper thresholds are flagged. Values equal to the threshold values are set to 0 (i.e., not flagged). If NA is provided for either threshold value, it is ignored and no values are flagged based on the thresholds. NA and NaN values in \code{Score} are given NA flag values.}
}
\value{
\code{data.frame} with one row per site with columns: \code{GroupID}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{PredictedCount}, \code{Flag}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Add columns flagging sites that represent possible statistical outliers when the Poisson statistical method is used.
}
\details{
This function flags sites based on the Poisson analysis result as part of \code{vignette("DataModel")}.
}
\section{Data Specification}{

\code{Flag_Poisson} is designed to support the input data (\code{dfAnalyzed}) generated from the \code{Analyze_Poisson} function. At a minimum, the input must define a \code{dfAnalyzed} data frame with a \code{Score} variable included and a \code{vThreshold}. These inputs will be used to identify possible statistical outliers in a new \code{Flag} column by comparing \code{Score} values to the specified thresholds.

The following columns are considered required:
\itemize{
\item \code{GroupID} - Group ID; default is \code{SiteID}
\item \code{GroupLevel} - Group Type
\item \code{Score} - Site residuals calculated from the rates of exposure provided to \code{Analyze_Poisson()}
}
}

\examples{
dfTransformed <- Transform_Rate(analyticsInput)

dfAnalyzed <- Analyze_Poisson(dfTransformed)

dfFlagged <- Flag_Poisson(dfAnalyzed, vThreshold = c(-7, -5, 5, 7))

}
