% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Enrolled.R
\name{Get_Enrolled}
\alias{Get_Enrolled}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
Get_Enrolled(dfSUBJ, dfConfig, lMapping, strUnit, strBy)
}
\arguments{
\item{dfSUBJ}{\code{data.frame} Typically a DM dataset.}

\item{dfConfig}{\code{data.frame} Dataset containing configuration parameters: \code{studyid}, \code{workflowid}, \code{gsm_version}, \code{param}, \code{index}, and \code{value}.}

\item{lMapping}{\code{list} Mappings for the assessment(s) being run.}

\item{strUnit}{\code{character} Type of enrollment; one of \code{participant} or \code{site}.}

\item{strBy}{\code{character} Domain of enrollment; one of \code{study} or \code{site}.}
}
\value{
\code{integer} or \code{data.frame}, depending on input parameters.
}
\description{
Derive number of enrolled participants or sites by a given denominator (study or site).
}
\details{
Get_Enrolled Calculated enrolled participants or sites.
}
\examples{
enrolled <- Get_Enrolled(
  dfSUBJ = clindata::rawplus_dm,
  dfConfig = gsm::config_param,
  lMapping = yaml::read_yaml(
    system.file("mappings",
      "mapping_rawplus.yaml",
      package = "gsm"
    )
  ),
  strUnit = "participant",
  strBy = "site"
)

}
