% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm-package.R
\docType{package}
\name{gsm-package}
\alias{gsm}
\alias{gsm-package}
\title{gsm: Gilead Statistical Monitoring}
\description{
The Gilead Statistical Monitoring or {gsm} R package provides a framework for statistical data monitoring.
}
\author{
\strong{Maintainer}: George Wu \email{george.wu@gilead.com}

Authors:
\itemize{
  \item Jeremy Wildfire \email{jeremy.wildfire@gilead.com}
}

}
\keyword{internal}
