\name{GSM-package}
\alias{GSM}
\alias{GSM-package}
\docType{package}
\title{
\packageTitle{GSM}
}
\description{
\packageDescription{GSM}
}
\details{
The DESCRIPTION file:
	\packageDESCRIPTION{GSM}
	\packageIndices{GSM}
	This package implements a Bayesian approach for estimation of a mixture of gamma distributions in which the mixing occurs over the shape parameter. This family provides a flexible and novel approach for modeling heavy-tailed distributions, it is computationally efficient, and it only requires to specify a prior distribution for a single parameter. See Venturini et al. (2008).
}
\references{
   Venturini, S., Dominici, F. and Parmigiani, G. (2008), "Gamma shape mixtures for heavy-tailed distributions". Annals of Applied Statistics, \bold{Volume 2}, Number 2, 756--776.
   \url{http://projecteuclid.org/euclid.aoas/1215118537}
}
\author{
\packageAuthor{GSM}

Maintainer: \packageMaintainer{GSM}
}
\seealso{
   \code{\link{estim.gsm}},
   \code{\link{estim.gsm_theta}}.
}
\keyword{distribution}
\keyword{mixture}
