% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE_Assess.R
\name{IE_Assess}
\alias{IE_Assess}
\title{Inclusion/Exclusion Assessment}
\usage{
IE_Assess(dfInput, nThreshold = 0.5, cLabel = "", bDataList = FALSE)
}
\arguments{
\item{dfInput}{input dataset with subject information containing : (required variables)}

\item{nThreshold}{threshold values Flagging, integer values greater than will be flagged.}

\item{cLabel}{Assessment label}

\item{bDataList}{Should all assessment datasets be returned as a list? If False (the default), only the finding data frame is returned}
}
\value{
Finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
Inclusion/Exclusion Assessment
}
