% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE_Assess.R
\name{IE_Assess}
\alias{IE_Assess}
\title{Inclusion/Exclusion Assessment}
\usage{
IE_Assess(
  dfInput,
  nThreshold = 0.5,
  lMapping = yaml::read_yaml(system.file("mappings", "IE_Assess.yaml", package = "gsm")),
  lLabels = list(workflowid = "", group = strGroup, abbreviation = "IE", metric =
    "Inclusion/Exclusion Issues", numerator = "Inclusion/Exclusion Issues", denominator =
    "", model = "Normal Approximation", score = "Adjusted Z-Score"),
  strGroup = "Site",
  nMinDenominator = NULL,
  bMakeCharts = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data, a data frame with one record per subject.}

\item{nThreshold}{\code{numeric} Threshold specification. Default: \code{0.5}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column. Default: package-defined Inclusion/Exclusion Assessment mapping.}

\item{lLabels}{\code{list} Labels used to populate chart labels.}

\item{strGroup}{\code{character} Grouping variable. \code{"Site"} (the default) uses the column named in \code{mapping$strSiteCol}. Other valid options using the default mapping are \code{"Study"}, \code{"Country"}, and \code{"CustomGroup"}.}

\item{nMinDenominator}{\code{numeric} Specifies the minimum denominator required to return a \code{score} and calculate a \code{flag}. Default: NULL}

\item{bMakeCharts}{\code{logical} Boolean value indicating whether to create charts.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} \code{lData}, a named list with:
\itemize{
\item each data frame in the data pipeline
\itemize{
\item \code{dfTransformed}, returned by \code{\link[=Transform_Count]{Transform_Count()}}
\item \code{dfAnalyzed}, returned by \code{\link[=Analyze_Identity]{Analyze_Identity()}}
\item \code{dfFlagged}, returned by \code{\link[=Flag]{Flag()}}
\item \code{dfSummary}, returned by \code{\link[=Summarize]{Summarize()}}
}
\item \code{list} \code{lCharts}, a named list with:
\itemize{
\item \code{barMetric}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}} using strType == "metric"
\item \code{barScore}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}} using strType == "score"
}
\item \code{list} \code{lChecks}, a named list with:
\itemize{
\item \code{dfInput}, a named list returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{status}, a boolean returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{mapping}, a named list that is provided as an argument to the \code{lMapping} parameter in \code{\link[=IE_Assess]{IE_Assess()}}
\item \code{spec}, a named list used to define variable specifications
}
}
}
\description{
Evaluates sites exhibiting aberrant or excessive rates of unmet or missing inclusion/exclusion (IE) criteria.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The Inclusion/Exclusion Assessment uses the standard \href{https://gilead-biostats.github.io/gsm/articles/DataPipeline.html}{GSM data pipeline} to flag sites with IE issues. This assessment detects sites with excessive rates of unmet or
missing IE criteria, as defined by \code{nThreshold}. The count returned in the summary represents the
number of subjects at a given site with at least one unmet or missing IE criterion.
}
\examples{
dfInput <- IE_Map_Raw()
ie_assessment <- IE_Assess(dfInput)

}
