% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE_Map_Raw.R
\name{IE_Map_Raw}
\alias{IE_Map_Raw}
\title{Inclusion/Exclusion Assessment Mapping from Raw Data- Make Input Data}
\usage{
IE_Map_Raw(dfIe = NULL)
}
\arguments{
\item{dfIe}{ie dataset with columns SUBJID IECAT IETEST IEORRES}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Total, Valid, Invalid, Missing, Details
}
\description{
Convert from raw data format to needed input format for Inclusion/Exclusion Assessment
}
