% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE_Map_Raw.R
\name{IE_Map_Raw}
\alias{IE_Map_Raw}
\title{Inclusion/Exclusion Assessment Mapping from Raw Data- Make Input Data}
\usage{
IE_Map_Raw(
  dfs = list(dfIE = clindata::rawplus_ie, dfSUBJ = clindata::rawplus_subj),
  lMapping = clindata::mapping_rawplus,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{list of data frames including:
\itemize{
\item \code{dfIE} ie dataset with columns SUBJID and values specified in strCategoryCol and strValueCol.
\item \code{dfSUBJ} Subject-level Raw Data required columns: SubjectID SiteID
}}

\item{lMapping}{List containing expected columns in each data set.}

\item{bReturnChecks}{Should input checks using \code{is_mapping_valid} be returned? Default is FALSE.}

\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}
}
\value{
Data frame with one record per participant giving the number of inclusion/exclusion criteria the participant did not meet as expected. Expected columns: SubjectID, SiteID, Count
}
\description{
Convert from raw data format to needed input format for Inclusion/Exclusion Assessment.
}
\details{
This function creates the required input for \code{\link{IE_Assess}}.
}
\section{Data Specification}{


The following columns are required:
\itemize{
\item \code{dfIE}
\itemize{
\item \code{SubjectID} - Unique subject ID
\item Value specified in \code{mapping} - IE Category; "IECAT_STD" by default
\item Value specified in \code{mapping} - Incl criteria not met Excl criteria met; "IEORRES" by default
}
\item \code{dfSUBJ}
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
}
}
}

\examples{

dfInput <- IE_Map_Raw() # Run with defaults
dfInput <- IE_Map_Raw(bReturnChecks=TRUE, bQuiet=FALSE) # Run with error checking and message log

}
