% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE_Map_Raw.R
\name{IE_Map_Raw}
\alias{IE_Map_Raw}
\title{Inclusion/Exclusion Assessment - Raw Mapping}
\usage{
IE_Map_Raw(
  dfs = list(dfIE = clindata::rawplus_ie, dfSUBJ = clindata::rawplus_subj),
  lMapping = clindata::mapping_rawplus,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfIE}: \code{data.frame} Criterion-level data with one record subject per criterion.
\item \code{dfSUBJ}: \code{data.frame} Subject-level data with one record per subject.
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name of the column.}

\item{bReturnChecks}{\code{logical} Return input checks from {gsm::is_mapping_valid()}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame with one record per subject, the input to {gsm::IE_Assess()}. If
\code{bReturnChecks} is \code{TRUE} \code{IE_Map_Raw} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert raw inclusion/exclusion (IE) data, typically processed case report form data, to formatted
input data to {gsm::IE_Assess()}.
}
\details{
\code{IE_Map_Raw} combines IE data with subject-level data to create formatted input data to
{gsm::IE_Assess()}. This function creates an input dataset for the IE Assessment
(${gsm::IE_Assess()}) by binding subject-level unmet IE criteria counts (derived from \code{dfIE}) to
subject-level data (from \code{dfSUBJ}). Note that the function can generate data summaries for
specific types of IE criteria by passing filtered IE data to \code{dfIE}.
}
\section{Data specification}{
\tabular{lllll}{
   domain \tab col_key \tab col_value \tab vRequired \tab vUniqueCols \cr
   dfSUBJ \tab strIDCol \tab SubjectID \tab TRUE \tab TRUE \cr
   dfSUBJ \tab strSiteCol \tab SiteID \tab TRUE \tab FALSE \cr
   dfIE \tab strIDCol \tab SubjectID \tab TRUE \tab NA \cr
   dfIE \tab strCategoryCol \tab IE_CATEGORY \tab TRUE \tab NA \cr
   dfIE \tab strValueCol \tab IE_VALUE \tab TRUE \tab NA \cr
}
}

\examples{
dfInput <- IE_Map_Raw() # Run with defaults
dfInput <- IE_Map_Raw(bReturnChecks = TRUE, bQuiet = FALSE) # Run with error checking and message log

}
