% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{input_data_schema}
\alias{input_data_schema}
\title{Input Data Schema}
\format{
A data frame with 76 rows and 7 variables.
#' \describe{
\item{Source}{clinical, ctms, or edc}
\item{Name}{Name of the Raw+ data source}
\item{Domain}{Abbreviated data domain name, e.g., dfSUBJ == dm}
\item{GSM_Domain_Key}{Name of the data domain as it is referenced in \code{gsm}, e.g., "dfNAME"}
\item{Column}{Raw+ column name.}
\item{GSM_Column_Key}{Name of \code{gsm} key that is paired with \code{column}}
\item{Description}{Description of \code{Column}}
}
}
\usage{
input_data_schema
}
\description{
A dataset containing key/column pairs for clinical, ctms, and edc mappings.
}
\keyword{datasets}
