% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_is_mapping_valid.R
\name{is_mapping_valid}
\alias{is_mapping_valid}
\title{Check that a data frame contains columns and fields specified in mapping}
\usage{
is_mapping_valid(df, mapping, unique_cols = NULL, na_cols = NULL)
}
\arguments{
\item{df}{dataframe to compare to mapping object}

\item{mapping}{named list specifying expected columns and fields in df}

\item{unique_cols}{list of columns expected to be unique. default = NULL (none)}

\item{no_cols}{list of columns where na values are acceptable default = NULL (none)}
}
\description{
Check that a data frame contains columns and fields specified in mapping
}
\examples{
rdsl_mapping<- list(id_col="SubjectID", site_col="SiteID", exposure_col="TimeOnTreatment")
is_mapping_valid(clindata::rawplus_rdsl, rdsl_mapping, unique_cols="SUBJID") #TRUE
is_mapping_valid(clindata::rawplus_rdsl, rdsl_mapping, unique_cols="SiteID") #FALSE - site_col not unique
rdsl_mapping$not_a_col<-"nope"
is_mapping_valid(clindata::rawplus_rdsl, rdsl_mapping, unique_cols="SUBJID") #FALSE - column not found

}
