% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-is_mapping_valid.R
\name{is_mapping_valid}
\alias{is_mapping_valid}
\title{Check that a data frame contains columns and fields specified in mapping}
\usage{
is_mapping_valid(
  df,
  mapping,
  vRequiredParams = NULL,
  vUniqueCols = NULL,
  vNACols = NULL,
  bQuiet = TRUE
)
}
\arguments{
\item{df}{data.frame to compare to mapping object.}

\item{mapping}{named list specifying expected columns and fields in df.}

\item{vRequiredParams}{character vector of names that must be present in \code{mapping}.}

\item{vUniqueCols}{list of columns expected to be unique. default = NULL (none).}

\item{vNACols}{list of columns where na values are acceptable default = NULL (none).}

\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}
}
\description{
Check that a data frame contains columns and fields specified in mapping
}
\examples{
rdsl_mapping <- list(strIDCol = "SubjectID",
                    strSiteCol = "SiteID",
                    strExposureCol = "TimeOnTreatment")

is_mapping_valid(df = clindata::rawplus_rdsl,
                 mapping = rdsl_mapping,
                 vUniqueCols = "SUBJID",
                 vRequiredParams = c("strIDCol", "strSiteCol", "strExposureCol"))

rdsl_mapping$not_a_col <- "nope"

is_mapping_valid(df = clindata::rawplus_rdsl,
                 mapping = rdsl_mapping,
                 vUniqueCols = "SUBJID",
                 vRequiredParams = c("strIDCol", "strSiteCol", "strExposureCol"))

}
