% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_workflow_valid.R
\name{is_workflow_valid}
\alias{is_workflow_valid}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
is_workflow_valid(lWorkflow)
}
\arguments{
\item{lWorkflow}{\code{list} of workflows to be run in \code{Study_Assess()} or \code{Make_Snapshot()}.}
}
\value{
\code{list} with TRUE/FALSE status for each workflow.
}
\description{
Check if workflow is valid/contains necessary metadata.
}
\examples{
workflow <- MakeWorkflowList()

}
