% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joining_map.R
\name{joining_map}
\alias{joining_map}
\title{a function for defining columns to join by within Distribution map function}
\usage{
joining_map(dist_list, lMapping, strDomain, strSubjCol, has_site)
}
\arguments{
\item{dist_list}{\code{list} a list of data and linked category columns}

\item{lMapping}{\code{list} a list of predefined columns to map over}

\item{strDomain}{\code{string} the name of the raw data domain to use}

\item{strSubjCol}{\code{string} the name of the subject id column found within specified domain}

\item{has_site}{\code{logical} indicator if siteid is found in mapping yaml for domain specified}
}
\description{
a function for defining columns to join by within Distribution map function
}
\keyword{internal}
