% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-kri_directionality_logo.R
\name{kri_directionality_logo}
\alias{kri_directionality_logo}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
kri_directionality_logo(flag_value)
}
\arguments{
\item{flag_value}{\code{numeric} Value between -2 and 2.}
}
\description{
Reporting utility function to create fontawesome directionality logos within \code{DT::datatable()}
}
\details{
KRI Directionality Logo.
}
