% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-kri_directionality_logo.R
\name{kri_directionality_logo}
\alias{kri_directionality_logo}
\title{KRI Directionality Logo.}
\usage{
kri_directionality_logo(flag_value, title = NULL)
}
\arguments{
\item{flag_value}{\code{numeric} Value between -2 and 2.}

\item{title}{\code{character} Data that will be passed to the \if{html}{\out{<title>}} tag.}
}
\description{
Reporting utility function to create fontawesome directionality logos within \code{DT::datatable()}
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
