% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB_Assess.R
\name{LB_Assess}
\alias{LB_Assess}
\title{Labs Assessment}
\usage{
LB_Assess(
  dfInput,
  vThreshold = NULL,
  strMethod = "chisq",
  strKRILabel = "\% Abnormal Labs",
  strGroup = "Site",
  lTags = list(Assessment = "Labs"),
  bChart = TRUE,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data, a data frame with one record per subject.}

\item{vThreshold}{\code{numeric} Threshold specification, a vector of length 2 that defaults to
\code{c(.05, NA)} for both Chi-square test (\code{strMethod} = "chisq") and Fisher's exact test (\code{strMethod} = "fisher").}

\item{strMethod}{\code{character} Statistical method. Valid values:
\itemize{
\item \code{"chisq"} (default)
\item \code{"fisher"}
\item \code{"identity"}
}}

\item{strKRILabel}{\code{character} KRI description. Default: \code{"\% Abnormal Labs"}}

\item{strGroup}{\code{character} Grouping variable. \code{"Site"} (the default) uses the column named in \code{mapping$strSiteCol}. Other valid options using the default mapping are \code{"Study"} and \code{"CustomGroup"}.}

\item{lTags}{\code{list} Assessment tags, a named list of tags describing the assessment that defaults
to \code{list(Assessment = "Labs")}. \code{lTags} is returned as part of the assessment (\code{lAssess$lTags})
and each tag is added as a column in \code{lAssess$dfSummary}.}

\item{bChart}{\code{logical} Generate data visualization? Default: \code{TRUE}}

\item{bReturnChecks}{\code{logical} Return input checks from \code{\link[=is_mapping_valid]{is_mapping_valid()}}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} Assessment, a named list with:
\itemize{
\item each data frame in the data pipeline
\itemize{
\item \code{dfInput}
\item \code{dfTransformed}, returned by \code{\link[=Transform_EventCount]{Transform_EventCount()}}
\item \code{dfAnalyzed}, returned by \code{\link[=Analyze_Chisq]{Analyze_Chisq()}} or \code{\link[=Analyze_Fisher]{Analyze_Fisher()}}
\item \code{dfFlagged}, returned by \code{\link[=Flag]{Flag()}}
\item \code{dfSummary}, returned by \code{\link[=Summarize]{Summarize()}}
}
\item assessment metadata
\itemize{
\item \code{strFunctionName}
\item \code{lParams}
\item \code{lTags}
}
\item output(s)
\itemize{
\item \code{chart}
}
}
}
\description{
Evaluates rate of reported Lab Abnormalities (LB).
}
\details{
The Lab Abnormality Assessment uses the standard \href{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}{GSM data pipeline} to flag possible outliers. Additional details regarding the data pipeline and statistical
methods are described below.
}
\section{Data specification}{
\tabular{llll}{
   \strong{Domain} \tab \strong{Column Key} \tab \strong{Default Value} \tab \strong{Required?} \cr
   dfInput \tab strIDCol \tab SubjectID \tab TRUE \cr
   dfInput \tab strCountCol \tab Count \tab TRUE \cr
   dfInput \tab strGroupCol \tab  \tab TRUE \cr
}
}

\examples{
dfInput <- LB_Map_Raw()
lb_assessment_chisq <- LB_Assess(dfInput)
lb_assessment_fisher <- LB_Assess(dfInput, strMethod = "fisher")
lb_assessment_identity <- LB_Assess(dfInput, strMethod = "identity")

}
