% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Covariate.R
\name{Make_Covariate}
\alias{Make_Covariate}
\title{Make Covariate Data}
\usage{
Make_Covariate(lSnapshot)
}
\arguments{
\item{lSnapshot}{\code{list} a snapshot object from \code{Make_Snapshot()}}
}
\value{
\code{list} containing covariate tables and charts for \link{Covariate_Report}.
}
\description{
Make Covariate Data
}
\examples{
\dontrun{
snapshot <- Make_Snapshot()
covariate <- Make_Covariate(snapshot)
}

}
