% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Snapshot.R
\name{Make_Snapshot}
\alias{Make_Snapshot}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
Make_Snapshot(
  lMeta = list(config_param = gsm::config_param, config_workflow = gsm::config_workflow,
    meta_params = gsm::meta_param, meta_site = clindata::ctms_site, meta_study =
    clindata::ctms_study, meta_workflow = gsm::meta_workflow),
  lData = list(dfSUBJ = clindata::rawplus_dm, dfAE = clindata::rawplus_ae, dfPD =
    clindata::ctms_protdev, dfCONSENT = clindata::rawplus_consent, dfIE =
    clindata::rawplus_ie, dfLB = clindata::rawplus_lb, dfSTUDCOMP =
    clindata::rawplus_studcomp, dfSDRGCOMP = clindata::rawplus_sdrgcomp \%>\%
    filter(.data$phase == "Blinded Study Drug Completion"), dfDATACHG =
    clindata::edc_data_points, dfDATAENT = clindata::edc_data_pages, dfQUERY =
    clindata::edc_queries, dfENROLL = clindata::rawplus_enroll),
  lMapping = Read_Mapping(),
  lAssessments = NULL,
  strAnalysisDate = NULL,
  bUpdateParams = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{lMeta}{\code{list} a named list of data frames containing metadata, configuration, and workflow parameters for a given study.
See the Data Model Vignette - Appendix 2 - Data Model Specifications for detailed specifications.}

\item{lData}{\code{list} a named list of domain-level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name of the column. Default: package-defined mapping for raw+.}

\item{lAssessments}{\code{list} a named list of metadata defining how each assessment should be run. By default, \code{MakeWorkflowList()} imports YAML specifications from \code{inst/workflow}.}

\item{strAnalysisDate}{\code{character} date that the data was pulled/wrangled/snapshot. Note: date should be provided in format: \code{YYYY-MM-DD}.}

\item{bUpdateParams}{\code{logical} if \code{TRUE}, configurable parameters found in \code{lMeta$config_param} will overwrite the default values in \code{lMeta$meta_params}. Default: \code{FALSE}.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}.}
}
\value{
\code{list} \code{lSnapshot}, a named list with a data.frame for each component of the {gsm} data model.
\itemize{
\item \code{status_study}
\item \code{status_site}
\item \code{status_workflow}
\item \code{status_param}
\item \code{results_summary}
\item \code{results_analysis}
\item \code{results_bounds}
\item \code{meta_workflow}
\item \code{meta_param}
}
}
\description{
\code{Make_Snapshot()} ingests data from a variety of sources, and runs KRIs and/or QTLs based on the \code{list} provided in \code{lAssessments}.

For more context about the inputs and outputs of \code{Make_Snapshot()}, refer to the \href{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}{GSM Data Pipeline Vignette}, specifically
Appendix 2 - Data Model Specifications
}
\details{
Make Snapshot - create and export data model.
}
\examples{
# run with default testing data
\dontrun{
snapshot <- Make_Snapshot()
}

}
