% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Snapshot.R
\name{Make_Snapshot}
\alias{Make_Snapshot}
\title{Make Snapshot - say cheese}
\usage{
Make_Snapshot(
  lMeta = list(meta_study = NULL, meta_site = NULL, config_schedule = NULL,
    config_workflow = gsm::config_workflow, config_params = gsm::config_param,
    meta_workflow = gsm::meta_workflow, meta_params = gsm::meta_param),
  lData = list(dfSUBJ = clindata::rawplus_dm \%>\% filter(!is.na(timeontreatment)), dfAE
    = clindata::rawplus_ae, dfPD = clindata::rawplus_protdev, dfCONSENT =
    clindata::rawplus_consent, dfIE = clindata::rawplus_ie, dfLB = clindata::rawplus_lb,
    dfSTUDCOMP = clindata::rawplus_studcomp, dfSDRGCOMP = clindata::rawplus_sdrgcomp
    \%>\% filter(.data$datapagename == "Blinded Study Drug Completion")),
  lMapping = yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package =
    "gsm")),
  cPath = NULL,
  bQuiet = TRUE
)
}
\arguments{
\item{bQuiet}{}
}
\description{
Make Snapshot - say cheese
}
\examples{

# run with default testing data
snapshot <- Make_Snapshot()


}
