% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Snapshot.R
\name{Make_Snapshot}
\alias{Make_Snapshot}
\title{Make Snapshot - create and export data model.}
\usage{
Make_Snapshot(
  lMeta = gsm::UseClindata(list(config_param = "gsm::config_param", config_workflow =
    "gsm::config_workflow", meta_params = "gsm::meta_param", meta_site =
    "clindata::ctms_site", meta_study = "clindata::ctms_study", meta_workflow =
    "gsm::meta_workflow")),
  lData = gsm::UseClindata(list(dfSUBJ = "clindata::rawplus_dm", dfAE =
    "clindata::rawplus_ae", dfPD = "clindata::ctms_protdev", dfCONSENT =
    "clindata::rawplus_consent", dfIE = "clindata::rawplus_ie", dfLB =
    "clindata::rawplus_lb", dfSTUDCOMP = "clindata::rawplus_studcomp", dfSDRGCOMP =
    "clindata::rawplus_sdrgcomp \%>\%\\n      filter(.data$phase == 'Blinded Study Drug Completion')",
    dfDATACHG = "clindata::edc_data_points", dfDATAENT = "clindata::edc_data_pages",
    dfQUERY = "clindata::edc_queries", dfENROLL = "clindata::rawplus_enroll")),
  lMapping = Read_Mapping(),
  lAssessments = MakeWorkflowList(lMeta = lMeta),
  lPrevSnapshot = NULL,
  strAnalysisDate = NULL,
  vAppendFiles = names(lPrevSnapshot$lSnapshot),
  bMakeCharts = TRUE,
  bQuiet = TRUE
)
}
\arguments{
\item{lMeta}{\code{list} a named list of data frames containing metadata, configuration, and workflow parameters for a given study.
See the Data Model Vignette - Appendix 2 - Data Model Specifications for detailed specifications.}

\item{lData}{\code{list} a named list of domain-level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name of the column. Default: package-defined mapping for raw+.}

\item{lAssessments}{\code{list} a named list of metadata defining how each assessment should be run. By default, \code{MakeWorkflowList()} imports YAML specifications from \code{inst/workflow}.}

\item{lPrevSnapshot}{\code{list} optional argument for the previous snapshot run to track longitudinal data. Default: \code{NULL}.}

\item{strAnalysisDate}{\code{character} date that the data was pulled/wrangled/snapshot. Note: date should be provided in format: \code{YYYY-MM-DD}.
If \code{NULL}, \code{gsm_analysis_date} will default to the \code{Sys.Date()} that \code{Make_Snapshot()} is run. Default: \code{NULL}.}

\item{vAppendFiles}{\code{vector} a character vector of log files to append. Default: all log files from \code{lPrevSnapshot} argument.}

\item{bMakeCharts}{\code{logical} boolean value indicating whether to create charts. Default: \code{TRUE}.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}.}
}
\description{
\code{Make_Snapshot()} ingests data from a variety of sources, and runs KRIs and/or QTLs based on the \code{list} provided in \code{lAssessments}.

For more context about the inputs and outputs of \code{Make_Snapshot()}, refer to the \href{https://gilead-biostats.github.io/gsm/articles/DataPipeline.html}{GSM Data Pipeline Vignette}, specifically
Appendix 2 - Data Model Specifications
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
# run with default testing data
\dontrun{
snapshot <- Make_Snapshot()
}

}
