% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Snapshot.R
\name{Make_Snapshot}
\alias{Make_Snapshot}
\title{Make Snapshot - create and export Gizmo data model.}
\usage{
Make_Snapshot(
  lMeta = list(config_param = gsm::config_param, config_schedule =
    clindata::config_schedule, config_workflow = gsm::config_workflow, meta_params =
    gsm::meta_param, meta_site = clindata::ctms_site, meta_study = clindata::ctms_study,
    meta_workflow = gsm::meta_workflow),
  lData = list(dfSUBJ = clindata::rawplus_dm, dfAE = clindata::rawplus_ae, dfLB =
    clindata::rawplus_lb, dfPD = clindata::rawplus_protdev, dfSTUDCOMP =
    clindata::rawplus_studcomp, dfSDRGCOMP = clindata::rawplus_sdrgcomp \%>\%
    filter(.data$datapagename == "Blinded Study Drug Completion")),
  lMapping = yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package =
    "gsm")),
  lAssessments = NULL,
  cPath = NULL,
  bQuiet = TRUE
)
}
\arguments{
\item{lMeta}{\code{list} a named list of data frames containing metadata, configuration, and workflow parameters for a given study. TODO: add details about expected lMeta input.}

\item{lData}{\code{list} a named list of domain level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{lMapping}{\code{list} a named list identifying the columns needed in each data domain.}

\item{lAssessments}{a named list of metadata defining how each assessment should be run. By default, \code{MakeWorkflowList()} imports YAML specifications from \code{inst/workflow}.}

\item{cPath}{\code{character} a character string indicating a working directory to save .csv files; the output of the snapshot.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\description{
Make Snapshot - create and export Gizmo data model.
}
\section{Return - Data Model Description}{
A named \code{list} of data.frames:
\itemize{
\item \code{status_study}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{enrolled_sites} - # of enrolled sites
\item \code{enrolled_participants} - # of enrolled participants
\item \code{planned_sites} - # of planned sites
\item \code{planned_participants} - # of planned participants
\item \code{title} - Protocol title
\item \code{nickname} - Protocol nickname
\item \code{indication} - Indication
\item \code{ta} - Therapeutic Area
\item \code{phase} - Phase
\item \code{status} - Study Status
\item \code{fpfv} - First-patient first visit date
\item \code{lplv} - Last-patient last visit date
\item \code{rbm_flag} - Risk-based monitoring flag
}
\item \code{status_site}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{siteid} - Unique Site ID
\item \code{institution} - Institution Name
\item \code{status} - Site Status
\item \code{enrolled_participants} - # of enrolled participants
\item \code{start_date} - Site Activation Date
\item \code{city} - City
\item \code{state} - State
\item \code{country} - Country
\item \code{invname} - Investigator name
}
\item \code{status_workflow}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{workflowid} - Unique Workflow ID
\item \code{gsm_version} - GSM version of workflow
\item \code{active} - Is workflow active?
\item \code{status} - Did workflow run?
\item \code{notes} - Status notes
}
\item \code{status_param}, with columns:
\itemize{
\item \code{study_id} - Unique Study ID
\item \code{workflowid} - Unique Workflow ID
\item \code{gsm_version} - GSM version of workflow
\item \code{param} - Parameter name
\item \code{index} - Index value for parameter
\item \code{value} - Value for parameter at index
}
\item \code{status_schedule}, with columns:
\itemize{
\item \code{study_id} - Unique Study ID
\item \code{snapshot_date} - Snapshot Date
}
\item \code{results_summary}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{workflowid} - Unique Workflow ID
\item \code{groupid} - Unique Group ID (e.g. SiteID for KRI workflows)
\item \code{numerator} - Metric numerator
\item \code{denominator} - Metric denominator
\item \code{metric} - Metric value
\item \code{score} - Statistical Score
\item \code{flag} - Flag
}
\item \code{results_bounds}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{workflowid} - Unique Workflow ID
\item \code{threshold} - Threshold
\item \code{numerator} - Y value
\item \code{denominator} - X value
\item \code{log_denominator} - Log X value
}
\item \code{meta_workflow}, with columns:
\itemize{
\item \code{workflowid} - Unique Workflow ID
\item \code{gsm_version} - GSM Version of workflow
\item \code{group} - Group for workflow - Site, Country, Region, or Study
\item \code{metric} - Metric Label
\item \code{numerator} - Numerator Label
\item \code{denominator} - Denominator Label
\item \code{outcome} - Metric Outcome Type (e.g. rate)
\item \code{model} - Statistical Model used to create score
\item \code{score} - Score Label
\item \code{data_inputs} - Data inputs for workflow
\item \code{data_filters} - Filters applied in workflow
}
\item \code{meta_param}, with columns:
\itemize{
\item \code{workflowid} - Unique Workflow ID
\item \code{gsm_version} - GSM version of workflow
\item \code{param} - Parameter name
\item \code{index} - Parameter index
\item \code{default} - Default value for parameter at index
\item \code{configurable} - Is parameter configurable?
}
}
}

\examples{
# run with default testing data
snapshot <- Make_Snapshot()

}
