% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Snapshot.R
\name{Make_Snapshot}
\alias{Make_Snapshot}
\title{Make Snapshot - create and export Gizmo data model.}
\usage{
Make_Snapshot(
  lMeta = list(config_params = gsm::config_param, config_schedule =
    clindata::ctms_schedule, config_workflow = gsm::config_workflow, meta_params =
    gsm::meta_param, meta_site = clindata::ctms_site, meta_study = clindata::ctms_study,
    meta_workflow = gsm::meta_workflow),
  lData = list(dfSUBJ = clindata::rawplus_dm \%>\% filter(!is.na(timeontreatment)), dfAE
    = clindata::rawplus_ae, dfCONSENT = clindata::rawplus_consent, dfIE =
    clindata::rawplus_ie, dfLB = clindata::rawplus_lb, dfPD = clindata::rawplus_protdev,
    dfSTUDCOMP = clindata::rawplus_studcomp, dfSDRGCOMP = clindata::rawplus_sdrgcomp
    \%>\% filter(.data$datapagename == "Blinded Study Drug Completion")),
  lMapping = yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package =
    "gsm")),
  lAssessments = NULL,
  cPath = NULL,
  bQuiet = TRUE
)
}
\arguments{
\item{lMeta}{\code{list} a named list of data frames containing metadata, configuration, and workflow parameters for a given study. TODO: add details about expected lMeta input.}

\item{lData}{\code{list} a named list of domain level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{lMapping}{\code{list} a named list identifying the columns needed in each data domain.}

\item{lAssessments}{a named list of metadata defining how each assessment should be run. By default, \code{MakeAssessmentList()} imports YAML specifications from \code{inst/workflow}.}

\item{cPath}{\code{character} a character string indicating a working directory to save .csv files; the output of the snapshot.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list}, \code{lSnapshot}
}
\description{
Make Snapshot - create and export Gizmo data model.
}
\examples{

# run with default testing data
snapshot <- Make_Snapshot()


}
