% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Snapshot.R
\name{Make_Snapshot}
\alias{Make_Snapshot}
\title{Make Snapshot - create and export Gizmo data model.}
\usage{
Make_Snapshot(
  lMeta = list(config_param = clindata::config_param, config_workflow =
    clindata::config_workflow, meta_params = gsm::meta_param, meta_site =
    clindata::ctms_site, meta_study = clindata::ctms_study, meta_workflow =
    gsm::meta_workflow),
  lData = list(dfSUBJ = clindata::rawplus_dm, dfAE = clindata::rawplus_ae, dfPD =
    clindata::rawplus_protdev, dfCONSENT = clindata::rawplus_consent, dfIE =
    clindata::rawplus_ie, dfLB = clindata::rawplus_lb, dfSTUDCOMP =
    clindata::rawplus_studcomp, dfSDRGCOMP = clindata::rawplus_sdrgcomp \%>\%
    filter(.data$datapagename == "Blinded Study Drug Completion"), dfDATACHG =
    clindata::edc_data_change_rate, dfDATAENT = clindata::edc_data_entry_lag, dfQUERY =
    clindata::edc_queries, dfENROLL = clindata::rawplus_enroll),
  lMapping = c(yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package =
    "gsm")), yaml::read_yaml(system.file("mappings", "mapping_adam.yaml", package =
    "gsm")), yaml::read_yaml(system.file("mappings", "mapping_edc.yaml", package =
    "gsm"))),
  lAssessments = NULL,
  bUpdateParams = FALSE,
  cPath = NULL,
  bQuiet = TRUE,
  bFlowchart = FALSE
)
}
\arguments{
\item{lMeta}{\code{list} a named list of data frames containing metadata, configuration, and workflow parameters for a given study.
See the Data Model Vignette - Appendix 2 - Data Model Specifications for detailed specifications.}

\item{lData}{\code{list} a named list of domain-level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name of the column. Default: package-defined mapping for raw+.}

\item{lAssessments}{\code{list} a named list of metadata defining how each assessment should be run. By default, \code{MakeWorkflowList()} imports YAML specifications from \code{inst/workflow}.}

\item{bUpdateParams}{\code{logical} if \code{TRUE}, configurable parameters found in \code{lMeta$config_param} will overwrite the default values in \code{lMeta$meta_params}. Default: \code{FALSE}.}

\item{cPath}{\code{character} a character string indicating a working directory to save .csv files; the output of the snapshot.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}

\item{bFlowchart}{\code{logical} Create flowchart to show data pipeline? Default: \code{FALSE}}
}
\value{
\code{list} \code{lSnapshot}, a named list with a data.frame for each component of the {gsm} data model.
\itemize{
\item \code{status_study}
\item \code{status_site}
\item \code{status_workflow}
\item \code{status_param}
\item \code{results_summary}
\item \code{results_bounds}
\item \code{meta_workflow}
\item \code{meta_param}
}
}
\description{
\code{Make_Snapshot()} ingests data from a variety of sources, and runs KRIs and/or QTLs based on the \code{list} provided in \code{lAssessments}. The output of \code{Make_Snapshot()} is used as the input data model
for the Gismo web application.

For more context about the inputs and outputs of \code{Make_Snapshot()}, refer to the \href{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}{GSM Data Pipeline Vignette}, specifically
Appendix 2 - Data Model Specifications
}
\section{Return - Data Model Description}{
A named \code{list} of data.frames:
\itemize{
\item \code{status_study}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{enrolled_sites} - # of enrolled sites
\item \code{enrolled_participants} - # of enrolled participants
\item \code{planned_sites} - # of planned sites
\item \code{planned_participants} - # of planned participants
\item \code{title} - Protocol title
\item \code{nickname} - Protocol nickname
\item \code{indication} - Indication
\item \code{ta} - Therapeutic Area
\item \code{phase} - Phase
\item \code{status} - Study Status
\item \code{fpfv} - First-patient first visit date
\item \code{lplv} - Last-patient last visit date
\item \code{rbm_flag} - Risk-based monitoring flag
}
\item \code{status_site}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{siteid} - Unique Site ID
\item \code{institution} - Institution Name
\item \code{status} - Site Status
\item \code{enrolled_participants} - # of enrolled participants
\item \code{start_date} - Site Activation Date
\item \code{city} - City
\item \code{state} - State
\item \code{country} - Country
\item \code{invname} - Investigator name
}
\item \code{status_workflow}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{workflowid} - Unique Workflow ID
\item \code{gsm_version} - GSM version of workflow
\item \code{active} - Is workflow active?
\item \code{status} - Did workflow run?
\item \code{notes} - Status notes
}
\item \code{status_param}, with columns:
\itemize{
\item \code{study_id} - Unique Study ID
\item \code{workflowid} - Unique Workflow ID
\item \code{gsm_version} - GSM version of workflow
\item \code{param} - Parameter name
\item \code{index} - Index value for parameter
\item \code{value} - Value for parameter at index
}
\item \code{status_schedule}, with columns:
\itemize{
\item \code{study_id} - Unique Study ID
\item \code{snapshot_date} - Snapshot Date
}
\item \code{results_summary}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{workflowid} - Unique Workflow ID
\item \code{groupid} - Unique Group ID (e.g. SiteID for KRI workflows)
\item \code{numerator} - Metric numerator
\item \code{denominator} - Metric denominator
\item \code{metric} - Metric value
\item \code{score} - Statistical Score
\item \code{flag} - Flag
}
\item \code{results_bounds}, with columns:
\itemize{
\item \code{studyid} - Unique Study ID
\item \code{workflowid} - Unique Workflow ID
\item \code{threshold} - Threshold
\item \code{numerator} - Y value
\item \code{denominator} - X value
\item \code{log_denominator} - Log X value
}
\item \code{meta_workflow}, with columns:
\itemize{
\item \code{workflowid} - Unique Workflow ID
\item \code{gsm_version} - GSM Version of workflow
\item \code{group} - Group for workflow - Site, Country, Region, or Study
\item \code{metric} - Metric Label
\item \code{numerator} - Numerator Label
\item \code{denominator} - Denominator Label
\item \code{outcome} - Metric Outcome Type (e.g. rate)
\item \code{model} - Statistical Model used to create score
\item \code{score} - Score Label
\item \code{data_inputs} - Data inputs for workflow
\item \code{data_filters} - Filters applied in workflow
}
\item \code{meta_param}, with columns:
\itemize{
\item \code{workflowid} - Unique Workflow ID
\item \code{gsm_version} - GSM version of workflow
\item \code{param} - Parameter name
\item \code{index} - Parameter index
\item \code{default} - Default value for parameter at index
\item \code{configurable} - Is parameter configurable?
}
}
}

\examples{
# run with default testing data
\dontrun{
snapshot <- Make_Snapshot()
}

}
