% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAnalysisDate.R
\name{MakeAnalysisDate}
\alias{MakeAnalysisDate}
\title{Make Analysis Date}
\usage{
MakeAnalysisDate(strAnalysisDate = NULL, bQuiet = FALSE)
}
\arguments{
\item{strAnalysisDate}{\code{character} Date that analysis was run. Formatted as YYYY-MM-DD.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{Date} A date in the format \code{YYYY-MM-DD}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
analysis_date <- MakeAnalysisDate()

}
