% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeAssessmentList.R
\name{MakeAssessmentList}
\alias{MakeAssessmentList}
\title{Load assessments from a package/directory}
\usage{
MakeAssessmentList(
  strNames = NULL,
  strPath = "workflow",
  strPackage = "gsm",
  bRecursive = FALSE
)
}
\arguments{
\item{strNames}{\verb{array of character} List of workflows to include. NULL (the default) includes all workflows in the specified locations.}

\item{strPath}{\code{character} The location of assessment YAML files. If package is specified, function will look in \verb{/inst} folder.}

\item{strPackage}{\code{character} The name of the package with assessments.}

\item{bRecursive}{\code{boolean} Find files in nested folders? Default FALSE.}
}
\value{
\code{list} A list of assessments with workflow and parameter metadata.
}
\description{
Load assessments from a package/directory
}
\details{
\code{MakeAssessmentList()} is a utility function that creates a workflow mapping for assessments used in \code{Study_Assess()}.
}
\examples{
MakeAssessmentList(strPath = "workflow", strPackage = "gsm")

}
