% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-makeAssessmentList.R
\name{makeAssessmentList}
\alias{makeAssessmentList}
\title{Load assessments from a package/directory}
\usage{
makeAssessmentList(path = "assessments", package = "gsm")
}
\arguments{
\item{lData}{list of data}

\item{lMapping}{mapping}

\item{lAssessments}{assessments}

\item{strPopFlags}{filter demog data?}

\item{lTags}{tags}
}
\value{
A list containing: dataChecks and results
}
\description{
Load assessments from a package/directory
}
\details{
Coming soon
}
\examples{
 NULL

}
