% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeAssessmentList.R
\name{MakeAssessmentList}
\alias{MakeAssessmentList}
\title{Load assessments from a package/directory}
\usage{
MakeAssessmentList(path = "assessments", package = "gsm")
}
\arguments{
\item{path}{\code{character} The location of assessment YAML files. If package is specified, function will look in \verb{/inst} folder.}

\item{package}{\code{character} package with assessments}
}
\value{
\code{list} A list of assessments with workflow and parameter metadata.
}
\description{
Load assessments from a package/directory
}
\details{
\code{MakeAssessmentList()} is a utility function that creates a workflow mapping for assessments used in \code{Study_Assess()}.
}
\examples{
MakeAssessmentList(path = "assessments", package = "gsm")

}
