% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeBounds.R
\name{MakeBounds}
\alias{MakeBounds}
\title{Calculate Bounds for display in scatterplots}
\usage{
MakeBounds(dfResults, dfMetrics, strMetrics = unique(dfResults$MetricID))
}
\arguments{
\item{dfResults}{Results data.}

\item{dfMetrics}{Metric metadata including thresholds and analysis type.}

\item{strMetrics}{Character vector of \code{MetricID}s to include in \code{dfBounds}.
All unique values from \code{dfResults$MetricID} used by default.}
}
\value{
A data frame.
}
\description{
Used to calculate bounds across a set of metrics/
}
\examples{
dfBounds <- MakeBounds(
  dfResults = reportingResults, dfMetrics = reportingMetrics
)

}
