% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeBounds.R
\name{MakeBounds}
\alias{MakeBounds}
\title{Calculate Bounds for display in scatterplots}
\usage{
MakeBounds(
  dfResults,
  dfMetrics,
  strMetrics = unique(dfResults$MetricID),
  dSnapshotDate = unique(dfResults$SnapshotDate),
  strStudyID = unique(dfResults$StudyID)
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[=BindResults]{BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID}, \code{SnapshotDate}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and reporting. Created by passing an \code{lWorkflow} object to \code{\link[=MakeMetric]{MakeMetric()}}. Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric}, \code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{Threshold}. For more details see the Data Model vignette: \code{vignette("DataModel", package = "gsm")}.}

\item{strMetrics}{Character vector of \code{MetricID}s to include in \code{dfBounds}.
All unique values from \code{dfResults$MetricID} used by default.}

\item{dSnapshotDate}{Snapshot date. Uses \code{dfResults$SnapshotDate} by default.
If more than one snapshot date is found in \code{dfResults}, a warning is thrown
and this function returns \code{NULL}.}

\item{strStudyID}{Study ID. Uses \code{dfResults$StudyID} by default. If more than
one snapshot date is found in \code{dfResults}, a warning is thrown and this
function returns \code{NULL}.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Calculate bounds across a set of metrics.
}
\examples{
dfBounds <- MakeBounds(
  dfResults = reportingResults,
  dfMetrics = reportingMetrics
)

}
