% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeCharts.R
\name{MakeCharts}
\alias{MakeCharts}
\title{Helper function create charts for multiple metrics}
\usage{
MakeCharts(dfSummary, dfBounds, dfGroups, dfMetrics)
}
\arguments{
\item{data}{The input data frame.}

\item{strGroupLevel}{A string specifying the group type.}

\item{strGroupCols}{A string specifying the group columns.}
}
\value{
A long format data frame.
}
\description{
Helper function create charts for multiple metrics
}
\examples{
df <- data.frame(GroupID = c(1, 2, 3), Param1 = c(10, 20, 30), Param2 = c(100, 200, 300))
MakeLongMeta(df, "GroupID")

}
