% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeCharts.R
\name{MakeCharts}
\alias{MakeCharts}
\title{Helper function create charts for multiple metrics}
\usage{
MakeCharts(dfResults, dfBounds, dfGroups, dfMetrics)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[=BindResults]{BindResults()}}.}

\item{dfBounds}{A data frame with the bounds of the simulation.}

\item{dfGroups}{A data frame with the groups of the simulation.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and reporting. Created by passing an \code{lWorkflow} object to \code{\link[=MakeMetric]{MakeMetric()}}.}
}
\value{
A list of charts for each metric.
}
\description{
Helper function create charts for multiple metrics
}
