% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeKRICharts.R
\name{MakeKRICharts}
\alias{MakeKRICharts}
\title{MakeKRICharts Function}
\usage{
MakeKRICharts(lData = NULL, lStackedSnapshots = NULL)
}
\arguments{
\item{lData}{\code{list} A list containing two data frames - dfSummary and dfBounds.}

\item{lStackedSnapshots}{\code{list} A list returned by \code{\link[=Make_Snapshot]{Make_Snapshot()}} containing flat files for snapshot/assessment data.}
}
\value{
A list (lCharts) containing three charts - scatterJS, barMetricJS, and barScoreJS.
}
\description{
The function MakeKRICharts creates three different types of charts (scatter plot and two bar charts) using the gsm package. It takes two inputs: lData and dfConfig, and returns a list of the created charts.
}
