% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeKRICharts.R
\name{MakeKRICharts}
\alias{MakeKRICharts}
\title{MakeKRICharts Function}
\usage{
MakeKRICharts(dfSummary, dfBounds, lLabels = NULL, lStackedSnapshots = NULL)
}
\arguments{
\item{dfSummary}{\code{data.frame} A data.frame returned by \code{\link[=Summarize]{Summarize()}}.}

\item{dfBounds}{\code{data.frame}, A data.frame returned by \code{\link[=Analyze_NormalApprox_PredictBounds]{Analyze_NormalApprox_PredictBounds()}} or \code{\link[=Analyze_Poisson_PredictBounds]{Analyze_Poisson_PredictBounds()}}}

\item{lLabels}{\code{list} Workflow metadata. See \link{meta_workflow}.}

\item{lStackedSnapshots}{\code{list} A list returned by \code{\link[=Make_Snapshot]{Make_Snapshot()}} containing flat files for snapshot/assessment data.}
}
\value{
A list (lCharts) containing three charts - scatterJS, barMetricJS, and barScoreJS.
}
\description{
The function MakeKRICharts creates three different types of charts (scatter plot and two bar charts) using the gsm package.
}
