% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{MakeKRIGlossary}
\alias{MakeKRIGlossary}
\title{Create KRI metadata table in KRIReport.Rmd}
\usage{
MakeKRIGlossary(strWorkflowIDs = NULL, dfMetaWorkflow = gsm::meta_workflow)
}
\arguments{
\item{dfMetaWorkflow}{\code{data.frame} Workflow metadata from \code{params} within \code{KRIReport.Rmd}}
}
\description{
Create KRI metadata table in KRIReport.Rmd
}
\keyword{internal}
