% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeMetric.R
\name{MakeMetric}
\alias{MakeMetric}
\title{Parse metadata from workflows to a data frame}
\usage{
MakeMetric(lWorkflows)
}
\arguments{
\item{lWorkflows}{A list of workflows, like the one returned by
\code{\link[=MakeWorkflowList]{MakeWorkflowList()}}.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Used to format metric metadata (\code{dfMetrics}) for use in charts and reports.
This function takes a list of workflows and returns a data frame with one row
per \code{MetricID}.
}
\examples{
lWorkflows <- MakeWorkflowList(strPath = "./inst/workflow/2_metrics", strNames = "kri")
dfMetrics <- MakeMetric(lWorkflows)

}
