% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Helpers.R
\name{MakeOverviewMessage}
\alias{MakeOverviewMessage}
\title{Create message describing study summary for Report}
\usage{
MakeOverviewMessage(report, status_study, overview_raw_table, red_kris)
}
\arguments{
\item{report}{\code{string} type of report being run}

\item{status_study}{\code{data.frame} the snapshot status study output created with \code{Make_Snapshot()$lSnapshot$status_study}}

\item{overview_raw_table}{\code{data.frame} non interactive output of \code{Overview_Table()} for the relevant report.}

\item{red_kris}{\code{string} a string or number containing the count of red flags in kri's}
}
\description{
Create message describing study summary for Report
}
\keyword{internal}
