% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeParamLabels.R
\name{MakeParamLabels}
\alias{MakeParamLabels}
\alias{MakeParamLabelsList}
\title{Create Labels for Parameters}
\usage{
MakeParamLabels(dfGroups, lParamLabels = NULL)

MakeParamLabelsList(chrParams, lParamLabels = NULL)
}
\arguments{
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. Created by
passing CTMS site and study data to \code{\link[=MakeLongMeta]{MakeLongMeta()}}. Expected columns:
\code{GroupID}, \code{GroupLevel}, \code{Param}, \code{Value}.}

\item{lParamLabels}{\code{list} Labels for parameters, with the parameters as
names, and the label as value.}

\item{chrParams}{A character vector of parameters, or a list that can be
coerced to a character vector.}
}
\value{
\code{dfGroups} with an added \code{Label} column, or a list of labeled
parameters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Convert a vector of parameters to labels in Title Case. \code{MakeParamLabels}
adds a \code{Labels} column to a \code{data.frame} that has a \code{Params} column (such
as \code{dfGroups}), while \code{MakeParamLabelsList} returns just the list of named
parameters.
}
\examples{
head(gsm::reportingGroups)
MakeParamLabels(head(gsm::reportingGroups))
MakeParamLabels(
  head(gsm::reportingGroups),
  list(ParticipantCount = "Number of Participants")
)
MakeParamLabelsList(head(gsm::reportingGroups$Param))

}
