% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeQTLCharts.R
\name{MakeQTLCharts}
\alias{MakeQTLCharts}
\title{MakeQTLCharts}
\usage{
MakeQTLCharts(strQtlName, dfResults, dfParam, dfAnalysis, lLabels)
}
\arguments{
\item{strQtlName}{\code{character} Name of QTL workflow, e.g., "qtl0004".}

\item{dfResults}{\code{data.frame} Longitudinal data, typically \code{rpt_site_kri_details} from \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}

\item{dfParam}{\code{data.frame} Longitudinal parameter/configuration data, typically \code{rpt_kri_threshold_param} from \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}

\item{dfAnalysis}{\code{data.frame} Longitudinal QTL analysis results, typically \code{rpt_qtl_analysis} from  \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}

\item{lLabels}{\code{list} Longitudinal workflow/metadata, typically \code{rpt_kri_details} from \code{\link[=Make_Snapshot]{Make_Snapshot()}}.}
}
\value{
\code{list} of HTML widgets.
}
\description{
Make cross-sectional timeseries charts in \verb{[gsm::Make_Snapshot()]}.
}
