% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeResultsSummary.R
\name{MakeResultsSummary}
\alias{MakeResultsSummary}
\title{Make \code{results_summary}}
\usage{
MakeResultsSummary(lResults, dfConfigWorkflow)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{\link[=Study_Assess]{Study_Assess()}}.}

\item{dfConfigWorkflow}{\code{data.frame} Workflow configuration data.}
}
\value{
\code{data.frame} Stacked results (a la \code{dplyr::bind_rows()}) of the \code{dfSummary} data.frames, from the
output of \code{\link[=Study_Assess]{Study_Assess()}}
}
\description{
Make \code{results_summary}
}
\examples{
\dontrun{
study <- Study_Assess()
results_summary <- MakeResultsSummary(study, gsm::config_workflow)
}

}
