% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeSnapshotHelpers.R
\name{MakeRptSiteDetails}
\alias{MakeRptSiteDetails}
\title{Create rpt_site_details output for \code{Make_Snapshot()}}
\usage{
MakeRptSiteDetails(lResults, status_site)
}
\arguments{
\item{lResults}{\code{list} the output from \code{Study_Assess()}}

\item{status_site}{\code{data.frame} the output from \code{Site_Map_Raw()}}
}
\description{
Create rpt_site_details output for \code{Make_Snapshot()}
}
\keyword{internal}
