% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeStratifiedAssessment.R
\name{MakeStratifiedAssessment}
\alias{MakeStratifiedAssessment}
\title{Create multiple Assessment workflows for a stratified assessment}
\usage{
MakeStratifiedAssessment(lAssessment, lMapping, lData)
}
\arguments{
\item{lAssessment}{a named list of metadata defining how an assessment should be run.}

\item{lMapping}{a named list identifying the columns needed in each data domain.}

\item{lData}{a named list of domain level data frames.}
}
\value{
\code{list} A list of assessments for each specified strata
}
\description{
Create multiple Assessment workflows for a stratified assessment
}
\examples{

StratifiedAE <- MakeStratifiedAssessment(
   lData=list(
     dfSUBJ = clindata::rawplus_subj,
     dfAE = clindata::rawplus_ae
   ), 
   lMapping=clindata::mapping_rawplus,
   lAssessment= MakeAssessmentList()$aeGrade
 )

}
