% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeStratifiedAssessment.R
\name{MakeStratifiedAssessment}
\alias{MakeStratifiedAssessment}
\title{Create multiple workflows from a single stratified workflow}
\usage{
MakeStratifiedAssessment(lWorkflow, lData, lMapping, bQuiet = TRUE)
}
\arguments{
\item{lWorkflow}{\code{list} A named list of metadata defining how an workflow should be run.}

\item{lData}{\code{list} A named list of domain-level data frames.}

\item{lMapping}{\code{list} A named list identifying the columns needed in each data domain.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} A list of workflows for each specified strata
}
\description{
Create multiple workflows from a single stratified workflow
}
\examples{
\dontrun{
lMapping <- yaml::read_yaml(
  system.file("mappings", "mapping_rawplus.yaml", package = "gsm")
)
lAssessmentList <- MakeAssessmentList()

# Adverse events by grade
StratifiedAE <- MakeStratifiedAssessment(
  lData = list(
    dfSUBJ = clindata::rawplus_subj,
    dfAE = clindata::rawplus_ae
  ),
  lMapping = lMapping,
  lWorkflow = lAssessmentList$aeGrade
)

StratifiedAEResult <- StratifiedAE \%>\%
  purrr::map(~ .x \%>\%
    RunAssessment(
      lData = list(
        dfSUBJ = clindata::rawplus_subj,
        dfAE = clindata::rawplus_ae
      ),
      lMapping = lMapping
    ))

# Protocol deviations by PD category
StratifiedPD <- MakeStratifiedAssessment(
  lData = list(
    dfSUBJ = clindata::rawplus_subj,
    dfPD = clindata::rawplus_pd
  ),
  lMapping = lMapping,
  lWorkflow = MakeAssessmentList()$pdCategory
)

StratifiedPDResult <- StratifiedPD \%>\%
  purrr::map(~ .x \%>\%
    RunAssessment(
      lData = list(
        dfSUBJ = clindata::rawplus_subj,
        dfPD = clindata::rawplus_pd
      ),
      lMapping = lMapping
    ))

# Labs by lab category
StratifiedLB <- MakeStratifiedAssessment(
  lData = list(
    dfSUBJ = clindata::rawplus_subj,
    dfLB = clindata::rawplus_lb
  ),
  lMapping = lMapping,
  lWorkflow = lAssessmentList$lbCategory
)

StratifiedLBResult <- StratifiedLB \%>\%
  purrr::map(~ .x \%>\%
    RunAssessment(
      lData = list(
        dfSUBJ = clindata::rawplus_subj,
        dfLB = clindata::rawplus_lb
      ),
      lMapping = lMapping
    ))
}

}
