% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeStudyInfo.R
\name{MakeStudyInfo}
\alias{MakeStudyInfo}
\title{Generate a study information data.frame for use in reports}
\usage{
MakeStudyInfo(lStudy, lStudyLabels = NULL)
}
\arguments{
\item{lStudy}{A list containing study information.}

\item{lStudyLabels}{A list containing study labels. Default is NULL.}
}
\value{
A data.frame containing study metadata.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generate a study info table summarizing study metadata.
}
\examples{
lStudy <- list(
  StudyID = "Unique Study ID",
  protocol_title = "Study Title",
  nickname = "Nickname",
  status = "Ongoing",
  phase = "Phase 1",
  therapeutic_area = "Therapeutic Area",
  protocol_indication = "Indication",
)

}
