% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeStudyInfo.R
\name{MakeStudyInfo}
\alias{MakeStudyInfo}
\title{Generate a study information data.frame for use in reports}
\usage{
MakeStudyInfo(dfGroups, lStudyLabels = NULL, lStudy = deprecated())
}
\arguments{
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. Created by
passing CTMS site and study data to \code{\link[=MakeLongMeta]{MakeLongMeta()}}. Expected columns:
\code{GroupID}, \code{GroupLevel}, \code{Param}, \code{Value}.}

\item{lStudyLabels}{\code{list} A list containing study labels. Default is NULL.}

\item{lStudy}{\code{deprecated} Study information as a named list.}
}
\value{
A data.frame containing study metadata.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Generate a study info table summarizing study metadata.
}
\examples{
MakeStudyInfo(gsm::reportingGroups)
MakeStudyInfo(gsm::reportingGroups, list(SiteCount = "# Sites"))

}
