% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{MakeStudyStatusTable}
\alias{MakeStudyStatusTable}
\title{Create Status Study table in KRIReport.Rmd}
\usage{
MakeStudyStatusTable(status_study)
}
\arguments{
\item{status_study}{\code{data.frame} from \code{params} within \code{KRIReport.Rmd}}
}
\description{
Create Status Study table in KRIReport.Rmd
}
\keyword{internal}
