% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReportHelpers.R
\name{MakeSummaryTable}
\alias{MakeSummaryTable}
\title{Create Summary table in KRIReport.Rmd for each KRI}
\usage{
MakeSummaryTable(lAssessment, dfSite = NULL)
}
\arguments{
\item{lAssessment}{\code{list} List of KRI assessments from \code{params} within \code{KRIReport.Rmd}.}

\item{dfSite}{\code{data.frame} Optional site-level metadata.}
}
\description{
Create Summary table in KRIReport.Rmd for each KRI
}
\keyword{internal}
