% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MakeWideGroups.R
\name{MakeWideGroups}
\alias{MakeWideGroups}
\title{Helper function to compile "wide" group metadata}
\usage{
MakeWideGroups(dfGroups, strGroupLevel)
}
\arguments{
<<<<<<< HEAD
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. Created by passing CTMS site and study data to \code{\link[gsm.mapping:MakeLongMeta]{gsm.mapping::MakeLongMeta()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Param}, \code{Value}.}
=======
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. Created by
passing CTMS site and study data to \code{\link[=MakeLongMeta]{MakeLongMeta()}}. Expected columns:
\code{GroupID}, \code{GroupLevel}, \code{Param}, \code{Value}.}
>>>>>>> dev

\item{strGroupLevel}{A string specifying the group level; used to filter
dfGroups$GroupLevel.}
}
\value{
A long format data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Used to convert metadata dictionary (dfGroups) to a wide format data frame for
use in charts and reports.
}
\examples{
MakeWideGroups(reportingGroups, "Site")

}
