% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-MergeSubjects.R
\name{MergeSubjects}
\alias{MergeSubjects}
\title{Merge Domain data with subject-level data}
\usage{
MergeSubjects(
  dfDomain,
  dfSUBJ,
  strIDCol = "SubjectID",
  vFillZero = NULL,
  bQuiet = TRUE
)
}
\arguments{
\item{dfDomain}{\code{data.frame} Subject-level domain data containing one record per participant.}

\item{dfSUBJ}{\code{data.frame} Subject level data often using ADSL-like data. Should include one record per participant for each participant included in the analysis population (all other participants should be dropped before calling mergeSubjects)}

\item{strIDCol}{\code{character} Vector of length 1. The name of ID Column - Default: \code{SubjectID}}

\item{vFillZero}{\code{vector} Column names from \code{dfDomain} to fill with zeros when no matching row is found for an ID in \code{dfSUBJ}.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} with one record per strIDCol.
}
\description{
\code{MergeSubjects} is a helper function used within mapping functions to join subject-level data with domain-level data, with the ability to impute
zeros when necessary.
}
\examples{
MergeSubjects(
  dfDomain = clindata::rawplus_consent,
  dfSUBJ = clindata::rawplus_dm,
  strIDCol = "subjid"
)

}
