% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mergeSubjects.R
\name{mergeSubjects}
\alias{mergeSubjects}
\title{Merge Domain data with subject-level data shell}
\usage{
mergeSubjects(
  dfSubjects,
  dfDomain,
  strIDCol = "SubjectID",
  vFillZero = NULL,
  bQuiet = FALSE
)
}
\arguments{
\item{dfDomain}{}

\item{strIDCol}{}

\item{dfRDSL}{}

\item{bFillZero}{Columns from dfDomain to fill with zeros when no matching row is found in for an ID in dfSubject}
}
\value{
data set with one record per IDCol
}
\description{
Merge Domain data with subject-level data shell
}
\examples{
NULL

}
