% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{meta_param}
\alias{meta_param}
\title{Workflow Parameterization Metadata}
\format{
A data frame with 8 rows and 12 variables:
\describe{
\item{workflowid}{Workflow ID}
\item{gsm_version}{Version of \code{gsm} that the workflow will be run on.}
\item{param}{Name of function parameter.}
\item{index}{Index that describes the position of the parameter values.}
\item{default}{Default parameter.}
\item{configurable}{Boolean denoting whether the parameter can be configured to use a different argument.}
}
}
\source{
\url{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}
}
\usage{
meta_param
}
\description{
A dataset containing the the parameters available in GSM workflows. Each row describes a configurable parameter for a workflow defined in \code{inst/workflows}.
}
\keyword{datasets}
