% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{meta_workflow}
\alias{meta_workflow}
\title{Workflow Metadata}
\format{
A data frame with 8 rows and 12 variables:
\describe{
\item{workflowid}{Workflow ID}
\item{gsm_version}{Version of \code{gsm} that the workflow will be run on.}
\item{group}{Description of grouping variable: \code{Site}, \code{Study}, or \code{Country}.}
\item{abbreviation}{Metric abbreviation - 3 characters max}
\item{metric}{Description of the assessment.}
\item{numerator}{The numerator used to calculate the KRI/QTL.}
\item{denominator}{The denominator used to calculate the KRI/QTL.}
\item{outcome}{Description of the outcome. Either \code{Rate} or \code{Percent}.}
\item{model}{Statistical model that is used for the assessment.}
\item{score}{Description of the statistical score that is derived. E.g., "Residual from Poisson Model" or "P-value from Fisher's Exact Test".}
\item{data_inputs}{Description of one or more data inputs needed to create the \code{dfInput} data.frame that is passed to the assessment function.}
\item{data_filters}{Description of parameters to use to filter when creating input data (\code{dfInput})}
}
}
\source{
\url{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}
}
\usage{
meta_workflow
}
\description{
A dataset containing the metadata describing the workflows available in GSM. Each row describes a workflow defined in \code{inst/workflows}.
}
\keyword{datasets}
