% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overview_Table.R
\name{Overview_Table}
\alias{Overview_Table}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
Overview_Table(lAssessments, bInteractive = TRUE)
}
\arguments{
\item{lAssessments}{\code{list} The output of running \code{\link[=Study_Assess]{Study_Assess()}}}

\item{bInteractive}{\code{logical} Display interactive widget? Default: \code{TRUE}.}
}
\description{
Overview Table - Create summary of red and amber KRIs for a study.
}
\examples{
\dontrun{
lAssessments <- Study_Assess()
Overview_Table(lAssessments)
}

}
