% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-parse_data_mapping.R
\name{parse_data_mapping}
\alias{parse_data_mapping}
\title{Parse Data Mapping}
\usage{
parse_data_mapping(content = NULL, file = NULL)
}
\arguments{
\item{content}{\code{list} data mapping}

\item{file}{\code{character} file path of .yaml file}
}
\description{
Transform nested data mapping to tabular structure
}
