% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-parse_data_spec.R
\name{parse_data_spec}
\alias{parse_data_spec}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
parse_data_spec(content = NULL, file = NULL)
}
\arguments{
\item{content}{\code{list} data specification}

\item{file}{\code{character} file path of .yaml file}
}
\description{
Parse Data Specification
}
\details{
Transform nested data specification to tabular structure for use in documentation.
}
