% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ParseThreshold.R
\name{ParseThreshold}
\alias{ParseThreshold}
\title{Parse a string into a numeric vector}
\usage{
ParseThreshold(strThreshold, bSort = TRUE)
}
\arguments{
\item{strThreshold}{\code{character} A comma-separated string of numeric values.}

\item{bSort}{\code{logical} Sort thresholds in ascending order? Default: \code{TRUE}.}
}
\value{
A numeric vector containing the parsed values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function takes a comma-separated string and parses it into a numeric
vector. It checks if all values in the string are numeric and returns the
parsed vector. If any value is not numeric, it throws an error.
}
\examples{
# standard thresholds
ParseThreshold("-3,-2,2,3")

# by default thresholds will be sorted in ascending order
ParseThreshold("3,2,-2,-3")

optionally disable the sort
ParseThreshold("0.9,0.85", bSort = FALSE)

}
