% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ParseThreshold.R
\name{ParseThreshold}
\alias{ParseThreshold}
\title{Parse a string into a numeric vector}
\usage{
ParseThreshold(strThreshold)
}
\arguments{
\item{strThreshold}{A comma-separated string of numeric values.}
}
\value{
A numeric vector containing the parsed values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function takes a comma-separated string and parses it into a numeric
vector. It checks if all values in the string are numeric and returns the
parsed vector. If any value is not numeric, it throws an error.
}
\examples{
ParseThreshold("1,2,3,4")
}
