% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ParseWarnings.R
\name{ParseWarnings}
\alias{ParseWarnings}
\title{{experimental} Parse warnings from the result of Study_Assess.}
\usage{
ParseWarnings(lResults)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{gsm::Study_Assess()}.}
}
\value{
\code{data.frame} With columns \code{notes} and \code{workflowid}.
}
\description{
\code{ParseWarnings} is used inside of \code{Make_Snapshot()} to summarize any issues with data needed to run a KRI or QTL. If there are any warnings for any
workflow run via \code{Make_Snapshot()}, they are appended to the \code{notes} column in the  \code{status_workflow} data.frame, which is included in the output
of \code{Make_Snapshot()}. If there are no warnings, the \code{notes} column contains \code{NA} for all KRIs/QTLs.
}
\examples{
# Set all subjid to NA to create warnings
lData <- list(
  dfAE = clindata::rawplus_ae \%>\%
    dplyr::mutate(subjid = NA),
  dfSUBJ = clindata::rawplus_dm
)

study <- Study_Assess(lData = lData)

warnings <- ParseWarnings(study)

}
