% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ParseWarnings.R
\name{ParseWarnings}
\alias{ParseWarnings}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ParseWarnings(lResults)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{\link[=Study_Assess]{Study_Assess()}}.}
}
\value{
\code{data.frame} With columns \code{workflowid}, \code{status}, and \code{notes}.
}
\description{
\code{ParseWarnings} is used inside of \code{\link[=Make_Snapshot]{Make_Snapshot()}} to summarize any issues with data needed to run a KRI or QTL. If there are any warnings for any
workflow run via \code{Make_Snapshot()}, they are appended to the \code{notes} column in the  \code{status_workflow} data.frame, which is included in the output
of \code{Make_Snapshot()}. If there are no warnings, the \code{notes} column contains \code{NA} for all KRIs/QTLs.
}
\details{
Parse warnings from the result of \code{\link[=Study_Assess]{Study_Assess()}}.
}
\examples{
# Set all subjid to NA to create warnings

\dontrun{
lData <- list(
  dfAE = clindata::rawplus_ae \%>\%
    dplyr::mutate(subjid = NA),
  dfSUBJ = clindata::rawplus_dm
)

study <- Study_Assess(lData = lData)

warnings <- ParseWarnings(study)
}

}
