% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ParseWarnings.R
\name{ParseWarnings}
\alias{ParseWarnings}
\title{Parse warnings from the result of Study_Assess.}
\usage{
ParseWarnings(lResults)
}
\arguments{
\item{lResults}{\code{list} List returned from \code{gsm::Study_Assess()}.}
}
\value{
\code{data.frame} With columns \code{notes} and \code{workflowid}.
}
\description{
Parse warnings from the result of Study_Assess.
}
\examples{

# Set all subjid to NA to create warnings
lData <- list(
         dfAE = clindata::rawplus_ae \%>\%
           dplyr::mutate(subjid = NA),
         dfSUBJ = clindata::rawplus_dm
         )

study <- Study_Assess(lData = lData)

warnings <- ParseWarnings(study)

}
