% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Assess.R
\name{PD_Assess}
\alias{PD_Assess}
\title{Protocol Deviation Assessment using Poisson Regression}
\usage{
PD_Assess(
  dfInput,
  lThreshold = NULL,
  nCutoff = 1,
  cLabel = "",
  bDataList = FALSE
)
}
\arguments{
\item{dfInput}{input data}

\item{lThreshold}{list of threshold values}

\item{nCutoff}{optional parameter to control the auto-thresholding}

\item{cLabel}{Assessment label}

\item{bDataList}{Should all assessment datasets be returned as a list? If False (the default), only the finding data frame is returned}
}
\value{
Finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
Protocol Deviation Assessment using Poisson Regression
}
