% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Assess.R
\name{PD_Assess}
\alias{PD_Assess}
\title{Protocol Deviation Assessment using Poisson Regression}
\usage{
PD_Assess(
  dfInput,
  vThreshold = NULL,
  nCutoff = 1,
  cLabel = "",
  cMethod = "poisson",
  cExposureCol = "TimeOnStudy",
  bDataList = FALSE
)
}
\arguments{
\item{dfInput}{input data}

\item{vThreshold}{list of threshold values default c(-5,5) for method = "poisson", c(.0001,NA) for method = Wilcoxon}

\item{nCutoff}{optional parameter to control the auto-thresholding}

\item{cLabel}{Assessment label}

\item{cMethod}{valid methods are "poisson" (the default), or  "wilcoxon"}

\item{cExposureCol}{select exposure column, default = 'TimeOnStudy' if using TimeOnStudy function before PD_Map function.}

\item{bDataList}{Should all assessment datasets be returned as a list? If False (the default), only the finding data frame is returned}
}
\value{
Finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
Protocol Deviation Assessment using Poisson Regression
}
