% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Assess.R
\name{PD_Assess}
\alias{PD_Assess}
\title{Protocol Deviation Assessment}
\usage{
PD_Assess(
  dfInput,
  vThreshold = NULL,
  strMethod = "poisson",
  lTags = list(Assessment = "PD")
)
}
\arguments{
\item{dfInput}{input data with one record per person and the following required columns: SubjectID, SiteID, Count, Exposure, Rate.}

\item{vThreshold}{list of threshold values default c(-5,5) for method = "poisson", c(.0001,NA) for method = Wilcoxon.}

\item{strMethod}{valid methods are "poisson" (the default), or  "wilcoxon".}

\item{lTags}{named list of tags describing the assessment. \code{lTags} is returned as part of the assessment (\code{lAssess$lTags}) and each tag is added as columns in \code{lassess$dfSummary}. Default is \code{list(Assessment="PD")}.}
}
\value{
A list containing all data and metadata in the standard data pipeline (\code{dfInput}, \code{dfTransformed}, \code{dfAnalyzed}, \code{dfFlagged}, \code{dfSummary}, \code{strFunctionName},  \code{lParams} and \code{lTags}) is returned.
}
\description{
Protocol Deviation Assessment
}
\details{
The Protocol Deviation Assessment uses the standard GSM data pipeline (TODO add link to data vignette) to flag possible outliers. More details regarding the data pipeline and statistical methods are described below.
}
\section{Data Specification}{


The input data (\code{dfInput}) for the PD Assessment is typically created using \code{\link{PD_Map_Raw}} and should be one record per person with columns for:
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{Count} - Number of protocol deviation events
\item \code{Exposure} - Number of days of exposure
\item \code{Rate} - Rate of Exposure (Count / Exposure)
}

The Assessment
\itemize{
\item \code{\link{Transform_EventCount}} creates \code{dfTransformed}.
\item \code{\link{Analyze_Poisson}} or \code{\link{Analyze_Wilcoxon}} creates \code{dfAnalyzed}.
\item \code{\link{Flag}} creates \code{dfFlagged}.
\item \code{\link{Summarize}} creates \code{dfSummary}.
}
}

\section{Statistical Assumptions}{


A Poisson or Wilcoxon model is used to generate estimates and p-values for each site (as specified with the \code{strMethod} parameter). Those model outputs are then used to flag possible outliers using the thresholds specified in \code{vThreshold}. In the Poisson model, sites with an estimand less than -5 are flagged as -1 and greater than 5 are flagged as 1 by default. For Wilcoxon, sites with p-values less than 0.0001 are flagged by default.

See \code{\link{Analyze_Poisson}} and \code{\link{Analyze_Wilcoxon}} for additional details about the statistical methods and thier assumptions.
}

\examples{
 dfInput <- PD_Map_Raw(clindata::raw_protdev \%>\% dplyr::filter(SUBJID != ""),
                       clindata::rawplus_rdsl)

SafetyPD <- PD_Assess( dfInput )

SafetyPD_Wilk <- PD_Assess( dfInput, strMethod="wilcoxon")$dfSummary

}
