% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Assess.R
\name{PD_Assess}
\alias{PD_Assess}
\title{Protocol Deviation Assessment}
\usage{
PD_Assess(
  dfInput,
  vThreshold = NULL,
  strMethod = "poisson",
  lMapping = yaml::read_yaml(system.file("mappings", "PD_Assess.yaml", package = "gsm")),
  strGroup = "Site",
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data, a data frame with one record per subject.}

\item{vThreshold}{\code{numeric} Threshold specification, a vector of length 4 that defaults to
\code{c(-7, -5, 5, 7)} for a Poisson model (\code{strMethod = "poisson"}) and a vector of length 2
that defaults to \code{c(0.000895, 0.003059)} for a nominal assessment (\code{strMethod = "identity"}).}

\item{strMethod}{\code{character} Statistical method. Valid values:
\itemize{
\item \code{"poisson"} (default)
\item \code{"identity"}
}}

\item{lMapping}{Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column.}

\item{strGroup}{\code{character} Grouping variable. \code{"Site"} (the default) uses the column named in \code{mapping$strSiteCol}. Other valid options using the default mapping are \code{"Study"} and \code{"CustomGroup"}.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} \code{lData}, a named list with:
\itemize{
\item each data frame in the data pipeline
\itemize{
\item \code{dfTransformed}, returned by \code{\link[=Transform_Rate]{Transform_Rate()}}
\item \code{dfAnalyzed}, returned by \code{\link[=Analyze_Poisson]{Analyze_Poisson()}} or \code{\link[=Analyze_Identity]{Analyze_Identity()}}
\item \code{dfFlagged}, returned by \code{\link[=Flag_Poisson]{Flag_Poisson()}} or \code{\link[=Flag]{Flag()}}
\item \code{dfSummary}, returned by \code{\link[=Summarize]{Summarize()}}
\item \code{dfBounds}, returned by \code{\link[=Analyze_Poisson_PredictBounds]{Analyze_Poisson_PredictBounds()}} only when strMethod == "poisson"
}
\item \code{list} \code{lCharts}, a named list with:
\itemize{
\item \code{scatter}, a ggplot2 object returned by \code{\link[=Visualize_Scatter]{Visualize_Scatter()}} only when strMethod != "identity"
\item \code{barMetric}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}}
\item \code{barScore}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}}
}
\item \code{list} \code{lChecks}, a named list with:
\itemize{
\item \code{dfInput}, a named list returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{status}, a boolean returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{mapping}, a named list that is provided as an argument to the \code{lMapping} parameter in \code{\link[=PD_Assess]{PD_Assess()}}
\item \code{spec}, a named list used to define variable specifications
}
}
}
\description{
Evaluates protocol deviation (PD) rates to identify sites that may be over- or under-reporting PDs.
}
\details{
The PD Assessment uses the standard \href{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}{GSM data pipeline} to flag possible outliers. Additional details regarding the data pipeline and statistical
methods are described below.
}
\section{Data specification}{
\tabular{llllll}{
   \strong{Domain} \tab \strong{Column Key} \tab \strong{Default Value} \tab \strong{Required?} \tab \strong{Require Unique Values?} \tab \strong{Accept NA/Empty Values?} \cr
   dfInput \tab strIDCol \tab SubjectID \tab TRUE \tab TRUE \tab FALSE \cr
   dfInput \tab strCountCol \tab Count \tab TRUE \tab FALSE \tab FALSE \cr
   dfInput \tab strExposureCol \tab Exposure \tab TRUE \tab FALSE \tab TRUE \cr
   dfInput \tab strRateCol \tab Rate \tab TRUE \tab FALSE \tab TRUE \cr
   dfInput \tab strGroupCol \tab SiteID \tab TRUE \tab FALSE \tab FALSE \cr
}
}

\section{Statistical assumptions}{
A Poisson model is used to generate estimates and p-values for each site
(as specified with the \code{strMethod} parameter). Those model outputs are
then used to flag possible outliers using the thresholds specified in
\code{vThreshold}. Sites with an estimate less than -5 are flagged as -1 and
greater than 5 are flagged as 1 by default.

See \code{\link[=Analyze_Poisson]{Analyze_Poisson()}} for additional
details about the statistical methods and their assumptions.
}

\examples{
dfInput <- PD_Map_Raw()
pd_assessment_poisson <- PD_Assess(dfInput)

}
