% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Assess_Binary.R
\name{PD_Assess_Binary}
\alias{PD_Assess_Binary}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
PD_Assess_Binary(
  dfInput,
  vThreshold = NULL,
  strMethod = "NormalApprox",
  lMapping = yaml::read_yaml(system.file("mappings", "PD_Assess_Binary.yaml", package =
    "gsm")),
  strGroup = "Site",
  nMinDenominator = NULL,
  nConfLevel = NULL,
  bQuiet = TRUE
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data, a data frame with one record per subject.}

\item{vThreshold}{\code{numeric} Threshold specification, a numeric vector that defaults to \code{c(-3, -2, 2, 3)} for a Normal Approximation (\code{strMethod = "NormalApprox"}),
\code{c(-7, -5, 5, 7)} for a Poisson model (\code{strMethod = "Poisson"}), \code{c(0.000895, 0.003059)} for a nominal assessment (\code{strMethod = "Identity"}), and \code{c(0.01)} for a QTL (\code{strMethod = "QTL"}).}

\item{strMethod}{\code{character} Statistical method. Valid values:
\itemize{
\item \code{"NormalApprox"} (default)
\item \code{"Poisson"}
\item \code{"Identity"}
\item \code{"QTL"}
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column. Default: package-defined Protocol Deviation Assessment mapping.}

\item{strGroup}{\code{character} Grouping variable. \code{"Site"} (the default) uses the column named in \code{mapping$strSiteCol}. Other valid options using the default mapping are \code{"Study"}, \code{"Country"}, and \code{"CustomGroup"}.}

\item{nMinDenominator}{\code{numeric} Specifies the minimum denominator required to return a \code{score} and calculate a \code{flag}. Default: NULL}

\item{nConfLevel}{\code{numeric} Confidence level for QTL analysis.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} \code{lData}, a named list with:
\itemize{
\item each data frame in the data pipeline
\itemize{
\item \code{dfTransformed}, returned by \code{\link[=Transform_Rate]{Transform_Rate()}}
\item \code{dfAnalyzed}, returned by \code{\link[=Analyze_NormalApprox]{Analyze_NormalApprox()}}, \code{\link[=Analyze_Poisson]{Analyze_Poisson()}} or \code{\link[=Analyze_Identity]{Analyze_Identity()}}
\item \code{dfFlagged}, returned by \code{\link[=Flag_NormalApprox]{Flag_NormalApprox()}}, \code{\link[=Flag_Poisson]{Flag_Poisson()}}, or \code{\link[=Flag]{Flag()}}
\item \code{dfSummary}, returned by \code{\link[=Summarize]{Summarize()}}
\item \code{dfBounds}, returned by \code{\link[=Analyze_NormalApprox_PredictBounds]{Analyze_NormalApprox_PredictBounds()}} when \code{strMethod == "NormalApprox"} or \code{\link[=Analyze_Poisson_PredictBounds]{Analyze_Poisson_PredictBounds()}} when \code{strMethod == "Poisson"}
}
\item \code{list} \code{lCharts}, a named list with:
\itemize{
\item \code{scatter}, a ggplot2 object returned by \code{\link[=Visualize_Scatter]{Visualize_Scatter()}} only when strMethod != "Identity"
\item \code{barMetric}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}}
\item \code{barScore}, a ggplot2 object returned by \code{\link[=Visualize_Score]{Visualize_Score()}}
}
\item \code{list} \code{lChecks}, a named list with:
\itemize{
\item \code{dfInput}, a named list returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{status}, a boolean returned by \code{\link[=is_mapping_valid]{is_mapping_valid()}}
\item \code{mapping}, a named list that is provided as an argument to the \code{lMapping} parameter in \code{\link[=PD_Assess_Rate]{PD_Assess_Rate()}}
\item \code{spec}, a named list used to define variable specifications
}
}
}
\description{
Evaluates Protocol Deviation (PD) rates to identify sites that may be over- or under-reporting PDs.
}
\details{
Protocol Deviation Assessment (Binary)

The Protocol Deviation Assessment uses the standard \href{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}{GSM data pipeline} to flag possible outliers. Additional details regarding the data pipeline and statistical
methods are described below.
}
\section{Data specification}{
\tabular{llllll}{
   \strong{Domain} \tab \strong{Column Key} \tab \strong{Default Value} \tab \strong{Required?} \tab \strong{Require Unique Values?} \tab \strong{Accept NA/Empty Values?} \cr
   dfInput \tab strIDCol \tab SubjectID \tab TRUE \tab TRUE \tab FALSE \cr
   dfInput \tab strCountCol \tab Count \tab TRUE \tab FALSE \tab FALSE \cr
   dfInput \tab strGroupCol \tab SiteID \tab TRUE \tab FALSE \tab FALSE \cr
   dfInput \tab strExposureCol \tab  \tab FALSE \tab FALSE \tab TRUE \cr
   dfInput \tab strRateCol \tab  \tab FALSE \tab FALSE \tab TRUE \cr
}
}

\section{Statistical assumptions}{
The default function \code{Analyze_NormalApprox} applies funnel plots using
asymptotic limits based on normal approximation of binomial distribution
for the binary outcome, or normal approximation of Poisson distribution
for the rate outcome with volume (the sample sizes or total exposure of
the sites) to assess data quality and safety.

Alternatively, a Poisson model is used to generate estimates and
p-values for each site (as specified with the \code{strMethod} parameter).
Those model outputs are then used to flag possible outliers using the
thresholds specified in \code{vThreshold}. Sites with an estimate less than
-5 are flagged as -1 and greater than 5 are flagged as 1 by default.

See \code{\link[=Analyze_NormalApprox]{Analyze_NormalApprox()}} or
\code{\link[=Analyze_Poisson]{Analyze_Poisson()}} for additional
details about the statistical methods and their assumptions.
}

\examples{
dfInput <- PD_Map_Raw_Binary()
pd_assessment_NormalApprox <- PD_Assess_Binary(dfInput)
pd_assessment_poisson <- PD_Assess_Binary(dfInput, strMethod = "Poisson")
pd_assessment_identity <- PD_Assess_Binary(dfInput, strMethod = "Identity")

}
