% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Map_Raw.R
\name{PD_Map_Raw}
\alias{PD_Map_Raw}
\title{Protocol Deviation Assessment Mapping from Raw Data- Make Input Data}
\usage{
PD_Map_Raw(dfTos = NULL, dfPd = NULL)
}
\arguments{
\item{dfTos}{data.frame Time on Study data in format produced by \code{\link{TimeOnStudy}}}

\item{dfPd}{PD dataset with columns SUBJID and rows for each Protocol Deviation}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count, Rate, Unit
}
\description{
Convert from raw data format to needed input format for Safety Assessment
Requires the following raw datasets: subid, ex, pd
}
