% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Map_Raw.R
\name{PD_Map_Raw}
\alias{PD_Map_Raw}
\title{Protocol Deviation Assessment Mapping from Raw Data- Make Input Data}
\usage{
PD_Map_Raw(
  dfs = list(dfPD = clindata::rawplus_pd, dfSUBJ = clindata::rawplus_subj),
  lMapping = NULL,
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}

\item{dfPD}{PD dataset with required column SUBJID and rows for each Protocol Deviation.}

\item{dfSUBJ}{Subject-level Raw Data required columns: SubjectID, SiteID, value specified in strTimeOnStudyCol.}

\item{mapping}{List containing expected columns in each data set.}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count, Exposure, Rate.
}
\description{
Convert from raw data format to needed input format for Safety Assessment
}
\details{
This function combines raw PD data with exposure data calculated by clindata::TimeOnStudy to create the required input for \code{\link{PD_Assess}}.
}
\section{Data Specification}{


This function creates an input dataset for the Protocol Deviation (\code{\link{PD_Assess}}) by adding Protocol Deviation Counts to basic subject-level time on study data from \code{clindata::TimeOnStudy}.

The following columns are required:
\itemize{
\item \code{dfPD}
\itemize{
\item \code{SubjectID} - Unique subject ID
}
\item \code{dfSUBJ}
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item \code{TimeOnStudy} - Time on Study in days.
}
}
}

\examples{
dfInput <- PD_Map_Raw(dfPD = clindata::rawplus_pd, dfSUBJ = clindata::rawplus_subj)

}
