% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Map_Raw.R
\name{PD_Map_Raw}
\alias{PD_Map_Raw}
\title{Protocol Deviation Assessment Mapping from Raw Data- Make Input Data}
\usage{
PD_Map_Raw(dfPD = NULL, dfTOS = NULL)
}
\arguments{
\item{dfPD}{PD dataset with SUBJID column and rows for each Protocol Deviation}

\item{dfTOS}{Time on Study data.set in format produced by \code{\link{TimeOnStudy}}}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count, Exposure, Rate
}
\description{
Convert from raw data format to needed input format for Safety Assessment
Requires the following raw datasets: subid, ex, pd
}
