% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_Map_Raw.R
\name{PD_Map_Raw}
\alias{PD_Map_Raw}
\title{Protocol Deviation Assessment Mapping from Raw Data- Make Input Data}
\usage{
PD_Map_Raw(dfPD = NULL, dfRDSL = NULL, strExposureCol = "TimeOnStudy")
}
\arguments{
\item{dfPD}{PD dataset with SUBJID column and rows for each Protocol Deviation}

\item{dfRDSL}{Subject-level Raw Data (RDSL) required columns: SubjectID, SiteID, value specified in strExposureCol}

\item{strExposureColumn}{Name of exposure column. 'TimeOnStudy' by default}
}
\value{
Data frame with one record per person data frame with columns: SubjectID, SiteID, Count, Exposure, Rate
}
\description{
Convert from raw data format to needed input format for Safety Assessment
}
\details{
This function combines raw PD data with exposure data calculated by clindata::TimeOnStudy to create the required input for \code{\link{AE_Assess}}.
}
\section{Data Specification}{


This function creates an input dataset for the Protocol Deviation (\code{\link{PD_Assess}}) by adding Protocol Deviation Counts to basic subject-level time on study data from \code{clindata::TimeOnStudy}.

The following columns are required:
\itemize{
\item \code{dfPD}
\itemize{
\item \code{SUBJID} - Unique subject ID
}
\item \code{dfRDSL}
\itemize{
\item \code{SubjectID} - Unique subject ID
\item \code{SiteID} - Site ID
\item Value specified in strExposureCol - Time on Study in days; "TimeOnStudy" by default
}
}
}

\examples{
 dfInput <- PD_Map_Raw(clindata::raw_protdev,clindata::rawplus_rdsl)

}
