% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryRate_Map_Raw.R
\name{QueryRate_Map_Raw}
\alias{QueryRate_Map_Raw}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
QueryRate_Map_Raw(
  dfs = list(dfSUBJ = clindata::rawplus_dm, dfQUERY = clindata::edc_queries, dfDATACHG =
    clindata::edc_data_points),
  lMapping = yaml::read_yaml(system.file("mappings", "mapping_edc.yaml", package =
    "gsm")),
  bReturnChecks = FALSE,
  bQuiet = TRUE
)
}
\arguments{
\item{dfs}{\code{list} Input data frames:
\itemize{
\item \code{dfSUBJ}: \code{data.frame} Subject-level data with one record per subject.
\item \code{dfQUERY}: \code{data.frame} Query-level data with one record per query.
\item \code{dfDATACHG}: \code{data.frame} Data-Point-level data with one record per data entry.
}}

\item{lMapping}{\code{list} Column metadata with structure \code{domain$key}, where \code{key} contains the name
of the column.}

\item{bReturnChecks}{\code{logical} Return input checks from \code{\link[=is_mapping_valid]{is_mapping_valid()}}? Default: \code{FALSE}}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{data.frame} Data frame with one record per subject, the input to \code{\link[=QueryRate_Assess]{QueryRate_Assess()}}. If
\code{bReturnChecks} is \code{TRUE} \code{QueryRate_Map_Raw} returns a named \code{list} with:
\itemize{
\item \code{df}: the data frame described above
\item \code{lChecks}: a named \code{list} of check results
}
}
\description{
Convert raw query data to formatted input data to \code{\link[=QueryRate_Assess]{QueryRate_Assess()}}.
}
\details{
Query Rate Assessment - Raw Mapping

\code{QueryRate_Map_Raw} combines query data with data points data and subject-level data to create
formatted input data to \code{\link[=QueryRate_Assess]{QueryRate_Assess()}}. This function creates an input dataset for the Query Rate Assessment
(\code{\link[=QueryRate_Assess]{QueryRate_Assess()}}) by binding subject-level query counts (derived from \code{dfQUERY}) and data point counts
(derived from \code{dfDATACHG}) to subject-level data (from \code{dfSUBJ}). Note that the function can generate data summaries for specific types of
queries by passing filtered query data to \code{dfQUERY}.
}
\section{Data specification}{
\tabular{lllll}{
   \strong{Domain} \tab \strong{Column Key} \tab \strong{Default Value} \tab \strong{Required?} \tab \strong{Require Unique Values?} \cr
   dfSUBJ \tab strSiteCol \tab siteid \tab TRUE \tab FALSE \cr
   dfSUBJ \tab strEDCIDCol \tab subject_nsv \tab TRUE \tab TRUE \cr
   dfDATACHG \tab strIDCol \tab subjectname \tab TRUE \tab FALSE \cr
   dfDATACHG \tab strNChangesCol \tab n_changes \tab TRUE \tab FALSE \cr
}
}

\examples{
# Run with defaults.
dfInput <- QueryRate_Map_Raw()

# Run with error checking and message log.
dfInput <- QueryRate_Map_Raw(bReturnChecks = TRUE, bQuiet = FALSE)

}
