% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-rank_chg.R
\name{rank_chg}
\alias{rank_chg}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
rank_chg(status)
}
\arguments{
\item{status}{\code{boolean} fontawesome emoji indicator that describes the status of error checking for all assessments run in \code{Study_Assess()}}
}
\description{
\code{rank_chg} - inserts icons for status in {gt} table.

Adopted from https://themockup.blog/posts/2020-10-31-embedding-custom-features-in-gt-tables/.
}
\details{
Report helper functions used to create fontawesome status icons for the report that is generated with \link{Study_Report}.
}
