% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{rbm_data_spec}
\alias{rbm_data_spec}
\title{RBM Data Specification}
\format{
A data frame with 125 rows and 4 variables.
\describe{
\item{System}{Gismo or GSM}
\item{Table}{Name of the table/data source created by \code{gsm::Make_Snapshot()}}
\item{Column}{Name of the column in the table}
\item{Description}{Description of \code{Column}}
\item{Order}{Order columns should appear in for standard export}
}
}
\source{
\url{https://silver-potato-cfe8c2fb.pages.github.io/articles/DataPipeline.html}
}
\usage{
rbm_data_spec
}
\description{
A dataset containing input/output specifications for gsm and Gismo.
}
\keyword{datasets}
