% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-RenderRmd.R
\name{RenderRmd}
\alias{RenderRmd}
\title{Custom Rmarkdown render function}
\usage{
RenderRmd(
  strInputPath,
  strOutputFile = basename(strInputPath),
  strOutputDir = getwd(),
  lParams
)
}
\arguments{
\item{strInputPath}{\code{string} or \code{fs_path} Path to the template \code{Rmd} file.}

\item{strOutputFile}{\code{string} Filename for the output.}

\item{strOutputDir}{\code{string} or \code{fs_path} Path to the directory where the output will be saved.}

\item{lParams}{\code{list} Parameters to pass to the template \code{Rmd} file.}
}
\value{
Rendered Rmarkdown file
}
\description{
Rmarkdown render function that defaults to rendering intermediate rmd files in a temporary directory
}
