% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_FlagOverTime.R
\name{Report_FlagOverTime}
\alias{Report_FlagOverTime}
\title{Summarize flags by SnapshotDate}
\usage{
Report_FlagOverTime(
  dfSummary,
  dfMetrics,
  strGroupLevel = c("site", "study", "country")
)
}
\arguments{
\item{dfSummary}{A summary of assessment results such as the ones generated
by \code{\link[=Summarize]{Summarize()}}.}

\item{dfMetrics}{Metric-specific metadata created by passing an \code{lWorkflow}
object to the \code{\link[=MakeMetric]{MakeMetric()}} function.}

\item{strGroupLevel}{A string specifying the group type.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Create a table of longitudinal study data by site, study, or country, showing
flags over time.
}
