% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_KRI.R
\name{Report_KRI}
\alias{Report_KRI}
\title{Report_KRI function}
\usage{
Report_KRI(
  lCharts = NULL,
  dfResults = NULL,
  dfGroups = NULL,
  dfMetrics = NULL,
  strOutpath = NULL
)
}
\arguments{
\item{lCharts}{A list of charts to include in the report.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[=BindResults]{BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID}, \code{SnapshotDate}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and reporting. Created by passing an \code{lWorkflow} object to \code{\link[=MakeMetric]{MakeMetric()}}. Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric}, \code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{Threshold}. For more details see the Data Pipeline vignette: \code{vignette("DataPipeline", package = "gsm")}.}

\item{strOutpath}{The output path for the generated report. If not provided, the report will be saved in the current working directory with the name "kri_report.html".}
}
\value{
None
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
This function generates a KRI report based on the provided inputs.
}
\examples{
 \dontrun{
# Run reports
lCharts <- MakeCharts(
  dfResults = reportingResults,
  dfGroups = reportingGroups,
  dfMetrics = reportingMetrics,
  dfBounds = reportingBounds
)

strOutpath <- "StandardSiteReport.html"
Report_KRI(
  lCharts = lCharts,
  dfResults =  reportingResults,
  dfGroups =  reportingGroups,
  dfMetrics = reportingMetrics
  strOutpath = strOutpath
)
}

}
\keyword{KRI}
\keyword{report}
