% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_MetricCharts.R
\name{Report_MetricCharts}
\alias{Report_MetricCharts}
\title{Render charts for a given metric to markdown}
\usage{
Report_MetricCharts(lCharts, strMetricID = "")
}
\arguments{
\item{lCharts}{A list of charts for the selected metric.}

\item{strMetricID}{\code{character} MetricID to subset the data.}
}
\value{
Markdown content with charts and a summary table for the metric
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function generates a markdown framework for charts
}
