% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_MetricTable.R
\name{Report_MetricTable}
\alias{Report_MetricTable}
\title{Generate a summary table for a report}
\usage{
Report_MetricTable(dfSummary, dfSite, strSnapshotDate = NULL)
}
\arguments{
\item{dfSummary}{The summary data frame}

\item{dfSite}{The site data frame}
}
\value{
A datatable containing the summary table
}
\description{
This function generates a summary table for a report by joining the provided
summary data frame with the site data frame. It then filters and arranges the
data based on certain conditions and displays the result in a datatable.
}
