% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_OverviewText.R
\name{Report_OverviewText}
\alias{Report_OverviewText}
\title{Create message describing study summary for Report}
\usage{
Report_OverviewText(lSetup, dfSummary, dfStudy)
}
\arguments{
\item{lSetup}{\code{list} that is produced by \code{Report_StudyInfo(dfStudy = params$dfStudy)}.}

\item{dfSummary}{\code{data.frame} containing summary information.}

\item{dfStudy}{\code{data.frame} contains site-level metadata.}
}
\description{
Create message describing study summary for Report
}
\keyword{internal}
