% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_OverviewText.R
\name{Report_OverviewText}
\alias{Report_OverviewText}
\title{Create message describing study summary for Report}
\usage{
Report_OverviewText(lSetup, dfResults, lStudy)
}
\arguments{
\item{lSetup}{\code{list} that is produced by \code{Report_StudyInfo()}.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to
\code{\link[gsm.mapping:BindResults]{gsm.mapping::BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator},
\code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID},
\code{SnapshotDate}.}

\item{lStudy}{\code{list} contains study-level metadata.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{internal}
