% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Setup.R
\name{Report_Setup}
\alias{Report_Setup}
\title{Calculate needed values for report}
\usage{
Report_Setup(dfGroups = NULL, dfMetrics = NULL, dfResults = NULL)
}
\arguments{
\item{dfGroups}{A data frame containing the group-level metadata.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and reporting. Created by passing an \code{lWorkflow} object to \code{\link[=MakeMetric]{MakeMetric()}}. Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric}, \code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{strThreshold}. For more details see the Data Pipeline vignette: \code{vignette("DataPipeline", package = "gsm")}.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[=BindResults]{BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID}, \code{SnapshotDate}.}
}
\value{
\code{list} with the following elements:
\itemize{
\item \code{GroupLevel} (character): The group level of the report.
\item \code{SnapshotDate} (Date): The date of the snapshot.
\item \code{lStudy} (list): Study-level metadata.
\item \code{StudyID} (character): The study ID.
\item \code{red_kris} (numeric): The number of red flags.
\item \code{amber_kris} (numeric): The number of amber flags.
}
}
\description{
Calculate needed values for report
}
\keyword{internal}
