% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_Setup.R
\name{Report_Setup}
\alias{Report_Setup}
\title{Calculate needed values for report}
\usage{
Report_Setup(dfGroups = NULL, dfMetrics = NULL, dfResults = NULL)
}
\arguments{
\item{dfGroups}{A data frame containing the group-level metadata.}

\item{dfMetrics}{A data frame containing the metric-level metadata.}

\item{dfResults}{A data frame containing the metric results.}
}
\value{
\code{list} with the following elements:
\itemize{
\item \code{GroupLevel} (character): The group level of the report.
\item \code{SnapshotDate} (Date): The date of the snapshot.
\item \code{lStudy} (list): Study-level metadata.
\item \code{StudyID} (character): The study ID.
\item \code{red_kris} (numeric): The number of red flags.
\item \code{amber_kris} (numeric): The number of amber flags.
}
}
\description{
Calculate needed values for report
}
\keyword{internal}
