% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_StudyInfo.R
\name{Report_StudyInfo}
\alias{Report_StudyInfo}
\title{Report Study Information}
\usage{
Report_StudyInfo(lStudy, lStudyLabels = NULL)
}
\arguments{
\item{lStudy}{A list containing study information.}

\item{lStudyLabels}{A list containing study labels. Default is NULL.}
}
\value{
None
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function generates a table summarizing study metadata.
}
\keyword{internal}
