% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevertSnapshotLogs.R
\name{RevertSnapshotLogs}
\alias{RevertSnapshotLogs}
\title{Function to revert to 1.8 data model}
\usage{
RevertSnapshotLogs(lSnapshot, lMeta = NULL, lData = NULL)
}
\arguments{
\item{lSnapshot}{\code{list} the snapshot object to revert}

\item{lMeta}{\code{list} the meta data to use for reversion}

\item{lData}{\code{list} the data to us for reversion}
}
\description{
Function to revert to 1.8 data model
}
\examples{
snapshot <- Make_Snapshot()
output <- RevertSnapshotLogs(snapshot)

}
\keyword{internal}
