% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-runAssessment.R
\name{runAssessment}
\alias{runAssessment}
\title{Run a single assessment}
\usage{
runAssessment(assessment, lData, lMapping, lTags, bQuiet = FALSE)
}
\arguments{
\item{assessment}{assessments}

\item{lData}{list of data}

\item{lMapping}{mapping}

\item{lTags}{tags}
}
\value{
A list containing: dataChecks and results
}
\description{
Run a single assessment
}
\details{
Coming soon
}
\examples{
 NULL

}
