% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-RunAssessment.R
\name{RunAssessment}
\alias{RunAssessment}
\title{Run a single assessment}
\usage{
RunAssessment(assessment, lData, lMapping, lTags = NULL, bQuiet = FALSE)
}
\arguments{
\item{assessment}{assessments}

\item{lData}{list of data}

\item{lMapping}{mapping}

\item{lTags}{tags}

\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}
}
\value{
A list containing: dataChecks and results
}
\description{
Run a single assessment
}
\details{
Coming soon
}
\examples{
 NULL

}
