% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunQTL.R
\name{RunQTL}
\alias{RunQTL}
\title{Run a QTL analysis for any supported \verb{*_Assess} function.}
\usage{
RunQTL(strName = NULL, lWorkflow = NULL, lData = NULL, lMapping = NULL)
}
\arguments{
\item{strName}{\code{character} Name of QTL YAML file found in \code{inst/workflow/experimental}, e.g. "qtl0003"}

\item{lWorkflow}{\code{character} Optional. Custom workflow to use in place of pre-defined workflow.}

\item{lData}{\code{list} List of data needed to run QTL analysis.}

\item{lMapping}{\code{list} List of data-domain mappings.}
}
\value{
\code{list} list of QTL analysis results.
}
\description{
Run a QTL based on a pre-defined or custom workflow outside of \code{Make_Snapshot()} or \code{Study_Assess()}. This is a helper function that makes it possible to
run a QTL for any supported \verb{*_Assess} function.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
qtl <- RunQTL("qtl0004")

}
