% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-RunStep.R
\name{RunStep}
\alias{RunStep}
\title{Run a single step in an assessment}
\usage{
RunStep(step, mapping, lTags, bQuiet)
}
\arguments{
\item{step}{list of data}

\item{mapping}{mapping}

\item{lTags}{tags}

\item{bQuiet}{Default is TRUE, which means warning messages are suppressed. Set to FALSE to see warning messages.}
}
\value{
A list containing: dataChecks and results
}
\description{
Run a single step in an assessment
}
\details{
Coming soon
}
\examples{
 NULL

}
