% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-RunStep.R
\name{RunStep}
\alias{RunStep}
\title{Run a single step in a workflow}
\usage{
RunStep(lStep, lMapping, lData, bQuiet)
}
\arguments{
\item{lStep}{\code{list} single workflow step (typically defined in \code{lAssessment$workflow}). Should
include the name of the function to run (\code{lStep$name}), data inputs (\code{lStep$inputs}), name of
output (\code{lStep$output}) and configurable parameters (\code{lStep$params}) (if any)}

\item{lMapping}{\code{list} List containing expected columns in each data set.}

\item{lData}{\code{list} a named list of domain level data frames. Names should match the values
specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs
from \code{X_Map_Raw}.}

\item{bQuiet}{\code{logical} Default is TRUE, which means warning messages are suppressed. Set to
FALSE to see warning messages.}
}
\value{
\code{list} containing the results of the \code{lStep$name} function call should contain \code{.$checks}
parameter with results from \code{is_mapping_vald} for each domain in \code{lStep$inputs}.
}
\description{
Runs a single step of an assessment workflow. Currently supports \code{Filter}, \code{Map}, and \code{Assess}
functions.
}
\examples{
lStep <- MakeAssessmentList()[["kri0001"]][["workflow"]][[1]]

lMapping <- yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package = "gsm"))

lData <- list(
  dfSUBJ = clindata::rawplus_dm,
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::rawplus_protdev,
  dfCONSENT = clindata::rawplus_consent,
  dfIE = clindata::rawplus_ie
)

lTags <- list(Study = "myStudy")

ae_step <- RunStep(lStep = lStep, lMapping = lMapping, lData = lData, lTags = lTags, bQuiet = FALSE)

}
