% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-RunStep.R
\name{RunStep}
\alias{RunStep}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
RunStep(lStep, lMapping, lData, bQuiet)
}
\arguments{
\item{lStep}{\code{list} single workflow step (typically defined in \code{lWorkflow$workflow}). Should
include the name of the function to run (\code{lStep$name}), data inputs (\code{lStep$inputs}), name of
output (\code{lStep$output}) and configurable parameters (\code{lStep$params}) (if any)}

\item{lMapping}{\code{list} List containing expected columns in each data set.}

\item{lData}{\code{list} a named list of domain level data frames. Names should match the values
specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs
from \code{X_Map_Raw}.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}.}
}
\value{
\code{list} containing the results of the \code{lStep$name} function call should contain \code{.$checks}
parameter with results from \code{is_mapping_vald} for each domain in \code{lStep$inputs}.
}
\description{
Runs a single step of an assessment workflow. Currently supports \code{Filter}, \code{Map}, and \code{Assess}
functions.
}
\details{
Run a single step in a workflow
}
\examples{
lStep <- MakeWorkflowList()[["kri0001"]][["steps"]][[1]]

lMapping <- yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package = "gsm"))

lData <- list(
  dfSUBJ = clindata::rawplus_dm,
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::ctms_protdev,
  dfCONSENT = clindata::rawplus_consent,
  dfIE = clindata::rawplus_ie
)


ae_step <- RunStep(lStep = lStep, lMapping = lMapping, lData = lData, bQuiet = FALSE)

}
