% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunStep.R
\name{RunStep}
\alias{RunStep}
\title{Run a single step in a workflow.}
\usage{
RunStep(lStep, lData, lMeta)
}
\arguments{
\item{lStep}{\code{list} single workflow step (typically defined in \code{lWorkflow$workflow}). Should
include the name of the function to run (\code{lStep$name}), data inputs (\code{lStep$inputs}), name of
output (\code{lStep$output}) and configurable parameters (\code{lStep$params}) (if any)}

\item{lData}{\code{list} a named list of domain level data frames.}

\item{lMeta}{\code{list} a named list of meta data.}
}
\value{
\code{list} containing the results of the \code{lStep$name} function call should contain \code{.$checks}
parameter with results from \code{is_mapping_vald} for each domain in \code{lStep$inputs}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Runs a single step of an assessment workflow. This function is called by \code{RunWorkflow} for each
step in the workflow. It prepares the parameters for the function call, including the metadata,
mapping, and data inputs. It then calls the function specified in \code{lStep$name} with the prepared
parameters.
}
\examples{
wf_mapping <- MakeWorkflowList("mapping")
lStep <- MakeWorkflowList("kri0001")[["kri0001"]][["steps"]][[1]]
lMeta <- MakeWorkflowList("kri0001")[["kri0001"]][["meta"]]
lData <- list(
  dfSUBJ = clindata::rawplus_dm,
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::ctms_protdev,
  dfCONSENT = clindata::rawplus_consent,
  dfIE = clindata::rawplus_ie
)
lMapped <- RunWorkflow(wf_mapping, lData)$lData

ae_step <- RunStep(lStep = lStep, lData = lMapped, lMeta = lMeta)

}
