% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-RunStratifiedWorkflow.R
\name{RunStratifiedWorkflow}
\alias{RunStratifiedWorkflow}
\title{Run a stratified workflow}
\usage{
RunStratifiedWorkflow(lWorkflow, lData, lMapping, bQuiet = TRUE)
}
\arguments{
\item{lWorkflow}{\code{list} A named list of metadata defining how the workflow should be run.
Properties should include: \code{label}, \code{tags} and \code{workflow}}

\item{lData}{\code{list} A named list of domain-level data frames. Names should match the values
specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from
\code{X_Map_Raw}.}

\item{lMapping}{\code{list} A named list identifying the columns needed in each data domain.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\value{
\code{list} \code{lWorkflow} along with \code{workflow}, \code{path}, \code{name}, \code{lData}, \code{lChecks},
\code{bStatus}, \code{checks}, and \code{lResults} added based on the results of the execution of
\code{lWorkflow$workflow}.
}
\description{
Attempts to run a stratified workflow (\code{lWorkflow}) using shared data (\code{lData}) and metadata (\code{lMapping}).
Calls \code{RunStep} for each item in \code{lWorkflow$workflow} and saves the results to \code{lWorkflow}
}
\examples{
lWorkflows <- MakeWorkflowList()
lData <- list(
  dfAE = clindata::rawplus_ae,
  dfCONSENT = clindata::rawplus_consent,
  dfDISP_Study = clindata::rawplus_studcomp,
  dfDISP_Treatment = clindata::rawplus_sdrgcomp,
  dfIE = clindata::rawplus_ie,
  dfLB = clindata::rawplus_lb,
  dfPD = clindata::rawplus_protdev,
  dfSUBJ = clindata::rawplus_dm
)
lMapping <- yaml::read_yaml(
  system.file("mappings", "mapping_rawplus.yaml", package = "gsm")
)

output <- RunStratifiedWorkflow(
  lWorkflows$aeGrade, # adverse event workflow, stratified by AE grade
  lData = lData,
  lMapping = lMapping
)

}
