% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunWorkflow.R
\name{RunWorkflow}
\alias{RunWorkflow}
\title{Run a workflow via it's YAML specification.}
\usage{
RunWorkflow(lWorkflow, lData, bReturnData = TRUE, bKeepInputData = FALSE)
}
\arguments{
\item{lWorkflow}{\code{list} A named list of metadata defining how the workflow should be run.}

\item{lData}{\code{list} A named list of domain-level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{bReturnData}{\code{boolean} should function return only lData or should meta and steps be included? Default is \code{TRUE}.}

\item{bKeepInputData}{\code{boolean} should the input data be returned? Default is \code{FALSE}.}
}
\value{
\code{list} containing objects named: \code{steps}, \code{path}, \code{name}, \code{lData}, \code{lChecks}, \code{bStatus}, \code{lWorkflowChecks}, and \code{lResults}.

\code{list} contains just lData if \code{bReturnData} is \code{TRUE}, otherwise returns the full \code{lWorkflow} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Attempts to run a single assessment (\code{lWorkflow}) using shared data (\code{lData}) and metadata (\code{lMapping}).
Calls \code{RunStep} for each item in \code{lWorkflow$workflow} and saves the results to \code{lWorkflow}.
}
\examples{
\dontrun{
lAssessments <- MakeWorkflowList("kri0001")
lData <- list(
  dfAE = clindata::rawplus_ae,
  dfCONSENT = clindata::rawplus_consent,
  dfDISP = clindata::rawplus_dm,
  dfIE = clindata::rawplus_ie,
  dfLB = clindata::rawplus_lb,
  dfPD = clindata::ctms_protdev,
  dfSUBJ = clindata::rawplus_dm
)
wf_mapping <- MakeWorkflowList("mapping")
lMapped <- RunWorkflow(wf_mapping, lData)$lData

output <- map(lAssessments, ~ RunWorkflow(., lMapped))
}
}
