% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunWorkflows.R
\name{RunWorkflows}
\alias{RunWorkflows}
\title{Convienience function to easily run multiple workflows}
\usage{
RunWorkflows(lWorkflows, lData)
}
\arguments{
\item{lWorkflows}{A list of workflows to be run}

\item{lData}{A list of data to be used in the workflows}
}
\value{
A named list of results, where the names correspond to the names of
the workflows
}
\description{
This function takes a list of workflows and a list of data as input. If there
are multiple workflows in the list, it runs each workflow and returns the
results as a named list. If there is only one workflow, it runs that workflow
and returns the result.
}
