% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunWorkflows.R
\name{RunWorkflows}
\alias{RunWorkflows}
\title{Convenience function to easily run multiple workflows}
\usage{
RunWorkflows(
  lWorkflows,
  lData = NULL,
  bKeepInputData = FALSE,
  bReturnResult = TRUE,
  strResultNames = c("Type", "ID"),
  strLogger = "gsm"
)
}
\arguments{
\item{lWorkflows}{\code{list} A named list of metadata defining how the workflow should be run.}

\item{lData}{\code{list} A named list of domain-level data frames.}

\item{bKeepInputData}{\code{boolean} should the input data be included in \code{lData} after the workflow is run? Only relevant when bReturnResult is FALSE. Default is \code{TRUE}.}

\item{bReturnResult}{\code{boolean} should \emph{only} the result from the last step (\code{lResults}) be returned? If false, the full workflow (including \code{lResults}) is returned. Default is \code{TRUE}.}

\item{strResultNames}{\code{string} vector of length two, which describes the meta fields used to name the output.}
}
\value{
A named list of results from \code{RunWorkflow()}, where the names correspond to the names of
the workflow ID
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function takes a list of workflows and a list of data as input.  It runs each workflow and returns the
results as a named list where the names of the list correspond to the workflow ID ($meta$ID).

Workflows are run in the order they are provided in the lWorkflows. The results from each workflow are passed as inputs (along with lData) for later workflows.
}
