% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunWorkflows.R
\name{RunWorkflows}
\alias{RunWorkflows}
\title{Convenience function to easily run multiple workflows}
\usage{
RunWorkflows(
  lWorkflows,
  lData = NULL,
  lInputConfig = NULL,
  bReturnData = TRUE,
  bKeepInputData = FALSE
)
}
\arguments{
\item{lWorkflows}{\code{list} A named list of metadata defining how the workflow should be run.}

\item{lData}{\code{list} A named list of domain-level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{lInputConfig}{\code{list} A named list of data sources to pull data from external source.}

\item{bReturnData}{\code{boolean} should function return only bData or should meta and steps be included? Default is \code{TRUE}.}

\item{bKeepInputData}{\code{boolean} should the input data be returned? Default is \code{FALSE}.}
}
\value{
A named list of results, where the names correspond to the names of
the workflows
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function takes a list of workflows and a list of data as input. If there
are multiple workflows in the list, it runs each workflow and returns the
results as a named list. If there is only one workflow, it runs that workflow
and returns the result.
}
