% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Snapshot.R
\name{Save_Snapshot}
\alias{Save_Snapshot}
\title{Save Snapshot}
\usage{
Save_Snapshot(
  lSnapshot,
  cPath,
  bCreateDefaultFolder = FALSE,
  bQuiet = TRUE,
  strFileFormat = "parquet"
)
}
\arguments{
\item{lSnapshot}{\code{list} the output of \code{\link[=Make_Snapshot]{Make_Snapshot()}} or \code{\link[=Augment_Snapshot]{Augment_Snapshot()}}.}

\item{cPath}{\code{character} a character string indicating a directory to save .csv files; the output of the snapshot.}

\item{bCreateDefaultFolder}{\code{logical} default: \code{FALSE}. If \code{TRUE}, creates a new folder in the \code{cPath} directory with the current date (YYYY-MM-DD).}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}

\item{strFileFormat}{\code{character} The file format used to save all flat files. Default: \code{parquet}. Valid options: \code{parquet} or \code{csv}.}
}
\value{
\code{.csv} files and \code{.rds} file saved to a specified directory.
}
\description{
Save the output of \code{\link[=Make_Snapshot]{Make_Snapshot()}} or \code{\link[=Augment_Snapshot]{Augment_Snapshot()}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{

snapshot <- Make_Snapshot()
Save_Snapshot(lSnapshot = snapshot, cPath = here::here("data"))
}

}
