% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveQTL.R
\name{SaveQTL}
\alias{SaveQTL}
\title{Title}
\usage{
SaveQTL(lSnapshot, strPath = "./inst/dummyqtldata.csv", bQuiet = TRUE)
}
\arguments{
\item{lSnapshot}{\code{list} List returned by \code{\link[=RunQTL]{RunQTL()}}}

\item{strPath}{\code{character} Path to historical QTL data.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}
}
\description{
Title
}
\examples{

\dontrun{
dispQTL <- RunQTL('qtl0007')
SaveQTL(dispQTL)
}

}
