% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot HTML widget from \code{{rbm-viz}}}
\usage{
scatterPlot(
  results,
  workflow,
  bounds,
  selectedGroupIDs = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{results}{TODO}

\item{workflow}{TODO}

\item{bounds}{TODO}

\item{selectedGroupIDs}{TODO}

\item{width}{TODO}

\item{height}{TODO}

\item{elementId}{TODO}
}
\value{
htmlwidget scatter plot
}
\description{
TODO: Update documentation
}
