% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_assess.R
\name{Study_Assess}
\alias{Study_Assess}
\title{Run multiple assessments on a single study}
\usage{
Study_Assess(
  lData = NULL,
  lMapping = NULL,
  lAssessments = NULL,
  lPopFlags = list(strRandFlagCol = "Y"),
  lTags = list(study = "myStudy"),
  bReturnInputs = FALSE,
  bQuiet = FALSE
)
}
\arguments{
\item{lData}{list of data}

\item{lMapping}{mapping}

\item{lAssessments}{assessments}

\item{lPopFlags}{filter demog data?}

\item{lTags}{tags}
}
\value{
A list containing: dataChecks and results
}
\description{
Run multiple assessments on a single study
}
\details{
Coming soon
}
\examples{
 NULL

}
