% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Assess.R
\name{Study_Assess}
\alias{Study_Assess}
\title{Run Multiple Assessments on a Study}
\usage{
Study_Assess(
  lData = NULL,
  lMapping = NULL,
  lAssessments = NULL,
  lSubjFilters = NULL,
  bQuiet = TRUE,
  bFlowchart = FALSE
)
}
\arguments{
\item{lData}{\code{list} a named list of domain level data frames. Names should match the values specified in \code{lMapping} and \code{lAssessments}, which are generally based on the expected inputs from \code{X_Map_Raw}.}

\item{lMapping}{\code{list} a named list identifying the columns needed in each data domain.}

\item{lAssessments}{\code{list} a named list of metadata defining how each assessment should be run. By default, \code{MakeWorkflowList()} imports YAML specifications from \code{inst/workflow}.}

\item{lSubjFilters}{\code{list} a named list of parameters to filter subject-level data on.}

\item{bQuiet}{\code{logical} Suppress warning messages? Default: \code{TRUE}}

\item{bFlowchart}{\code{logical} Create flowchart to show data pipeline? Default: \code{FALSE}}
}
\value{
\code{list} of assessments containing status information and results.
}
\description{
Attempts to run one or more assessments (\code{lAssessments}) using shared data (\code{lData}) and metadata (\code{lMapping}). By default, the sample \code{rawplus} data from the {clindata} package is used, and all assessments defined in \code{inst/workflow} are evaluated. Individual assessments are run using \code{gsm::RunAssessment()}
}
\examples{
\dontrun{
results <- Study_Assess() # run using defaults
}

}
