% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_AssessmentReport.R
\name{Study_AssessmentReport}
\alias{Study_AssessmentReport}
\title{Make Summary of 1 or more assessment Data checks}
\usage{
Study_AssessmentReport(lAssessments)
}
\arguments{
\item{lAssessments}{List of 1+ assessments like those created by \code{runAssessment()} or \code{Assess_Study()}}
}
\value{
returns a list containing a data.frame summarizing the checks \code{dfSummary} and a dataframe listing all checks (`dfAllChecks``)
}
\description{
Make overview table with one row per assessment and one column per site showing flagged assessments.
}
