% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Report.R
\name{Study_Report}
\alias{Study_Report}
\title{Study Report}
\usage{
Study_Report(assessments, meta, outpath = NULL)
}
\arguments{
\item{assessments}{character vector listing assessments}

\item{meta}{list of metadata related to study}

\item{outpath}{path to safe the report}
}
\description{
Pulls needed study data and runs one or more assessments
}
