% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Report.R
\name{Study_Report}
\alias{Study_Report}
\title{{experimental} Study Report}
\usage{
Study_Report(
  lAssessments,
  lMeta = list(Project = "My Project"),
  strOutpath = NULL
)
}
\arguments{
\item{lAssessments}{\code{list} The results of multiple assessments run using \code{Study_Assess}.}

\item{lMeta}{\code{list} Metadata related to study.}

\item{strOutpath}{\code{character} File path; location where the report will be saved.}
}
\value{
HTML report of study data.
}
\description{
Create HTML summary report using the results of \code{Study_Assess}, including tables, charts, and error checking.
}
\examples{
\dontrun{
lWorkflow <- Study_Assess()
Study_Report(lWorkflow, lMeta = list(study = "my study name"))
}

}
