% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Report.R
\name{Study_Report}
\alias{Study_Report}
\title{{experimental} Study Report}
\usage{
Study_Report(lAssessments, strOutpath = NULL, strReportType = "kri")
}
\arguments{
\item{lAssessments}{\code{list} The results of multiple assessments run using \code{Study_Assess}.}

\item{strOutpath}{\code{character} File path; location where the report will be saved.}

\item{strReportType}{\code{character} Type of report to generate. Default: \code{kri}. Other options include: \code{study}.}
}
\value{
HTML report of study data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Create HTML summary report using the results of \code{Study_Assess}, including tables, charts, and error checking.
}
\examples{
\dontrun{
lWorkflow <- Study_Assess()
Study_Report(lWorkflow, lMeta = list(study = "my study name"))
}

}
