% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Report.R
\name{Study_Report}
\alias{Study_Report}
\title{Study Report}
\usage{
Study_Report(
  lSnapshot,
  dfStudy = NULL,
  dfSite = NULL,
  strOutpath = NULL,
  strReportType = "site"
)
}
\arguments{
\item{lSnapshot}{\code{list} The results of multiple assessments run using \code{Study_Assess} or \code{Make_Snapshot}.}

\item{dfStudy}{\code{data.frame} A data.frame containing study status metadata. Typically output from \code{Make_Snapshot()$lSnapshot$status_study}}

\item{dfSite}{\code{data.frame} A data.frame containing site status metadata. Typically output from \code{Make_Snapshot()$lSnapshot$status_site}}

\item{strOutpath}{\code{character} File path; location where the report will be saved.}

\item{strReportType}{\code{character} The type of report to be generated. Valid values:
\itemize{
\item \code{"site"} for site-level KRI summary (default). The site-level report is stable and is under active development.
\item \code{"country"} for country-level KRI summary. The country-level report is stable and is under active development.
\item \code{"QTL"} for QTL summary. The QTL report is in early draft mode and is under active development.
}}
}
\value{
HTML report of study data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Create HTML summary report using the results of \code{Study_Assess}, including tables, charts, and error checking.
}
\examples{
\dontrun{
# Using `Study_Assess()`
study <- Study_Assess()
Study_Report(study)
Study_Report(study, strReportType = "country")

# Adding metadata for a snapshot
one_snapshot <- Make_Snapshot()
Study_Report(
  lSnapshot = one_snapshot
)
}


}
