% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Report.R
\name{Study_Report}
\alias{Study_Report}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
Study_Report(
  lSnapshot,
  dfStudy = NULL,
  dfSite = NULL,
  strOutpath = NULL,
  strReportType = "site"
)
}
\arguments{
\item{lSnapshot}{\code{list} The results of multiple assessments run using \code{Study_Assess}, \code{Make_Snapshot}, or \code{Augment_Snapshot}.}

\item{dfStudy}{\code{data.frame} A data.frame containing study status metadata. Typically output from \code{Make_Snapshot()$lSnapshot$status_study}}

\item{dfSite}{\code{data.frame} A data.frame containing site status metadata. Typically output from \code{Make_Snapshot()$lSnapshot$status_site}}

\item{strOutpath}{\code{character} File path; location where the report will be saved.}

\item{strReportType}{\code{character} The type of report to be generated. Valid values:
\itemize{
\item \code{"site"} for site-level KRI summary (default)
\item \code{"country"} for country-level KRI summary
\item \code{"QTL"} for QTL summary
}}
}
\value{
HTML report of study data.
}
\description{
Study Report
}
\details{
Create HTML summary report using the results of \code{Study_Assess}, including tables, charts, and error checking.
}
\examples{
\dontrun{
# Using `Study_Assess()`
study <- Study_Assess()
Study_Report(study)
Study_Report(study, strReportType = "country")

# Adding metadata for a single snapshot
one_snapshot <- Make_Snapshot()
Study_Report(
  lAssessments = one_snapshot$lStudyAssessResults,
  dfStudy = one_snapshot$lSnapshot$status_study
)

# Longitudinal Data
snapshot <- Make_Snapshot()

longitudinal <- Augment_Snapshot(
  snapshot,
  system.file("snapshots", "AA-AA-000-0000", package = "gsm")
)

Study_Report(
  lSnapshot = longitudinal,
  dfStudy = longitudinal$lSnapshot$status_study
)
}

}
