% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_Table.R
\name{Study_Table}
\alias{Study_Table}
\title{Make Assessment overview table}
\usage{
Study_Table(
  dfFindings,
  bFormat = TRUE,
  showCounts = TRUE,
  showSiteScore = TRUE,
  siteScoreThreshold = 1,
  colCollapse = TRUE
)
}
\arguments{
\item{dfFindings}{dataframe containing one or more stacked findings. Findings are one record per assessment per site and have the following columns: Assessment, Label, SiteID, N, PValue, Flag. PValue is ignored in the summary table.}

\item{bFormat}{Use html-friendly icons in table cells. -1 is converted to a down arrow. 1 is converted to an up arrow. 0 is not shown. Other values are left as is.}

\item{showCounts}{Show site counts? Uses first value of N for each site given in dfFindings.}

\item{showSiteScore}{Show a "Score" row with total number of flagged assessments for each site. TODO:  add method for custom scoring in future release)}

\item{siteScoreThreshold}{Hide sites with a site score less than this value (1 by default).}

\item{colCollapse}{Combine the Assessment and Label columns into a single "Title Column"}
}
\value{
returns a dataframe giving assessment status (rows) by Site (column)
}
\description{
Make overview table with one row per assessment and one column per site showing flagged assessments.
}
