% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetStackedSnapshots.R
\name{SubsetStackedSnapshots}
\alias{SubsetStackedSnapshots}
\title{SubsetStackedSnapshots}
\usage{
SubsetStackedSnapshots(workflowid)
}
\arguments{
\item{workflowid}{\code{character} workflow ID or \code{kri_id}.}
}
\description{
This function is used to subset a list of stacked snapshots within \code{\link[=Make_Snapshot]{Make_Snapshot()}}.
}
\keyword{internal}
