% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.R
\name{Summarize}
\alias{Summarize}
\title{Make Summary Data Frame}
\usage{
Summarize(dfFlagged, cAssessment = "", cLabel = "")
}
\arguments{
\item{dfFlagged}{data frame in format produced by \code{\link{Flag}}}

\item{cAssessment}{brief description of current assessment}

\item{cLabel}{brief description of line item in current assessment}
}
\value{
Simplified finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
Adds columns flagging sites that represent possible statistical outliers
}
