% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.R
\name{Summarize}
\alias{Summarize}
\title{Make Summary Data Frame}
\usage{
Summarize(dfFlagged, strScoreCol = "Score", lTags = NULL)
}
\arguments{
\item{dfFlagged}{data.frame in format produced by \code{\link{Flag}}}

\item{strScoreCol}{column from analysis results to be copied to \code{dfSummary$Score}}

\item{lTags}{List of tags containing metadata to add to the data frame.}
}
\value{
Simplified finding data.frame with columns for GroupID, GroupLabel, N, KRI, KRILabel, Score, ScoreLabel, Flag, and any associated metadata
when associated with a workflow.
}
\description{
Create a concise summary of assessment results that is easy to aggregate across assessments
}
\details{
\code{Summarize} supports the input data (\code{dfFlagged}) from the \code{Flag} function.
}
\section{Data Specification}{


(\code{dfFlagged}) has the following required columns:
\itemize{
\item \code{GroupID} - Group ID
\item \code{N} - Total number of participants at site
\item \code{Flag} - Flagging value of -1, 0, or 1
\item \code{strScoreCol} - Column from analysis results.
}
}

\examples{
dfInput <- AE_Map_Adam()

dfTransformed <- Transform_EventCount(dfInput,
  strCountCol = "Count",
  strExposureCol = "Exposure",
  strKRILabel = "AEs/Week"
)

dfAnalyzed <- Analyze_Wilcoxon(dfTransformed)

dfFlagged <- Flag(dfAnalyzed, strColumn = "Score", strValueColumn = "Estimate")

dfSummary <- Summarize(dfFlagged)

}
