% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.R
\name{Summarize}
\alias{Summarize}
\title{Make Summary Data Frame}
\usage{
Summarize(dfFlagged, cAssessment = "", cLabel = "")
}
\arguments{
\item{dfFlagged}{data frame in format produced by \code{\link{Flag}}}

\item{cAssessment}{brief description of current assessment}

\item{cLabel}{brief description of line item in current assessment}
}
\value{
Simplified finding data frame with columns for "SiteID", "N", "PValue", "Flag".
}
\description{
Create a concise summary of assessment results that is easy to aggregate across assessments
}
\examples{
dfInput <- AE_Map_Adam( safetyData::adam_adsl, safetyData::adam_adae )
dfTransformed <- Transform_EventCount( dfInput, cCountCol = 'Count', cExposureCol = "Exposure" )
dfAnalyzed <- Analyze_Wilcoxon( dfTransformed)
dfFlagged <- Flag( dfAnalyzed ,  strColumn = 'PValue', strValueColumn = 'Rate')
dfSummary <- Summarize(dfFlagged, cAssessment="Safety", cLabel= "")

}
