% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.R
\name{Summarize}
\alias{Summarize}
\title{Make Summary Data Frame}
\usage{
Summarize(dfFlagged, nMinDenominator = NULL)
}
\arguments{
\item{dfFlagged}{data.frame in format produced by \code{\link[=Flag]{Flag()}}.}

\item{nMinDenominator}{\code{numeric} Specifies the minimum denominator required to return a \code{score} and calculate a \code{flag}. Default: NULL}
}
\value{
Simplified finding data.frame with columns for GroupID, GroupType, Metric, Score, Flag
when associated with a workflow.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Create a concise summary of assessment results that is easy to aggregate across assessments
}
\details{
\code{Summarize} supports the input data (\code{dfFlagged}) from the \code{Flag} function.
}
\section{Data Specification}{


(\code{dfFlagged}) has the following required columns:
\itemize{
\item \code{GroupID} - Group ID
\item \code{GroupLevel} - Group Type
\item \code{Flag} - Flagging value of -1, 0, or 1
\item \code{Score} - Column from analysis results.
}
}

\examples{

dfTransformed <- Transform_Rate(analyticsInput)
dfAnalyzed <- Analyze_NormalApprox(dfTransformed)
dfFlagged <- Flag(dfAnalyzed)
dfSummary <- Summarize(dfFlagged)

}
