% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-TimeOnStudy.R
\name{TimeOnStudy}
\alias{TimeOnStudy}
\title{Utility Function to Calculate Time on Study}
\usage{
TimeOnStudy(dfVisit = NULL, dfStud = NULL, dtSnapshot = NULL)
}
\arguments{
\item{dfVisit}{data frame of visit information with required columns SUBJID INVID FOLDERNAME RECORDDATE}

\item{dfStud}{optional, if no data frame is supplied will assume no subject completed study, otherwise
requires data frame of study completion information with columns SUBJID COMPYN_STD COMPREAS}

\item{dtSnapshot}{date of data snapshot, if NULL, will impute to be current date}
}
\value{
dataframe of time on study with a column for site ID, subject ID, and a column for time on study in weeks
}
\description{
Calculates time-on-study for subjects in a study using raw visit dataset
}
