% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_Count.R
\name{Transform_Count}
\alias{Transform_Count}
\title{Transform Count}
\usage{
Transform_Count(dfInput, strCountCol, strGroupCol = "SiteID")
}
\arguments{
\item{dfInput}{A data.frame with one record per person.}

\item{strCountCol}{Required. Numerical or logical. Column to be counted.}

\item{strGroupCol}{\code{character} Name of column for grouping variable. Default: \code{"SiteID"}}
}
\value{
\code{data.frame} with one row per site with columns SiteID, TotalCount with additional columns Exposure and Rate if strExposureCol is used.
}
\description{
Convert from input data format to needed input format to derive KRI for an Assessment.
}
\details{
This function transforms data to prepare it for the Analysis step
}
\section{Data Specification}{


The input data (\code{dfInput}) for the AE Assessment is typically created using any of these functions:
\code{\link{AE_Map_Raw}}
\code{\link{AE_Map_Adam}}
\code{\link{Consent_Map_Raw}}
\code{\link{Disp_Map_Raw}}
\code{\link{IE_Map_Raw}}
\code{\link{LB_Map_Raw}}
\code{\link{PD_Map_Raw}}

(\code{dfInput}) has the following required and optional columns:
Required:
\itemize{
\item \code{SiteID} - Site ID
\item \code{StudyID} - Study ID
\item \code{CustomGroupID} - Custom Group ID, currently implemented as a placeholder for Country ID.
\item \code{Count} - Number of Adverse Events the actual name of this column is specified by the parameter strCountCol.
Optional
\item \code{Exposure} - Number of days of exposure, name specified by strExposureCol.
}

The input data has one or more rows per site. Transform_Count sums strCountCol for a TotalCount for each site.
For data with an optional strExposureCol, a summed exposure is calculated for each site.
}

\examples{
dfInput <- Disp_Map_Raw()
dfTransformed <- Transform_Count(dfInput, strCountCol = "Count")

}
