% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_Count.R
\name{Transform_Count}
\alias{Transform_Count}
\title{Transform Count}
\usage{
Transform_Count(dfInput, strCountCol, strGroupCol = "SiteID")
}
\arguments{
\item{dfInput}{A data.frame with one record per subject.}

\item{strCountCol}{Required. Numerical or logical. Column to be counted.}

\item{strGroupCol}{\code{character} Name of column for grouping variable. Default: \code{"SiteID"}}
}
\value{
\code{data.frame} with one row per site with columns \code{GroupID}, \code{TotalCount}, and \code{Metric.}
}
\description{
Convert from input data format to needed input format to derive KRI for an Assessment. Calculate site-level count.
}
\details{
This function transforms data to prepare it for the analysis step. It is currently only sourced for the Consent and IE Assessments.
}
\section{Data Specification}{


The input data (\code{dfInput}) for the Consent and IE Assessments is typically created using any of these functions:
\code{\link{Consent_Map_Raw}}
\code{\link{IE_Map_Raw}}

(\code{dfInput}) must include the columns specified by \code{strCountCol} and \code{strGroupCol}.
Required columns include:
\itemize{
\item \code{SiteID} - Site ID
\item \code{StudyID} - Study ID
\item \code{CustomGroupID} - Custom Group ID
\item \code{Count} - Number of events of interest; the actual name of this column is specified by the parameter \code{strCountCol.}
}

The input data has one or more rows per site. \code{Transform_Count()} sums \code{strCountCol} for a \code{TotalCount} for each site. \code{Metric} is set to \code{TotalCount} to be used downstream in the workflow.
}

\examples{
dfInput <- Consent_Map_Raw()
dfTransformed <- Transform_Count(dfInput, strCountCol = "Count")

}
