% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_EventCount.R
\name{Transform_EventCount}
\alias{Transform_EventCount}
\title{Transform_EventCount}
\usage{
Transform_EventCount(dfInput)
}
\arguments{
\item{dfInput}{input data frame with one record per person containing the following columns - SiteID, Count, Exposure, Unit}
}
\value{
dataframe one record per site and columns for Site ("SiteID"), Number of participants ("N"), Number of events ("TotalCount"),  Total Exposure ("TotalExposure"), Exposure per person ("Rate"), Unit of measure for exposure ("Unit").
}
\description{
Transform_EventCount
}
