% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_EventCount.R
\name{Transform_EventCount}
\alias{Transform_EventCount}
\title{Transform Event Count}
\usage{
Transform_EventCount(dfInput, cCountCol, cExposureCol = NULL)
}
\arguments{
\item{dfInput}{A data frame with one Record per person}

\item{cCountCol}{required. numerical or logical. Column to be counted.}

\item{cExposureCol}{Optional, numerical Exposure Column}
}
\description{
Transform Event Count
}
