% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_EventCount.R
\name{Transform_EventCount}
\alias{Transform_EventCount}
\title{Transform Event Count}
\usage{
Transform_EventCount(dfInput, cCountCol, cExposureCol = NULL, cUnitCol = NULL)
}
\arguments{
\item{dfInput}{A data frame with one Record per person}

\item{cCountCol}{required. Column to be counted. "InvalidorMissing" = Invalid>0 | Missing > 0 for IE Assessment.}

\item{cExposureCol}{Optional, Exposure Column}

\item{cUnitCol}{Optional, Unit Column (currently unused)}
}
\description{
Transform Event Count
}
