% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_EventCount.R
\name{Transform_EventCount}
\alias{Transform_EventCount}
\title{Transform Event Count}
\usage{
Transform_EventCount(dfInput, strCountCol, strExposureCol = NULL, strKRILabel)
}
\arguments{
\item{dfInput}{A data.frame with one record per person.}

\item{strCountCol}{Required. Numerical or logical. Column to be counted.}

\item{strExposureCol}{Optional. Numerical \code{Exposure} column.}

\item{strKRILabel}{Optional. Character vector to describe the \code{KRI} column.}
}
\value{
data.frame with one row per site with columns SiteID, N, TotalCount with additional columns Exposure and Rate if strExposureCol is used.
}
\description{
Convert from ADaM format to needed input format for Safety Assessment
}
\details{
This function transforms data to prepare it for the Analysis step
}
\section{Data Specification}{


The input data (\code{dfInput}) for the AE Assessment is typically created using any of these functions:
\code{\link{AE_Map_Raw}}
\code{\link{AE_Map_Adam}}
\code{\link{PD_Map_Raw}}
\code{\link{IE_Map_Raw}}
\code{\link{Consent_Map_Raw}}

(\code{dfInput}) has the following required and optional columns:
Required:
\itemize{
\item \code{SiteID} - Site ID
\item \code{Count} - Number of Adverse Events the actual name of this column is specified by the parameter strCountCol.
Optional
\item \code{Exposure} - Number of days of exposure, name specified by strExposureCol.
}

The input data has one or more rows per site. Transform_EventCount sums strCountCol for a TotalCount for each site.
For data with an optional strExposureCol, a summed exposure is calculated for each site.
}

\examples{
dfInput <- AE_Map_Adam()
dfTransformed <- Transform_EventCount(dfInput, strCountCol = "Count", strExposureCol = "Exposure", strKRILabel = "AEs/Week")

}
