% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_Rate.R
\name{Transform_Rate}
\alias{Transform_Rate}
\title{Transform - Rate}
\usage{
Transform_Rate(
  dfInput,
  strNumeratorCol,
  strDenominatorCol = NULL,
  strGroupCol = "SiteID"
)
}
\arguments{
\item{dfInput}{A data.frame with one record per person.}

\item{strNumeratorCol}{Required. Numerical or logical. Column to be counted.}

\item{strDenominatorCol}{Optional. Numerical \code{Exposure} column.}

\item{strGroupCol}{\code{character} Required. Name of column for grouping variable. Default: \code{"SiteID"}}
}
\value{
\code{data.frame} with one row per site with columns SiteID, TotalCount with additional columns Exposure and Rate if strExposureCol is used.
}
\description{
Calculate a site-level Rate
}
\details{
This function transforms data to prepare it for the Analysis step
}
\section{Data Specification}{


The input data (\code{dfInput}) for the AE Assessment is typically created using any of these functions:
\code{\link{AE_Map_Raw}}
\code{\link{AE_Map_Adam}}
\code{\link{Disp_Map_Raw}}
\code{\link{LB_Map_Raw}}
\code{\link{PD_Map_Raw}}

(\code{dfInput}) must include the columns specified by \code{strNumeratorCol}, \code{strDenominatorCol} and \code{strGroupCol}
}

\examples{
dfInput <- AE_Map_Raw()
dfTransformed <- Transform_Rate(dfInput,
                                strNumeratorCol = "Count",
                                strDenominatorCol = "Exposure")

}
