% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_Rate.R
\name{Transform_Rate}
\alias{Transform_Rate}
\title{Transform Rate}
\usage{
Transform_Rate(
  dfInput,
  strNumeratorCol = "Numerator",
  strDenominatorCol = "Denominator",
  strGroupCol = "SiteID"
)
}
\arguments{
\item{dfInput}{A data.frame with one record per subject.}

\item{strNumeratorCol}{Required. Numerical or logical. Column to be counted.}

\item{strDenominatorCol}{\code{numeric} Required. Numerical \code{Exposure} column.}

\item{strGroupCol}{\code{character} Required. Name of column for grouping variable. Default: \code{"SiteID"}}
}
\value{
\code{data.frame} with one row per site with columns \code{GroupID}, \code{Numerator}, \code{Denominator}, and \code{Metric}.
}
\description{
Convert from input data format to needed input format to derive KRI for an Assessment. Calculate a site-level rate.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function transforms data to prepare it for the analysis step. It is currently only sourced for the Adverse Event, Disposition, Labs, and Protocol Deviations Assessments.
}
\section{Data Specification}{


(\code{dfInput}) must include the columns specified by \code{strNumeratorCol}, \code{strDenominatorCol} and \code{strGroupCol}.
Required columns include:
\itemize{
\item \code{SiteID} - Site ID
\item \code{StudyID} - Study ID
\item \code{CustomGroupID} - Custom Group ID
\item \code{Count} - Number of events of interest; the actual name of this column is specified by the parameter \code{strNumeratorCol}
\item \code{Exposure} - Number of days on treatment; the actual name of this column is specified by the parameter \code{strDenominatorCol}
}
}

\examples{
dfInput <- tibble::tribble(
  ~SubjectID, ~SiteID, ~StudyID, ~CountryID, ~CustomGroupID, ~Exposure, ~Count, ~Rate,
  "0496", "5", "AA-AA-000-0000", "US", "0X167", 730, 5, 5/720,
  "1350", "78", "AA-AA-000-0000", "US", "0X002", 50, 2, 2/50,
  "0539", "139", "AA-AA-000-0000", "US", "0X052", 901, 5, 5/901,
  "0329", "162", "AA-AA-000-0000", "US", "0X049", 370, 3, 3/370,
  "0429", "29", "AA-AA-000-0000", "Japan", "0X116", 450, 2, 2/450,
  "1218", "143", "AA-AA-000-0000", "US", "0X153", 170, 3, 3/170,
  "0808", "173", "AA-AA-000-0000", "US", "0X124", 680, 6, 6/680,
  "1314", "189", "AA-AA-000-0000", "US", "0X093", 815, 4, 4/815,
  "1236", "58", "AA-AA-000-0000", "China", "0X091", 225, 1, 1/225,
  "0163", "167", "AA-AA-000-0000", "US", "0X059", 360, 3, 3/360
)

dfTransformed <- Transform_Rate(dfInput,
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)

}
