% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_Rate.R
\name{Transform_Rate}
\alias{Transform_Rate}
\title{Transform Rate}
\usage{
Transform_Rate(
  dfInput,
  strNumeratorCol = "Numerator",
  strDenominatorCol = "Denominator"
)
}
\arguments{
\item{dfInput}{\code{data.frame} Input data with one record per subject. Created
by passing Raw+ data into \code{\link[=Input_Rate]{Input_Rate()}}. Expected columns: \code{GroupID},
\code{GroupLevel}, \code{Numerator}, \code{Denominator} and/or columns specified in
\code{strCountCol} and \code{strGroupCol}.}

\item{strNumeratorCol}{\code{string} Column to be counted. Defaults to "Numerator".}

\item{strDenominatorCol}{\code{string} Column name for the numerical \code{Exposure} column. Defaults to "Denominator".}
}
\value{
\code{data.frame} with one row per site with columns \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, and \code{Metric}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Convert from input data format to needed input format to derive KRI for an Assessment. Calculate a site-level rate.
}
\details{
This function transforms data to prepare it for the analysis step. It is currently only sourced for the Adverse Event, Disposition, Labs, and Protocol Deviations Assessments.
}
\section{Data Specification}{


(\code{dfInput}) must include the columns specified by \code{strNumeratorCol} and \code{strDenominatorCol}.
Required columns include:
\itemize{
\item \code{GroupID} - Group ID
\item \code{GroupLevel} - Group Type
\item \code{Numerator} - Number of events of interest; the actual name of this column is specified by the parameter \code{strNumeratorCol}
\item \code{Denominator} - Number of days on treatment; the actual name of this column is specified by the parameter \code{strDenominatorCol}
}
}

\examples{
dfTransformed <- Transform_Rate(
  analyticsInput,
  strNumeratorCol = "Numerator",
  strDenominatorCol = "Denominator"
)

}
