% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-TreatmentExposure.R
\name{TreatmentExposure}
\alias{TreatmentExposure}
\title{Utility Function to Calculate Treatment Exposure}
\usage{
TreatmentExposure(dfEx = NULL, dfSdrg = NULL, dtSnapshot = NULL)
}
\arguments{
\item{dfEx}{data frame of treatment information with required columns SUBJID INVID EXSTDAT EXENDAT. If
multiple treatments and only want to focus on one treatment then input data frame will need to be subset
by user prior to input}

\item{dfSdrg}{optional, if no data frame is supplied will assume no subject completed treatment, otherwise
requires data frame of treatment completion information with columns SUBJID SDRGYN_STD. If multiple treatments
and only want to focus on one treatment then input data frame will need to be subset
by user prior to input}

\item{dtSnapshot}{date of data snapshot, if NULL, will impute to be current date}
}
\value{
dataframe of time on study with a column for site ID, subject ID, and a column for time on study in days
}
\description{
Calculates treatment exposure duration for subjects in a study using raw ex dataset
}
\details{
This output of this function is used by assessments as a standardized measurement of time on treatment.
}
\section{Data Specification}{


The following columns are required:
\itemize{
\item \code{dfEX}
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{INVID} - Unique Investigator ID
\item \code{EXSTDAT} - Start date, dose date
\item \code{EXENDAT} - Stop date
}
}

The following columns are optional
\itemize{
\item \code{dfSdrg}
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{SDRGYN_STD} - Y/N Did subject complete study drug closing
}
}
}

\examples{

dfTos <- TreatmentExposure(dfEx = clindata::raw_ex, dfSdrg = clindata::raw_sdrgcom2)

}
