% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParams.R
\name{UpdateParams}
\alias{UpdateParams}
\title{Update parameters for workflow}
\usage{
UpdateParams(lWorkflow, dfConfig, dfMeta)
}
\arguments{
\item{lWorkflow}{\code{list} A named list of metadata defining how the workflow should be run.}

\item{dfConfig}{\code{data.frame} Configuration parameters data.frame; \code{gsm::config_param}}

\item{dfMeta}{\code{data.frame} Metadata parameters data.frame; \code{gsm::meta_param}}
}
\value{
\code{list} lWorkflow - modified list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
