% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParams.R
\name{UpdateParams}
\alias{UpdateParams}
\title{Title}
\usage{
UpdateParams(lWorkflow, dfConfig, dfMeta)
}
\arguments{
\item{lWorkflow}{\code{list} Object returned by \code{MakeWorkflowList()}}

\item{dfConfig}{\code{data.frame} Configuration parameters data.frame; \code{clindata::config_param}}

\item{dfMeta}{\code{data.frame} Metadata parameters data.frame; \code{gsm::meta_param}}
}
\value{
\code{list} lWorkflow - modified list.
}
\description{
Title
}
