% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-templates.R
\name{use_gsm_test}
\alias{use_gsm_test}
\title{Create testthat skeleton}
\usage{
use_gsm_test(mapping = NULL)
}
\arguments{
\item{mapping}{coming soon...}
}
\description{
Create testthat skeleton
}
\examples{
\dontrun{
use_gsm_test() # default is "map"
use_gsm_test("ae_map")
use_gsm_test("ie_map")
}
}
