% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Count.R
\name{Visualize_Count}
\alias{Visualize_Count}
\title{Site-level visualization of site-level Inclusion/Exclusion results}
\usage{
Visualize_Count(
  dfAnalyzed,
  strTotalCol = "N",
  strCountCol = "TotalCount",
  strTitle = ""
)
}
\arguments{
\item{dfAnalyzed}{Map results from IE or Consent assessments.}

\item{strTotalCol}{Column containing total of site-level participants. Default is "N" from \code{\link{Transform_EventCount}}.}

\item{strCountCol}{Column containing total number of site-level occurrences. Default is "TotalCount" from \code{\link{Transform_EventCount}}.}

\item{strTitle}{Title of plot. NULL by default.}
}
\value{
site level plot object
}
\description{
Site-level visualization of site-level Inclusion/Exclusion results
}
\examples{

IE_Input <- IE_Map_Raw(
   clindata::rawplus_ie,
   clindata::rawplus_subj,
   vCategoryValues= c("EXCL","INCL"),
   vExpectedResultValues=c(0,1)
)

IE_Assess <- IE_Assess(IE_Input)

Visualize_Count(IE_Assess$dfAnalyzed)

Consent_Input <- Consent_Map_Raw(
   dfCONSENT = clindata::rawplus_consent,
   dfSUBJ = clindata::rawplus_subj,
   strConsentTypeValue = "MAINCONSENT",
   strConsentStatusValue="Y"
)

Consent_Assess <- Consent_Assess(Consent_Input)
Visualize_Count(Consent_Assess$dfAnalyzed)

}
