% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Count.R
\name{Visualize_Count}
\alias{Visualize_Count}
\title{Site-level visualization of site-level Inclusion/Exclusion results}
\usage{
Visualize_Count(
  dfAnalyzed,
  strTotalCol = "N",
  strCountCol = "TotalCount",
  strFlagCol = "Flag"
)
}
\arguments{
\item{dfInput}{Map results from IE or Consent assessments.}

\item{strTitle}{Title of plot. NULL by default.}
}
\value{
site level plot object
}
\description{
Site-level visualization of site-level Inclusion/Exclusion results
}
\examples{
ie_input <- IE_Map_Raw(
   clindata::raw_ie_all , 
   clindata::rawplus_rdsl,
   strCategoryCol = 'IECAT_STD', 
   strResultCol = 'IEORRES'
)
ie_assess <- IE_Assess(ie_input, bDataList=TRUE)
VisualizeCount(ie_assess$dfAnalyzed)

consent_input <- Consent_Map_Raw(
  dfConsent = clindata::raw_consent,
  dfRDSL = clindata::rawplus_rdsl,
  strConsentReason = NULL
)

consent_assess<-Consent_Assess(dfInput, bDataList=TRUE)
VisualizeCount(consent_assess$dfAnalyzed)

}
