% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Count.R
\name{Visualize_Count}
\alias{Visualize_Count}
\title{Site-level visualization of site-level Inclusion/Exclusion results}
\usage{
Visualize_Count(dfInput, strTitle = NULL)
}
\arguments{
\item{dfInput}{Map results from IE or Consent assessments.}

\item{strTitle}{Title of plot. NULL by default.}
}
\value{
site level plot object
}
\description{
Site-level visualization of site-level Inclusion/Exclusion results
}
\examples{
dfInput <- IE_Map_Raw(clindata::raw_ie_all,
                      clindata::rawplus_rdsl,
                      strCategoryCol = "IECAT_STD",
                      strResultCol = "IEORRES")

}
