% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Count.R
\name{Visualize_Count}
\alias{Visualize_Count}
\title{Group-level visualization of group-level Inclusion/Exclusion results}
\usage{
Visualize_Count(
  dfAnalyzed,
  strGroupCol = NULL,
  strTotalCol = "N",
  strCountCol = "TotalCount",
  strTitle = ""
)
}
\arguments{
\item{dfAnalyzed}{Map results from IE or Consent assessments.}

\item{strGroupCol}{name of stratification column for facet wrap (default=NULL)}

\item{strTotalCol}{Column containing total of group-level participants. Default is "N" from \code{\link{Transform_EventCount}}.}

\item{strCountCol}{Column containing total number of group-level occurrences. Default is "TotalCount" from \code{\link{Transform_EventCount}}.}

\item{strTitle}{Title of plot. NULL by default.}
}
\value{
group-level plot object.
}
\description{
Group-level visualization of group-level Inclusion/Exclusion results
}
\examples{
IE_Input <- IE_Map_Raw()
IE_Assess <- IE_Assess(IE_Input)
Visualize_Count(IE_Assess$dfAnalyzed)

Consent_Input <- Consent_Map_Raw()
Consent_Assess <- Consent_Assess(Consent_Input)
Visualize_Count(Consent_Assess$dfAnalyzed)

}
