% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Count.R
\name{Visualize_Count}
\alias{Visualize_Count}
\title{Site-level visualization of site-level Inclusion/Exclusion results}
\usage{
Visualize_Count(
  dfAnalyzed,
  strTotalCol = "N",
  strCountCol = "TotalCount",
  strTitle = ""
)
}
\arguments{
\item{dfAnalyzed}{Map results from IE or Consent assessments.}

\item{strTotalCol}{Column containing total of site-level participants. Default is "N" from \code{\link{Transform_EventCount}}.}

\item{strCountCol}{Column containing total number of site-level occurrences. Default is "TotalCount" from \code{\link{Transform_EventCount}}.}

\item{strTitle}{Title of plot. NULL by default.}
}
\value{
site level plot object
}
\description{
Site-level visualization of site-level Inclusion/Exclusion results
}
\examples{

IE_Input <- IE_Map_Raw(
   clindata::raw_ie_all \%>\% dplyr::filter(SUBJID != "" ),
   clindata::rawplus_rdsl,
   vCategoryValues= c("EXCL","INCL"),
   vExpectedResultValues=c(0,1)
)

IE_Assess <- IE_Assess(IE_Input)

Visualize_Count(IE_Assess$dfAnalyzed)

library(dplyr)
raw_consent <- clindata::raw_ic_elig \%>\%
   select( c("SUBJID","DSSTDAT_RAW") )\%>\%
   mutate( CONSCAT_STD = "MAINCONSENT", CONSYN="Y") \%>\%
   rename( CONSDAT = DSSTDAT_RAW ) \%>\%
   mutate( CONSDAT = as.Date(CONSDAT, format="\%d \%B \%Y") ) \%>\%
   filter(SUBJID != "")

Consent_Input <- Consent_Map_Raw(
   dfConsent = raw_consent,
   dfRDSL = clindata::rawplus_rdsl,
   strConsentTypeValue = "MAINCONSENT",
   strConsentStatusValue="Y"
)

Consent_Assess <- Consent_Assess(Consent_Input)
Visualize_Count(Consent_Assess$dfAnalyzed)

}
