% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_KRI.R
\name{Visualize_KRI}
\alias{Visualize_KRI}
\title{Visualize_KRI Function}
\usage{
Visualize_KRI(
  dfSummary,
  dfSite = NULL,
  dfBounds = NULL,
  dfParams = NULL,
  lLabels = NULL,
  strSnapshotDate = NULL
)
}
\arguments{
\item{dfSummary}{\code{data.frame} A data.frame returned by \code{\link[=Summarize]{Summarize()}}.}

\item{dfSite}{\code{data.frame} Site metadata returned by \code{\link[=Site_Map_Raw]{Site_Map_Raw()}}.}

\item{dfBounds}{\code{data.frame}, A data.frame returned by \code{\link[=Analyze_NormalApprox_PredictBounds]{Analyze_NormalApprox_PredictBounds()}} or \code{\link[=Analyze_Poisson_PredictBounds]{Analyze_Poisson_PredictBounds()}}}

\item{lLabels}{\code{list} Workflow metadata. See \link{meta_workflow}.}
}
\value{
A list (lCharts) containing three charts - scatterJS, barMetricJS, and barScoreJS.
}
\description{
The function Visualize_KRI creates three different types of charts (scatter plot and two bar charts) using the gsm package.
}
