% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Metric.R
\name{Visualize_Metric}
\alias{Visualize_Metric}
\title{Visualize_Metric Function}
\usage{
Visualize_Metric(
  dfResults = dfResults,
  dfBounds = NULL,
  dfGroups = NULL,
  dfMetrics = NULL,
  strMetricID = NULL,
  strSnapshotDate = NULL,
  bDebug = FALSE
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[=BindResults]{BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID}, \code{SnapshotDate}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and reporting. Created by passing an \code{lWorkflow} object to \code{\link[=MakeMetric]{MakeMetric()}}. Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric}, \code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{strThreshold}. For more details see the Data Pipeline vignette: \code{vignette("DataPipeline", package = "gsm")}.}

\item{strMetricID}{\code{character} MetricID to subset the data.}

\item{strSnapshotDate}{\code{character} Snapshot date to subset the data.}

\item{bDebug}{\code{logical} Display console in html viewer for debugging. Default is \code{FALSE}.}
}
\value{
A list containing the following charts:
\itemize{
\item scatterJS: A scatter plot using JavaScript.
\item scatter: A scatter plot using ggplot2.
\item barMetricJS: A bar chart using JavaScript with metric on the y-axis.
\item barScoreJS: A bar chart using JavaScript with score on the y-axis.
\item barMetric: A bar chart using ggplot2 with metric on the y-axis.
\item barScore: A bar chart using ggplot2 with score on the y-axis.
\item timeSeriesContinuousScoreJS: A time series chart using JavaScript with score on the y-axis.
\item timeSeriesContinuousMetricJS: A time series chart using JavaScript with metric on the y-axis.
\item timeSeriesContinuousNumeratorJS: A time series chart using JavaScript with numerator on the y-axis.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
The function creates all available charts for a metric using the data provided
}
\examples{
 \dontrun{
 charts <- Visualize_Metric(
   dfResults = reportingResults,
   dfBounds = reportingBounds,
   dfGroups = reportingGroups,
   dfMetrics = reportingMetrics,
   strMetricID = "kri0001"
 )
 }

}
