% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Metric.R
\name{Visualize_Metric}
\alias{Visualize_Metric}
\title{Visualize_Metric Function}
\usage{
Visualize_Metric(
  dfResults = dfResults,
  dfBounds = NULL,
  dfGroups = NULL,
  dfMetrics = NULL,
  strMetricID = NULL,
  strSnapshotDate = NULL,
  bDebug = FALSE
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to
\code{\link[gsm.mapping:BindResults]{gsm.mapping::BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator},
\code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID},
\code{SnapshotDate}.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting. Created by passing \code{dfResults} and \code{dfMetrics} to
\code{\link[gsm.mapping:MakeBounds]{gsm.mapping::MakeBounds()}}. Expected columns: \code{Threshold}, \code{Denominator}, \code{Numerator},
\code{Metric}, \code{MetricID}, \code{StudyID}, \code{SnapshotDate}.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. Created by
passing CTMS site and study data to \code{\link[gsm.mapping:MakeLongMeta]{gsm.mapping::MakeLongMeta()}}. Expected columns:
\code{GroupID}, \code{GroupLevel}, \code{Param}, \code{Value}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting. Created by passing an \code{lWorkflow} object to \code{\link[gsm.mapping:MakeMetric]{gsm.mapping::MakeMetric()}}.
Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric},
\code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{Threshold}. For more
details see the Data Model vignette: \code{vignette("DataModel", package = "gsm")}.}

\item{strMetricID}{\code{character} MetricID to subset the data.}

\item{strSnapshotDate}{\code{character} Snapshot date to subset the data.}

\item{bDebug}{\code{logical} Display console in html viewer for debugging. Default is \code{FALSE}.}
}
\value{
A list containing the following charts:
\itemize{
\item scatterPlot: A scatter plot using JavaScript.
\item barChart: A bar chart using JavaScript with metric on the y-axis.
\item timeSeries: A time series chart using JavaScript with score on the y-axis.
\item metricTable: A table containing all
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The function creates all available charts for a metric using the data provided
}
\examples{
lCharts <- Visualize_Metric(
  dfResults = reportingResults,
  dfBounds = reportingBounds,
  dfGroups = reportingGroups,
  dfMetrics = reportingMetrics,
  strMetricID = "Analysis_kri0001"
)

}
