% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Scatter.R
\name{Visualize_Scatter}
\alias{Visualize_Scatter}
\title{Site-level visualization of site-level results using a Poisson or Wilcoxon model.}
\usage{
Visualize_Scatter(dfFlagged, dfBounds = NULL, strUnit = "days")
}
\arguments{
\item{dfFlagged}{analyze_poisson results with flags added.}

\item{dfBounds}{data.frame giving prediction bounds for range of dfFlagged.}

\item{strUnit}{exposure time unit. Defaults to "days".}
}
\value{
site-level plot object.
}
\description{
Site-level visualization of site-level results using a Poisson or Wilcoxon model.
}
\examples{
dfInput <- AE_Map_Adam()
SafetyAE <- AE_Assess(dfInput)
dfBounds <- Analyze_Poisson_PredictBounds(SafetyAE$dfTransformed, c(-5, 5))
Visualize_Scatter(SafetyAE$dfFlagged, dfBounds)

SafetyAE_wilk <- AE_Assess(dfInput, strMethod = "wilcoxon")
Visualize_Scatter(SafetyAE_wilk$dfFlagged)

}
