% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Scatter.R
\name{Visualize_Scatter}
\alias{Visualize_Scatter}
\title{Group-level visualization of group-level results using a Poisson or Wilcoxon model.}
\usage{
Visualize_Scatter(
  dfFlagged,
  dfBounds = NULL,
  strGroupCol = NULL,
  strGroupLabel = NULL,
  strUnit = "days"
)
}
\arguments{
\item{dfFlagged}{analyze_poisson results with flags added.}

\item{dfBounds}{data.frame giving prediction bounds for range of dfFlagged.}

\item{strGroupCol}{name of stratification column for facet wrap (default=NULL)}

\item{strGroupLabel}{name of group, used for labelling axes.}

\item{strUnit}{exposure time unit. Defaults to "days".}
}
\value{
group-level plot object.
}
\description{
Group-level visualization of group-level results using a Poisson or Wilcoxon model.
}
\examples{
dfInput <- AE_Map_Adam()
SafetyAE <- AE_Assess(dfInput)
dfBounds <- Analyze_Poisson_PredictBounds(SafetyAE$dfTransformed, c(-5, 5))
Visualize_Scatter(SafetyAE$dfFlagged, dfBounds)

SafetyAE_wilk <- AE_Assess(dfInput, strMethod = "wilcoxon")
Visualize_Scatter(SafetyAE_wilk$dfFlagged)

# TODO: add stratified example

}
