% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Score.R
\name{Visualize_Score}
\alias{Visualize_Score}
\title{Group-level visualization of scores.}
\usage{
Visualize_Score(
  dfFlagged,
  vThreshold = NULL,
  strType = "metric",
  bFlagFilter = FALSE,
  strTitle = ""
)
}
\arguments{
\item{dfFlagged}{\code{data.frame} returned by \code{\link[=Flag]{Flag()}}}

\item{vThreshold}{\code{numeric} Threshold specification, a vector of length 2 that defaults to NULL.}

\item{strType}{\code{character} One of \code{"KRI"} or \code{"score"}.}

\item{bFlagFilter}{\code{logical} Filter out non-flagged groups? Default: \code{FALSE}}

\item{strTitle}{Title of plot. NULL by default.}
}
\value{
group-level plot object.
}
\description{
Group-level visualization of scores.
}
\examples{
ae <- AE_Map_Raw() \%>\% AE_Assess(vThreshold = c(-8, 8))

Visualize_Score(ae$lData$dfFlagged) # visualize KRI (default)
Visualize_Score(ae$lData$dfFlagged, bFlagFilter = TRUE) # drop non-flagged groups

consent <- Consent_Map_Raw() \%>\% Consent_Assess()
Visualize_Score(consent$lData$dfFlagged, strType = "score") # visualize score

}
