% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Score.R
\name{Visualize_Score}
\alias{Visualize_Score}
\title{Group-level visualization of scores.}
\usage{
Visualize_Score(
  dfSummary,
  vThreshold = NULL,
  strType = "metric",
  bFlagFilter = FALSE,
  strTitle = ""
)
}
\arguments{
\item{dfSummary}{\code{data.frame} returned by \code{\link[=Summarize]{Summarize()}}}

\item{vThreshold}{\code{numeric} Threshold specification, a vector of length 2 that defaults to NULL.}

\item{strType}{\code{character} One of \code{"KRI"} or \code{"score"}.}

\item{bFlagFilter}{\code{logical} Filter out non-flagged groups? Default: \code{FALSE}}

\item{strTitle}{\code{character} Title of plot.}
}
\value{
group-level ggplot2 object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
ae <- AE_Map_Raw() \%>\% AE_Assess()

Visualize_Score(ae$lData$dfSummary) # visualize KRI (default)
Visualize_Score(ae$lData$dfSummary, bFlagFilter = TRUE) # drop non-flagged groups

consent <- Consent_Map_Raw() \%>\% Consent_Assess()
Visualize_Score(consent$lData$dfSummary, strType = "score") # visualize score

}
