% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Score.R
\name{Visualize_Score}
\alias{Visualize_Score}
\title{Group-level visualization of scores.}
\usage{
Visualize_Score(
  dfSummary,
  vThreshold = NULL,
  strType = "metric",
  bFlagFilter = FALSE,
  strTitle = ""
)
}
\arguments{
\item{dfSummary}{\code{data.frame} returned by \code{\link[=Summarize]{Summarize()}}}

\item{vThreshold}{\code{numeric} Threshold specification, a vector of length 2 that defaults to NULL.}

\item{strType}{\code{character} One of \code{"KRI"} or \code{"score"}.}

\item{bFlagFilter}{\code{logical} Filter out non-flagged groups? Default: \code{FALSE}}

\item{strTitle}{\code{character} Title of plot.}
}
\value{
group-level ggplot2 object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
lData <- gsm::UseClindata(
list(
  "dfSUBJ" = "clindata::rawplus_dm",
  "dfAE" = "clindata::rawplus_ae",
  "dfPD" = "clindata::ctms_protdev",
  "dfCONSENT" = "clindata::rawplus_consent",
  "dfIE" = "clindata::rawplus_ie",
  "dfLB" = "clindata::rawplus_lb",
  "dfSTUDCOMP" = "clindata::rawplus_studcomp",
  "dfSDRGCOMP" = "clindata::rawplus_sdrgcomp \%>\%
          dplyr::filter(.data$phase == 'Blinded Study Drug Completion')",
  "dfDATACHG" = "clindata::edc_data_points",
  "dfDATAENT" = "clindata::edc_data_pages",
  "dfQUERY" = "clindata::edc_queries",
  "dfENROLL" = "clindata::rawplus_enroll"
)
)
wf_mapping <- MakeWorkflowList("mapping")
ae_workflow <- MakeWorkflowList(strNames = "kri0001")
lMapped <- RunWorkflow(wf_mapping, lData)$mapping$lResults
AE <- RunWorkflow(ae_workflow, lMapped)
Visualize_Score(AE$kri0001$lResults$dfSummary, dfBounds)
}
}
