% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Score.R
\name{Visualize_Score}
\alias{Visualize_Score}
\title{Group-level visualization of scores.}
\usage{
Visualize_Score(
  dfResults,
  vThreshold = NULL,
  strType = "Metric",
  bFlagFilter = FALSE,
  strTitle = ""
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to
\code{\link[gsm.mapping:BindResults]{gsm.mapping::BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator},
\code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID},
\code{SnapshotDate}.}

\item{vThreshold}{\code{numeric} Threshold specification, a vector of length 2 that defaults to NULL.}

\item{strType}{\code{character} One of \code{"KRI"} or \code{"score"}.}

\item{bFlagFilter}{\code{logical} Filter out non-flagged groups? Default: \code{FALSE}}

\item{strTitle}{\code{character} Title of plot.}
}
\value{
group-level ggplot2 object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
## Filter data to one metric and snapshot
reportingResults_filter <- reportingResults \%>\%
  dplyr::filter(MetricID == "Analysis_kri0001" & SnapshotDate == max(SnapshotDate))

Visualize_Score(dfResults = reportingResults_filter)

## Only show Flagged Groups
Visualize_Score(
  dfResults = reportingResults_filter,
  bFlagFilter = TRUE
)

## Custom Title
Visualize_Score(
  dfResults = reportingResults_filter,
  strTitle = "Custom Title"
)

}
