% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_StudyOverview.R
\name{Visualize_StudyOverview}
\alias{Visualize_StudyOverview}
\title{Title}
\usage{
Visualize_StudyOverview(
  dfResults,
  dfMetrics,
  dfGroups,
  strGroupLevel = NULL,
  strGroupSubset = "red",
  strGroupLabelKey = "InvestigatorLastName",
  bDebug = FALSE
)
}
\arguments{
\item{dfResults}{a data.frame}

\item{dfMetrics}{a data.frame}

\item{dfGroups}{a data.frame}

\item{strGroupLevel}{a character}

\item{strGroupSubset}{a character}

\item{strGroupLabelKey}{a character}

\item{bDebug}{a logical}
}
\value{
a list of visualizations
}
\description{
Title
}
