% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Workflow.R
\name{Visualize_Workflow}
\alias{Visualize_Workflow}
\title{Flowchart visualization of data pipeline steps from filtering to summary data for an assessment workflow.}
\usage{
Visualize_Workflow(lAssessments)
}
\arguments{
\item{lAssessments}{\code{list} A list of assessment-specific metadata.}
}
\value{
A flowchart of type \code{grViz}/\code{htmlwidget}.
}
\description{
Flowchart visualization of data pipeline steps from filtering to summary data for an assessment workflow.
}
\examples{
lAssessments <- MakeAssessmentList()
lData <- list(
  dfSUBJ = clindata::rawplus_dm,
  dfAE = clindata::rawplus_ae,
  dfPD = clindata::rawplus_protdev,
  dfCONSENT = clindata::rawplus_consent,
  dfIE = clindata::rawplus_ie
)
lTags <- list(
  Study = "myStudy"
)
lMapping <- yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package = "gsm"))

ae_assessment <- RunAssessment(lAssessments$ae, lData = lData, lMapping = lMapping, lTags = lTags)

Visualize_Workflow(list(ae = ae_assessment))

}
