% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_BarChart.R
\name{Widget_BarChart}
\alias{Widget_BarChart}
\title{KRI Bar Chart}
\usage{
Widget_BarChart(
  dfSummary,
  lLabels,
  dfSite = NULL,
  dfThreshold = NULL,
  strYAxisType = "score",
  selectedGroupIDs = NULL,
  addSiteSelect = TRUE,
  bHideDropdown = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dfSummary}{data with column names:
\itemize{
\item{\code{studyid}}
\item{\code{workflowid}}
\item{\code{groupid}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{metric}}
\item{\code{score}}
\item{\code{flag}}
}}

\item{lLabels}{configuration data with columns:}

\item{dfSite}{\code{data.frame} Site metadata.}

\item{dfThreshold}{a one row data frame containing columns:
\itemize{
\item{\code{workflowid}}
\item{\code{gsm_version}}
\item{\code{param}}
\item{\code{index}}
\item{\code{default}}
\item{\code{configurable}}
}}

\item{strYAxisType}{either \code{'score'} or \code{'metric'}}

\item{selectedGroupIDs}{group IDs to highlight, \code{NULL} by default, can be a single site or a vector.}

\item{addSiteSelect}{\code{logical} add a dropdown to highlight sites? Default: \code{TRUE}.}

\item{bHideDropdown}{\code{logical} should the dropdown be hidden? Default: \code{FALSE}. This is primarily used for the Shiny app \code{{gsmApp}} to hide the drop-down site selector,
since it is an additional element that needs to be updated based on user interactivity.}

\item{width}{width of widget, full screen by default}

\item{height}{height of widget, calculated based on width}

\item{elementId}{ID of container HTML element}
}
\description{
A widget that displays a group-level bar chart based on the output of a KRI analysis.
Bar charts are provided by default in any Assess function, and are suffixed with "JS" to indicate that they are an \code{htmlwidget} ported from the \code{rbm-viz} JavaScript library.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
ae_transform <- Transform_Rate(
  dfInput = sampleInput,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)

ae_analyze <- Analyze_NormalApprox(
  dfTransformed = ae_transform,
  strType = "rate"
)

ae_flag <- Flag_NormalApprox(
  ae_analyze,
  vThreshold = c(-3, -2, 2, 3)
)

ae_summary <- Summarize(
  ae_flag
)

# labels list
lLabels <- list(
  workflowid = "",
  group = "Site",
  abbreviation = "AE",
  metric = "Adverse Event Rate",
  numerator = "Adverse Events",
  denominator = "Days on Study",
  model = "Normal Approximation",
  score = "Adjusted Z-Score"
)

plot <- Widget_BarChart(
  dfSummary = ae_summary,
  lLabels = lLabels,
  strYAxisType = "metric",
  elementId = "aeAssessMetric"
)

}
