% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_FlagOverTime.R
\name{Widget_FlagOverTime}
\alias{Widget_FlagOverTime}
\title{Flag Over Time Widget}
\usage{
Widget_FlagOverTime(
  dfResults,
  dfMetrics,
  strGroupLevel = c("Site", "Study", "Country"),
  strFootnote = NULL,
  bExcludeEver = FALSE,
  bDebug = FALSE
)
}
\arguments{
<<<<<<< HEAD
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[gsm.mapping:BindResults]{gsm.mapping::BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID}, \code{SnapshotDate}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and reporting. Created by passing an \code{lWorkflow} object to \code{\link[gsm.mapping:MakeMetric]{gsm.mapping::MakeMetric()}}. Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric}, \code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{Threshold}. For more details see the Data Model vignette: \code{vignette("DataModel", package = "gsm")}.}
=======
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to
\code{\link[=BindResults]{BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator},
\code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID},
\code{SnapshotDate}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting. Created by passing an \code{lWorkflow} object to \code{\link[=MakeMetric]{MakeMetric()}}.
Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric},
\code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{Threshold}. For more
details see the Data Model vignette: \code{vignette("DataModel", package = "gsm")}.}
>>>>>>> fix-1972v2

\item{strGroupLevel}{\code{character} Value for the group level. Default: "Site".}

\item{strFootnote}{\code{character} Text to insert for figure}

\item{bExcludeEver}{\code{logical} Exclude options in widget dropdown that include the string "ever".
Default: \code{FALSE}.}

\item{bDebug}{\code{logical} Print debug messages? Default: \code{FALSE}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A widget that generates a table of flags over time using
\code{\link[=Report_FlagOverTime]{Report_FlagOverTime()}}.
}
\examples{
# Include all risk signals, irrespective flag value.
Widget_FlagOverTime(
  dfResults = reportingResults,
  dfMetrics = reportingMetrics
)

# Include risk signals that were ever flagged.
Widget_FlagOverTime(
  dfResults = FilterByFlags(
    reportingResults
  ),
  dfMetrics = reportingMetrics
)

# Include risk signals that were only flagged in the most recent snapshot.
Widget_FlagOverTime(
  dfResults = FilterByFlags(
    reportingResults,
    bCurrentlyFlagged = TRUE
  ),
  dfMetrics = reportingMetrics,
  bExcludeEver = TRUE
)

}
