% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_GroupOverview.R
\name{Widget_GroupOverview}
\alias{Widget_GroupOverview}
\title{Group Overview Widget}
\usage{
Widget_GroupOverview(
  dfResults,
  dfMetrics,
  dfGroups,
  strGroupLevel = NULL,
  strGroupSubset = "red",
  strGroupLabelKey = "InvestigatorLastName",
  bDebug = FALSE
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to
\code{\link[=BindResults]{BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator},
\code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID},
\code{SnapshotDate}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting. Created by passing an \code{lWorkflow} object to \code{\link[=MakeMetric]{MakeMetric()}}.
Expected columns: \code{File}, \code{MetricID}, \code{Group}, \code{Abbreviation}, \code{Metric},
\code{Numerator}, \code{Denominator}, \code{Model}, \code{Score}, and \code{Threshold}. For more
details see the Data Model vignette: \code{vignette("DataModel", package = "gsm")}.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. Created by
passing CTMS site and study data to \code{\link[gsm.mapping:MakeLongMeta]{gsm.mapping::MakeLongMeta()}}. Expected columns:
\code{GroupID}, \code{GroupLevel}, \code{Param}, \code{Value}.}

\item{strGroupLevel}{\code{character} Value for the group level. Default: NULL and taken from \code{dfMetrics$GroupLevel} if available.}

\item{strGroupSubset}{\code{character} Subset of groups to include in the table. Default: 'red'. Options:
\itemize{
\item 'all': All groups.
\item 'red': Groups with 1+ red flags.
\item 'red/amber': Groups with 1+ red/amber flag.
\item 'amber': Groups with 1+ amber flag.
}}

\item{strGroupLabelKey}{\code{character} Value for the group label key. Default: 'InvestigatorLastName'.}

\item{bDebug}{\code{logical} Print debug messages? Default: \code{FALSE}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A widget that generates a group overview table of group-level metric results across one or more
metrics.
}
\examples{
# site-level report
Widget_GroupOverview(
  dfResults = FilterByLatestSnapshotDate(reportingResults),
  dfMetrics = reportingMetrics,
  dfGroups = reportingGroups
)

# filter site-level report to all flags
Widget_GroupOverview(
  dfResults = FilterByLatestSnapshotDate(reportingResults),
  dfMetrics = reportingMetrics,
  dfGroups = reportingGroups,
  strGroupSubset = "all"
)

# country-level report
reportingMetrics$GroupLevel <- "Country"
Widget_GroupOverview(
  dfResults = FilterByLatestSnapshotDate(reportingResults),
  dfMetrics = reportingMetrics,
  dfGroups = reportingGroups,
  strGroupLevel = "Country"
)

}
