% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_ScatterPlot.R
\name{Widget_ScatterPlot}
\alias{Widget_ScatterPlot}
\title{KRI Scatter Plot}
\usage{
Widget_ScatterPlot(
  dfSummary,
  lLabels,
  dfSite = NULL,
  dfBounds = NULL,
  selectedGroupIDs = NULL,
  addSiteSelect = TRUE,
  bHideDropdown = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dfSummary}{\code{data.frame} A data.frame returned by \code{\link[=Summarize]{Summarize()}}.}

\item{lLabels}{\code{list} Metrics metadata.}

\item{dfSite}{\code{data.frame} Site metadata.}

\item{dfBounds}{\code{data.frame}, A data.frame returned by \code{\link[=Analyze_NormalApprox_PredictBounds]{Analyze_NormalApprox_PredictBounds()}} or \code{\link[=Analyze_Poisson_PredictBounds]{Analyze_Poisson_PredictBounds()}}}

\item{selectedGroupIDs}{group IDs to highlight, \code{NULL} by default, can be a single site or a vector.}

\item{addSiteSelect}{\code{logical} add a dropdown to highlight sites? Default: \code{TRUE}.}

\item{bHideDropdown}{\code{logical} should the dropdown be hidden? Default: \code{FALSE}. This is primarily used for the Shiny app \code{{gsmApp}} to hide the drop-down site selector,
since it is an additional element that needs to be updated based on user interactivity.}

\item{width}{width of widget, full screen by default}

\item{height}{height of widget, calculated based on width}

\item{elementId}{ID of container HTML element}
}
\description{
A widget that displays a group-level scatter plot based on the output of a KRI analysis.
Scatter plots are provided by default in any Assess function, and are suffixed with "JS" to indicate that they are an \code{htmlwidget} ported from the \code{rbm-viz} JavaScript library.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
ae_transform <- Transform_Rate(
  dfInput = sampleInput)

ae_analyze <- Analyze_NormalApprox(
  dfTransformed = ae_transform,
  strType = "rate"
)

dfBounds <- Analyze_NormalApprox_PredictBounds(
  dfTransformed = ae_transform,
  vThreshold = c(-3, -2, 2, 3),
  strType = "rate"
)

ae_flag <- Flag_NormalApprox(
  ae_analyze,
  vThreshold = c(-3, -2, 2, 3)
)

ae_summary <- Summarize(
  ae_flag
)

lLabels <- list(
  workflowid = "",
  Group = "Site",
  Abbreviation = "AE",
  Metric = "Adverse Event Rate",
  Numerator = "Adverse Events",
  Denominator = "Days on Study",
  Model = "Normal Approximation",
  Score = "Adjusted Z-Score"
)

plot <- Widget_ScatterPlot(
  dfSummary = ae_summary,
  lLabels = lLabels,
  dfBounds = dfBounds,
  elementId = "aeAssessScatter"
)

}
