% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_ScatterPlot.R
\name{Widget_ScatterPlot}
\alias{Widget_ScatterPlot}
\title{Scatter Plot Widget}
\usage{
Widget_ScatterPlot(
  dfSummary,
  lMetric,
  dfGroups = NULL,
  dfBounds = NULL,
  bAddGroupSelect = TRUE,
  bDebug = FALSE
)
}
\arguments{
\item{dfSummary}{\code{data.frame} Output of \code{\link[=Summarize]{Summarize()}}.}

\item{lMetric}{\code{list} Metric metadata, captured at the top of metric workflows and returned by
\code{\link[=MakeMetricInfo]{MakeMetricInfo()}}.}

\item{dfGroups}{\code{data.frame} Group metadata.}

\item{dfBounds}{\code{data.frame} Output of \code{\link[=Analyze_NormalApprox_PredictBounds]{Analyze_NormalApprox_PredictBounds()}} or
\code{\link[=Analyze_Poisson_PredictBounds]{Analyze_Poisson_PredictBounds()}}}

\item{bAddGroupSelect}{\code{logical} Add a dropdown to highlight sites? Default: \code{TRUE}.}

\item{bDebug}{\code{logical} Print debug messages? Default: \code{FALSE}.}
}
\description{
A widget that generates a scatter plot of group-level metric results, plotting the denominator
on the x-axis and the numerator on the y-axis.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
lDataRaw <- list(
    dfSTUDY = clindata::ctms_study,
    dfSITE = clindata::ctms_site,
    dfSUBJ = clindata::rawplus_dm,
    dfAE = clindata::rawplus_ae
)

lMappingWorkflow <- MakeWorkflowList('mapping')$mapping

lMappingWorkflow$steps <- lMappingWorkflow$steps \%>\%
    purrr::keep(~ .x$params$df \%in\% names(lDataRaw))

lDataMapped <- RunWorkflow(
    lMappingWorkflow,
    lDataRaw
)$lData

strMetricID <- 'kri0001'
lMetricWorkflow <- MakeWorkflowList(strMetricID)[[ strMetricID ]]

lResults <- RunWorkflow(
    lMetricWorkflow,
    lDataMapped
)

dfGroups <- clindata::ctms_site \%>\%
    left_join(
        lDataMapped$dfEnrolled \%>\%
            group_by(siteid) \%>\%
            tally(name = 'enrolled_participants'),
        c('site_num' = 'siteid')
    ) \%>\%
    rename(
        SiteID = site_num,
        status = site_status
    )

Widget_ScatterPlot(
    dfSummary = lResults$lData$dfSummary,
    lMetric = lMetricWorkflow$meta,
    dfGroups = dfGroups,
    dfBounds = lResults$lData$dfBounds
)
}
}
