% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_ScatterPlot.R
\name{Widget_ScatterPlot}
\alias{Widget_ScatterPlot}
\title{KRI Scatter Plot}
\usage{
Widget_ScatterPlot(
  dfSummary,
  lLabels,
  dfBounds,
  selectedGroupIDs = NULL,
  addSiteSelect = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dfSummary}{data with columns:
\itemize{
\item{\code{studyid}}
\item{\code{workflowid}}
\item{\code{groupid}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{metric}}
\item{\code{score}}
\item{\code{flag}}
}}

\item{lLabels}{configuration with columns:
\itemize{
\item{\code{workflow}}
\item{\code{gsm_version}}
\item{\code{group}}
\item{\code{metric}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{outcome}}
\item{\code{model}}
\item{\code{score}}
\item{\code{data_inputs}}
\item{\code{data_filters}}
}}

\item{dfBounds}{chart bounds data with columns:
\itemize{
\item{\code{threshold}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{metric}}
\item{\code{logdenominator}}
}}

\item{selectedGroupIDs}{group IDs to highlight, \code{NULL} by default, can be a single site or a vector.}

\item{addSiteSelect}{\code{logical} add a dropdown to highlight sites? Default: \code{TRUE}.}

\item{width}{width of widget, full screen by default}

\item{height}{height of widget, calculated based on width}

\item{elementId}{ID of container HTML element}
}
\description{
A widget that displays a group-level scatter plot based on the output of a KRI analysis.
Scatter plots are provided by default in any Assess function, and are suffixed with "JS" to indicate that they are an \code{htmlwidget} ported from the \code{rbm-viz} JavaScript library.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
ae <- AE_Map_Raw()

ae_transform <- Transform_Rate(
  dfInput = ae,
  strGroupCol = "SiteID",
  strNumeratorCol = "Count",
  strDenominatorCol = "Exposure"
)

ae_analyze <- Analyze_NormalApprox(
  dfTransformed = ae_transform,
  strType = "rate"
)

dfBounds <- Analyze_NormalApprox_PredictBounds(
  dfTransformed = ae_transform,
  vThreshold = c(-3, -2, 2, 3),
  strType = "rate"
)

ae_flag <- Flag_NormalApprox(
  ae_analyze,
  vThreshold = c(-3, -2, 2, 3)
)

ae_summary <- Summarize(
  ae_flag
)

lLabels <- list(
  workflowid = "",
  group = "Site",
  abbreviation = "AE",
  metric = "Adverse Event Rate",
  numerator = "Adverse Events",
  denominator = "Days on Study",
  model = "Normal Approximation",
  score = "Adjusted Z-Score"
)

plot <- Widget_ScatterPlot(
  dfSummary = ae_summary,
  lLabels = lLabels,
  dfBounds = dfBounds,
  elementId = "aeAssessScatter"
)

}
