% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_ScatterPlot.R
\name{Widget_ScatterPlot}
\alias{Widget_ScatterPlot}
\title{Scatter Plot Widget}
\usage{
Widget_ScatterPlot(
  dfResults,
  lMetric = list(),
  dfGroups = NULL,
  dfBounds = NULL,
  bAddGroupSelect = TRUE,
  bDebug = FALSE
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output. Created by passing a list of results returned by \code{\link[=Summarize]{Summarize()}} to \code{\link[=BindResults]{BindResults()}}. Expected columns: \code{GroupID}, \code{GroupLevel}, \code{Numerator}, \code{Denominator}, \code{Metric}, \code{Score}, \code{Flag}, \code{MetricID}, \code{StudyID}, \code{SnapshotDate}.}

\item{lMetric}{\code{list} Metric metadata, captured at the top of metric workflows and returned by
\code{\link[=MakeMetric]{MakeMetric()}}.}

\item{dfGroups}{\code{data.frame} Group metadata.}

\item{dfBounds}{\code{data.frame} Output of \code{\link[=Analyze_NormalApprox_PredictBounds]{Analyze_NormalApprox_PredictBounds()}} or
\code{\link[=Analyze_Poisson_PredictBounds]{Analyze_Poisson_PredictBounds()}}}

\item{bAddGroupSelect}{\code{logical} Add a dropdown to highlight sites? Default: \code{TRUE}.}

\item{bDebug}{\code{logical} Print debug messages? Default: \code{FALSE}.}
}
\description{
A widget that generates a scatter plot of group-level metric results, plotting the denominator
on the x-axis and the numerator on the y-axis.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
strMetricID <- 'kri0001'
lMetricWorkflow <- MakeWorkflowList()[[ strMetricID ]]

lData <- list(
    dfEnrolled = clindata::rawplus_dm \%>\% filter(enrollyn == 'Y'),
    dfAE = clindata::rawplus_ae
)

lResults <- lMetricWorkflow \%>\%
    RunWorkflow(lData)

dfGroups <- bind_rows(
    "SELECT pi_number as GroupID, site_status as Status, pi_first_name as InvestigatorFirstName, pi_last_name as InvestigatorLastName, city as City, state as State, country as Country, * FROM df" \%>\%
        RunQuery(clindata::ctms_site) \%>\%
        MakeLongMeta('Site'),
    "SELECT invid as GroupID, COUNT(DISTINCT subjectid) as ParticipantCount, COUNT(DISTINCT invid) as SiteCount FROM df GROUP BY invid" \%>\%
        RunQuery(lData$dfEnrolled) \%>\%
        MakeLongMeta('Site'),
    "SELECT country as GroupID, COUNT(DISTINCT subjectid) as ParticipantCount, COUNT(DISTINCT invid) as SiteCount FROM df GROUP BY country" \%>\%
        RunQuery(lData$dfEnrolled) \%>\%
        MakeLongMeta('Country')
)

dfBounds <- lResults$dfTransformed \%>\%
    Analyze_NormalApprox_PredictBounds(
        lMetricWorkflow$meta$strThreshold \%>\%
            stringr::str_split_1(',') \%>\%
            as.numeric(),
        strType = lMetricWorkflow$meta$Type
    )

Widget_ScatterPlot(
    dfResults = lResults$dfSummary,
    lMetric = lMetricWorkflow$meta,
    dfGroups = dfGroups,
    dfBounds = dfBounds
)
}
}
