% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_SiteOverview.R
\name{Widget_SiteOverview}
\alias{Widget_SiteOverview}
\title{Site Overview Table}
\usage{
Widget_SiteOverview(
  dfSummary,
  lConfig = list(),
  dfSite = clindata::ctms_site,
  dfWorkflow = gsm::meta_workflow,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dfSummary}{data with columns:
\itemize{
\item{\code{studyid}}
\item{\code{workflowid}}
\item{\code{groupid}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{metric}}
\item{\code{score}}
\item{\code{flag}}
}}

\item{lConfig}{configuration with columns:
\itemize{
\item{\code{workflow}}
\item{\code{gsm_version}}
\item{\code{group}}
\item{\code{metric}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{outcome}}
\item{\code{model}}
\item{\code{score}}
\item{\code{data_inputs}}
\item{\code{data_filters}}
}}

\item{dfSite}{\code{data.frame} Site metadata.}

\item{dfWorkflow}{\code{data.frame} Workflow metadata with columns:
\itemize{
\item{\code{workflowid}}
\item{\code{description}}
\item{\code{abbreviation}}
}}

\item{width}{width of widget, full screen by default}

\item{height}{height of widget, calculated based on width}

\item{elementId}{ID of container HTML element}
}
\description{
A widget that displays a site overview table based on the output of \code{\link[=Study_Assess]{Study_Assess()}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
wf_mapping <- MakeWorkflowList(strNames="mapping")[[1]]
wf_metrics <- MakeWorkflowList(strNames=paste0("kri",sprintf("\%04d", 1:2)))
dfMetrics <- wf_metrics \%>\% map_df(function(wf){
  wf$meta$vThreshold <- paste(wf$meta$vThreshold, collapse = ",")
  return(wf$meta)
})

# Import Site+Study Metadata
dfStudy<-clindata::ctms_study \%>\% rename(StudyID = protocol_number)
dfSite<- clindata::ctms_site \%>\% rename(SiteID = site_num)

# Pull Raw Data - this will overwrite the previous data pull
lRaw <- gsm::UseClindata(
  list(
    "dfSUBJ" = "clindata::rawplus_dm",
    "dfAE" = "clindata::rawplus_ae",
    "dfPD" = "clindata::ctms_protdev",
    "dfLB" = "clindata::rawplus_lb",
    "dfSTUDCOMP" = "clindata::rawplus_studcomp",
    "dfSDRGCOMP" = "clindata::rawplus_sdrgcomp",
    "dfDATACHG" = "clindata::edc_data_points",
    "dfDATAENT" = "clindata::edc_data_pages",
    "dfQUERY" = "clindata::edc_queries",
    "dfENROLL" = "clindata::rawplus_enroll"
  )
)

# Create Mapped Data
lMapped <- RunWorkflow(lWorkflow = wf_mapping, lData = lRaw)$lData

# Run Metrics
lResults <-wf_metrics \%>\% map(~RunWorkflow(lWorkflow=.x, lData=lMapped)) <-wf_metrics \%>\% map(~RunWorkflow(lWorkflow=.x, lData=lMapped))

dfSummary <- lResults \%>\%
imap_dfr(~.x$lData$dfSummary \%>\% mutate(MetricID = .y)) \%>\%
  mutate(StudyID = "ABC-123") \%>\%
  mutate(SnapshotDate = Sys.Date())

Widget_SiteOverview(dfSummary = dfSummary)

}
