% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_SiteOverview.R
\name{Widget_SiteOverview}
\alias{Widget_SiteOverview}
\title{Site Overview Table}
\usage{
Widget_SiteOverview(
  dfSummary,
  lConfig = list(),
  dfSite = Site_Map_Raw(),
  dfWorkflow = gsm::meta_workflow,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dfSummary}{data with columns:
\itemize{
\item{\code{studyid}}
\item{\code{workflowid}}
\item{\code{groupid}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{metric}}
\item{\code{score}}
\item{\code{flag}}
}}

\item{lConfig}{configuration with columns:
\itemize{
\item{\code{workflow}}
\item{\code{gsm_version}}
\item{\code{group}}
\item{\code{metric}}
\item{\code{numerator}}
\item{\code{denominator}}
\item{\code{outcome}}
\item{\code{model}}
\item{\code{score}}
\item{\code{data_inputs}}
\item{\code{data_filters}}
}}

\item{dfSite}{\code{data.frame} Site metadata returned by \code{\link[=Site_Map_Raw]{Site_Map_Raw()}}.}

\item{dfWorkflow}{\code{data.frame} Workflow metadata with columns:
\itemize{
\item{\code{workflowid}}
\item{\code{description}}
\item{\code{abbreviation}}
}}

\item{width}{width of widget, full screen by default}

\item{height}{height of widget, calculated based on width}

\item{elementId}{ID of container HTML element}
}
\description{
A widget that displays a site overview table based on the output of \code{\link[=Study_Assess]{Study_Assess()}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
study_assessment <- Study_Assess()

kri_results <- study_assessment \%>\%
    imap_dfr(function(kri, workflowid) {
        kri$lResults$lData$dfSummary \%>\%
            rename_with(tolower) \%>\%
            mutate(
                workflowid = !!workflowid
            )
    }) \%>\%
    filter(grepl('^kri', workflowid))

Widget_SiteOverview(kri_results)

}
