% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_SiteOverview.R
\name{Widget_SiteOverview}
\alias{Widget_SiteOverview}
\title{Site Overview Widget}
\usage{
Widget_SiteOverview(
  dfSummary,
  dfMetrics = meta_workflow,
  dfGroups = clindata::ctms_site \%>\% dplyr::left_join(clindata::rawplus_dm \%>\%
    dplyr::group_by("siteid") \%>\% dplyr::tally(name = "enrolled_participants"),
    c(site_num = "siteid")) \%>\% dplyr::rename(SiteID = "site_num", status =
    "site_status"),
  strGroupSubset = "red",
  bDebug = FALSE
)
}
\arguments{
\item{dfSummary}{\code{data.frame} Output of \code{\link[=Summarize]{Summarize()}}.}

\item{dfMetrics}{\code{list} Metric metadata, captured at the top of metric workflows and returned by
\code{\link[=MakeMetricInfo]{MakeMetricInfo()}}.}

\item{dfGroups}{\code{data.frame} Group metadata.}

\item{strGroupSubset}{\code{character} Subset of groups to include in the table. Default: 'red'. Options:
\itemize{
\item 'all': All groups.
\item 'red': Groups with 1+ red flags.
\item 'red/amber': Groups with 1+ red/amber flag.
\item 'amber': Groups with 1+ amber flag.
}}

\item{bDebug}{\code{logical} Print debug messages? Default: \code{FALSE}.}
}
\description{
A widget that generates a site overview table of group-level metric results across one or more
metrics.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
data_raw <- UseClindata(
    list(
        "dfSUBJ" = "clindata::rawplus_dm",
        "dfAE" = "clindata::rawplus_ae",
        "dfPD" = "clindata::ctms_protdev",
        "dfLB" = "clindata::rawplus_lb",
        "dfSTUDCOMP" = "clindata::rawplus_studcomp",
        "dfSDRGCOMP" = "clindata::rawplus_sdrgcomp \%>\%
      dplyr::filter(.data$phase == 'Blinded Study Drug Completion')",
        "dfDATACHG" = "clindata::edc_data_points",
        "dfDATAENT" = "clindata::edc_data_pages",
        "dfQUERY" = "clindata::edc_queries",
        "dfENROLL" = "clindata::rawplus_enroll"
    )
)

workflows <- MakeWorkflowList()

data_mapped <- RunWorkflow(workflows$mapping, data_raw)$lData

results <- workflows[ grepl('^kri', names(workflows) ) ] \%>\%
    map(~ RunWorkflow(.x, data_mapped))

dfSummary <- results \%>\% map_dfr(~ {
    data <- .x$lData$dfSummary
    data$MetricID <- .x$name
    data
})

counts <- RunWorkflow(workflows$counts, data_mapped)

# TODO: use [ clindata::rawplus_dm$invid ] and [ clindata::ctms_site$pi_number ] insstead of
# [ clindata::rawplus_dm$siteid ] and [ clindata::ctms_site$site_num ].
dfGroups <- clindata::ctms_site \%>\%
  dplyr::left_join(
    clindata::rawplus_dm \%>\%
      dplyr::group_by(siteid) \%>\%
      dplyr::tally(name = "enrolled_participants"),
    c('site_num' = 'siteid')
  ) \%>\%
  dplyr::rename(
    SiteID = site_num,
    status = site_status
  )

dfMetrics <- results \%>\% map_dfr(~ {
    metric <- .x$meta
    metric$vThreshold <- paste(metric$vThreshold, collapse = ',')
    metric
})

Widget_SiteOverview(
  dfSummary,
  dfMetrics,
  dfGroups
)
}
}
