% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_TimeSeries.R
\name{Widget_TimeSeries}
\alias{Widget_TimeSeries}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
Widget_TimeSeries(
  results,
  workflow,
  parameters,
  selectedGroupIDs = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  addSiteSelect = TRUE,
  siteSelectLabelValue = NULL
)
}
\arguments{
\item{results}{\code{data.frame} the stacked output of \code{Make_Snapshot()$lSnapshot$results_summary}, containing a minimum of two unique values for \code{gsm_analysis_date}.}

\item{workflow}{\code{data.frame} the output of \code{Make_Snapshot()$lSnapshot$meta_workflow}.}

\item{parameters}{\code{data.frame} the stacked output of \code{Make_Snapshot()$lSnapshot$meta_param}.}

\item{selectedGroupIDs}{\code{character} group IDs to highlight, \code{NULL} by default, can be a single site or a vector.}

\item{width}{\code{numeric} width of widget.}

\item{height}{\code{numeric} height of widget.}

\item{elementId}{\code{character} ID of container HTML element.}

\item{addSiteSelect}{\code{logical} add a dropdown to highlight sites? Default: \code{TRUE}.}

\item{siteSelectLabelValue}{Label used to populate the HTML drop-down menu. Constructed as: 'Highlighted {siteSelectLabelValue}: '.}
}
\description{
A widget that displays a time-series plot based on longitudinal snapshots using \code{{gsm}}.
}
\details{
Time-Series Continuous Plot
}
