% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_TimeSeries.R
\name{Widget_TimeSeries}
\alias{Widget_TimeSeries}
\title{Time-Series Continuous Plot}
\usage{
Widget_TimeSeries(
  dfSummary,
  lLabels,
  dfSite = NULL,
  dfParams = NULL,
  yAxis = "score",
  selectedGroupIDs = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  addSiteSelect = TRUE,
  siteSelectLabelValue = "Site"
)
}
\arguments{
\item{dfSummary}{\code{data.frame} summary data containing a minimum of two unique values for \code{gsm_analysis_date}.}

\item{lLabels}{\code{list} chart labels.}

\item{dfSite}{\code{data.frame} Site metadata.}

\item{dfParams}{\code{data.frame} the stacked output.}

\item{yAxis}{\code{character} the name of a column from \code{lLabels} to be passed to the y-axis on the widget plot.}

\item{selectedGroupIDs}{\code{character} group IDs to highlight, \code{NULL} by default, can be a single site or a vector.}

\item{width}{\code{numeric} width of widget.}

\item{height}{\code{numeric} height of widget.}

\item{elementId}{\code{character} ID of container HTML element.}

\item{addSiteSelect}{\code{logical} add a dropdown to highlight sites? Default: \code{TRUE}.}

\item{siteSelectLabelValue}{Label used to populate the HTML drop-down menu.}
}
\description{
A widget that displays a time-series plot based on longitudinal snapshots using \code{{gsm}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
