% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_TimeSeriesQTL.R
\name{Widget_TimeSeriesQTL}
\alias{Widget_TimeSeriesQTL}
\title{Time Series QTL}
\usage{
Widget_TimeSeriesQTL(
  qtl,
  raw_results,
  raw_workflow,
  raw_param,
  raw_analysis,
  selectedGroupIDs = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{qtl}{specific qtl to filter to}

\item{raw_results}{TODO}

\item{raw_workflow}{TODO}

\item{raw_param}{TODO}

\item{raw_analysis}{TODO}

\item{selectedGroupIDs}{TODO}

\item{width}{the width of the widget}

\item{height}{the hight of the widget}

\item{elementId}{id of widget, automatically generated if not supplied}
}
\description{
A Time Series graphic for qtl data
}
